OpenRepGrid: an R package for the analysis of repertory grids.

The OpenRepGrid package provides tools for the analysis of repertory grid data. The repertory grid is a method devised by George Alexander Kelly in his seminal work "The Psychology of Personal Constructs" published in 1955. The repertory grid has been used in and outside the context of Personal Construct Psychology (PCP) in a broad range of fields. For an introduction into the technique see e.g. Fransella, Bell and Bannister (2003).

Note

To get started with OpenRepGrid visit the project's home under openrepgrid.org. On this site you will find tutorials, explanation about the theory, the analysis methods and the corresponding R code.

To see how to cite the OpenRepGrid package, type citation("OpenRepGrid") into the R console.

References

Fransella, F., Bell, R. C., & Bannister, D. (2003). A Manual for Repertory Grid Technique (2. Ed.). Chichester: John Wiley & Sons.

Kelly, G. A. (1955). The psychology of personal constructs. Vol. I, II. New York: Norton, (2nd printing: 1991, Routledge, London, New York).

See also

Author

Maintainer: Mark Heckmann (@markheckmann)

Contributors: Richard C. Bell, Alejandro García Gutiérrez (@j4n7), Diego Vitali (@artoo-git), José Antonio González Del Puerto (@MindCartographer), Jonathan D. Raskin

How to contribute: You can contribute in various ways. The OpenRepGrid code is hosted on GitHub, where you can issue bug reports or feature requests. You may email your request to the package maintainer.