OpenRepGrid: Annotated overview of package functions.

This documentation page contains an overview over the package functions ordered by topics. The best place to start learning OpenRepGrid will be the package website https://openrepgrid.org though.

Functions sorted by topic

Manipulating grids

left()	Move construct(s) to the left

right()	Move construct(s) to the right

up()	Move construct(s) upwards

down()	Move construct(s) downwards

Loading and saving data

importGridcor()	Import GRIDCOR data files

importGridstat()	Import Gridstat data files

importGridsuite()	Import Gridsuite data files

importScivesco()	Import sci:vesco data files

importTxt()	Import grid data from a text file

saveAsTxt()	Save grid in a text file (txt)

Analyzing constructs

Descriptive statistics of constructs Construct correlations distance Root mean square of inter-construct correlations Somers' D Principal component analysis (PCA) of construct correlation matrix Cluster analysis of constructs

Analyzing elements

Visual representation

Bertin plots

bertin()	Make Bertin display of grid data

bertinCluster()	Bertin display with corresponding cluster analysis

Biplots

biplot2d()	Draw a two-dimensional biplot

biplotEsa2d()	Plot an eigenstructure analysis (ESA) biplot in 2D

biplotSlater2d()	Draws Slater's INGRID biplot in 2D

biplotPseudo3d()	See 'biplotPseudo3d' for its use. Draws a biplot of the grid in 2D with depth impression (pseudo 3D)

biplotEsaPseudo3d()	Plot an eigenstructure analysis (ESA) in 2D grid with 3D impression (pseudo 3D)

biplotSlaterPseudo3d()	Draws Slater's biplot in 2D with depth impression (pseudo 3D)

biplot3d()	Draw grid in rgl (3D device)

biplotEsa3d()	Draw the eigenstructure analysis (ESA) biplot in rgl (3D device)

biplotSlater3d()	Draw the Slater's INGRID biplot in rgl (3D device)

biplotSimple()	A graphically unsophisticated version of a biplot

Index measures

indexConflict1()	Conflict measure for grids (Slade & Sheehan, 1979) based on correlations

indexConflict2()	Conflict measure for grids (Bassler et al., 1992) based on correlations

indexConflict3()	Conflict or inconsistency measure for grids (Bell, 2004) based on distances

indexDilemma()	Detect implicative dilemmas (conflicts)

indexIntensity()	Intensity index

indexPvaff()	Percentage of Variance Accounted for by the First Factor (PVAFF)

indexBias()	Calculate 'bias' of grid as defined by Slater (1977)

indexVariability()	Calculate 'variability' of a grid as defined by Slater (1977)

Special features

alignByIdeal()	Align constructs using the ideal element to gain pole preferences

alignByLoadings()	Align constructs by loadings on first principal component

reorder2d()	Order grid by angles between construct and/or elements in 2D

OpenRepGrid uses several default settings e.g. to determine how many construct characters to display by default when displaying a grid. The function settings can be used to show and change these settings. Also it is possible to store the settings to a file and load the settings file to restore the settings.

settings()	Show and modify global settings for OpenRepGrid

settingsSave()	Save OpenRepGrid settings to file

settingsLoad()	Load OpenRepGrid settings from file

Grid datasets

OpenRepGrid already contains some ready to use grid data sets. Most of the datasets are taken from the literature. To output the data simply type Type the name of the dataset to the console and press enter.

Single grids

bell2010()	Grid data from a study by Haritos et al. (2004) on role titles; used for demonstration of construct alignment in Bell (2010, p. 46).

bellmcgorry1992()	Grid from a psychotic patient used in Bell (1997, p. 6). Data originated from a study by Bell and McGorry (1992).

boeker()	Grid from seventeen year old female schizophrenic patient undergoing last stage of psychoanalytically oriented psychotherapy (Boeker, 1996, p. 163).

fbb2003()	Dataset used in A manual for Repertory Grid Technique (Fransella, Bell, & Bannister, 2003b, p. 60).

feixas2004()	Grid from a 22 year old Spanish girl suffering self-worth problems (Feixas & Saul, 2004, p. 77).

mackay1992()	Dataset Grid C used in Mackay's paper on inter-element correlation (1992, p. 65).

leach2001a(), leach2001b()	Pre- (a) and post-therapy (b) dataset from sexual child abuse survivor (Leach, Freshwater, Aldridge, & Sunderland, 2001, p. 227).

raeithel()	Grid data to demonstrate the use of Bertin diagrams (Raeithel, 1998, p. 223). The context of its administration is unknown.

slater1977a()	Drug addict grid dataset from (Slater, 1977, p. 32).

slater1977b()	Grid dataset (ranked) from a seventeen year old female psychiatric patient (Slater, 1977, p. 110) showing depression, anxiety and self-mutilation. The data was originally reported by Watson (1970).

Multiple grids

NOT YET AVAILABLE

Functions for developers

OpenRepGrid: internal functions overview for developers.

Below you find a guide for developers: these functions are usually not needed by the casual user. The internal functions have a twofold goal

to provide means for advanced numerical grid analysis and 2) to facilitate function development. The function for these purposes are internal, i.e. they are not visible in the package documentation. Nonetheless they do have a documentation that can be accesses in the same way as for other functions. More in the details section.

Functions for advanced grid analysis

The package provides functions to facilitate numerical research for grids. These comprise the generation of random data, permutation of grids etc. to facilitate Monte Carlo simulations, batch analysis of grids and other methods. With R as an underlying framework, the results of grid analysis easily lend themselves to further statistical processing and analysis within R. This is one of the central advantages for researchers compared to other standard grid software. The following table lists several functions for these purposes.

randomGrid()

randomGrids()

permuteConstructs()

permuteGrid()

quasiDistributionDistanceSlater()

Modules for function development

Beside the advanced analysis feature the developer's functions comprise low-level modules to create new functions for grid analysis. Though the internal structure of a repgrid object in R is simple (type e.g. str(bell2010, 2) to get an impression), it is convenient to not have to deal with access on this level. Several function like e.g. getElementNames are convenient wrappers that perform standard tasks needed when implementing new functions. The following table lists several functions for these purposes.

getRatingLayer()	Retrieve grid scores from grid object.

getNoOfConstructs()	Get the number of constructs in a grid object.

getNoOfElements()	Get the number of elements in a grid object.

dim()	Get grid dimensions, i.e. constructs x elements.

getScale()	Get minimum and maximum scale value used in grid.

getScaleMidpoint()	Get midpoint of the grid rating scale.

getConstructNames()	Get construct names.

getConstructNames2()	Get construct names (another newer version).

getElementNames()	Retrieve element names of repgrid object.

bindConstructs()	Concatenate the constructs of two grids.

doubleEntry()	Join the constructs of a grid with the same reversed constructs.

Other internal functions

importTxtInternal()

See also

Author

Current members of the OpenRepGrid development team: Mark Heckmann. Everyone who is interested in developing the package is invited to join.

       The \pkg{OpenRepGrid} package development is hosted on github (<https://github.com/markheckmann/OpenRepGrid>).
       The github site provides information and allows to file bug reports or feature requests.
       Bug reports can also be emailed to the package maintainer or issued on
       <https://openrepgrid.org> under section *Suggestions/Issues*.
       The package maintainer is Mark Heckmann <heckmann(dot)mark(at)gmail(dot)com>.