Add a new average element

A new element is added to the grid. The scores are the arithmetic means across all selected elements.

Usage

addAvgElement(x, name = "avg", i, digits = Inf)

Arguments

x

A repgrid object.

name

Name of new element.

i

Indexes of elements to be averaged across. Negative indexes can be used to exclude elements from the complete set. Duplicate indexes are allowed but a warning is issued.

digits

Digits to round mean value to. By default no rounding is used (digits = Inf). Use digits = 0 to round to nearest integer, i.e. only using original integer rating scores.

Value

A repgrid object with additional new element.

Examples

addAvgElement(feixas2004, "others", i = 2:12)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                       Friend 2 - 7 8 - Non-grata                               
#>                     Friend 1 - 6 | | 9 - Friend 3                              
#>                  Boyfriend - 5 | | | | 10 - Cousin                             
#>                  Brother - 4 | | | | | | 11 - Godmother                        
#>                 Father - 3 | | | | | | | | 12 - Friend 4                       
#>               Mother - 2 | | | | | | | | | | 13 - Ideal Self                   
#>           Self now - 1 | | | | | | | | | | | | 14 - others                     
#>                      | | | | | | | | | | | | | |                               
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7 3.72727272727273   (1) Optimisti
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4 4.54545454545455   (2) Takes it 
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5                4   (3) Enterpris
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7 4.45454545454545   (4) Works to 
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7 3.09090909090909   (5) Tolerant 
#>        Teasing (6)   2 7 1 6 4 3 4 6 3 3 5 6 3 4.36363636363636   (6) Touchy   
#> eciates others (7)   2 6 6 6 1 5 4 7 4 2 2 5 1 4.36363636363636   (7) Does not 
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7 4.36363636363636   (8) Calm     
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2 3.81818181818182   (9) Selfish  
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7 3.54545454545455   (10) Generous
#>     Sensitive (11)   1 5 7 7 1 4 5 7 1 4 3 4 1 4.36363636363636   (11) Material
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7 5.09090909090909   (12) Respectf
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7 4.36363636363636   (13) Sincere 
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7                4   (14) Non blac
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6 3.90909090909091   (15) Natural 
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7 3.45454545454545   (16) Looks af
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7 2.90909090909091   (17) Accessib
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5 4.54545454545455   (18) Extrover
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6 3.90909090909091   (19) Does not
#>  good in thin (20)   6 6 4 6 1 5 2 7 6 3 3 5 1 4.36363636363636   (20) Sees onl
#> 
#>     
#>     
#>     
#>     
#>     
#>     
#>     
#>     
#> c   
#> easy
#> ing 
#> live
#> with
#>     
#> appr
#>     
#>     
#>     
#> isti
#> ul  
#>     
#> kmai
#>     
#> ter 
#> le  
#> ted 
#>  get
#> y th
#> 
addAvgElement(feixas2004, "others", i = 2:12, digits = 0) # integers
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                       Friend 2 - 7 8 - Non-grata                    
#>                     Friend 1 - 6 | | 9 - Friend 3                   
#>                  Boyfriend - 5 | | | | 10 - Cousin                  
#>                  Brother - 4 | | | | | | 11 - Godmother             
#>                 Father - 3 | | | | | | | | 12 - Friend 4            
#>               Mother - 2 | | | | | | | | | | 13 - Ideal Self        
#>           Self now - 1 | | | | | | | | | | | | 14 - others          
#>                      | | | | | | | | | | | | | |                    
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7 4   (1) Optimistic   
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4 5   (2) Takes it easy
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5 4   (3) Enterprising 
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7 4   (4) Works to live
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7 3   (5) Tolerant with
#>        Teasing (6)   2 7 1 6 4 3 4 6 3 3 5 6 3 4   (6) Touchy       
#> eciates others (7)   2 6 6 6 1 5 4 7 4 2 2 5 1 4   (7) Does not appr
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7 4   (8) Calm         
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2 4   (9) Selfish      
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7 4   (10) Generous    
#>     Sensitive (11)   1 5 7 7 1 4 5 7 1 4 3 4 1 4   (11) Materialisti
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7 5   (12) Respectful  
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7 4   (13) Sincere     
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7 4   (14) Non blackmai
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6 4   (15) Natural     
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7 3   (16) Looks after 
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7 3   (17) Accessible  
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5 5   (18) Extroverted 
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6 4   (19) Does not get
#>  good in thin (20)   6 6 4 6 1 5 2 7 6 3 3 5 1 4   (20) Sees only th
#> 

# exluding elements via negative indexes
addAvgElement(feixas2004, "others", i = c(-1, -13))
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                       Friend 2 - 7 8 - Non-grata                               
#>                     Friend 1 - 6 | | 9 - Friend 3                              
#>                  Boyfriend - 5 | | | | 10 - Cousin                             
#>                  Brother - 4 | | | | | | 11 - Godmother                        
#>                 Father - 3 | | | | | | | | 12 - Friend 4                       
#>               Mother - 2 | | | | | | | | | | 13 - Ideal Self                   
#>           Self now - 1 | | | | | | | | | | | | 14 - others                     
#>                      | | | | | | | | | | | | | |                               
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7 3.72727272727273   (1) Optimisti
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4 4.54545454545455   (2) Takes it 
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5                4   (3) Enterpris
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7 4.45454545454545   (4) Works to 
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7 3.09090909090909   (5) Tolerant 
#>        Teasing (6)   2 7 1 6 4 3 4 6 3 3 5 6 3 4.36363636363636   (6) Touchy   
#> eciates others (7)   2 6 6 6 1 5 4 7 4 2 2 5 1 4.36363636363636   (7) Does not 
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7 4.36363636363636   (8) Calm     
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2 3.81818181818182   (9) Selfish  
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7 3.54545454545455   (10) Generous
#>     Sensitive (11)   1 5 7 7 1 4 5 7 1 4 3 4 1 4.36363636363636   (11) Material
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7 5.09090909090909   (12) Respectf
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7 4.36363636363636   (13) Sincere 
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7                4   (14) Non blac
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6 3.90909090909091   (15) Natural 
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7 3.45454545454545   (16) Looks af
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7 2.90909090909091   (17) Accessib
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5 4.54545454545455   (18) Extrover
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6 3.90909090909091   (19) Does not
#>  good in thin (20)   6 6 4 6 1 5 2 7 6 3 3 5 1 4.36363636363636   (20) Sees onl
#> 
#>     
#>     
#>     
#>     
#>     
#>     
#>     
#>     
#> c   
#> easy
#> ing 
#> live
#> with
#>     
#> appr
#>     
#>     
#>     
#> isti
#> ul  
#>     
#> kmai
#>     
#> ter 
#> le  
#> ted 
#>  get
#> y th
#>