Add a new construct to an existing grid object.

Add a new construct to an existing grid object.

Usage

addConstruct(
  x,
  l.name = NA,
  r.name = NA,
  scores = NA,
  l.preferred = NA,
  r.preferred = NA,
  l.emerged = NA,
  r.emerged = NA,
  position = NA,
  side = "pre"
)

Arguments

x

repgrid object.

l.name

Name of the left pole (character string).

r.name

Name of the right pole (character string).

scores

Numerical ratings for the new construct row (length must match number of elements in the grid).

l.preferred

Is the left one the preferred pole? (logical).

r.preferred

Is the right one the preferred pole? (logical).

l.emerged

Is the left one the emergent pole? (logical).

r.emerged

Is the right one the emergent pole? (logical).

position

An integer at which row the construct will be added. TODO. Does not work properly.

side

Not yet in use.

Value

repgrid object.

See also

Examples

if (FALSE) { # \dontrun{

# show grid
bell2010
addConstruct(bell2010, "left pole", "pole right", c(3, 1, 3, 2, 5, 4, 6, 3, 7, 1))
} # }