Adds the percentage of the sum-of-squares explained by each axis to the plot.

Adds the percentage of the sum-of-squares explained by each axis to the plot.

Usage

addVarianceExplainedToBiplot2d(
  x,
  dim = c(1, 2, 3),
  var.cex = 0.7,
  var.show = TRUE,
  var.col = grey(0.1),
  axis.ext = 0.1,
  center = 1,
  normalize = 0,
  g = 0,
  h = 1 - g,
  col.active = NA,
  col.passive = NA,
  ...
)

Arguments

x

repgrid object containing the biplot coords, i.e. after having called calcBiplotCoords() and prepareBiplotData().

dim

The dimensions to be printed.

var.cex

The cex value for the percentages shown in the plot.

var.show

Show explained sum-of-squares in biplot? (default TRUE).

var.col

The color value of the percentages shown in the plot.

axis.ext

Axis extension factor (default is .1). A bigger value will zoom out the plot.

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). The default is 1 (row centering).

normalize

A numeric value indicating along what direction (rows, columns) to normalize by standard deviations. 0 = none, 1= rows, 2 = columns (default is 0).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

col.active

Columns (elements) that are no supplementary points, i.e. they are used in the SVD to find principal components. default is to use all elements.

col.passive

Columns (elements) that are supplementary points, i.e. they are NOT used in the SVD but projected into the component space afterwards. They do not determine the solution. Default is NA, i.e. no elements are set supplementary.

...

Not evaluated.