Align constructs using the ideal element to gain pole preferences.

The direction of the constructs in a grid is arbitrary and a reflection of a scale does not affect the information contained in the grid. Nonetheless, the direction of a scale has an effect on inter-element correlations (Mackay, 1992) and on the spatial representation and clustering of the grid (Bell, 2010). Hence, it is desirable to follow a protocol to align constructs that will render unique results. A common approach is to align constructs by pole preference, i. e. aligning all positive and negative poles. This can e. g. be achieved using swapPoles(). If an ideal element is present, this element can be used to identify the positive and negative pole. The function alignByIdeal will align the constructs accordingly. Note that this approach does not always yield definite results as sometimes ratings do not show a clear preference for one pole (Winter, Bell & Watson, 2010). If a preference cannot be determined definitely, the construct direction remains unchanged (a warning is issued in that case).

Usage

alignByIdeal(x, ideal, high = TRUE)

Arguments

x

repgrid object

ideal

Number of the element that is used for alignment (the ideal).

high

Logical. Whether to align the constructs so the ideal will have high ratings on the constructs (i.e. TRUE, default) or low ratings (FALSE). High scores will lead to the preference pole on the right side, low scores will align the preference pole on the left side.

Value

repgrid object with aligned constructs.

References

Bell, R. C. (2010). A note on aligning constructs. Personal Construct Theory & Practice, 7, 42-48.

Mackay, N. (1992). Identification, Reflection, and Correlation: Problems in the bases of repertory grid measures. International Journal of Personal Construct Psychology, 5(1), 57-75.

Winter, D. A., Bell, R. C., & Watson, S. (2010). Midpoint ratings on personal constructs: Constriction or the middle way? Journal of Constructivist Psychology, 23(4), 337-356.

See also

Examples

feixas2004 # original grid
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Friend 2                     
#>                     Friend 1 - 6 | 8 - Non-grata                  
#>                  Boyfriend - 5 | | | 9 - Friend 3                 
#>                  Brother - 4 | | | | | 10 - Cousin                
#>                 Father - 3 | | | | | | | 11 - Godmother           
#>               Mother - 2 | | | | | | | | | 12 - Friend 4          
#>           Self now - 1 | | | | | | | | | | | 13 - Ideal Self      
#>                      | | | | | | | | | | | | |                    
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7   (1) Optimistic   
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4   (2) Takes it easy
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5   (3) Enterprising 
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7   (4) Works to live
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7   (5) Tolerant with
#>        Teasing (6)   2 7 1 6 4 3 4 6 3 3 5 6 3   (6) Touchy       
#> eciates others (7)   2 6 6 6 1 5 4 7 4 2 2 5 1   (7) Does not appr
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7   (8) Calm         
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2   (9) Selfish      
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7   (10) Generous    
#>     Sensitive (11)   1 5 7 7 1 4 5 7 1 4 3 4 1   (11) Materialisti
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7   (12) Respectful  
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7   (13) Sincere     
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7   (14) Non blackmai
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6   (15) Natural     
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7   (16) Looks after 
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7   (17) Accessible  
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5   (18) Extroverted 
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6   (19) Does not get
#>  good in thin (20)   6 6 4 6 1 5 2 7 6 3 3 5 1   (20) Sees only th
#> 
alignByIdeal(feixas2004, 13) # aligned with preference pole on the right
#> Warning: The following constructs do not show a preference for either poleand have thus not been aligned: 2
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Friend 2                     
#>                     Friend 1 - 6 | 8 - Non-grata                  
#>                  Boyfriend - 5 | | | 9 - Friend 3                 
#>                  Brother - 4 | | | | | 10 - Cousin                
#>                 Father - 3 | | | | | | | 11 - Godmother           
#>               Mother - 2 | | | | | | | | | 12 - Friend 4          
#>           Self now - 1 | | | | | | | | | | | 13 - Ideal Self      
#>                      | | | | | | | | | | | | |                    
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7   (1) Optimistic   
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4   (2) Takes it easy
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5   (3) Enterprising 
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7   (4) Works to live
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7   (5) Tolerant with
#>         Touchy (6)   6 1 7 2 4 5 4 2 5 5 3 2 5   (6) Teasing      
#> reciate others (7)   6 2 2 2 7 3 4 1 4 6 6 3 7   (7) Appreciates o
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7   (8) Calm         
#>        Selfish (9)   6 6 2 1 6 5 3 1 5 5 6 6 6   (9) Concerned abo
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7   (10) Generous    
#> , superficial (11)   7 3 1 1 7 4 3 1 7 4 5 4 7   (11) Sensitive   
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7   (12) Respectful  
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7   (13) Sincere     
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7   (14) Non blackmai
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6   (15) Natural     
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7   (16) Looks after 
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7   (17) Accessible  
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5   (18) Extroverted 
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6   (19) Does not get
#>  the negative (20)   2 2 4 2 7 3 6 1 2 5 5 3 7   (20) Tries to fin
#> 

raeithel # original grid
#> 
#> META DATA:
#> Number of constructs:  13 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from -1 (left pole) to 1 (right pole)
#> 
#> RATINGS:
#>                                        7 - Mutter                              
#>                             Virus - 6  |  8 - Schwester                        
#>                     Ich vorher - 5  |  |  |  9 - Partner                       
#>                        Arzt - 4  |  |  |  |  |  10 - Vater                     
#>                    Ideal - 3  |  |  |  |  |  |  |  11 - Neg. Person            
#>             Therapeut - 2  |  |  |  |  |  |  |  |  |  12 - Ich                 
#>             Freund - 1  |  |  |  |  |  |  |  |  |  |  |  13 - Freundin         
#>                      |  |  |  |  |  |  |  |  |  |  |  |  |                     
#>   charakterlos (1)    1  1  1  1  1 -1 -1 -1  1 -1 -1  1  1   (1) Charakter hab
#> uninteressiert (2)    1  1  1  1  1  1 -1 -1  1  1 -1  1  1   (2) interessiert 
#> h gehen lassen (3)    1  1  1  1  1  1 -1 -1  1 -1 -1  1  1   (3) zielstrebig  
#>     unerfahren (4)    1  1  1  1  1  1 -1 -1 -1  0 -1  1  1   (4) erfahren     
#>   abschaetzbar (5)   -1  1  1  1  1  1  1 -1 -1 -1 -1  1  0   (5) unberechenbar
#>          mutig (6)   -1 -1 -1 -1  1  1  1  1 -1  1  1  1 -1   (6) feige        
#>  leichtfuessig (7)   -1 -1 -1 -1 -1 -1  1  1 -1  1  0  1 -1   (7) schwerfaellig
#>          kuehl (8)    1  1  1  0 -1 -1 -1  1  1  1  1  0 -1   (8) warmherzig   
#> h eigennuetzig (9)    1  1  1  1 -1 -1 -1  0  1  1 -1 -1  0   (9) lebensbejahen
#>     angepasst (10)    1 -1  1  0  1  1 -1 -1 -1 -1 -1 -1 -1   (10) unangepasst 
#>         offen (11)   -1 -1 -1 -1  1  1  1 -1 -1 -1  1 -1 -1   (11) hinterhaelti
#>  unvorsichtig (12)   -1 -1  1  1 -1  1  0  1 -1  1 -1  1 -1   (12) disziplinier
#> undschaftlich (13)    0 -1 -1  1 -1  0  0 -1  0  1  0  0 -1   (13) vaeterlich a
#> 
alignByIdeal(raeithel, 3, high = FALSE) # aligned with preference pole on the left
#> 
#> META DATA:
#> Number of constructs:  13 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from -1 (left pole) to 1 (right pole)
#> 
#> RATINGS:
#>                                        7 - Mutter                              
#>                             Virus - 6  |  8 - Schwester                        
#>                     Ich vorher - 5  |  |  |  9 - Partner                       
#>                        Arzt - 4  |  |  |  |  |  10 - Vater                     
#>                    Ideal - 3  |  |  |  |  |  |  |  11 - Neg. Person            
#>             Therapeut - 2  |  |  |  |  |  |  |  |  |  12 - Ich                 
#>             Freund - 1  |  |  |  |  |  |  |  |  |  |  |  13 - Freundin         
#>                      |  |  |  |  |  |  |  |  |  |  |  |  |                     
#> harakter haben (1)   -1 -1 -1 -1 -1  1  1  1 -1  1  1 -1 -1   (1) charakterlos 
#>   interessiert (2)   -1 -1 -1 -1 -1 -1  1  1 -1 -1  1 -1 -1   (2) uninteressier
#>    zielstrebig (3)   -1 -1 -1 -1 -1 -1  1  1 -1  1  1 -1 -1   (3) sich gehen la
#>       erfahren (4)   -1 -1 -1 -1 -1 -1  1  1  1  0  1 -1 -1   (4) unerfahren   
#>  unberechenbar (5)    1 -1 -1 -1 -1 -1 -1  1  1  1  1 -1  0   (5) abschaetzbar 
#>          mutig (6)   -1 -1 -1 -1  1  1  1  1 -1  1  1  1 -1   (6) feige        
#>  leichtfuessig (7)   -1 -1 -1 -1 -1 -1  1  1 -1  1  0  1 -1   (7) schwerfaellig
#>     warmherzig (8)   -1 -1 -1  0  1  1  1 -1 -1 -1 -1  0  1   (8) kuehl        
#> ejahend sozial (9)   -1 -1 -1 -1  1  1  1  0 -1 -1  1  1  0   (9) egoistisch ei
#>   unangepasst (10)   -1  1 -1  0 -1 -1  1  1  1  1  1  1  1   (10) angepasst   
#>         offen (11)   -1 -1 -1 -1  1  1  1 -1 -1 -1  1 -1 -1   (11) hinterhaelti
#> iniert gesund (12)    1  1 -1 -1  1 -1  0 -1  1 -1  1 -1  1   (12) unvorsichtig
#> undschaftlich (13)    0 -1 -1  1 -1  0  0 -1  0  1  0  0 -1   (13) vaeterlich a
#>