Align constructs by preferred pole

The direction of the constructs in a grid is arbitrary. While their reversal (see reverse()) does not affect the information contained in the grid, it is often useful to align constructs for easier interpretation. One way of alignment is placing all positive poles on the same side. Note that this this is only possible if the preferred poles are defined (see preferredPoles()).

Usage

alignByPreferredPole(x, side_positive = "right")

Arguments

x

A repgrid object.

side_positive

Align all positoive poles on '

Value

A repgrid object with aligned constructs.

See also

Examples

x <- preferredPolesByIdeal(boeker, "ideal self")
x <- alignByPreferredPole(x)
x
#> 
#> META DATA:
#> Number of elements:  15 
#> Number of constructs:  14 
#> Preferred poles defined:  14/14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                     8 - martin                              
#>                          george - 7 | 9 - elizabeth                         
#>                          karl - 6 | | | 10 - therapist                      
#>                        kurt - 5 | | | | | 11 - irene                        
#>                    father - 4 | | | | | | | 12 - childhood self             
#>                  mother - 3 | | | | | | | | | 13 - self before illness      
#>            ideal self - 2 | | | | | | | | | | | 14 - self with delusion     
#>                self - 1 | | | | | | | | | | | | | 15 - self as dreamer      
#>                       | | | | | | | | | | | | | | |                         
#>       balanced (-1)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (1+) get along with co
#>       isolated (-2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2+) sociable         
#>       excluded (-3)   5 5 5 4 2 4 5 4 5 4 4 3 3 2 4   (3+) closely integrate
#>        passive (-4)   3 6 4 6 5 3 5 4 4 5 4 4 4 2 3   (4+) discursive       
#>    indifferent (-5)   5 6 5 6 5 3 3 5 3 5 1 4 5 5 4   (5+) open minded      
#>         dreamy (-6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6+) dispassionate    
#>      depressed (-7)   5 6 4 5 4 4 4 5 5 4 5 4 4 4 4   (7+) practically orien
#>        playful (-8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8+) serious          
#>        selfish (-9)   5 6 4 5 3 2 3 6 4 5 1 4 4 4 4   (9+) socially minded  
#>   quarrelsome (-10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10+) peaceful        
#>     technical (-11)   2 6 5 3 4 2 4 5 3 4 4 3 3 3 3   (11+) artistic        
#>     emotional (-12)   5 6 2 4 3 3 2 4 3 6 1 3 5 4 4   (12+) scientific      
#>     introvert (-13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13+) extrovert       
#> home oriented (-14)   6 6 3 5 3 2 5 2 2 4 1 6 6 5 6   (14+) wanderlust      
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined