Concatenate the constructs of two grids.

I.e. the constructs are combined to form one long grid. This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

Usage

bind(x, y, match = TRUE, index = TRUE)

Arguments

x

repgrid object

y

repgrid object

match

If the elements do not have the same order they are reordered to match the element order of the first grid 'x' (if test=TRUE, default). If set to FALSE an error occurs if the element order is not identical in both grids.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

if (FALSE) { # \dontrun{

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)

bindConstructs(a, b, m = F) # no binding
} # }