Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the Mr. then                   
#>                           back - 7 | 9 - initially he                 
#>  Harry is an essentially any - 6 | | | 10 - good                      
#>                   realizes - 5 | | | | | 11 - happened as the back hap
#> oney initially before is - 4 | | | | | | | 12 - the she               
#>          happened back - 3 | | | | | | | | | 13 - an without          
#>           the tricks - 2 | | | | | | | | | | | 14 - economic airplane 
#>         control is - 1 | | | | | | | | | | | | | 15 - any away        
#>                      | | | | | | | | | | | | | | |                    
#> what the the f (1)   3 3 3 4 1 1 2 2 1 1 3 5 5 5 2   (1) back         
#>           then (2)   1 3 3 4 5 4 3 3 4 3 4 3 5 1 5   (2) she then     
#>       his then (3)   4 3 2 1 1 1 3 4 4 3 1 5 5 1 1   (3) slipping char
#>  economic into (4)   4 5 2 1 4 4 3 1 1 5 2 3 3 5 1   (4) initially by 
#> is the then by (5)   5 1 1 1 5 1 5 5 4 3 4 2 3 5 1   (5) into good    
#> lly without as (6)   4 3 3 2 5 3 2 3 2 5 3 1 4 4 5   (6) before       
#>        him the (7)   1 5 3 4 1 5 3 5 5 3 2 1 1 1 5   (7) as back the H
#>     Brownstone (8)   4 5 5 3 3 4 3 4 1 5 4 2 4 2 2   (8) beyond I cust
#>         tricks (9)   3 4 3 5 4 2 2 4 2 5 4 1 2 5 1   (9) his the      
#> ket charter m (10)   3 5 5 3 2 4 2 3 4 4 2 4 1 5 1   (10) to his owes 
#> ing he tricks (11)   3 5 2 1 4 5 2 5 4 5 1 5 5 3 3   (11) before is es
#> picts novel o (12)   2 1 5 2 2 1 3 5 2 5 5 4 3 4 1   (12) Johnson away
#> lizes depicts (13)   2 2 4 3 5 5 4 1 4 2 4 3 1 3 3   (13) essentially 
#>  him tricks I (14)   2 1 2 4 2 3 1 1 5 3 1 5 3 4 2   (14) is Johnson t
#>  he Mr. happe (15)   3 4 2 5 3 1 2 5 5 5 5 1 1 4 4   (15) he depicts g
#>      airplane (16)   5 1 2 1 1 2 1 2 2 5 1 5 2 5 5   (16) has blackmar
#> ces control w (17)   5 3 3 3 3 4 5 1 5 2 1 1 4 2 3   (17) he man into 
#>        forces (18)   3 3 2 5 1 1 4 3 3 4 3 2 3 4 3   (18) to charter t
#> wes tricks Ma (19)   1 4 5 4 1 2 5 4 1 5 4 1 3 1 3   (19) the his she 
#> hing customer (20)   2 5 4 2 5 5 2 4 3 1 1 5 2 5 1   (20) his slipping
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the Mr. then                   
#>                           back - 7 | 9 - initially he                 
#>  Harry is an essentially any - 6 | | | 10 - good                      
#>                   realizes - 5 | | | | | 11 - happened as the back hap
#> oney initially before is - 4 | | | | | | | 12 - the she               
#>          happened back - 3 | | | | | | | | | 13 - an without          
#>           the tricks - 2 | | | | | | | | | | | 14 - economic airplane 
#>         control is - 1 | | | | | | | | | | | | | 15 - any away        
#>                      | | | | | | | | | | | | | | |                    
#> what the the f (1)   3 3 3 4 1 1 2 2 1 1 3 5 5 5 2   (1) back         
#>           then (2)   1 3 3 4 5 4 3 3 4 3 4 3 5 1 5   (2) she then     
#>       his then (3)   4 3 2 1 1 1 3 4 4 3 1 5 5 1 1   (3) slipping char
#>  economic into (4)   4 5 2 1 4 4 3 1 1 5 2 3 3 5 1   (4) initially by 
#> is the then by (5)   5 1 1 1 5 1 5 5 4 3 4 2 3 5 1   (5) into good    
#> lly without as (6)   4 3 3 2 5 3 2 3 2 5 3 1 4 4 5   (6) before       
#>        him the (7)   1 5 3 4 1 5 3 5 5 3 2 1 1 1 5   (7) as back the H
#>     Brownstone (8)   4 5 5 3 3 4 3 4 1 5 4 2 4 2 2   (8) beyond I cust
#>         tricks (9)   3 4 3 5 4 2 2 4 2 5 4 1 2 5 1   (9) his the      
#> ket charter m (10)   3 5 5 3 2 4 2 3 4 4 2 4 1 5 1   (10) to his owes 
#> ing he tricks (11)   3 5 2 1 4 5 2 5 4 5 1 5 5 3 3   (11) before is es
#> picts novel o (12)   2 1 5 2 2 1 3 5 2 5 5 4 3 4 1   (12) Johnson away
#> lizes depicts (13)   2 2 4 3 5 5 4 1 4 2 4 3 1 3 3   (13) essentially 
#>  him tricks I (14)   2 1 2 4 2 3 1 1 5 3 1 5 3 4 2   (14) is Johnson t
#>  he Mr. happe (15)   3 4 2 5 3 1 2 5 5 5 5 1 1 4 4   (15) he depicts g
#>      airplane (16)   5 1 2 1 1 2 1 2 2 5 1 5 2 5 5   (16) has blackmar
#> ces control w (17)   5 3 3 3 3 4 5 1 5 2 1 1 4 2 3   (17) he man into 
#>        forces (18)   3 3 2 5 1 1 4 3 3 4 3 2 3 4 3   (18) to charter t
#> wes tricks Ma (19)   1 4 5 4 1 2 5 4 1 5 4 1 3 1 3   (19) the his she 
#> hing customer (20)   2 5 4 2 5 5 2 4 3 1 1 5 2 5 1   (20) his slipping
#> hat the the f (21)   3 3 3 4 1 1 2 2 1 1 3 5 5 5 2   (21) back        
#>          then (22)   1 3 3 4 5 4 3 3 4 3 4 3 5 1 5   (22) she then    
#>      his then (23)   4 3 2 1 1 1 3 4 4 3 1 5 5 1 1   (23) slipping cha
#> economic into (24)   4 5 2 1 4 4 3 1 1 5 2 3 3 5 1   (24) initially by
#> s the then by (25)   5 1 1 1 5 1 5 5 4 3 4 2 3 5 1   (25) into good   
#> ly without as (26)   4 3 3 2 5 3 2 3 2 5 3 1 4 4 5   (26) before      
#>       him the (27)   1 5 3 4 1 5 3 5 5 3 2 1 1 1 5   (27) as back the 
#>    Brownstone (28)   4 5 5 3 3 4 3 4 1 5 4 2 4 2 2   (28) beyond I cus
#>        tricks (29)   3 4 3 5 4 2 2 4 2 5 4 1 2 5 1   (29) his the     
#> ket charter m (30)   3 5 5 3 2 4 2 3 4 4 2 4 1 5 1   (30) to his owes 
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the Mr. then                   
#>                           back - 7 | 9 - initially he                 
#>  Harry is an essentially any - 6 | | | 10 - good                      
#>                   realizes - 5 | | | | | 11 - happened as the back hap
#> oney initially before is - 4 | | | | | | | 12 - the she               
#>          happened back - 3 | | | | | | | | | 13 - an without          
#>           the tricks - 2 | | | | | | | | | | | 14 - economic airplane 
#>         control is - 1 | | | | | | | | | | | | | 15 - any away        
#>                      | | | | | | | | | | | | | | |                    
#> what the the f (1)   3 3 3 4 1 1 2 2 1 1 3 5 5 5 2   (1) back         
#>           then (2)   1 3 3 4 5 4 3 3 4 3 4 3 5 1 5   (2) she then     
#>       his then (3)   4 3 2 1 1 1 3 4 4 3 1 5 5 1 1   (3) slipping char
#>  economic into (4)   4 5 2 1 4 4 3 1 1 5 2 3 3 5 1   (4) initially by 
#> is the then by (5)   5 1 1 1 5 1 5 5 4 3 4 2 3 5 1   (5) into good    
#> lly without as (6)   4 3 3 2 5 3 2 3 2 5 3 1 4 4 5   (6) before       
#>        him the (7)   1 5 3 4 1 5 3 5 5 3 2 1 1 1 5   (7) as back the H
#>     Brownstone (8)   4 5 5 3 3 4 3 4 1 5 4 2 4 2 2   (8) beyond I cust
#>         tricks (9)   3 4 3 5 4 2 2 4 2 5 4 1 2 5 1   (9) his the      
#> ket charter m (10)   3 5 5 3 2 4 2 3 4 4 2 4 1 5 1   (10) to his owes 
#> hat the the f (11)   3 3 3 4 1 1 2 2 1 1 3 5 5 5 2   (11) back        
#>          then (12)   1 3 3 4 5 4 3 3 4 3 4 3 5 1 5   (12) she then    
#>      his then (13)   4 3 2 1 1 1 3 4 4 3 1 5 5 1 1   (13) slipping cha
#> economic into (14)   4 5 2 1 4 4 3 1 1 5 2 3 3 5 1   (14) initially by
#> s the then by (15)   5 1 1 1 5 1 5 5 4 3 4 2 3 5 1   (15) into good   
#> ly without as (16)   4 3 3 2 5 3 2 3 2 5 3 1 4 4 5   (16) before      
#>       him the (17)   1 5 3 4 1 5 3 5 5 3 2 1 1 1 5   (17) as back the 
#>    Brownstone (18)   4 5 5 3 3 4 3 4 1 5 4 2 4 2 2   (18) beyond I cus
#>        tricks (19)   3 4 3 5 4 2 2 4 2 5 4 1 2 5 1   (19) his the     
#> ket charter m (20)   3 5 5 3 2 4 2 3 4 4 2 4 1 5 1   (20) to his owes 
#> ing he tricks (21)   3 5 2 1 4 5 2 5 4 5 1 5 5 3 3   (21) before is es
#> picts novel o (22)   2 1 5 2 2 1 3 5 2 5 5 4 3 4 1   (22) Johnson away
#> lizes depicts (23)   2 2 4 3 5 5 4 1 4 2 4 3 1 3 3   (23) essentially 
#>  him tricks I (24)   2 1 2 4 2 3 1 1 5 3 1 5 3 4 2   (24) is Johnson t
#>  he Mr. happe (25)   3 4 2 5 3 1 2 5 5 5 5 1 1 4 4   (25) he depicts g
#>      airplane (26)   5 1 2 1 1 2 1 2 2 5 1 5 2 5 5   (26) has blackmar
#> ces control w (27)   5 3 3 3 3 4 5 1 5 2 1 1 4 2 3   (27) he man into 
#>        forces (28)   3 3 2 5 1 1 4 3 3 4 3 2 3 4 3   (28) to charter t
#> wes tricks Ma (29)   1 4 5 4 1 2 5 4 1 5 4 1 3 1 3   (29) the his she 
#> hing customer (30)   2 5 4 2 5 5 2 4 3 1 1 5 2 5 1   (30) his slipping
#>