Draw grid in rgl (3D device).

The 3D biplot opens an interactive 3D device that can be rotated and zoomed using the mouse. A 3D device facilitates the exploration of grid data as significant proportions of the sum-of-squares are often represented beyond the first two dimensions. Also, in a lot of cases it may be of interest to explore the grid space from a certain angle, e.g. to gain an optimal view onto the set of elements under investigation (e.g. Raeithel, 1998).

Usage

biplot3d(
  x,
  dim = 1:3,
  labels.e = TRUE,
  labels.c = TRUE,
  lines.c = TRUE,
  lef = 1.3,
  center = 1,
  normalize = 0,
  g = 0,
  h = 1,
  col.active = NA,
  col.passive = NA,
  c.sphere.col = grey(0.4),
  c.cex = 0.6,
  c.text.col = grey(0.4),
  e.sphere.col = grey(0),
  e.cex = 0.6,
  e.text.col = grey(0),
  alpha.sphere = 0.05,
  col.sphere = "black",
  unity = FALSE,
  unity3d = FALSE,
  scale.e = 0.9,
  zoom = 1,
  ...
)

Arguments

x

repgrid object.

dim

Dimensions to display.

labels.e

Logical. whether element labels are displayed.

labels.c

Logical. whether construct labels are displayed.

lines.c

Numeric. The way lines are drawn through the construct vectors. 0 = no lines, 1 = lines from constructs to outer frame, 2 = lines from the center to outer frame.

lef

Construct lines extension factor

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). Default is 1 (row centering).

normalize

A numeric value indicating along what direction (rows, columns) to normalize by standard deviations. 0 = none, 1= rows, 2 = columns (default is 0).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

col.active

Columns (elements) that are no supplementary points, i.e. they are used in the SVD to find principal components. default is to use all elements.

col.passive

Columns (elements) that are supplementary points, i.e. they are NOT used in the SVD but projected into the component space afterwards. They do not determine the solution. Default is NA, i.e. no elements are set supplementary.

c.sphere.col

Color of construct spheres.

c.cex

Size of construct text.

c.text.col

Color for construct text.

e.sphere.col

Color of elements.

e.cex

Size of element labels.

e.text.col

Color of element labels.

alpha.sphere

Numeric. alpha blending of the surrounding sphere (default".05").

col.sphere

Color of surrounding sphere (default"black").

unity

Scale elements and constructs coordinates to unit scale (maximum of 1) so they are printed more neatly (default TRUE).

unity3d

To come.

scale.e

Scaling factor for element vectors. Will cause element points to move a bit more to the center (but only if unity or unity3d is TRUE). This argument is for visual appeal only.

zoom

Not yet used. Scaling factor for all vectors. Can be used to zoom the plot in and out (default 1).

...

Parameters to be passed on.

References

Raeithel, A. (1998). Kooperative Modellproduktion von Professionellen und Klienten - erlauetert am Beispiel des Repertory Grid. Selbstorganisation, Kooperation, Zeichenprozess: Arbeiten zu einer kulturwissenschaftlichen, anwendungsbezogenen Psychologie (pp. 209-254). Opladen: Westdeutscher Verlag.

See also

Examples

if (FALSE) {

biplot3d(boeker)
biplot3d(boeker, unity3d = T)

biplot3d(boeker,
  e.sphere.col = "red",
  c.text.col = "blue"
)
biplot3d(boeker, e.cex = 1)
biplot3d(boeker, col.sphere = "red")

biplot3d(boeker, g = 1, h = 1) # INGRID biplot
biplot3d(boeker,
  g = 1, h = 1, # ESA biplot
  center = 4
)
}