Draw the eigenstructure analysis (ESA) biplot in rgl (3D device).

The 3D biplot opens an interactive 3D device that can be rotated and zoomed using the mouse. A 3D device facilitates the exploration of grid data as significant proportions of the sum-of-squares are often represented beyond the first two dimensions. Also, in a lot of cases it may be of interest to explore the grid space from a certain angle, e.g. to gain an optimal view onto the set of elements under investigation (e.g. Raeithel, 1998). Note that the eigenstructure analysis just a special case of a biplot that can also be produced using the biplot3d() function with the arguments center=4, g=1, h=1.

Usage

biplotEsa3d(x, center = 1, g = 1, h = 1, ...)

Arguments

x

repgrid object.

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). Default is 4 (scale midpoint centering).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

...

Additional arguments to be passed to biplot3d().

See also

Examples

if (FALSE) { # \dontrun{

biplotEsa3d(boeker)
biplotEsa3d(boeker, unity3d = T)

biplotEsa3d(boeker,
  e.sphere.col = "red",
  c.text.col = "blue"
)
biplotEsa3d(boeker, e.cex = 1)
biplotEsa3d(boeker, col.sphere = "red")
} # }