Draws a biplot of the grid in 2D with depth impression (pseudo 3D).

This version is basically a 2D biplot. It only modifies color and size of the symbols in order to create a 3D impression of the data points. This function will call the standard biplot2d() function with some modified arguments. For the whole set of arguments that can be used see biplot2d(). Here only the arguments special to biplotPseudo3d are outlined.

Usage

biplotPseudo3d(
  x,
  dim = 1:2,
  map.dim = 3,
  e.point.col = c("white", "black"),
  e.point.cex = c(0.6, 1.2),
  e.label.col = c("white", "black"),
  e.label.cex = c(0.6, 0.8),
  e.color.map = c(0.4, 1),
  c.point.col = c("white", "darkred"),
  c.point.cex = c(0.6, 1.2),
  c.label.col = c("white", "darkred"),
  c.label.cex = c(0.6, 0.8),
  c.color.map = c(0.4, 1),
  ...
)

Arguments

x

repgrid object.

dim

Dimensions (i.e. principal components) to be used for biplot (default is c(1,2)).

map.dim

Third dimension (depth) used to map aesthetic attributes to (default is 3).

e.point.col

Color(s) of the element symbols. Two values can be entered that will create a color ramp. The values of map.dim are mapped onto the ramp. The default is c("white", "black"). If only one color color value is supplied (e.g. "black") no mapping occurs and all elements will have the same color irrespective of their value on the map.dim dimension.

e.point.cex

Size of the element symbols. Two values can be entered that will represents the lower and upper size of a range of cex the values of map.dim are mapped onto. The default is c(.6, 1.2). If only one cex value is supplied (e.g. .7) no mapping occurs and all elements will have the same size irrespective of their value on the map.dim dimension.

e.label.col

Color(s) of the element labels. Two values can be entered that will create a color ramp. The values of map.dim are mapped onto the ramp. The default is c("white", "black"). If only one color color value is supplied (e.g. "black") no mapping occurs and all element labels will have the same color irrespective of their value on the map.dim dimension.

e.label.cex

Size of the element labels. Two values can be entered that will represents the lower and upper size of a range of cex the values of map.dim are mapped onto. The default is c(.6, .8). If only one cex value is supplied (e.g. .7) no mapping occurs and all element labels will have the same size irrespective of their value on the map.dim dimension.

e.color.map

Value range to determine what range of the color ramp defined in e.color will be used for mapping the colors. Default is c(.4, ,1). Usually not important for the user.

c.point.col

Color(s) of the construct symbols. Two values can be entered that will create a color ramp. The values of map.dim are mapped onto the ramp. The default is c("white", "darkred"). If only one color color value is supplied (e.g. "black") no mapping occurs and all elements will have the same color irrespective of their value on the map.dim dimension.

c.point.cex

Size of the construct symbols. Two values can be entered that will represents the lower and upper size of a range of cex the values of map.dim are mapped onto. The default is c(.6, 1.2). If only one cex value is supplied (e.g. .7) no mapping occurs and all elements will have the same size irrespective of their value on the map.dim dimension.

c.label.col

Color(s) of the construct labels. Two values can be entered that will create a color ramp. The values of map.dim are mapped onto the ramp. The default is c("white", "black"). If only one color color value is supplied (e.g. "black") no mapping occurs and all construct labels will have the same color irrespective of their value on the map.dim dimension.

c.label.cex

Size of the construct labels. Two values can be entered that will represents the lower and upper size of a range of cex the values of map.dim are mapped onto. The default is c(.6, .9). If only one cex value is supplied (e.g. .7) no mapping occurs and all construct labels will have the same size irrespective of their value on the map.dim dimension.

c.color.map

Value range to determine what range of the color ramp defined in c.color will be used for mapping. Default is c(.4, ,1). Usually not important for the user.

...

Additional parameters passed to biplot2d().

See also

Examples

if (FALSE) { # \dontrun{
# biplot with 3D impression
biplotPseudo3d(boeker)
# Slater's biplot with 3D impression
biplotPseudo3d(boeker, g = 1, h = 1, center = 1)

# show 2nd and 3rd dim. and map 4th
biplotPseudo3d(boeker, dim = 2:3, map.dim = 4)

# change elem. colors
biplotPseudo3d(boeker, e.color = c("white", "darkgreen"))
# change con. colors
biplotPseudo3d(boeker, c.color = c("white", "darkgreen"))
# change color mapping range
biplotPseudo3d(boeker, c.colors.map = c(0, 1))

# set uniform con. text size
biplotPseudo3d(boeker, c.cex = 1)
# change text size mapping range
biplotPseudo3d(boeker, c.cex = c(.4, 1.2))
} # }