A graphically unsophisticated version of a biplot.

It will draw elements and constructs vectors using similar arguments as biplot2d(). It is a version for quick exploration used during development.

Usage

biplotSimple(
  x,
  dim = 1:2,
  center = 1,
  normalize = 0,
  g = 0,
  h = 1 - g,
  unity = T,
  col.active = NA,
  col.passive = NA,
  scale.e = 0.9,
  zoom = 1,
  e.point.col = "black",
  e.point.cex = 1,
  e.label.col = "black",
  e.label.cex = 0.7,
  c.point.col = grey(0.6),
  c.label.col = grey(0.6),
  c.label.cex = 0.6,
  ...
)

Arguments

x

repgrid object.

dim

Dimensions (i.e. principal components) to be used for biplot (default is c(1,2)).

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). The default is 1 (row centering).

normalize

A numeric value indicating along what direction (rows, columns) to normalize by standard deviations. 0 = none, 1= rows, 2 = columns (default is 0).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

unity

Scale elements and constructs coordinates to unit scale in 2D (maximum of 1) so they are printed more neatly (default TRUE).

col.active

Columns (elements) that are no supplementary points, i.e. they are used in the SVD to find principal components. default is to use all elements.

col.passive

Columns (elements) that are supplementary points, i.e. they are NOT used in the SVD but projected into the component space afterwards. They do not determine the solution. Default is NA, i.e. no elements are set supplementary.

scale.e

Scaling factor for element vectors. Will cause element points to move a bit more to the center. This argument is for visual appeal only.

zoom

Scaling factor for all vectors. Can be used to zoom the plot in and out (default 1).

e.point.col

Color of the element symbols (default is "black".

e.point.cex

Size of the element symbol (default is 1.

e.label.col

Color of the element labels (default is "black".

e.label.cex

Size of the element labels (default is .7.

c.point.col

Color of the construct lines (default is grey(.6).

c.label.col

Color of the construct labels (default is grey(.6).

c.label.cex

Size of the construct labels (default is .6.

...

Parameters to be passed on to center() and normalize.

Value

repgrid object.

See also

Examples

if (FALSE) { # \dontrun{

biplotSimple(boeker)
biplotSimple(boeker, unity = F)

biplotSimple(boeker, g = 1, h = 1) # INGRID biplot
biplotSimple(boeker, g = 1, h = 1, center = 4) # ESA biplot

biplotSimple(boeker, zoom = .9) # zooming out
biplotSimple(boeker, scale.e = .6) # scale element vectors

biplotSimple(boeker, e.point.col = "brown") # change colors
biplotSimple(boeker,
  e.point.col = "brown",
  c.label.col = "darkblue"
)
} # }