Draws Slater's INGRID biplot in 2D.

The default is to use row centering and no normalization. Note that Slater's biplot is just a special case of a biplot that can be produced using the biplot2d() function with the arguments center=1, g=1, h=1. The arguments that can be used in this function are the same as in biplot2d(). Here, only the arguments that are set for Slater's biplot are described. To see all the parameters that can be changed see biplot2d().

Usage

biplotSlater2d(x, center = 1, g = 1, h = 1, ...)

Arguments

x

repgrid object.

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). Slater's biplot uses 1 (row centering).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

...

Additional parameters for be passed to biplot2d().

See also

Examples

if (FALSE) { # \dontrun{
# See examples in [biplot2d()] as the same arguments
# can used for this function.
} # }