Concatenate the elements of two grids

Concatenate the elements of two grids

Usage

# S3 method for class 'repgrid'
cbind(..., .reorder = TRUE, .unique = FALSE)

Arguments

...

repgrid objects or list of objects.

.reorder

If TRUE (default), matches construct order of y to x.

.unique

If FALSE (default), x and y may have common elements. If FALSE, they must be mutually exclusive.

Value

repgrid with combined elements.

Examples

x <- boeker[, 1:2]
y <- boeker[, 5:7]
cbind(x, y)
#> 
#> META DATA:
#> Number of elements:  5 
#> Number of constructs:  14 
#> Preferred poles defined:  0/14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                               3 - kurt                      
#>                ideal self - 2 | 4 - karl                    
#>                    self - 1 | | | 5 - george                
#>                           | | | | |                         
#>           balanced (.1)   1 4 3 5 2   (1.) get along with co
#>           isolated (.2)   3 6 5 4 5   (2.) sociable         
#> closely integrated (.3)   2 2 5 3 2   (3.) excluded         
#>         discursive (.4)   4 1 2 4 2   (4.) passive          
#>        open minded (.5)   2 1 2 4 4   (5.) indifferent      
#>             dreamy (.6)   4 5 4 5 4   (6.) dispassionate    
#> actically oriented (.7)   2 1 3 3 3   (7.) depressed        
#>            playful (.8)   4 5 4 3 2   (8.) serious          
#>    socially minded (.9)   2 1 4 5 4   (9.) selfish          
#>       quarrelsome (.10)   5 5 5 2 5   (10.) peaceful        
#>          artistic (.11)   5 1 3 5 3   (11.) technical       
#>        scientific (.12)   2 1 4 4 5   (12.) emotional       
#>         introvert (.13)   4 5 5 3 5   (13.) extrovert       
#>        wanderlust (.14)   1 1 4 5 2   (14.) home oriented   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
x / y
#> 
#> META DATA:
#> Number of elements:  5 
#> Number of constructs:  14 
#> Preferred poles defined:  0/14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                               3 - kurt                      
#>                ideal self - 2 | 4 - karl                    
#>                    self - 1 | | | 5 - george                
#>                           | | | | |                         
#>           balanced (.1)   1 4 3 5 2   (1.) get along with co
#>           isolated (.2)   3 6 5 4 5   (2.) sociable         
#> closely integrated (.3)   2 2 5 3 2   (3.) excluded         
#>         discursive (.4)   4 1 2 4 2   (4.) passive          
#>        open minded (.5)   2 1 2 4 4   (5.) indifferent      
#>             dreamy (.6)   4 5 4 5 4   (6.) dispassionate    
#> actically oriented (.7)   2 1 3 3 3   (7.) depressed        
#>            playful (.8)   4 5 4 3 2   (8.) serious          
#>    socially minded (.9)   2 1 4 5 4   (9.) selfish          
#>       quarrelsome (.10)   5 5 5 2 5   (10.) peaceful        
#>          artistic (.11)   5 1 3 5 3   (11.) technical       
#>        scientific (.12)   2 1 4 4 5   (12.) emotional       
#>         introvert (.13)   4 5 5 3 5   (13.) extrovert       
#>        wanderlust (.14)   1 1 4 5 2   (14.) home oriented   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
y_reordered <- y[sample(nrow(y)), ]
cbind(x, y, y_reordered)
#> 
#> META DATA:
#> Number of elements:  8 
#> Number of constructs:  14 
#> Preferred poles defined:  0/14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                          karl - 4 5 - george                      
#>                        kurt - 3 | | 6 - kurt                      
#>                ideal self - 2 | | | | 7 - karl                    
#>                    self - 1 | | | | | | 8 - george                
#>                           | | | | | | | |                         
#>           balanced (.1)   1 4 3 5 2 3 5 2   (1.) get along with co
#>           isolated (.2)   3 6 5 4 5 5 4 5   (2.) sociable         
#> closely integrated (.3)   2 2 5 3 2 5 3 2   (3.) excluded         
#>         discursive (.4)   4 1 2 4 2 2 4 2   (4.) passive          
#>        open minded (.5)   2 1 2 4 4 2 4 4   (5.) indifferent      
#>             dreamy (.6)   4 5 4 5 4 4 5 4   (6.) dispassionate    
#> actically oriented (.7)   2 1 3 3 3 3 3 3   (7.) depressed        
#>            playful (.8)   4 5 4 3 2 4 3 2   (8.) serious          
#>    socially minded (.9)   2 1 4 5 4 4 5 4   (9.) selfish          
#>       quarrelsome (.10)   5 5 5 2 5 5 2 5   (10.) peaceful        
#>          artistic (.11)   5 1 3 5 3 3 5 3   (11.) technical       
#>        scientific (.12)   2 1 4 4 5 4 4 5   (12.) emotional       
#>         introvert (.13)   4 5 5 3 5 5 3 5   (13.) extrovert       
#>        wanderlust (.14)   1 1 4 5 2 4 5 2   (14.) home oriented   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined