Principal component analysis (PCA) of inter-construct correlations.

Various methods for rotation and methods for the calculation of the correlations are available. Note that the number of factors has to be specified. For more information on the PCA function itself type ?principal.

Usage

constructPca(
  x,
  nfactors = 3,
  rotate = "varimax",
  method = "pearson",
  trim = NA
)

Arguments

x

repgrid object.

nfactors

Number of components to extract (default is 3).

rotate

"none", "varimax", "promax" and "cluster" are possible rotations (default is none).

method

A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall" or "spearman", can be abbreviated. The default is "pearson".

trim

The number of characters a construct is trimmed to (default is 7). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs with long names.

Value

Returns an object of class constructPca.

References

Fransella, F., Bell, R. & Bannister, D. (2003). A Manual for Repertory Grid Technique (2. Ed.). Chichester: John Wiley & Sons.

See also

Examples

constructPca(bell2010)
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 3
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                                    RC1   RC2   RC3  
#> relaxed - worried & tense                          -0.26  0.14 -0.90
#> not so smart (academically) - smart (academically)  0.26 -0.87  0.01
#> dislikes sports - loves sports                      0.20  0.67 -0.10
#> not interactive - loves people                      0.06  0.09  0.98
#> not transparent - transparent                      -0.04  0.92  0.08
#> insensitive - sensitive                            -0.81  0.36  0.17
#> fearful&timid - fearless                            0.90 -0.10  0.29
#> rough - gentle                                     -0.79 -0.03 -0.49
#> accept as it is - loves to argue                    0.72  0.26  0.10
#> 
#>                 RC1  RC2  RC3
#> SS loadings    2.79 2.27 2.14
#> Proportion Var 0.31 0.25 0.24
#> Cumulative Var 0.31 0.56 0.80

# data from grid manual by Fransella et al. (2003, p. 87)
# note that the construct order is different
constructPca(fbb2003, nfactors = 2)
#> Warning: Matrix was not positive definite, smoothing was done
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 2
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                    RC1   RC2  
#> clever - not bright                 0.98  0.13
#> disorganized - organized           -0.79 -0.40
#> listens - doesn't hear              0.95 -0.17
#> no clear view - clear view of life -0.57 -0.54
#> understands me - no understanding   0.84 -0.13
#> ambitious - no ambition             0.20  0.64
#> respected - not respected           0.93  0.09
#> distant - warm                     -0.16  0.75
#> rather aggressive - not aggressive -0.03  0.79
#> 
#>                 RC1  RC2
#> SS loadings    4.47 2.13
#> Proportion Var 0.50 0.24
#> Cumulative Var 0.50 0.73

# no rotation
constructPca(fbb2003, rotate = "none")
#> Warning: Matrix was not positive definite, smoothing was done
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 3
#> Type of rotation: none 
#> 
#> Loadings:
#>                                    PC1   PC2   PC3  
#> clever - not bright                 0.99 -0.09  0.03
#> disorganized - organized           -0.86 -0.22 -0.28
#> listens - doesn't hear              0.89 -0.37 -0.03
#> no clear view - clear view of life -0.68 -0.41  0.42
#> understands me - no understanding   0.79 -0.31  0.22
#> ambitious - no ambition             0.33  0.58 -0.68
#> respected - not respected           0.93 -0.12  0.02
#> distant - warm                      0.01  0.77  0.18
#> rather aggressive - not aggressive  0.15  0.78  0.55
#> 
#>                 PC1  PC2  PC3
#> SS loadings    4.60 2.00 1.10
#> Proportion Var 0.51 0.22 0.12
#> Cumulative Var 0.51 0.73 0.86

# use a different type of correlation (Spearman)
constructPca(fbb2003, method = "spearman")
#> Warning: NaNs produced
#> Warning: NaNs produced
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 3
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                    RC1   RC2   RC3  
#> clever - not bright                 0.97 -0.02  0.21
#> disorganized - organized           -0.82 -0.47 -0.09
#> listens - doesn't hear              0.95 -0.16  0.17
#> no clear view - clear view of life -0.47 -0.23 -0.75
#> understands me - no understanding   0.82 -0.23  0.00
#> ambitious - no ambition             0.06  0.09  0.95
#> respected - not respected           0.89 -0.02  0.21
#> distant - warm                     -0.10  0.76  0.23
#> rather aggressive - not aggressive -0.04  0.97  0.01
#> 
#>                 RC1  RC2  RC3
#> SS loadings    4.20 1.87 1.65
#> Proportion Var 0.47 0.21 0.18
#> Cumulative Var 0.47 0.67 0.86

# save output to object
m <- constructPca(fbb2003, nfactors = 2)
#> Warning: Matrix was not positive definite, smoothing was done
m
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 2
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                    RC1   RC2  
#> clever - not bright                 0.98  0.13
#> disorganized - organized           -0.79 -0.40
#> listens - doesn't hear              0.95 -0.17
#> no clear view - clear view of life -0.57 -0.54
#> understands me - no understanding   0.84 -0.13
#> ambitious - no ambition             0.20  0.64
#> respected - not respected           0.93  0.09
#> distant - warm                     -0.16  0.75
#> rather aggressive - not aggressive -0.03  0.79
#> 
#>                 RC1  RC2
#> SS loadings    4.47 2.13
#> Proportion Var 0.50 0.24
#> Cumulative Var 0.50 0.73

# different printing options
print(m, digits = 5)
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 2
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                    RC1      RC2     
#> clever - not bright                 0.98491  0.13119
#> disorganized - organized           -0.79127 -0.40068
#> listens - doesn't hear              0.95289 -0.16812
#> no clear view - clear view of life -0.57148 -0.54458
#> understands me - no understanding   0.83965 -0.12645
#> ambitious - no ambition             0.19932  0.63501
#> respected - not respected           0.93351  0.08899
#> distant - warm                     -0.15730  0.75377
#> rather aggressive - not aggressive -0.02937  0.79235
#> 
#>                    RC1     RC2
#> SS loadings    4.47254 2.12573
#> Proportion Var 0.49695 0.23619
#> Cumulative Var 0.49695 0.73314
print(m, cutoff = .3)
#> 
#> #################
#> PCA of constructs
#> #################
#> 
#> Number of components extracted: 2
#> Type of rotation: varimax 
#> 
#> Loadings:
#>                                    RC1   RC2  
#> clever - not bright                 0.98      
#> disorganized - organized           -0.79 -0.40
#> listens - doesn't hear              0.95      
#> no clear view - clear view of life -0.57 -0.54
#> understands me - no understanding   0.84      
#> ambitious - no ambition                   0.64
#> respected - not respected           0.93      
#> distant - warm                            0.75
#> rather aggressive - not aggressive        0.79
#> 
#>                 RC1  RC2
#> SS loadings    4.47 2.13
#> Proportion Var 0.50 0.24
#> Cumulative Var 0.50 0.73