Root mean square (RMS) of inter-construct correlations.

The RMS is also known as 'quadratic mean' of the inter-construct correlations. The RMS serves as a simplification of the correlation table. It reflects the average relation of one construct to all other constructs. Note that as the correlations are squared during its calculation, the RMS is not affected by the sign of the correlation (cf. Fransella, Bell & Bannister, 2003, p. 86).

Usage

constructRmsCor(x, method = "pearson", trim = NA)

Arguments

x

repgrid object

method

A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall" or "spearman", can be abbreviated. The default is "pearson".

trim

The number of characters a construct is trimmed to (default is NA). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs with long names.

Value

dataframe of the RMS of inter-construct correlations

References

Fransella, F., Bell, R. C., & Bannister, D. (2003). A Manual for Repertory Grid Technique (2. Ed.). Chichester: John Wiley & Sons.

See also

Examples

# data from grid manual by Fransella, Bell and Bannister
constructRmsCor(fbb2003)
#> 
#> ##########################################
#> Root-mean-square correlation of constructs
#> ##########################################
#> 
#>                                         RMS
#> (1) clever - not bright                0.66
#> (2) disorganized - organized           0.58
#> (3) listens - doesn't hear             0.61
#> (4) no clear view - clear view of life 0.46
#> (5) understands me - no understanding  0.53
#> (6) ambitious - no ambition            0.30
#> (7) respected - not respected          0.62
#> (8) distant - warm                     0.25
#> (9) rather aggressive - not aggressive 0.29
#> 
#> Average of statistic 0.48 
#> Standard deviation of statistic 0.15 
constructRmsCor(fbb2003, trim = 20)
#> 
#> ##########################################
#> Root-mean-square correlation of constructs
#> ##########################################
#> 
#>                              RMS
#> (1) clever - not bright     0.66
#> (2) disorganiz - organized  0.58
#> (3) listens - doesn't he    0.61
#> (4) no clear v - clear view 0.46
#> (5) understand - no underst 0.53
#> (6) ambitious - no ambitio  0.30
#> (7) respected - not respec  0.62
#> (8) distant - warm          0.25
#> (9) rather agg - not aggres 0.29
#> 
#> Average of statistic 0.48 
#> Standard deviation of statistic 0.15 

# modify output
r <- constructRmsCor(fbb2003)
print(r, digits = 5)
#> 
#> ##########################################
#> Root-mean-square correlation of constructs
#> ##########################################
#> 
#>                                            RMS
#> (1) clever - not bright                0.66354
#> (2) disorganized - organized           0.57855
#> (3) listens - doesn't hear             0.61088
#> (4) no clear view - clear view of life 0.46125
#> (5) understands me - no understanding  0.53241
#> (6) ambitious - no ambition            0.29853
#> (7) respected - not respected          0.61928
#> (8) distant - warm                     0.24817
#> (9) rather aggressive - not aggressive 0.29197
#> 
#> Average of statistic 0.47829 
#> Standard deviation of statistic 0.15093 
   # access calculation results
   r[2, 1]
#> [1] 0.5785465