Grid data from Boeker (1996).

Grid data from a schizophrenic patient undergoing psychoanalytically oriented psychotherapy. The data was taken during the last stage of therapy (Boeker, 1996, p. 163).

References

Boeker, H. (1996). The reconstruction of the self in the psychotherapy of chronic schizophrenia: a case study with the Repertory Grid Technique. In: Scheer, J. W., Catina, A. (Eds.): Empirical Constructivism in Europe - The Personal Construct Approach (p. 160-167). Giessen: Psychosozial-Verlag.

Examples

boeker
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#>       balanced (1)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (1) get along wit
#>       isolated (2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2) sociable     
#> ely integrated (3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3) excluded     
#>     discursive (4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4) passive      
#>    open minded (5)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (5) indifferent  
#>         dreamy (6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6) dispassionate
#> cally oriented (7)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (7) depressed    
#>        playful (8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8) serious      
#> ocially minded (9)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (9) selfish      
#>   quarrelsome (10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10) peaceful    
#>      artistic (11)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (11) technical   
#>    scientific (12)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (12) emotional   
#>     introvert (13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13) extrovert   
#>    wanderlust (14)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (14) home oriente
#>