Grid data from Feixas and Saul (2004).

A description by the authors: "When Teresa, 22 years old, was seen by the second author (LAS) at the psychological services of the University of Salamanca, she was in the final year of her studies in chemical sciences. Although Teresa proves to be an excellent student, she reveals serious doubts about her self worth. She cries frequently, and has great difficulty in meeting others, even though she has a boyfriend who is extremely supportive. Teresa is anxiously hesitant about accepting a new job which would involve moving to another city 600 Km away from home." (Feixas & Saul, 2004, p. 77).

References

Feixas, G., & Saul, L. A. (2004). The Multi-Center Dilemma Project: an investigation on the role of cognitive conflicts in health. The Spanish Journal of Psychology, 7(1), 69-78.

Examples

feixas2004
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Friend 2                     
#>                     Friend 1 - 6 | 8 - Non-grata                  
#>                  Boyfriend - 5 | | | 9 - Friend 3                 
#>                  Brother - 4 | | | | | 10 - Cousin                
#>                 Father - 3 | | | | | | | 11 - Godmother           
#>               Mother - 2 | | | | | | | | | 12 - Friend 4          
#>           Self now - 1 | | | | | | | | | | | 13 - Ideal Self      
#>                      | | | | | | | | | | | | |                    
#>    Pessimistic (1)   1 1 5 2 7 3 6 2 6 4 3 2 7   (1) Optimistic   
#> Self-demanding (2)   1 6 6 2 2 5 6 3 5 6 4 5 4   (2) Takes it easy
#>        Fearful (3)   2 2 6 2 4 5 6 5 2 3 4 5 5   (3) Enterprising 
#>  Lives to work (4)   5 1 2 2 6 6 6 1 6 7 6 6 7   (4) Works to live
#> his/her wishes (5)   6 2 1 1 4 3 6 1 7 3 4 2 7   (5) Tolerant with
#>        Teasing (6)   2 7 1 6 4 3 4 6 3 3 5 6 3   (6) Touchy       
#> eciates others (7)   2 6 6 6 1 5 4 7 4 2 2 5 1   (7) Does not appr
#>     Aggressive (8)   6 4 2 2 7 4 6 2 6 6 6 3 7   (8) Calm         
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2   (9) Selfish      
#>    Avaricious (10)   6 1 1 1 7 5 5 1 6 3 3 6 7   (10) Generous    
#>     Sensitive (11)   1 5 7 7 1 4 5 7 1 4 3 4 1   (11) Materialisti
#>        Cheeky (12)   6 6 5 4 6 6 6 1 6 5 6 5 7   (12) Respectful  
#>  Hypocritical (13)   5 4 4 2 6 5 5 1 6 6 5 4 7   (13) Sincere     
#>   Blackmailer (14)   3 2 2 1 5 6 6 1 6 6 6 3 7   (14) Non blackmai
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6   (15) Natural     
#> er the friend (16)   6 3 3 3 6 2 1 2 4 4 6 4 7   (16) Looks after 
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7   (17) Accessible  
#>   Introverted (18)   1 2 6 2 4 5 7 5 2 6 6 5 5   (18) Extroverted 
#> ressed easily (19)   1 2 6 3 6 3 7 6 1 3 3 3 6   (19) Does not get
#>  good in thin (20)   6 6 4 6 1 5 2 7 6 3 3 5 1   (20) Sees only th
#>