Pre- and post therapy dataset from Leach et al. (2001).

Case as described by the authors: "Sarah, aged 32, was referred with problems of depression and sexual difficulties relating to childhood sexual abuse. She had three children and was living with her male partner. From the age of 9, her brother, an adult, had sexually abused Sarah. She attended a group for survivors of child sexual abuse and completed repertory grids prior to the group, immediately after the group and at 3- and 6-month follow-up." (Leach et al. 2001, p. 230).

Details

leach2001a is the pre-therapy, leach2001b is the post-therapy therapy dataset. The construct and elements are identical.

References

Leach, C., Freshwater, K., Aldridge, J., & Sunderland, J. (2001). Analysis of repertory grids in clinical practice. The British Journalof Clinical Psychology, 40, 225-248.

Examples

leach2001a
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                              5 - Father                   
#>           Men in general - 4 | 6 - Partner                
#>       Women in general - 3 | | | 7 - Ideal self           
#>             Self now - 2 | | | | | 8 - Mother             
#>         Child self - 1 | | | | | | | 9 - Abuser in childho
#>                      | | | | | | | | |                    
#>      assertive (1)   2 7 4 2 3 5 3 1 1   (1) not assertive
#>      confident (2)   1 7 3 2 2 4 2 1 1   (2) unconfident  
#> ot feel guilty (3)   1 6 4 2 1 1 1 1 1   (3) feels guilty 
#>        abusive (4)   7 7 4 6 7 6 7 3 1   (4) not abusive  
#>    frightening (5)   7 7 4 5 7 7 7 3 2   (5) not frighteni
#>  untrustworthy (6)   7 7 6 5 7 7 7 3 1   (6) trustworthy  
#>       powerful (7)   7 5 4 2 3 5 2 1 1   (7) powerless    
#>     big headed (8)   7 5 4 2 6 6 4 2 1   (8) not big heade
#>    independent (9)   5 6 3 2 2 4 1 3 1   (9) dependent    
#>     confusing (10)   7 2 4 4 7 6 7 1 2   (10) not confusin
#>        guilty (11)   7 3 4 4 7 6 7 4 1   (11) not guilty  
#>          cold (12)   7 3 5 4 7 7 6 2 6   (12) shows feelin
#>     masculine (13)   7 7 5 1 1 2 5 2 1   (13) feminine    
#> rested in sex (14)   7 5 3 1 1 1 2 7 1   (14) not interest
#> 
leach2001b
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                              5 - Father                   
#>           Men in general - 4 | 6 - Partner                
#>       Women in general - 3 | | | 7 - Ideal self           
#>             Self now - 2 | | | | | 8 - Mother             
#>         Child self - 1 | | | | | | | 9 - Abuser in childho
#>                      | | | | | | | | |                    
#>      assertive (1)   4 5 5 3 6 6 2 1 1   (1) not assertive
#>      confident (2)   3 6 4 3 3 5 1 1 1   (2) unconfident  
#> ot feel guilty (3)   2 4 4 2 1 2 1 1 1   (3) feels guilty 
#>        abusive (4)   7 5 4 4 7 7 7 3 1   (4) not abusive  
#>  untrustworthy (5)   7 7 6 5 7 7 7 3 1   (5) trustworthy  
#>         guilty (6)   7 7 4 4 7 7 7 5 1   (6) not guilty   
#>     big headed (7)   6 6 4 4 7 5 7 4 1   (7) not big heade
#>    frightening (8)   7 6 4 4 7 7 7 2 4   (8) not frighteni
#>           cold (9)   5 6 6 4 7 7 7 2 5   (9) shows feeling
#>      powerful (10)   6 3 3 2 3 5 1 1 1   (10) powerless   
#>     confusing (11)   7 3 6 6 6 6 7 1 3   (11) not confusin
#> rested in sex (12)   1 4 4 4 5 6 6 1 7   (12) interested i
#>     dependent (13)   3 2 6 5 5 4 6 3 6   (13) independent 
#>     masculine (14)   6 6 7 1 2 1 7 4 1   (14) feminine    
#>