Grid data from Slater (1977).

Grid data (ranked) from a seventeen year old female psychiatric patient (Slater, 1977, p. 110). She was depressed, anxious and took to cutting herself. The data was originally reported by Watson (1970).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Watson, J. P. (1970). The relationship between a self-mutilating patient and her doctor. Psychotherapy and Psychosomatics, 18(1), 67-73.

Examples

slater1977b
#> 
#> META DATA:
#> Number of constructs:  10 
#> Number of elements:  10 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 10 (right pole)
#> 
#> RATINGS:
#>               Being at home - 5 6 - Being in hospital           
#>                 Seeing G. - 4 | | 7 - Being with my mother      
#>       Being in a crowd - 3  | | | |  8 - Being with Dr. W.      
#>  same thoughts for a - 2 |  | | | |  |  9 - Being with Mrs. M.  
#> alk to someone and - 1 | |  | | | |  |  |  10 - Being with my fa
#>                      | | |  | | | |  |  |  |                    
#>                (1)   2 1  3 6 4 5  7  8 10 9   (1) Make me cut m
#>                (2)   1 3  6 2 4 7  5  8 10 9   (2) Make me think
#>                (3)   2 5  3 1 4 6  7 10  9 8   (3) Make me feel 
#>                (4)   1 2  4 3 7 6 10  5  8 9   (4) Make me feel 
#>                (5)   2 3  5 1 9 7 10  4  6 8   (5) Make me feel 
#>                (6)   5 4  7 1 2 6  3  9 10 8   (6) Make me feel 
#>                (7)   2 7  5 9 1 8  3  6 10 4   (7) Make me feel 
#>                (8)   9 8  7 1 5 6  4  2 10 3   (8) Make me feel 
#>                (9)   5 9 10 1 8 2  6  3  4 7   (9) Help me in th
#>               (10)   9 8 10 1 5 6  3  4  7 2   (10) Make me feel
#>