Sample dataframe with grid data (elements are columns)

This dataframe can be converted into a repgrid object via importDataframe(). The dataframe column names are the minimum of the rating scale (1), the element names (element_1 to element_4), the maximum of the rating scale (5), and optionally a column indicating the preferred pole. Each row contains the constructs' entries (left pole, ratings, right pole, preferred pole). The preferred pole must be one of left, right, none, NA (see preferredPoles()). See sample data df_element_columns.

Details

1	element_1	element_2	element_3	element_4	5	preferred

left_pole_1	1	5	3	4	right_pole_1	left

left_pole_2	3	1	1	3	right_pole_2	right

left_pole_3	4	2	5	1	right_pole_3	none

See also

Examples

df_element_columns
#>             1 element 1 element 2 element 3 element 4            5
#> 1 left pole 1         5         3         1         4 right pole 1
#> 2 left pole 2         3         3         5         3 right pole 2
#> 3 left pole 3         2         4         2         3 right pole 3
#>   preferred_pole
#> 1           left
#> 2          right
#> 3           none
importDataframe(df_element_columns)
#> 
#> META DATA:
#> Number of elements:  4 
#> Number of constructs:  3 
#> Preferred poles defined:  3/3 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>          element 2 - 2 3 - element 3          
#>        element 1 - 1 | | 4 - element 4        
#>                    | | | |                    
#> left pole 1 (+1)   5 3 1 4   (1-) right pole 1
#> left pole 2 (-2)   3 3 5 3   (2+) right pole 2
#> left pole 3 (/3)   2 4 2 3   (3/) right pole 3
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined