Sample dataframe with grid data (long)

This dataframe can be converted into a repgrid object via importDataframe(). The long format has this name because it has few columns and many rows. It is a common format in data analytics. Here, each row contains a different element-construct combination and the corresponding rating value. The format looks like this:

Details

element	left_pole	right_pole	rating	preferred_pole	rmin	rmax

element 1	left pole 1	right pole 1	5	left	1	5

element_2	left pole 1	right pole 1	3	left	1	5

element_3	left pole 1	right pole 1	1	left	1	5

The columns element, left_pole, right_pole, and rating are mandatory, the columns preferred_pole, rmin, and rmax are optional. rmin and rmax contain the min and max of the rating scale. Alternatively, you may pass rmin and rmax as arguments in the function call.

See also

Examples

df_long
#> # A tibble: 12 × 7
#>    element   left_pole   right_pole   rating preferred_pole  rmin  rmax
#>    <chr>     <chr>       <chr>        <chr>  <chr>          <dbl> <dbl>
#>  1 element 1 left pole 1 right pole 1 5      left               1     5
#>  2 element 2 left pole 1 right pole 1 3      left               1     5
#>  3 element 3 left pole 1 right pole 1 1      left               1     5
#>  4 element 4 left pole 1 right pole 1 4      left               1     5
#>  5 element 1 left pole 2 right pole 2 3      right              1     5
#>  6 element 2 left pole 2 right pole 2 3      right              1     5
#>  7 element 3 left pole 2 right pole 2 5      right              1     5
#>  8 element 4 left pole 2 right pole 2 3      right              1     5
#>  9 element 1 left pole 3 right pole 3 2      none               1     5
#> 10 element 2 left pole 3 right pole 3 4      none               1     5
#> 11 element 3 left pole 3 right pole 3 2      none               1     5
#> 12 element 4 left pole 3 right pole 3 3      none               1     5
importDataframe(df_long, format = "long")
#> 
#> META DATA:
#> Number of elements:  4 
#> Number of constructs:  3 
#> Preferred poles defined:  3/3 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>          element 2 - 2 3 - element 3          
#>        element 1 - 1 | | 4 - element 4        
#>                    | | | |                    
#> left pole 1 (+1)   5 3 1 4   (1-) right pole 1
#> left pole 2 (-2)   3 3 5 3   (2+) right pole 2
#> left pole 3 (/3)   2 4 2 3   (3/) right pole 3
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined