Slater distances (standardized Euclidean distances).

The euclidean distance is often used as a measure of similarity between elements (see distance(). A drawback of this measure is that it depends on the range of the rating scale and the number of constructs used, i. e. on the size of a grid.
An approach to standardize the euclidean distance to make it independent from size and range of ratings and was proposed by Slater (1977, pp. 94). The 'Slater distance' is the Euclidean distance divided by the expected distance. Slater distances bigger than 1 are greater than expected, lesser than 1 are smaller than expected. The minimum value is 0 and values bigger than 2 are rarely found. Slater distances have been be used to compare inter-element distances between different grids, where the grids do not need to have the same constructs or elements. Hartmann (1992) showed that Slater distance is not independent of grid size. Also the distribution of the Slater distances is asymmetric. Hence, the upper and lower limit to infer 'significance' of distance is not symmetric. The practical relevance of Hartmann's findings have been demonstrated by Schoeneich and Klapp (1998). To calculate Hartmann's version of the standardized distances see distanceHartmann()

Usage

distanceSlater(x, trim = 20, index = TRUE)

Arguments

x

repgrid object.

trim

The number of characters a construct or element is trimmed to (default is 20). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs with long names.

index

Whether to print the number of the construct or element in front of the name (default is TRUE). This is useful to avoid identical row names, which may cause an error.

Value

A matrix with Slater distances.

Calculation

The Slater distance is calculated as follows. For a derivation see Slater (1977, p.94).
Let matrix \(D\) contain the row centered ratings. Then $$P = D^TD$$ and $$S = tr(P)$$ The expected 'unit of expected distance' results as
$$U = (2S/(m-1))^{1/2}$$ where \(m\) denotes the number of elements of the grid. The standardized Slater distances is the matrix of Euclidean distances \(E\) divided by the expected distance \(U\). $$E/U$$

References

Hartmann, A. (1992). Element comparisons in repertory grid technique: Results and consequences of a Monte Carlo study. International Journal of Personal Construct Psychology, 5(1), 41-56.

Schoeneich, F., & Klapp, B. F. (1998). Standardization of interelement distances in repertory grid technique and its consequences for psychological interpretation of self-identity plots: An empirical study. Journal of Constructivist Psychology, 11(1), 49-58.

Slater, P. (1977). The measurement of intrapersonal space by Grid technique. Vol. II. London: Wiley.

See also

Examples

distanceSlater(bell2010)
#> 
#> ##########################
#> Distances between elements
#> ##########################
#> 
#> Distance method:  Slater (standardized Euclidean)
#> Normalized:
#>                                 1    2    3    4    5    6    7    8    9   10
#> (1) self                   1      0.91 0.69 0.62 0.84 1.12 0.93 1.09 0.63 1.18
#> (2) closest friend of th   2           0.96 1.11 0.61 1.03 0.61 1.04 1.02 0.96
#> (3) the unhappiest perso   3                0.75 1.02 0.87 0.93 1.27 0.93 1.05
#> (4) A person of the oppo   4                     1.09 1.10 1.01 1.02 0.71 1.22
#> (5) A teacher you respec   5                          1.24 0.89 0.89 0.94 0.85
#> (6) Mother (or the perso   6                               0.86 1.44 1.12 1.18
#> (7) A person of the oppo   7                                    1.11 1.06 0.94
#> (8) The most confident p   8                                         0.96 1.12
#> (9) A person you work we   9                                              1.27
#> (10) A teacher you did no 10                                                  
#> 
#> Note that Slater distances cannot be compared across grids with a different number of constructs (see Hartmann, 1992).
#> 
distanceSlater(bell2010, trim = 40)
#> 
#> ##########################
#> Distances between elements
#> ##########################
#> 
#> Distance method:  Slater (standardized Euclidean)
#> Normalized:
#>                                 1    2    3    4    5    6    7    8    9   10
#> (1) self                   1      0.91 0.69 0.62 0.84 1.12 0.93 1.09 0.63 1.18
#> (2) closest friend of th   2           0.96 1.11 0.61 1.03 0.61 1.04 1.02 0.96
#> (3) the unhappiest perso   3                0.75 1.02 0.87 0.93 1.27 0.93 1.05
#> (4) A person of the oppo   4                     1.09 1.10 1.01 1.02 0.71 1.22
#> (5) A teacher you respec   5                          1.24 0.89 0.89 0.94 0.85
#> (6) Mother (or the perso   6                               0.86 1.44 1.12 1.18
#> (7) A person of the oppo   7                                    1.11 1.06 0.94
#> (8) The most confident p   8                                         0.96 1.12
#> (9) A person you work we   9                                              1.27
#> (10) A teacher you did no 10                                                  
#> 
#> Note that Slater distances cannot be compared across grids with a different number of constructs (see Hartmann, 1992).
#> 

d <- distanceSlater(bell2010)
print(d)
#> 
#> ##########################
#> Distances between elements
#> ##########################
#> 
#> Distance method:  Slater (standardized Euclidean)
#> Normalized:
#>                                 1    2    3    4    5    6    7    8    9   10
#> (1) self                   1      0.91 0.69 0.62 0.84 1.12 0.93 1.09 0.63 1.18
#> (2) closest friend of th   2           0.96 1.11 0.61 1.03 0.61 1.04 1.02 0.96
#> (3) the unhappiest perso   3                0.75 1.02 0.87 0.93 1.27 0.93 1.05
#> (4) A person of the oppo   4                     1.09 1.10 1.01 1.02 0.71 1.22
#> (5) A teacher you respec   5                          1.24 0.89 0.89 0.94 0.85
#> (6) Mother (or the perso   6                               0.86 1.44 1.12 1.18
#> (7) A person of the oppo   7                                    1.11 1.06 0.94
#> (8) The most confident p   8                                         0.96 1.12
#> (9) A person you work we   9                                              1.27
#> (10) A teacher you did no 10                                                  
#> 
#> Note that Slater distances cannot be compared across grids with a different number of constructs (see Hartmann, 1992).
#> 
print(d, digits = 4)
#> 
#> ##########################
#> Distances between elements
#> ##########################
#> 
#> Distance method:  Slater (standardized Euclidean)
#> Normalized:
#>                                   1      2      3      4      5      6      7
#> (1) self                   1        0.9066 0.6901 0.6227 0.8440 1.1163 0.9290
#> (2) closest friend of th   2               0.9615 1.1134 0.6058 1.0347 0.6058
#> (3) the unhappiest perso   3                      0.7546 1.0158 0.8737 0.9310
#> (4) A person of the oppo   4                             1.0950 1.1032 1.0125
#> (5) A teacher you respec   5                                    1.2449 0.8887
#> (6) Mother (or the perso   6                                           0.8552
#> (7) A person of the oppo   7                                                 
#> (8) The most confident p   8                                                 
#> (9) A person you work we   9                                                 
#> (10) A teacher you did no 10                                                 
#>                                8      9     10
#> (1) self                  1.0925 0.6327 1.1782
#> (2) closest friend of th  1.0403 1.0219 0.9625
#> (3) the unhappiest perso  1.2735 0.9285 1.0482
#> (4) A person of the oppo  1.0174 0.7052 1.2188
#> (5) A teacher you respec  0.8852 0.9401 0.8500
#> (6) Mother (or the perso  1.4424 1.1248 1.1767
#> (7) A person of the oppo  1.1085 1.0559 0.9361
#> (8) The most confident p         0.9639 1.1205
#> (9) A person you work we                1.2726
#> (10) A teacher you did no                     
#> 
#> Note that Slater distances cannot be compared across grids with a different number of constructs (see Hartmann, 1992).
#> 

# using Norris and Makhlouf-Norris (problematic) cutoffs
print(d, cutoffs = c(.8, 1.2))
#> 
#> ##########################
#> Distances between elements
#> ##########################
#> 
#> Distance method:  Slater (standardized Euclidean)
#> Normalized:
#>                                 1    2    3    4    5    6    7    8    9   10
#> (1) self                   1           0.69 0.62                     0.63     
#> (2) closest friend of th   2                     0.61      0.61               
#> (3) the unhappiest perso   3                0.75                1.27          
#> (4) A person of the oppo   4                                         0.71 1.22
#> (5) A teacher you respec   5                          1.24                    
#> (6) Mother (or the perso   6                                    1.44          
#> (7) A person of the oppo   7                                                  
#> (8) The most confident p   8                                                  
#> (9) A person you work we   9                                              1.27
#> (10) A teacher you did no 10                                                  
#> 
#> Note that Slater distances cannot be compared across grids with a different number of constructs (see Hartmann, 1992).
#>