Root mean square (RMS) of inter-element correlations.

The RMS is also known as 'quadratic mean' of the inter-element correlations. The RMS serves as a simplification of the correlation table. It reflects the average relation of one element with all other elements. Note that as the correlations are squared during its calculation, the RMS is not affected by the sign of the correlation (cf. Fransella, Bell & Bannister, 2003, p. 86).

Usage

elementRmsCor(x, rc = TRUE, method = "pearson", trim = NA)

Arguments

x

repgrid object.

rc

Whether to use Cohen's rc which is invariant to construct reflection (see description above). It is used as the default.

method

A character string indicating which correlation coefficient to be computed. One of "pearson" (default), "kendall" or "spearman", can be abbreviated. The default is "pearson".

trim

The number of characters an element is trimmed to (default is NA). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs with long names.

Value

dataframe of the RMS of inter-element correlations.

Details

Note that simple element correlations as a measure of similarity are flawed as they are not invariant to construct reflection (Mackay, 1992; Bell, 2010). A correlation index invariant to construct reflection is Cohen's rc measure (1969), which can be calculated using the argument rc=TRUE which is the default option in this function.

References

Fransella, F., Bell, R. C., & Bannister, D. (2003). A Manual for Repertory Grid Technique (2. Ed.). Chichester: John Wiley & Sons.

See also

Examples

# data from grid manual by Fransella, Bell and Bannister
elementRmsCor(fbb2003)
#> 
#> ##########################################
#> Root-mean-square correlation of elements
#> ##########################################
#> 
#>                            RMS
#> (1) self                  0.51
#> (2) my father             0.65
#> (3) an old flame          0.40
#> (4) an ethical person     0.52
#> (5) my mother             0.26
#> (6) a rejected teacher    0.42
#> (7) as I would love to be 0.51
#> (8) a pitied person       0.47
#> 
#> Average of statistic 0.47 
#> Standard deviation of statistic 0.11 
elementRmsCor(fbb2003, trim = 10)
#> 
#> ##########################################
#> Root-mean-square correlation of elements
#> ##########################################
#> 
#>                 RMS
#> (1) self       0.51
#> (2) my father  0.65
#> (3) an old fla 0.40
#> (4) an ethical 0.52
#> (5) my mother  0.26
#> (6) a rejected 0.42
#> (7) as I would 0.51
#> (8) a pitied p 0.47
#> 
#> Average of statistic 0.47 
#> Standard deviation of statistic 0.11 

# modify output
r <- elementRmsCor(fbb2003)
print(r, digits = 5)
#> 
#> ##########################################
#> Root-mean-square correlation of elements
#> ##########################################
#> 
#>                               RMS
#> (1) self                  0.50967
#> (2) my father             0.64574
#> (3) an old flame          0.40418
#> (4) an ethical person     0.51928
#> (5) my mother             0.26293
#> (6) a rejected teacher    0.42362
#> (7) as I would love to be 0.50950
#> (8) a pitied person       0.47085
#> 
#> Average of statistic 0.46822 
#> Standard deviation of statistic 0.10355 

# access second row of calculation results
r[2, "RMS"]
#> [1] 0.6457385