Extract parts of the repgrid object.

Methods for "[", i.e., subsetting of repgrid objects.

Usage

# S4 method for class 'repgrid'
x[i, j, ..., drop = TRUE]

Arguments

x

A repgrid object.

i

Row index (numeric)

j

Column index. Either numeric or character (elements names).

...

Not evaluated.

drop

Not used.

Examples

x <- boeker
x[1:4, ]
#> 
#> META DATA:
#> Number of elements:  15 
#> Number of constructs:  4 
#> Preferred poles defined:  0/4 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                         8 - martin                             
#>                              george - 7 | 9 - elizabeth                        
#>                              karl - 6 | | | 10 - therapist                     
#>                            kurt - 5 | | | | | 11 - irene                       
#>                        father - 4 | | | | | | | 12 - childhood self            
#>                      mother - 3 | | | | | | | | | 13 - self before illness     
#>                ideal self - 2 | | | | | | | | | | | 14 - self with delusion    
#>                    self - 1 | | | | | | | | | | | | | 15 - self as dreamer     
#>                           | | | | | | | | | | | | | | |                        
#>           balanced (.1)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (1.) get along with c
#>           isolated (.2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2.) sociable        
#> closely integrated (.3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3.) excluded        
#>         discursive (.4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4.) passive         
#> 
#>  
#>  
#>  
#>  
#>  
#>  
#>  
#>  
#>  
#> o
#>  
#>  
#>  
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
x[, 1:3]
#> 
#> META DATA:
#> Number of elements:  3 
#> Number of constructs:  14 
#> Preferred poles defined:  0/14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                             2 - ideal self              
#>                    self - 1 | 3 - mother                
#>                           | | |                         
#>           balanced (.1)   1 4 2   (1.) get along with co
#>           isolated (.2)   3 6 3   (2.) sociable         
#> closely integrated (.3)   2 2 2   (3.) excluded         
#>         discursive (.4)   4 1 3   (4.) passive          
#>        open minded (.5)   2 1 2   (5.) indifferent      
#>             dreamy (.6)   4 5 3   (6.) dispassionate    
#> actically oriented (.7)   2 1 3   (7.) depressed        
#>            playful (.8)   4 5 4   (8.) serious          
#>    socially minded (.9)   2 1 3   (9.) selfish          
#>       quarrelsome (.10)   5 5 5   (10.) peaceful        
#>          artistic (.11)   5 1 2   (11.) technical       
#>        scientific (.12)   2 1 5   (12.) emotional       
#>         introvert (.13)   4 5 4   (13.) extrovert       
#>        wanderlust (.14)   1 1 4   (14.) home oriented   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
x[1:4, 1:3]
#> 
#> META DATA:
#> Number of elements:  3 
#> Number of constructs:  4 
#> Preferred poles defined:  0/4 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                             2 - ideal self              
#>                    self - 1 | 3 - mother                
#>                           | | |                         
#>           balanced (.1)   1 4 2   (1.) get along with co
#>           isolated (.2)   3 6 3   (2.) sociable         
#> closely integrated (.3)   2 2 2   (3.) excluded         
#>         discursive (.4)   4 1 3   (4.) passive          
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
x[1:4, c("self", "ideal self", "mother")]
#> 
#> META DATA:
#> Number of elements:  3 
#> Number of constructs:  4 
#> Preferred poles defined:  0/4 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                             2 - ideal self              
#>                    self - 1 | 3 - mother                
#>                           | | |                         
#>           balanced (.1)   1 4 2   (1.) get along with co
#>           isolated (.2)   3 6 3   (2.) sociable         
#> closely integrated (.3)   2 2 2   (3.) excluded         
#>         discursive (.4)   4 1 3   (4.) passive          
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined