Retrieve element names in needed format.

Retrieves the element names from a repgrid.

Usage

getElementNames2(x, trim = 20, index = F, pre = "(", post = ") ")

Arguments

x

repgrid object.

trim

Number of characters to trim the construct names to (default NA). NA will suppress trimming. The length of index is not included in the trimming.

index

Logical. Whether to add a index number before the construct names (default TRUE).

pre

String before index number (default ().

post

String after index number (default ) ).

Value

Vector with (trimmed) element names.

Details

Different features like trimming, indexing and choices of separators allow to return the kind of format that is needed.

Examples

if (FALSE) { # \dontrun{

getElementNames2(bell2010)
getElementNames2(bell2010, mode = 2)
getElementNames2(bell2010, index = T)
getElementNames2(bell2010, index = T, trim = 30)
} # }