Rotate the interactive 3D device to default views.

Rotate the interactive 3D device to a default viewpoint or to a position defined by theta and phi in Euler angles. Three default viewpoints are implemented rendering a view so that two axes span a plane and the third axis is pointing out of the screen.

Usage

home(view = 1, theta = NULL, phi = NULL)

Arguments

view

Numeric. Specifying one of three default views. 1 = XY, 2=XZ and 3=YZ-plane.

theta

Numeric. Euler angle. Overrides view setting.

phi

Numeric. Euler angle. Overrides view setting.

return NULL.

See also

Examples

if (FALSE) { # \dontrun{

biplot3d(boeker)
home(2)
home(3)
home(1)
home(theta = 45, phi = 45)
} # }