Import GRIDCOR data files.

Reads the file format that is used by the grid program GRIDCOR (Feixas & Cornejo, 2002).

Usage

importGridcor(file, dir = NULL)

Arguments

file

filename including path if file is not in current working directory. File can also be a complete URL. The fileformat is .dat.

dir

alternative way to supply the directory where the file is located (default NULL).

Value

a single repgrid object in case one file and a list of repgrid objects in case multiple files are imported.

Note

Note that the GRIDCOR data sets the minimum ratings scale range to 1. The maximum value can differ and is defined in the data file.

Also note that both Gridcor and Gridstat data files do have the same suffix .dat. Make sure not to mix them up.

References

Feixas, G., & Cornejo, J. M. (2002). GRIDCOR: Correspondence Analysis for Grid Data (version 4.0). Barcelona: Centro de Terapia Cognitiva. Retrieved from https://repertorygrid.net/en/.

See also

Examples

if (FALSE) { # \dontrun{

# supposing that the data file gridcor.dat is in the current directory
file <- "gridcor.dat"
rg <- importGridcor(file)

# specifying a directory (arbitrary example directory)
dir <- "/Users/markheckmann/data"
rg <- importGridcor(file, dir)

# using a full path
rg <- importGridcor("/Users/markheckmann/data/gridcor.dat")
} # }