Import Gridsuite data files.

Import Gridsuite data files.

Usage

importGridsuite(file, dir = NULL)

Arguments

file

Filename including path if file is not in current working directory. File can also be a complete URL. The fileformat is .dat.

dir

Alternative way to supply the directory where the file is located (default NULL).

Value

A single repgrid object in case one file and a list of repgrid objects in case multiple files are imported.

Note

The developers of Gridsuite have proposed to use an XML scheme as a standard exchange format for repertory grid data (Walter, Bacher & Fromm, 2004).

TODO: The element and construct IDs are not used yet. Thus, if the output should be in different order the current mechanism will cause false assignments.

References

http://www.gridsuite.de/

Walter, O. B., Bacher, A., & Fromm, M. (2004). A proposal for a common data exchange format for repertory grid data.Journal of Constructivist Psychology, 17(3), 247. doi:10.1080/10720530490447167

See also

Examples

if (FALSE) { # \dontrun{

# supposing that the data file gridsuite.xml is in the current directory
file <- "gridsuite.xml"
rg <- importGridsuite(file)

# specifying a directory (arbitrary example directory)
dir <- "/Users/markheckmann/data"
rg <- importGridsuite(file, dir)

# using a full path
rg <- importGridsuite("/Users/markheckmann/data/gridsuite.xml")
} # }