Conflict measure for grids (Slade & Sheehan, 1979) based on correlations.

Conflict measure as proposed by Slade and Sheehan (1979)

Usage

indexConflict1(x)

Arguments

x

repgrid object.

Value

A list with the following elements:

total: Total number of triads

imbalanced: Number of imbalanced triads

prop.balanced: Proportion of balanced triads

prop.imbalanced: Proportion of imbalanced triads

Details

The first approach to mathematically derive a conflict measure based on grid data was presented by Slade and Sheehan (1979). Their operationalization is based on an approach by Lauterbach (1975) who applied the balance theory (Heider, 1958) for a quantitative assessment of psychological conflict. It is based on a count of balanced and imbalanced triads of construct correlations. A triad is imbalanced if one or all three of the correlations are negative, i. e. leading to contrary implications. This approach was shown by Winter (1982) to be flawed. An improved version was proposed by Bassler et al. (1992) and has been implemented in the function indexConflict2.

The table below shows when a triad made up of the constructs A, B, and C is balanced and imbalanced:

cor(A,B)	cor(A,C)	cor(B,C)	Triad characteristic

+	+	+	balanced

+	+	-	imbalanced

+	-	+	imbalanced

+	-	-	balanced

-	+	+	imbalanced

-	+	-	balanced

-	-	+	balanced

-	-	-	imbalanced

References

Bassler, M., Krauthauser, H., & Hoffmann, S. O. (1992). A new approach to the identification of cognitive conflicts in the repertory grid: An illustrative case study. Journal of Constructivist Psychology, 5(1), 95-111.

Heider, F. (1958). The Psychology of Interpersonal Relation. John Wiley & Sons.

Lauterbach, W. (1975). Assessing psychological conflict. The British Journal of Social and Clinical Psychology, 14(1), 43-47.

Slade, P. D., & Sheehan, M. J. (1979). The measurement of 'conflict' in repertory grids. British Journal of Psychology, 70(4), 519-524.

Winter, D. A. (1982). Construct relationships, psychological disorder and therapeutic change. The British Journal of Medical Psychology, 55 (Pt 3), 257-269.

See also

Examples

indexConflict1(feixas2004)
#> 
#> ################################
#> Conflicts based on correlations
#> ################################
#> 
#> As devised by Slade & Sheehan (1979)
#> 
#> Total number of triads: 1140
#> Number of imbalanced triads: 266
#> 
#> Proportion of balanced triads: 76.7 %
#> Proportion of imbalanced triads: 23.3 %
indexConflict1(boeker)
#> 
#> ################################
#> Conflicts based on correlations
#> ################################
#> 
#> As devised by Slade & Sheehan (1979)
#> 
#> Total number of triads: 364
#> Number of imbalanced triads: 106
#> 
#> Proportion of balanced triads: 70.9 %
#> Proportion of imbalanced triads: 29.1 %