Conflict measure for grids (Bassler et al., 1992) based on correlations.

The function calculates the conflict measure as devised by Bassler et al. (1992). It is an improved version of the ideas by Slade and Sheehan (1979) that have been implemented in the function indexConflict1(). The new approach also takes into account the magnitude of the correlations in a trait to assess whether it is balanced or imbalanced. As a result, small correlations that are psychologically meaningless are considered accordingly. Also, correlations with a small magnitude, i. e. near zero, which may be positive or negative due to chance alone will no longer distort the measure (Bassler et al., 1992).

Usage

indexConflict2(x, crit = 0.03)

Arguments

x

A repgrid object.

crit

Sensitivity criterion with which triads are marked as unbalanced. A bigger values will lead to less imbalanced triads. The default is 0.03. The value should be adjusted with regard to the researchers interest.

Details

Description of the balance / imbalance assessment:

Order correlations of the triad by absolute magnitude, so that \( r_{max} > r_{mdn} > r_{min} \).

Apply Fisher's Z-transformation and division by 3 to yield values between 1 and -1 (\( Z_{max} > Z_{mdn} > Z_{min} \)).

Check whether the triad is balanced by assessing if the following relation holds:

If \( Z_{max} Z_{mdn} > 0 \), the triad is balanced if \( Z_{max} Z_{mdn} - Z_{min} <= crit \) Z_max x Z_mdn - Z_min <= crit .

If \( Z_{max} Z_{mdn} < 0 \), the triad is balanced if \( Z_{min} - Z_{max} Z_{mdn} <= crit \) Z_min - Z_max x Z_mdn <= crit .

Personal remarks (MH)

I am a bit suspicious about step 2 from above. To devide by 3 appears pretty arbitrary. The r for a z-values of 3 is 0.9950548 and not 1. The r for 4 is 0.9993293. Hence, why not a value of 4, 5, or 6? Denoting the value to devide by with a, the relation for the first case translates into \( a Z_{max} Z_{mdn} <= \frac{crit}{a} + Z_{min} \) a x Z_max x Z_mdn =< crit/a + Z_min. This shows that a bigger value of a will make it more improbable that the relation will hold.

References

Bassler, M., Krauthauser, H., & Hoffmann, S. O. (1992). A new approach to the identification of cognitive conflicts in the repertory grid: An illustrative case study. Journal of Constructivist Psychology, 5(1), 95-111.

Slade, P. D., & Sheehan, M. J. (1979). The measurement of 'conflict' in repertory grids. British Journal of Psychology, 70(4), 519-524.

See also

Examples

indexConflict2(bell2010)
#> 
#> ###############################
#> Conflicts based on correlations
#> ###############################
#> 
#> As devised by Bassler et al. (1992)
#> 
#> Total number of triads: 84
#> Number of imbalanced triads: 63
#> 
#> Proportion of balanced triads: 25 %
#> Proportion of imbalanced triads: 75 %

x <- indexConflict2(bell2010)
print(x)
#> 
#> ###############################
#> Conflicts based on correlations
#> ###############################
#> 
#> As devised by Bassler et al. (1992)
#> 
#> Total number of triads: 84
#> Number of imbalanced triads: 63
#> 
#> Proportion of balanced triads: 25 %
#> Proportion of imbalanced triads: 75 %

# show conflictive triads
print(x, output = 2)
#> 
#> ###############################
#> Conflicts based on correlations
#> ###############################
#> 
#> As devised by Bassler et al. (1992)
#> 
#> Total number of triads: 84
#> Number of imbalanced triads: 63
#> 
#> Proportion of balanced triads: 25 %
#> Proportion of imbalanced triads: 75 %
#> 
#> Constructs that form imbalanced triads:
#>            
#> 1  1  2   4
#> 2  1  2   7
#> 3  1  2   8
#> 4  1  3   8
#> 5  1  3   9
#> 6  1  4   5
#> 7  1  4   6
#> 8  1  4   7
#> 9  1  4   8
#> 10 1  4   9
#> 11 1  5   8
#> 12 1  5   9
#> 13 1  6   7
#> 14 1  6   8
#> 15 1  7   8
#> 16 1  7   9
#> 17 1  8   9
#> 18 2  3   5
#> 19 2  3   6
#> 20 2  3   7
#> 21 2  3   8
#> 22 2  3   9
#> 23 2  4   5
#> 24 2  4   6
#> 25 2  4   7
#> 26 2  4   8
#> 27 2  4   9
#> 28 2  5   6
#> 29 2  5   7
#> 30 2  5   8
#> 31 2  5   9
#> 32 2  6   7
#> 33 2  6   8
#> 34 2  7   8
#> 35 2  7   9
#> 36 3  4   5
#> 37 3  4   6
#> 38 3  4   8
#> 39 3  4   9
#> 40 3  5   6
#> 41 3  5   8
#> 42 3  6   8
#> 43 3  6   9
#> 44 3  7   8
#> 45 3  7   9
#> 46 3  8   9
#> 47 4  5   7
#> 48 4  5   8
#> 49 4  6   7
#> 50 4  6   8
#> 51 4  6   9
#> 52 4  7   8
#> 53 4  8   9
#> 54 5  6   7
#> 55 5  6   8
#> 56 5  6   9
#> 57 5  7   8
#> 58 5  7   9
#> 59 5  8   9
#> 60 6  7   8
#> 61 6  7   9
#> 62 6  8   9
#> 63 7  8   9

# accessing the calculations for further use
x$total
#> [1] 84
x$imbalanced
#> [1] 63
x$prop.balanced
#> [1] 0.25
x$prop.imbalanced
#> [1] 0.75
x$triads.imbalanced
#>       [,1] [,2] [,3]
#>  [1,]    1    2    4
#>  [2,]    1    2    7
#>  [3,]    1    2    8
#>  [4,]    1    3    8
#>  [5,]    1    3    9
#>  [6,]    1    4    5
#>  [7,]    1    4    6
#>  [8,]    1    4    7
#>  [9,]    1    4    8
#> [10,]    1    4    9
#> [11,]    1    5    8
#> [12,]    1    5    9
#> [13,]    1    6    7
#> [14,]    1    6    8
#> [15,]    1    7    8
#> [16,]    1    7    9
#> [17,]    1    8    9
#> [18,]    2    3    5
#> [19,]    2    3    6
#> [20,]    2    3    7
#> [21,]    2    3    8
#> [22,]    2    3    9
#> [23,]    2    4    5
#> [24,]    2    4    6
#> [25,]    2    4    7
#> [26,]    2    4    8
#> [27,]    2    4    9
#> [28,]    2    5    6
#> [29,]    2    5    7
#> [30,]    2    5    8
#> [31,]    2    5    9
#> [32,]    2    6    7
#> [33,]    2    6    8
#> [34,]    2    7    8
#> [35,]    2    7    9
#> [36,]    3    4    5
#> [37,]    3    4    6
#> [38,]    3    4    8
#> [39,]    3    4    9
#> [40,]    3    5    6
#> [41,]    3    5    8
#> [42,]    3    6    8
#> [43,]    3    6    9
#> [44,]    3    7    8
#> [45,]    3    7    9
#> [46,]    3    8    9
#> [47,]    4    5    7
#> [48,]    4    5    8
#> [49,]    4    6    7
#> [50,]    4    6    8
#> [51,]    4    6    9
#> [52,]    4    7    8
#> [53,]    4    8    9
#> [54,]    5    6    7
#> [55,]    5    6    8
#> [56,]    5    6    9
#> [57,]    5    7    8
#> [58,]    5    7    9
#> [59,]    5    8    9
#> [60,]    6    7    8
#> [61,]    6    7    9
#> [62,]    6    8    9
#> [63,]    7    8    9