Conflict or inconsistency measure for grids (Bell, 2004) based on distances.

Measure of conflict or inconsistency as proposed by Bell (2004). The identification of conflict is based on distances rather than correlations as in other measures of conflict indexConflict1() and indexConflict2(). It assesses if the distances between all components of a triad, made up of one element and two constructs, satisfies the "triangle inequality" (cf. Bell, 2004). If not, a triad is regarded as conflictive. An advantage of the measure is that it can be interpreted not only as a global measure for a grid but also on an element, construct, and element by construct level making it valuable for detailed feedback. Also, differences in conflict can be submitted to statistical testing procedures.

Usage

indexConflict3(
  x,
  p = 2,
  e.out = NA,
  e.threshold = NA,
  c.out = NA,
  c.threshold = NA,
  trim = 20
)

Arguments

x

repgrid object.

p

The power of the Minkowski distance. p=2 (default) will result in euclidean distances, p=1 in city block distances.

e.out

Numeric. A vector giving the indexes of the elements for which detailed stats (number of conflicts per element, discrepancies for triangles etc.) are prompted (default NA, i.e. no detailed stats for any element).

e.threshold

Numeric. Detailed stats are prompted for those elements with a an attributable percentage to the overall conflicts higher than the supplied threshold (default NA).

c.out

Numeric. A vector giving the indexes of the constructs for which detailed stats (discrepancies for triangles etc.) are prompted (default NA, i. e. no detailed stats).

c.threshold

Numeric. Detailed stats are prompted for those constructs with a an attributable percentage to the overall conflicts higher than the supplied threshold (default NA).

trim

The number of characters a construct (element) is trimmed to (default is 10). If NA no trimming is done. Trimming simply saves space when displaying the output.

Value

A list (invisibly) containing:

potential: number of potential conflicts

actual: count of actual conflicts

overall: percentage of conflictive relations

e.count: number of involvements of each element in conflictive relations

e.perc: percentage of involvement of each element in total of conflictive relations

c.count: number of involvements of each construct in conflictive relation

c.perc: percentage of involvement of each construct in total of conflictive relations

e.stats: detailed statistics for prompted elements

c.stats: detailed statistics for prompted constructs

e.threshold: threshold percentage. Used by print method

c.threshold: threshold percentage. Used by print method

enames: trimmed element names. Used by print method

cnames: trimmed construct names. Used by print method

Details

Status: working; output for euclidean and manhattan distance checked against Gridstat output.
TODO: standardization and z-test for discrepancies; Index of Conflict Variation.

output

For further control over the output see print.indexConflict3().

References

Bell, R. C. (2004). A new approach to measuring inconsistency or conflict in grids. Personal Construct Theory & Practice, (1), 53-59.

See also

Examples

# calculate conflicts
indexConflict3(bell2010)
#> 
#> ##########################################################
#> CONFLICT OR INCONSISTENCIES BASED ON TRIANGLE INEQUALITIES
#> ##########################################################
#> 
#> Potential conflicts in grid:  360
#> Actual conflicts in grid:  123
#> Overall percentage of conflict in grid:  34.17 %
#> 
#> ELEMENTS
#> ########
#> 
#> Percent of conflict attributable to element:
#> 
#>                         percentage
#> 1 self                      162.60
#> 2 closest friend of th     1382.11
#> 3 the unhappiest perso      406.50
#> 4 A person of the oppo      650.41
#> 5 A teacher you respec     1056.91
#> 6 Mother (or the perso     1138.21
#> 7 A person of the oppo      813.01
#> 8 The most confident p     1788.62
#> 9 A person you work we     1138.21
#> 10 A teacher you did no    1463.41
#> 
#> Chi-square test of equal count of conflicts for elements.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$e.count
#> X-squared = 27.488, df = 9, p-value = 0.001161
#> 
#> 
#> CONSTRUCTS
#> ##########
#> 
#> Percent of conflict attributable to construct:
#> 
#>                           percentage
#> 1 relaxed - worried &          10.16
#> 2 not so sma - smart (aca      11.38
#> 3 dislikes s - loves spor      13.01
#> 4 not intera - loves peop      10.57
#> 5 not transp - transparen      12.60
#> 6 insensitiv - sensitive       12.20
#> 7 fearful&ti - fearless        10.16
#> 8 rough - gentle               11.38
#> 9 accept as  - loves to a       8.54
#> 
#> Chi-square test of equal count of conflicts for constructs.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$c.count
#> X-squared = 3.5122, df = 8, p-value = 0.8982
#> 

# show additional stats for elements 1 to 3
indexConflict3(bell2010, e.out = 1:3)
#> 
#> ##########################################################
#> CONFLICT OR INCONSISTENCIES BASED ON TRIANGLE INEQUALITIES
#> ##########################################################
#> 
#> Potential conflicts in grid:  360
#> Actual conflicts in grid:  123
#> Overall percentage of conflict in grid:  34.17 %
#> 
#> ELEMENTS
#> ########
#> 
#> Percent of conflict attributable to element:
#> 
#>                         percentage
#> 1 self                      162.60
#> 2 closest friend of th     1382.11
#> 3 the unhappiest perso      406.50
#> 4 A person of the oppo      650.41
#> 5 A teacher you respec     1056.91
#> 6 Mother (or the perso     1138.21
#> 7 A person of the oppo      813.01
#> 8 The most confident p     1788.62
#> 9 A person you work we     1138.21
#> 10 A teacher you did no    1463.41
#> 
#> Chi-square test of equal count of conflicts for elements.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$e.count
#> X-squared = 27.488, df = 9, p-value = 0.001161
#> 
#> 
#> CONSTRUCTS
#> ##########
#> 
#> Percent of conflict attributable to construct:
#> 
#>                           percentage
#> 1 relaxed - worried &          10.16
#> 2 not so sma - smart (aca      11.38
#> 3 dislikes s - loves spor      13.01
#> 4 not intera - loves peop      10.57
#> 5 not transp - transparen      12.60
#> 6 insensitiv - sensitive       12.20
#> 7 fearful&ti - fearless        10.16
#> 8 rough - gentle               11.38
#> 9 accept as  - loves to a       8.54
#> 
#> Chi-square test of equal count of conflicts for constructs.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$c.count
#> X-squared = 3.5122, df = 8, p-value = 0.8982
#> 
#> 
#> 
#> CONFLICTS BY ELEMENT
#> ####################
#> 
#> 
#> ### Element:  1 self 
#> 
#> Number of conflicting construct pairs:  2 
#> 
#> Construct conflict discrepancies:
#> 
#>       1    2    3    4    5    6    7    8    9
#> 1                                              
#> 2  <NA>                                        
#> 3  <NA> <NA>                                   
#> 4  <NA> <NA> <NA>                              
#> 5  <NA> <NA> <NA> <NA>                         
#> 6  <NA> <NA> 0.03 <NA> 0.33                    
#> 7  <NA> <NA> <NA> <NA> <NA> <NA>               
#> 8  <NA> <NA> <NA> <NA> <NA> <NA> <NA>          
#> 9  <NA> <NA> <NA> <NA> <NA> <NA> <NA> <NA>     
#> 
#> Percent of conflict attributable to each construct:
#> 
#>                           percentage
#> 1 relaxed - worried &              0
#> 2 not so sma - smart (aca          0
#> 3 dislikes s - loves spor         25
#> 4 not intera - loves peop          0
#> 5 not transp - transparen         25
#> 6 insensitiv - sensitive          50
#> 7 fearful&ti - fearless            0
#> 8 rough - gentle                   0
#> 9 accept as  - loves to a          0
#> 
#> Av. level of discrepancy:    0.01 
#> 
#> Std. dev. of discrepancies:  0.176 
#> 
#> 
#> ### Element:  2 closest friend of th 
#> 
#> Number of conflicting construct pairs:  17 
#> 
#> Construct conflict discrepancies:
#> 
#>       1    2    3    4    5    6    7    8    9
#> 1                                              
#> 2  0.53                                        
#> 3  <NA> 1.35                                   
#> 4  0.19 <NA> 1.81                              
#> 5  <NA> 0.55 <NA> 1.51                         
#> 6  0.37 <NA> 1.17 <NA> 0.33                    
#> 7  <NA> 1.59 <NA> 1.75 <NA> 0.03               
#> 8  0.47 <NA> 0.62 <NA> 0.17 <NA> <NA>          
#> 9  <NA> 0.55 <NA>  0.8 <NA> <NA> <NA> <NA>     
#> 
#> Percent of conflict attributable to each construct:
#> 
#>                           percentage
#> 1 relaxed - worried &          11.76
#> 2 not so sma - smart (aca      14.71
#> 3 dislikes s - loves spor      11.76
#> 4 not intera - loves peop      14.71
#> 5 not transp - transparen      11.76
#> 6 insensitiv - sensitive       11.76
#> 7 fearful&ti - fearless         8.82
#> 8 rough - gentle                8.82
#> 9 accept as  - loves to a       5.88
#> 
#> Av. level of discrepancy:    0.34 
#> 
#> Std. dev. of discrepancies:  0.583 
#> 
#> 
#> ### Element:  3 the unhappiest perso 
#> 
#> Number of conflicting construct pairs:  5 
#> 
#> Construct conflict discrepancies:
#> 
#>       1    2    3    4    5    6    7    8    9
#> 1                                              
#> 2  <NA>                                        
#> 3  <NA> <NA>                                   
#> 4  <NA> <NA> <NA>                              
#> 5  <NA> <NA> 0.11 <NA>                         
#> 6  <NA> <NA> 0.03 <NA> <NA>                    
#> 7  <NA>  0.4 1.27 <NA> <NA> <NA>               
#> 8  <NA> <NA> <NA> <NA> <NA> <NA> <NA>          
#> 9  <NA> <NA> <NA> <NA> <NA> <NA> 1.18 <NA>     
#> 
#> Percent of conflict attributable to each construct:
#> 
#>                           percentage
#> 1 relaxed - worried &              0
#> 2 not so sma - smart (aca         10
#> 3 dislikes s - loves spor         30
#> 4 not intera - loves peop          0
#> 5 not transp - transparen         10
#> 6 insensitiv - sensitive          10
#> 7 fearful&ti - fearless           30
#> 8 rough - gentle                   0
#> 9 accept as  - loves to a         10
#> 
#> Av. level of discrepancy:    0.07 
#> 
#> Std. dev. of discrepancies:  0.556 

# show additional stats for constructs 1 and 5
indexConflict3(bell2010, c.out = c(1, 5))
#> 
#> ##########################################################
#> CONFLICT OR INCONSISTENCIES BASED ON TRIANGLE INEQUALITIES
#> ##########################################################
#> 
#> Potential conflicts in grid:  360
#> Actual conflicts in grid:  123
#> Overall percentage of conflict in grid:  34.17 %
#> 
#> ELEMENTS
#> ########
#> 
#> Percent of conflict attributable to element:
#> 
#>                         percentage
#> 1 self                      162.60
#> 2 closest friend of th     1382.11
#> 3 the unhappiest perso      406.50
#> 4 A person of the oppo      650.41
#> 5 A teacher you respec     1056.91
#> 6 Mother (or the perso     1138.21
#> 7 A person of the oppo      813.01
#> 8 The most confident p     1788.62
#> 9 A person you work we     1138.21
#> 10 A teacher you did no    1463.41
#> 
#> Chi-square test of equal count of conflicts for elements.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$e.count
#> X-squared = 27.488, df = 9, p-value = 0.001161
#> 
#> 
#> CONSTRUCTS
#> ##########
#> 
#> Percent of conflict attributable to construct:
#> 
#>                           percentage
#> 1 relaxed - worried &          10.16
#> 2 not so sma - smart (aca      11.38
#> 3 dislikes s - loves spor      13.01
#> 4 not intera - loves peop      10.57
#> 5 not transp - transparen      12.60
#> 6 insensitiv - sensitive       12.20
#> 7 fearful&ti - fearless        10.16
#> 8 rough - gentle               11.38
#> 9 accept as  - loves to a       8.54
#> 
#> Chi-square test of equal count of conflicts for constructs.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$c.count
#> X-squared = 3.5122, df = 8, p-value = 0.8982
#> 
#> 
#> 
#> CONFLICTS BY CONSTRUCT
#> ######################
#> 
#> 
#> ### Construct:  1 relaxed - worried &  
#> 
#> Element-construct conflict discrepancies:
#> 
#>      e1   e2   e3   e4   e5   e6   e7   e8   e9  e10
#> c1                                                  
#> c2 <NA>                                             
#> c3 <NA> <NA>                                        
#> c4 <NA> 0.19 <NA>                                   
#> c5 <NA> <NA> <NA> <NA>                              
#> c6 <NA> 0.37 <NA> <NA> <NA>                         
#> c7 <NA> <NA> <NA> <NA> <NA>    1                    
#> c8 <NA> 0.47 <NA> <NA> 0.47 <NA> 0.47               
#> c9 <NA> <NA> <NA> <NA> <NA> <NA> <NA> 3.44          
#> 
#> Av. level of discrepancy:    0.38 
#> 
#> Std. dev. of discrepancies:  0.973 
#> 
#> 
#> ### Construct:  5 not transp - transparen 
#> 
#> Element-construct conflict discrepancies:
#> 
#>      e1   e2   e3   e4   e5   e6   e7   e8   e9  e10
#> c1                                                  
#> c2 <NA>                                             
#> c3 <NA> <NA>                                        
#> c4 <NA> 1.51 <NA>                                   
#> c5 <NA> <NA> <NA> <NA>                              
#> c6 0.33 0.33 <NA> 1.51 <NA>                         
#> c7 <NA> <NA> <NA> <NA>  0.3 2.75                    
#> c8 <NA> 0.17 <NA> 1.33 0.17 <NA> <NA>               
#> c9 <NA> <NA> <NA> <NA> <NA>  1.8 <NA> 0.59          
#> 
#> Av. level of discrepancy:    0.32 
#> 
#> Std. dev. of discrepancies:  0.761 

# finetune output
## change number of digits
x <- indexConflict3(bell2010)
print(x, digits = 4)
#> 
#> ##########################################################
#> CONFLICT OR INCONSISTENCIES BASED ON TRIANGLE INEQUALITIES
#> ##########################################################
#> 
#> Potential conflicts in grid:  360
#> Actual conflicts in grid:  123
#> Overall percentage of conflict in grid:  34.1667 %
#> 
#> ELEMENTS
#> ########
#> 
#> Percent of conflict attributable to element:
#> 
#>                         percentage
#> 1 self                    162.6016
#> 2 closest friend of th   1382.1138
#> 3 the unhappiest perso    406.5041
#> 4 A person of the oppo    650.4065
#> 5 A teacher you respec   1056.9106
#> 6 Mother (or the perso   1138.2114
#> 7 A person of the oppo    813.0081
#> 8 The most confident p   1788.6179
#> 9 A person you work we   1138.2114
#> 10 A teacher you did no  1463.4146
#> 
#> Chi-square test of equal count of conflicts for elements.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$e.count
#> X-squared = 27.488, df = 9, p-value = 0.001161
#> 
#> 
#> CONSTRUCTS
#> ##########
#> 
#> Percent of conflict attributable to construct:
#> 
#>                           percentage
#> 1 relaxed - worried &        10.1626
#> 2 not so sma - smart (aca    11.3821
#> 3 dislikes s - loves spor    13.0081
#> 4 not intera - loves peop    10.5691
#> 5 not transp - transparen    12.6016
#> 6 insensitiv - sensitive     12.1951
#> 7 fearful&ti - fearless      10.1626
#> 8 rough - gentle             11.3821
#> 9 accept as  - loves to a     8.5366
#> 
#> Chi-square test of equal count of conflicts for constructs.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$c.count
#> X-squared = 3.5122, df = 8, p-value = 0.8982
#> 
#> NULL

## omit discrepancy matrices for constructs
x <- indexConflict3(bell2010, c.out = 5:6)
print(x, discrepancies = FALSE)
#> 
#> ##########################################################
#> CONFLICT OR INCONSISTENCIES BASED ON TRIANGLE INEQUALITIES
#> ##########################################################
#> 
#> Potential conflicts in grid:  360
#> Actual conflicts in grid:  123
#> Overall percentage of conflict in grid:  34.17 %
#> 
#> ELEMENTS
#> ########
#> 
#> Percent of conflict attributable to element:
#> 
#>                         percentage
#> 1 self                      162.60
#> 2 closest friend of th     1382.11
#> 3 the unhappiest perso      406.50
#> 4 A person of the oppo      650.41
#> 5 A teacher you respec     1056.91
#> 6 Mother (or the perso     1138.21
#> 7 A person of the oppo      813.01
#> 8 The most confident p     1788.62
#> 9 A person you work we     1138.21
#> 10 A teacher you did no    1463.41
#> 
#> Chi-square test of equal count of conflicts for elements.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$e.count
#> X-squared = 27.488, df = 9, p-value = 0.001161
#> 
#> 
#> CONSTRUCTS
#> ##########
#> 
#> Percent of conflict attributable to construct:
#> 
#>                           percentage
#> 1 relaxed - worried &          10.16
#> 2 not so sma - smart (aca      11.38
#> 3 dislikes s - loves spor      13.01
#> 4 not intera - loves peop      10.57
#> 5 not transp - transparen      12.60
#> 6 insensitiv - sensitive       12.20
#> 7 fearful&ti - fearless        10.16
#> 8 rough - gentle               11.38
#> 9 accept as  - loves to a       8.54
#> 
#> Chi-square test of equal count of conflicts for constructs.
#> 
#> 	Chi-squared test for given probabilities
#> 
#> data:  x$c.count
#> X-squared = 3.5122, df = 8, p-value = 0.8982
#> 
#> 
#> 
#> CONFLICTS BY CONSTRUCT
#> ######################
#> 
#> 
#> ### Construct:  5 not transp - transparen 
#> 
#> Av. level of discrepancy:    0.32 
#> 
#> Std. dev. of discrepancies:  0.761 
#> 
#> 
#> ### Construct:  6 insensitiv - sensitive 
#> 
#> Av. level of discrepancy:    0.32 
#> 
#> Std. dev. of discrepancies:  0.929