Polarization (percentage of extreme ratings)

Polarization is the percentage of extreme ratings, e.g. the values 1 and 7 for a grid with a 7-point ratings scale.

Usage

indexPolarization(x, deviation = 0)

Arguments

x

A repgrid object.

deviation

The maximal deviation from the end of the rating scale for values to be considered an 'extreme' rating. By default only values that lie directly on ends of the ratings scales are considered 'extreme' (default = 0).

Value

List of class indexPolarization:

scale: Minimum and maximum of grid rating scale.

lower,upper Lower and upper value to decide which ratings are considered extreme.

polarization_total: Grid's overall polarization.

polarization_constructs: Polarization per construct.

polarization_elements: Polarization per element.

Examples

p <- indexPolarization(boeker)
p
#> 
#> ##################
#> Polarization index
#> ##################
#> 
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> Extreme ratings are ratings <= 1 or >= 6
#> 
#> 
#> POLARIZATION OVERALL
#> 
#>   Ratings Extreme Polarization
#> 1     210      26        12.4%
#> 
#> 
#> POLARIZATION BY CONSTRUCT
#> 
#>                              Construct Ratings Extremes Polarization
#> 1  balanced - get along with conflicts      15        2        13.3%
#> 2                  isolated - sociable      15        1         6.7%
#> 3        closely integrated - excluded      15        0         0.0%
#> 4                 discursive - passive      15        2        13.3%
#> 5            open minded - indifferent      15        3        20.0%
#> 6               dreamy - dispassionate      15        1         6.7%
#> 7     practically oriented - depressed      15        1         6.7%
#> 8                    playful - serious      15        0         0.0%
#> 9            socially minded - selfish      15        3        20.0%
#> 10              quarrelsome - peaceful      15        2        13.3%
#> 11                artistic - technical      15        1         6.7%
#> 12              scientific - emotional      15        3        20.0%
#> 13               introvert - extrovert      15        1         6.7%
#> 14          wanderlust - home oriented      15        6        40.0%
#> 
#> 
#> POLARIZATION BY ELEMENT
#> 
#>                Element Ratings Extremes Polarization
#> 1                 self      14        2        14.3%
#> 2           ideal self      14        8        57.1%
#> 3               mother      14        0         0.0%
#> 4               father      14        3        21.4%
#> 5                 kurt      14        0         0.0%
#> 6                 karl      14        0         0.0%
#> 7               george      14        0         0.0%
#> 8               martin      14        1         7.1%
#> 9            elizabeth      14        0         0.0%
#> 10           therapist      14        1         7.1%
#> 11               irene      14        7        50.0%
#> 12      childhood self      14        2        14.3%
#> 13 self before illness      14        1         7.1%
#> 14  self with delusion      14        0         0.0%
#> 15     self as dreamer      14        1         7.1%

# control the output
print(p, output = "T")  # total polarization
#> 
#> ##################
#> Polarization index
#> ##################
#> 
#> 
#> 
#> POLARIZATION OVERALL
#> 
#>   Ratings Extreme Polarization
#> 1     210      26        12.4%
print(p, output = "C")  # construct polarization
#> 
#> ##################
#> Polarization index
#> ##################
#> 
#> 
#> POLARIZATION BY CONSTRUCT
#> 
#>                              Construct Ratings Extremes Polarization
#> 1  balanced - get along with conflicts      15        2        13.3%
#> 2                  isolated - sociable      15        1         6.7%
#> 3        closely integrated - excluded      15        0         0.0%
#> 4                 discursive - passive      15        2        13.3%
#> 5            open minded - indifferent      15        3        20.0%
#> 6               dreamy - dispassionate      15        1         6.7%
#> 7     practically oriented - depressed      15        1         6.7%
#> 8                    playful - serious      15        0         0.0%
#> 9            socially minded - selfish      15        3        20.0%
#> 10              quarrelsome - peaceful      15        2        13.3%
#> 11                artistic - technical      15        1         6.7%
#> 12              scientific - emotional      15        3        20.0%
#> 13               introvert - extrovert      15        1         6.7%
#> 14          wanderlust - home oriented      15        6        40.0%
print(p, output = "E")  # element polarization
#> 
#> ##################
#> Polarization index
#> ##################
#> 
#> 
#> POLARIZATION BY ELEMENT
#> 
#>                Element Ratings Extremes Polarization
#> 1                 self      14        2        14.3%
#> 2           ideal self      14        8        57.1%
#> 3               mother      14        0         0.0%
#> 4               father      14        3        21.4%
#> 5                 kurt      14        0         0.0%
#> 6                 karl      14        0         0.0%
#> 7               george      14        0         0.0%
#> 8               martin      14        1         7.1%
#> 9            elizabeth      14        0         0.0%
#> 10           therapist      14        1         7.1%
#> 11               irene      14        7        50.0%
#> 12      childhood self      14        2        14.3%
#> 13 self before illness      14        1         7.1%
#> 14  self with delusion      14        0         0.0%
#> 15     self as dreamer      14        1         7.1%