Percentage of Variance Accounted for by the First Factor (PVAFF)

The PVAFF is used as a measure of cognitive complexity. It was introduced in an unpublished PhD thesis by Jones (1954, cit. Bonarius, 1965). To calculate it, the 'first factor' two different methods may be used. One applies principal component analysis (PCA) to the construct centered raw data (default), the second applies SVD to the construct correlation matrix. The PVAFF reflects the amount of variation that is accounted for by a single linear component. If a single latent component is able to explain the variation in the grid, the cognitive complexity is said to be low. In this case the construct system is regarded as 'simple' (Bell, 2003).

Usage

indexPvaff(x, method = 1)

Arguments

x

repgrid object.

method

Method to compute PVAFF: 1 = PCA is applied to raw data with centered constructs (default), 2 = SVD of construct correlation matrix.

References

Bell, R. C. (2003). An evaluation of indices used to represent construct structure. In G. Chiari & M. L. Nuzzo (Eds.), Psychological Constructivism and the Social World (pp. 297-305). Milan: FrancoAngeli.

Bonarius, J. C. J. (1965). Research in the personal construct theory of George A. Kelly: role construct repertory test and basic theory. In B. A. Maher (Ed.), Progress in experimental personality research (Vol. 2). New York: Academic Press.

James, R. E. (1954). Identification in terms of personal constructs (Unpublished doctoral thesis). Ohio State University, Columbus, OH.

Examples

indexPvaff(bell2010)
#> Note: As of v0.1.14 PVAFF is derived using PCA of the construct centered ratings by default.Before that the construct correlation matrix was used (see method=2).
#> 
#> [1] 0.3402002