Uncertainty index

A measure for the degree of dispersion of dependency in a dependency grid (Bell, 2001). It is normalized measure with a value range between 0 and 1. The index is applicable to dependency grids (e.g., situation-resource) only, i.e., all grid ratings must be 0 or 1.

Usage

indexUncertainty(x)

Arguments

x

A repgrid object with 0/1 ratings only, where 1 indicates a dependency.

Details

Theoretical Background: Dispersion of Dependency: Kelly (1969) proposed that it is problematic to view people as either independent or dependent because everyone is, to greater or lesser degrees, dependent upon others in life. What Kelly felt was important was how well people disperse their dependencies across different people. Whereas young children tend to have their dependencies concentrated on a small number of people (typically parents), adults are more likely to spread their dependencies across a variety of others. Dispersing one's dependencies is generally considered more psychologically adjusted for adults (Walker et al., 1988).

References

Bell, R. C. (2001). Some new Measures of the Dispersion of Dependency in a Situation-Resource Grid. Journal of Constructivist Psychology, 14(3), 227-234, doi:10.1080/713840106 .

See also

Examples

# sample grid from Bell (2001, p.231)
file <- system.file("extdata", "dep_grid_bell_2001.xlsx" , package = "OpenRepGrid")
x <- importExcel(file)

indexUncertainty(x) 
#> Uncertainty Index 
#>         0.9909569