Map arbitrary numeric vector to a given range of values.

From a given numeric vector z the range is determined and the values are linearly mapped onto the interval given by val.range. This function can be used in order to map arbitrary vectors to a given range of values.

Usage

mapCoordinatesToValue(z, val.range = c(0.5, 1))

Arguments

z

numeric vector.

val.range

numeric vector of lengths two (default c(.5, 1)).

Value

numeric vector