Number of matches in ratings

Count the number of matches, i.e. (near) identical ratings between two elements or constructs. Matches are used as the basis for the calculation of grid indexes.

Usage

matches(x, deviation = 0, diag.na = TRUE)

Arguments

x

A repgrid object.

deviation

Maximal difference between ratings to be considered a match (default 0 = only identical rating scores are a match). Especially useful for long rating scale (e.g. 0 to 100).

diag.na

Whether to set the diagonal of the matrices to NA (default is TRUE).

Value

A list of class org.matches with:

grid: The grid used to calculate the matches.

deviation The deviation parameter.

max_constructs Maximum possible number of matches across constructs.

max_elements Maximum possible number of matches across elements.

total_constructs Total number of matches across constructs.

total_elements Total number of matches across elements.

constructs: Matrix with no. of matches for constructs.

elements: Matrix with no. of matches for elements.

Examples

m <- matches(boeker)

# several output options
print(m, index = FALSE, names = FALSE, upper = FALSE)
#> 
#> ##############
#> RATING MATCHES
#> ##############
#> 
#> Maximal rating difference to count as match:  0
#> 
#> Total no. of matches between constructs:  327
#> Maximum possible no. of matches between constructs:  1365
#> 
#> Total no. of matches between elements:  343
#> Maximum possible no. of matches between elements:  1470
#> 
#> ELEMENTS
#> 
#>     1  2  3  4  5  6  7  8  9 10 11 12 13 14 15
#> 1  NA  3  6  3  4  2  3  2  4  4  1  1  4  3  4
#> 2   3 NA  2  4  2  1  3  2  3  1  2  1  2  1  2
#> 3   6  2 NA  2  5  1  5  3  4  4  1  5  7  6  4
#> 4   3  4  2 NA  2  3  4  5  3  3  2  3  3  4  5
#> 5   4  2  5  2 NA  2  8  1  5  6  2  2  3  6  3
#> 6   2  1  1  3  2 NA  2  8  3  3  2  4  1  1  5
#> 7   3  3  5  4  8  2 NA  0  5  5  2  2  4  3  2
#> 8   2  2  3  5  1  8  0 NA  3  3  4  3  2  2  4
#> 9   4  3  4  3  5  3  5  3 NA  3  3  4  3  3  2
#> 10  4  1  4  3  6  3  5  3  3 NA  3  4  4  4  2
#> 11  1  2  1  2  2  2  2  4  3  3 NA  2  1  0  1
#> 12  1  1  5  3  2  4  2  3  4  4  2 NA  9  5  6
#> 13  4  2  7  3  3  1  4  2  3  4  1  9 NA  8  5
#> 14  3  1  6  4  6  1  3  2  3  4  0  5  8 NA  6
#> 15  4  2  4  5  3  5  2  4  2  2  1  6  5  6 NA
#> 
#> CONSTRUCTS
#> 
#>     1  2  3  4  5  6  7  8  9 10 11 12 13 14
#> 1  NA  2  3  2  6  5  4  4  6  1  4  5  4  6
#> 2   2 NA  3  2  4  3  2  3  2  4  1  5  6  2
#> 3   3  3 NA  4  3  0  5  5  2  2  5  5  4  2
#> 4   2  2  4 NA  6  4  4  4  6  1  3  3  3  2
#> 5   6  4  3  6 NA  4  5  4  7  2  4  7  4  3
#> 6   5  3  0  4  4 NA  4  6  8  4  2  4  2  4
#> 7   4  2  5  4  5  4 NA  3  8  1  5  4  2  3
#> 8   4  3  5  4  4  6  3 NA  3  3  2  6  8  3
#> 9   6  2  2  6  7  8  8  3 NA  0  2  6  2  5
#> 10  1  4  2  1  2  4  1  3  0 NA  3  3  3  0
#> 11  4  1  5  3  4  2  5  2  2  3 NA  3  0  3
#> 12  5  5  5  3  7  4  4  6  6  3  3 NA  4  3
#> 13  4  6  4  3  4  2  2  8  2  3  0  4 NA  3
#> 14  6  2  2  2  3  4  3  3  5  0  3  3  3 NA
print(m, output = "C")  # construct matches
#> 
#> ##############
#> RATING MATCHES
#> ##############
#> 
#> CONSTRUCTS
#> 
#>                                            1  2  3  4  5  6  7  8  9 10 11 12
#> 1  balanced - get along with conflicts  1     2  3  2  6  5  4  4  6  1  4  5
#> 2                  isolated - sociable  2        3  2  4  3  2  3  2  4  1  5
#> 3        closely integrated - excluded  3           4  3  0  5  5  2  2  5  5
#> 4                 discursive - passive  4              6  4  4  4  6  1  3  3
#> 5            open minded - indifferent  5                 4  5  4  7  2  4  7
#> 6               dreamy - dispassionate  6                    4  6  8  4  2  4
#> 7     practically oriented - depressed  7                       3  8  1  5  4
#> 8                    playful - serious  8                          3  3  2  6
#> 9            socially minded - selfish  9                             0  2  6
#> 10              quarrelsome - peaceful 10                                3  3
#> 11                artistic - technical 11                                   3
#> 12              scientific - emotional 12                                    
#> 13               introvert - extrovert 13                                    
#> 14          wanderlust - home oriented 14                                    
#>    13 14
#> 1   4  6
#> 2   6  2
#> 3   4  2
#> 4   3  2
#> 5   4  3
#> 6   2  4
#> 7   2  3
#> 8   8  3
#> 9   2  5
#> 10  3  0
#> 11  0  3
#> 12  4  3
#> 13     3
#> 14      
print(m, output = "E")  # element matches
#> 
#> ##############
#> RATING MATCHES
#> ##############
#> 
#> ELEMENTS
#> 
#>                            1  2  3  4  5  6  7  8  9 10 11 12 13 14 15
#> 1                 self  1     3  6  3  4  2  3  2  4  4  1  1  4  3  4
#> 2           ideal self  2        2  4  2  1  3  2  3  1  2  1  2  1  2
#> 3               mother  3           2  5  1  5  3  4  4  1  5  7  6  4
#> 4               father  4              2  3  4  5  3  3  2  3  3  4  5
#> 5                 kurt  5                 2  8  1  5  6  2  2  3  6  3
#> 6                 karl  6                    2  8  3  3  2  4  1  1  5
#> 7               george  7                       0  5  5  2  2  4  3  2
#> 8               martin  8                          3  3  4  3  2  2  4
#> 9            elizabeth  9                             3  3  4  3  3  2
#> 10           therapist 10                                3  4  4  4  2
#> 11               irene 11                                   2  1  0  1
#> 12      childhood self 12                                      9  5  6
#> 13 self before illness 13                                         8  5
#> 14  self with delusion 14                                            6
#> 15     self as dreamer 15                                             

# extract the matrices
m$constructs
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   NA    2    3    2    6    5    4    4    6     1     4     5     4
#>  [2,]    2   NA    3    2    4    3    2    3    2     4     1     5     6
#>  [3,]    3    3   NA    4    3    0    5    5    2     2     5     5     4
#>  [4,]    2    2    4   NA    6    4    4    4    6     1     3     3     3
#>  [5,]    6    4    3    6   NA    4    5    4    7     2     4     7     4
#>  [6,]    5    3    0    4    4   NA    4    6    8     4     2     4     2
#>  [7,]    4    2    5    4    5    4   NA    3    8     1     5     4     2
#>  [8,]    4    3    5    4    4    6    3   NA    3     3     2     6     8
#>  [9,]    6    2    2    6    7    8    8    3   NA     0     2     6     2
#> [10,]    1    4    2    1    2    4    1    3    0    NA     3     3     3
#> [11,]    4    1    5    3    4    2    5    2    2     3    NA     3     0
#> [12,]    5    5    5    3    7    4    4    6    6     3     3    NA     4
#> [13,]    4    6    4    3    4    2    2    8    2     3     0     4    NA
#> [14,]    6    2    2    2    3    4    3    3    5     0     3     3     3
#>       [,14]
#>  [1,]     6
#>  [2,]     2
#>  [3,]     2
#>  [4,]     2
#>  [5,]     3
#>  [6,]     4
#>  [7,]     3
#>  [8,]     3
#>  [9,]     5
#> [10,]     0
#> [11,]     3
#> [12,]     3
#> [13,]     3
#> [14,]    NA
m$elements
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   NA    3    6    3    4    2    3    2    4     4     1     1     4
#>  [2,]    3   NA    2    4    2    1    3    2    3     1     2     1     2
#>  [3,]    6    2   NA    2    5    1    5    3    4     4     1     5     7
#>  [4,]    3    4    2   NA    2    3    4    5    3     3     2     3     3
#>  [5,]    4    2    5    2   NA    2    8    1    5     6     2     2     3
#>  [6,]    2    1    1    3    2   NA    2    8    3     3     2     4     1
#>  [7,]    3    3    5    4    8    2   NA    0    5     5     2     2     4
#>  [8,]    2    2    3    5    1    8    0   NA    3     3     4     3     2
#>  [9,]    4    3    4    3    5    3    5    3   NA     3     3     4     3
#> [10,]    4    1    4    3    6    3    5    3    3    NA     3     4     4
#> [11,]    1    2    1    2    2    2    2    4    3     3    NA     2     1
#> [12,]    1    1    5    3    2    4    2    3    4     4     2    NA     9
#> [13,]    4    2    7    3    3    1    4    2    3     4     1     9    NA
#> [14,]    3    1    6    4    6    1    3    2    3     4     0     5     8
#> [15,]    4    2    4    5    3    5    2    4    2     2     1     6     5
#>       [,14] [,15]
#>  [1,]     3     4
#>  [2,]     1     2
#>  [3,]     6     4
#>  [4,]     4     5
#>  [5,]     6     3
#>  [6,]     1     5
#>  [7,]     3     2
#>  [8,]     2     4
#>  [9,]     3     2
#> [10,]     4     2
#> [11,]     0     1
#> [12,]     5     6
#> [13,]     8     5
#> [14,]    NA     6
#> [15,]     6    NA