Concatenate repgrid objects.

Simple concatenation of repgrid objects or list containing repgrid objects using the '+' operator.

Usage

# S4 method for repgrid,repgrid
+(e1, e2)

# S4 method for list,repgrid
+(e1, e2)

# S4 method for repgrid,list
+(e1, e2)

Arguments

e1, e2

A repgrid object.

Details

Methods for "+" function.

Examples

x <- bell2010
x + x
#> 
#> META DATA:
#> Number of constructs:  18 
#> Number of elements:  10 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>    A teacher you respected - 5 6 - Mother (or the person who
#> on of the opposite sex t - 4 | | 7 - A person of the opposit
#> ppiest person you know - 3 | | | | 8 - The most confident pe
#> iend of the same sex - 2 | | | | | | 9 - A person you work w
#>               self - 1 | | | | | | | | 10 - A teacher you di
#>                      | | | | | | | | | |                    
#>        relaxed (1)   4 4 6 5 3 6 5 2 2 6   (1) worried & ten
#> (academically) (2)   6 7 6 5 7 4 6 7 4 7   (2) smart (academ
#> islikes sports (3)   6 3 7 6 4 4 2 3 6 3   (3) loves sports 
#> ot interactive (4)   6 7 5 6 6 5 6 7 7 4   (4) loves people 
#> ot transparent (5)   6 4 5 7 3 7 6 5 6 3   (5) transparent  
#>    insensitive (6)   4 6 5 4 4 6 5 3 4 2   (6) sensitive    
#>  fearful&timid (7)   5 4 4 6 5 3 5 6 5 5   (7) fearless     
#>          rough (8)   5 6 6 4 5 7 7 3 5 6   (8) gentle       
#> ccept as it is (9)   5 5 6 7 4 4 6 7 5 5   (9) loves to argu
#>       relaxed (10)   4 4 6 5 3 6 5 2 2 6   (10) worried & te
#> academically) (11)   6 7 6 5 7 4 6 7 4 7   (11) smart (acade
#> slikes sports (12)   6 3 7 6 4 4 2 3 6 3   (12) loves sports
#> t interactive (13)   6 7 5 6 6 5 6 7 7 4   (13) loves people
#> t transparent (14)   6 4 5 7 3 7 6 5 6 3   (14) transparent 
#>   insensitive (15)   4 6 5 4 4 6 5 3 4 2   (15) sensitive   
#> fearful&timid (16)   5 4 4 6 5 3 5 6 5 5   (16) fearless    
#>         rough (17)   5 6 6 4 5 7 7 3 5 6   (17) gentle      
#> cept as it is (18)   5 5 6 7 4 4 6 7 5 5   (18) loves to arg
#> 
x + list(x, x)
#> 
#> META DATA:
#> Number of constructs:  27 
#> Number of elements:  10 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>    A teacher you respected - 5 6 - Mother (or the person who
#> on of the opposite sex t - 4 | | 7 - A person of the opposit
#> ppiest person you know - 3 | | | | 8 - The most confident pe
#> iend of the same sex - 2 | | | | | | 9 - A person you work w
#>               self - 1 | | | | | | | | 10 - A teacher you di
#>                      | | | | | | | | | |                    
#>        relaxed (1)   4 4 6 5 3 6 5 2 2 6   (1) worried & ten
#> (academically) (2)   6 7 6 5 7 4 6 7 4 7   (2) smart (academ
#> islikes sports (3)   6 3 7 6 4 4 2 3 6 3   (3) loves sports 
#> ot interactive (4)   6 7 5 6 6 5 6 7 7 4   (4) loves people 
#> ot transparent (5)   6 4 5 7 3 7 6 5 6 3   (5) transparent  
#>    insensitive (6)   4 6 5 4 4 6 5 3 4 2   (6) sensitive    
#>  fearful&timid (7)   5 4 4 6 5 3 5 6 5 5   (7) fearless     
#>          rough (8)   5 6 6 4 5 7 7 3 5 6   (8) gentle       
#> ccept as it is (9)   5 5 6 7 4 4 6 7 5 5   (9) loves to argu
#>       relaxed (10)   4 4 6 5 3 6 5 2 2 6   (10) worried & te
#> academically) (11)   6 7 6 5 7 4 6 7 4 7   (11) smart (acade
#> slikes sports (12)   6 3 7 6 4 4 2 3 6 3   (12) loves sports
#> t interactive (13)   6 7 5 6 6 5 6 7 7 4   (13) loves people
#> t transparent (14)   6 4 5 7 3 7 6 5 6 3   (14) transparent 
#>   insensitive (15)   4 6 5 4 4 6 5 3 4 2   (15) sensitive   
#> fearful&timid (16)   5 4 4 6 5 3 5 6 5 5   (16) fearless    
#>         rough (17)   5 6 6 4 5 7 7 3 5 6   (17) gentle      
#> cept as it is (18)   5 5 6 7 4 4 6 7 5 5   (18) loves to arg
#>       relaxed (19)   4 4 6 5 3 6 5 2 2 6   (19) worried & te
#> academically) (20)   6 7 6 5 7 4 6 7 4 7   (20) smart (acade
#> slikes sports (21)   6 3 7 6 4 4 2 3 6 3   (21) loves sports
#> t interactive (22)   6 7 5 6 6 5 6 7 7 4   (22) loves people
#> t transparent (23)   6 4 5 7 3 7 6 5 6 3   (23) transparent 
#>   insensitive (24)   4 6 5 4 4 6 5 3 4 2   (24) sensitive   
#> fearful&timid (25)   5 4 4 6 5 3 5 6 5 5   (25) fearless    
#>         rough (26)   5 6 6 4 5 7 7 3 5 6   (26) gentle      
#> cept as it is (27)   5 5 6 7 4 4 6 7 5 5   (27) loves to arg
#> 
list(x, x) + x
#> 
#> META DATA:
#> Number of constructs:  27 
#> Number of elements:  10 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>    A teacher you respected - 5 6 - Mother (or the person who
#> on of the opposite sex t - 4 | | 7 - A person of the opposit
#> ppiest person you know - 3 | | | | 8 - The most confident pe
#> iend of the same sex - 2 | | | | | | 9 - A person you work w
#>               self - 1 | | | | | | | | 10 - A teacher you di
#>                      | | | | | | | | | |                    
#>        relaxed (1)   4 4 6 5 3 6 5 2 2 6   (1) worried & ten
#> (academically) (2)   6 7 6 5 7 4 6 7 4 7   (2) smart (academ
#> islikes sports (3)   6 3 7 6 4 4 2 3 6 3   (3) loves sports 
#> ot interactive (4)   6 7 5 6 6 5 6 7 7 4   (4) loves people 
#> ot transparent (5)   6 4 5 7 3 7 6 5 6 3   (5) transparent  
#>    insensitive (6)   4 6 5 4 4 6 5 3 4 2   (6) sensitive    
#>  fearful&timid (7)   5 4 4 6 5 3 5 6 5 5   (7) fearless     
#>          rough (8)   5 6 6 4 5 7 7 3 5 6   (8) gentle       
#> ccept as it is (9)   5 5 6 7 4 4 6 7 5 5   (9) loves to argu
#>       relaxed (10)   4 4 6 5 3 6 5 2 2 6   (10) worried & te
#> academically) (11)   6 7 6 5 7 4 6 7 4 7   (11) smart (acade
#> slikes sports (12)   6 3 7 6 4 4 2 3 6 3   (12) loves sports
#> t interactive (13)   6 7 5 6 6 5 6 7 7 4   (13) loves people
#> t transparent (14)   6 4 5 7 3 7 6 5 6 3   (14) transparent 
#>   insensitive (15)   4 6 5 4 4 6 5 3 4 2   (15) sensitive   
#> fearful&timid (16)   5 4 4 6 5 3 5 6 5 5   (16) fearless    
#>         rough (17)   5 6 6 4 5 7 7 3 5 6   (17) gentle      
#> cept as it is (18)   5 5 6 7 4 4 6 7 5 5   (18) loves to arg
#>       relaxed (19)   4 4 6 5 3 6 5 2 2 6   (19) worried & te
#> academically) (20)   6 7 6 5 7 4 6 7 4 7   (20) smart (acade
#> slikes sports (21)   6 3 7 6 4 4 2 3 6 3   (21) loves sports
#> t interactive (22)   6 7 5 6 6 5 6 7 7 4   (22) loves people
#> t transparent (23)   6 4 5 7 3 7 6 5 6 3   (23) transparent 
#>   insensitive (24)   4 6 5 4 4 6 5 3 4 2   (24) sensitive   
#> fearful&timid (25)   5 4 4 6 5 3 5 6 5 5   (25) fearless    
#>         rough (26)   5 6 6 4 5 7 7 3 5 6   (26) gentle      
#> cept as it is (27)   5 5 6 7 4 4 6 7 5 5   (27) loves to arg
#>