Plot method for indexDilemma (network graph)

Produces a network graph using of the detected implicative dilemmas using the igraph package.

Usage

# S3 method for indexDilemma
plot(
  x,
  layout = "rows",
  both.poles = TRUE,
  node.size = 50,
  node.text.cex = 1,
  node.label.color = "black",
  node.color.discrepant = "darkolivegreen3",
  node.color.congruent = "lightcoral",
  edge.label.color = grey(0.4),
  edge.label.cex = 1,
  edge.digits = 2,
  edge.arrow.size = 0.5,
  edge.color = grey(0.6),
  edge.lty = 2,
  ...
)

Arguments

x

Object returned by indexDilemma.

layout

Name of layout. One of rows, circle, star, or nicely or a igraph layout function.

both.poles

Show both construct poles? (default TRUE). If FALSE only the poles corresponding to the implied undesired changes are shown.

node.size

Size of nodes (default 50).

node.text.cex

Text size of construct labels.

node.label.color

Color of construct labels.

node.color.discrepant, node.color.congruent

Color of discrepant and congruent constructs nodes.

edge.label.color, edge.label.cex

Color and size of correlation labels.

edge.color, edge.arrow.size

Color and Size of arrow.

edge.lty

Linetype of arrow.

digits

Number of digits for correlations.