Get / set preferred construct poles

Constructs are bipolar, usually with one pole being preferred (positive). Setting the preferred poles may is useful in some analyses.

Usage

preferredPoles(x)

preferredPoles(x) <- value

Arguments

x

A repgrid object.

value

Vector of with preferred poles. One of 'left', 'right', 'none' or NA. Abbreviations are allowsd ('l' for 'left').

Examples

x <- fbb2003

preferredPoles(x) # no preferences assigned yet
#> [1] NA NA NA NA NA NA NA NA NA

# set preference by ideal rating
x <- preferredPolesByIdeal(x, ideal = "as I would love to be")
x <- preferredPolesByIdeal(x, ideal = 7) # same with element index
x
#> 
#> META DATA:
#> Number of elements:  8 
#> Number of constructs:  9 
#> Preferred poles defined:  9/9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>            an ethical person - 4 5 - my mother                   
#>               an old flame - 3 | | 6 - a rejected teacher        
#>                my father - 2 | | | | 7 - as I would love to be   
#>                   self - 1 | | | | | | 8 - a pitied person       
#>                          | | | | | | | |                         
#>            clever (+1)   2 1 6 3 5 7 1 5   (1-) not bright       
#>      disorganized (-2)   6 6 4 5 2 2 5 2   (2+) organized        
#>           listens (+3)   3 1 6 3 3 7 1 4   (3-) doesn't hear     
#>     no clear view (-4)   5 6 3 3 3 5 7 3   (4+) clear view of lif
#>    understands me (+5)   3 2 6 2 2 6 2 5   (5-) no understanding 
#>         ambitious (+6)   6 3 5 4 7 3 3 5   (6-) no ambition      
#>         respected (+7)   2 2 4 2 5 6 1 4   (7-) not respected    
#>           distant (-8)   3 3 7 3 5 1 6 5   (8+) warm             
#> rather aggressive (-9)   1 3 3 3 5 2 5 7   (9+) not aggressive   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined

# set preferred poles manually
preferredPoles(x) <- c("left", "right", "left", "r", "l", "l", "l", "r", "r")
x
#> 
#> META DATA:
#> Number of elements:  8 
#> Number of constructs:  9 
#> Preferred poles defined:  9/9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>            an ethical person - 4 5 - my mother                   
#>               an old flame - 3 | | 6 - a rejected teacher        
#>                my father - 2 | | | | 7 - as I would love to be   
#>                   self - 1 | | | | | | 8 - a pitied person       
#>                          | | | | | | | |                         
#>            clever (+1)   2 1 6 3 5 7 1 5   (1-) not bright       
#>      disorganized (-2)   6 6 4 5 2 2 5 2   (2+) organized        
#>           listens (+3)   3 1 6 3 3 7 1 4   (3-) doesn't hear     
#>     no clear view (-4)   5 6 3 3 3 5 7 3   (4+) clear view of lif
#>    understands me (+5)   3 2 6 2 2 6 2 5   (5-) no understanding 
#>         ambitious (+6)   6 3 5 4 7 3 3 5   (6-) no ambition      
#>         respected (+7)   2 2 4 2 5 6 1 4   (7-) not respected    
#>           distant (-8)   3 3 7 3 5 1 6 5   (8+) warm             
#> rather aggressive (-9)   1 3 3 3 5 2 5 7   (9+) not aggressive   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined

# change preferance for constructs 1 and 5
preferredPoles(x)[2] <- "left"
x
#> 
#> META DATA:
#> Number of elements:  8 
#> Number of constructs:  9 
#> Preferred poles defined:  9/9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>            an ethical person - 4 5 - my mother                   
#>               an old flame - 3 | | 6 - a rejected teacher        
#>                my father - 2 | | | | 7 - as I would love to be   
#>                   self - 1 | | | | | | 8 - a pitied person       
#>                          | | | | | | | |                         
#>            clever (+1)   2 1 6 3 5 7 1 5   (1-) not bright       
#>      disorganized (+2)   6 6 4 5 2 2 5 2   (2-) organized        
#>           listens (+3)   3 1 6 3 3 7 1 4   (3-) doesn't hear     
#>     no clear view (-4)   5 6 3 3 3 5 7 3   (4+) clear view of lif
#>    understands me (+5)   3 2 6 2 2 6 2 5   (5-) no understanding 
#>         ambitious (+6)   6 3 5 4 7 3 3 5   (6-) no ambition      
#>         respected (+7)   2 2 4 2 5 6 1 4   (7-) not respected    
#>           distant (-8)   3 3 7 3 5 1 6 5   (8+) warm             
#> rather aggressive (-9)   1 3 3 3 5 2 5 7   (9+) not aggressive   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined

# remove prefernces
preferredPoles(x) <- NA
x
#> 
#> META DATA:
#> Number of elements:  8 
#> Number of constructs:  9 
#> Preferred poles defined:  0/9 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>            an ethical person - 4 5 - my mother                   
#>               an old flame - 3 | | 6 - a rejected teacher        
#>                my father - 2 | | | | 7 - as I would love to be   
#>                   self - 1 | | | | | | 8 - a pitied person       
#>                          | | | | | | | |                         
#>            clever (.1)   2 1 6 3 5 7 1 5   (1.) not bright       
#>      disorganized (.2)   6 6 4 5 2 2 5 2   (2.) organized        
#>           listens (.3)   3 1 6 3 3 7 1 4   (3.) doesn't hear     
#>     no clear view (.4)   5 6 3 3 3 5 7 3   (4.) clear view of lif
#>    understands me (.5)   3 2 6 2 2 6 2 5   (5.) no understanding 
#>         ambitious (.6)   6 3 5 4 7 3 3 5   (6.) no ambition      
#>         respected (.7)   2 2 4 2 5 6 1 4   (7.) not respected    
#>           distant (.8)   3 3 7 3 5 1 6 5   (8.) warm             
#> rather aggressive (.9)   1 3 3 3 5 2 5 7   (9.) not aggressive   
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined