Invert construct and element order

Invert construct and element order

Usage

# S3 method for repgrid
reorder(x, what = "CE", ...)

Arguments

x

A repgrid object.

what

A string or numeric to indicate if constructs ("C", 1) or elements ("C", 1), or both ("CE", 12) should be reversed.

...

Ignored.

Examples

# invert order of constructs
reorder(boeker, "C")
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#>     wanderlust (1)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (1) home oriented
#>      introvert (2)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (2) extrovert    
#>     scientific (3)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (3) emotional    
#>       artistic (4)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (4) technical    
#>    quarrelsome (5)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (5) peaceful     
#> ocially minded (6)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (6) selfish      
#>        playful (7)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (7) serious      
#> cally oriented (8)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (8) depressed    
#>         dreamy (9)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (9) dispassionate
#>   open minded (10)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (10) indifferent 
#>    discursive (11)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (11) passive     
#> ly integrated (12)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (12) excluded    
#>      isolated (13)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (13) sociable    
#>      balanced (14)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (14) get along wi
#> 
reorder(boeker, 1)
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#>     wanderlust (1)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (1) home oriented
#>      introvert (2)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (2) extrovert    
#>     scientific (3)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (3) emotional    
#>       artistic (4)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (4) technical    
#>    quarrelsome (5)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (5) peaceful     
#> ocially minded (6)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (6) selfish      
#>        playful (7)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (7) serious      
#> cally oriented (8)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (8) depressed    
#>         dreamy (9)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (9) dispassionate
#>   open minded (10)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (10) indifferent 
#>    discursive (11)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (11) passive     
#> ly integrated (12)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (12) excluded    
#>      isolated (13)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (13) sociable    
#>      balanced (14)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (14) get along wi
#> 

# invert order of elements
reorder(boeker, "E")
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                      elizabeth - 7 | 9 - george                       
#>                    therapist - 6 | | | 10 - karl                      
#>                      irene - 5 | | | | | 11 - kurt                    
#>           childhood self - 4 | | | | | | | 12 - father                
#>    self before illness - 3 | | | | | | | | | 13 - mother              
#>   self with delusion - 2 | | | | | | | | | | | 14 - ideal self        
#>    self as dreamer - 1 | | | | | | | | | | | | | 15 - self            
#>                      | | | | | | | | | | | | | | |                    
#>       balanced (1)   3 3 2 2 6 2 4 5 2 5 3 2 2 4 1   (1) get along wit
#>       isolated (2)   3 2 2 4 4 4 5 4 5 4 5 5 3 6 3   (2) sociable     
#> ely integrated (3)   3 5 4 4 3 3 2 3 2 3 5 3 2 2 2   (3) excluded     
#>     discursive (4)   4 5 3 3 3 2 3 3 2 4 2 1 3 1 4   (4) passive      
#>    open minded (5)   3 2 2 3 6 2 4 2 4 4 2 1 2 1 2   (5) indifferent  
#>         dreamy (6)   2 3 3 3 6 4 4 5 4 5 4 5 3 5 4   (6) dispassionate
#> cally oriented (7)   3 3 3 3 2 3 2 2 3 3 3 2 3 1 2   (7) depressed    
#>        playful (8)   3 4 2 3 5 4 4 3 2 3 4 3 4 5 4   (8) serious      
#> ocially minded (9)   3 3 3 3 6 2 3 1 4 5 4 2 3 1 2   (9) selfish      
#>   quarrelsome (10)   5 5 5 6 1 4 4 2 5 2 5 5 5 5 5   (10) peaceful    
#>      artistic (11)   4 4 4 4 3 3 4 2 3 5 3 4 2 1 5   (11) technical   
#>    scientific (12)   3 3 2 4 6 1 4 3 5 4 4 3 5 1 2   (12) emotional   
#>     introvert (13)   3 2 2 2 5 2 5 3 5 3 5 6 4 5 4   (13) extrovert   
#>    wanderlust (14)   1 2 1 1 6 3 5 5 2 5 4 2 4 1 1   (14) home oriente
#> 
reorder(boeker, 2)
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                      elizabeth - 7 | 9 - george                       
#>                    therapist - 6 | | | 10 - karl                      
#>                      irene - 5 | | | | | 11 - kurt                    
#>           childhood self - 4 | | | | | | | 12 - father                
#>    self before illness - 3 | | | | | | | | | 13 - mother              
#>   self with delusion - 2 | | | | | | | | | | | 14 - ideal self        
#>    self as dreamer - 1 | | | | | | | | | | | | | 15 - self            
#>                      | | | | | | | | | | | | | | |                    
#>       balanced (1)   3 3 2 2 6 2 4 5 2 5 3 2 2 4 1   (1) get along wit
#>       isolated (2)   3 2 2 4 4 4 5 4 5 4 5 5 3 6 3   (2) sociable     
#> ely integrated (3)   3 5 4 4 3 3 2 3 2 3 5 3 2 2 2   (3) excluded     
#>     discursive (4)   4 5 3 3 3 2 3 3 2 4 2 1 3 1 4   (4) passive      
#>    open minded (5)   3 2 2 3 6 2 4 2 4 4 2 1 2 1 2   (5) indifferent  
#>         dreamy (6)   2 3 3 3 6 4 4 5 4 5 4 5 3 5 4   (6) dispassionate
#> cally oriented (7)   3 3 3 3 2 3 2 2 3 3 3 2 3 1 2   (7) depressed    
#>        playful (8)   3 4 2 3 5 4 4 3 2 3 4 3 4 5 4   (8) serious      
#> ocially minded (9)   3 3 3 3 6 2 3 1 4 5 4 2 3 1 2   (9) selfish      
#>   quarrelsome (10)   5 5 5 6 1 4 4 2 5 2 5 5 5 5 5   (10) peaceful    
#>      artistic (11)   4 4 4 4 3 3 4 2 3 5 3 4 2 1 5   (11) technical   
#>    scientific (12)   3 3 2 4 6 1 4 3 5 4 4 3 5 1 2   (12) emotional   
#>     introvert (13)   3 2 2 2 5 2 5 3 5 3 5 6 4 5 4   (13) extrovert   
#>    wanderlust (14)   1 2 1 1 6 3 5 5 2 5 4 2 4 1 1   (14) home oriente
#> 

# invert both (default)
reorder(boeker)
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                      elizabeth - 7 | 9 - george                       
#>                    therapist - 6 | | | 10 - karl                      
#>                      irene - 5 | | | | | 11 - kurt                    
#>           childhood self - 4 | | | | | | | 12 - father                
#>    self before illness - 3 | | | | | | | | | 13 - mother              
#>   self with delusion - 2 | | | | | | | | | | | 14 - ideal self        
#>    self as dreamer - 1 | | | | | | | | | | | | | 15 - self            
#>                      | | | | | | | | | | | | | | |                    
#>     wanderlust (1)   1 2 1 1 6 3 5 5 2 5 4 2 4 1 1   (1) home oriented
#>      introvert (2)   3 2 2 2 5 2 5 3 5 3 5 6 4 5 4   (2) extrovert    
#>     scientific (3)   3 3 2 4 6 1 4 3 5 4 4 3 5 1 2   (3) emotional    
#>       artistic (4)   4 4 4 4 3 3 4 2 3 5 3 4 2 1 5   (4) technical    
#>    quarrelsome (5)   5 5 5 6 1 4 4 2 5 2 5 5 5 5 5   (5) peaceful     
#> ocially minded (6)   3 3 3 3 6 2 3 1 4 5 4 2 3 1 2   (6) selfish      
#>        playful (7)   3 4 2 3 5 4 4 3 2 3 4 3 4 5 4   (7) serious      
#> cally oriented (8)   3 3 3 3 2 3 2 2 3 3 3 2 3 1 2   (8) depressed    
#>         dreamy (9)   2 3 3 3 6 4 4 5 4 5 4 5 3 5 4   (9) dispassionate
#>   open minded (10)   3 2 2 3 6 2 4 2 4 4 2 1 2 1 2   (10) indifferent 
#>    discursive (11)   4 5 3 3 3 2 3 3 2 4 2 1 3 1 4   (11) passive     
#> ly integrated (12)   3 5 4 4 3 3 2 3 2 3 5 3 2 2 2   (12) excluded    
#>      isolated (13)   3 2 2 4 4 4 5 4 5 4 5 5 3 6 3   (13) sociable    
#>      balanced (14)   3 3 2 2 6 2 4 5 2 5 3 2 2 4 1   (14) get along wi
#> 
reorder(boeker, "CE")
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                      elizabeth - 7 | 9 - george                       
#>                    therapist - 6 | | | 10 - karl                      
#>                      irene - 5 | | | | | 11 - kurt                    
#>           childhood self - 4 | | | | | | | 12 - father                
#>    self before illness - 3 | | | | | | | | | 13 - mother              
#>   self with delusion - 2 | | | | | | | | | | | 14 - ideal self        
#>    self as dreamer - 1 | | | | | | | | | | | | | 15 - self            
#>                      | | | | | | | | | | | | | | |                    
#>     wanderlust (1)   1 2 1 1 6 3 5 5 2 5 4 2 4 1 1   (1) home oriented
#>      introvert (2)   3 2 2 2 5 2 5 3 5 3 5 6 4 5 4   (2) extrovert    
#>     scientific (3)   3 3 2 4 6 1 4 3 5 4 4 3 5 1 2   (3) emotional    
#>       artistic (4)   4 4 4 4 3 3 4 2 3 5 3 4 2 1 5   (4) technical    
#>    quarrelsome (5)   5 5 5 6 1 4 4 2 5 2 5 5 5 5 5   (5) peaceful     
#> ocially minded (6)   3 3 3 3 6 2 3 1 4 5 4 2 3 1 2   (6) selfish      
#>        playful (7)   3 4 2 3 5 4 4 3 2 3 4 3 4 5 4   (7) serious      
#> cally oriented (8)   3 3 3 3 2 3 2 2 3 3 3 2 3 1 2   (8) depressed    
#>         dreamy (9)   2 3 3 3 6 4 4 5 4 5 4 5 3 5 4   (9) dispassionate
#>   open minded (10)   3 2 2 3 6 2 4 2 4 4 2 1 2 1 2   (10) indifferent 
#>    discursive (11)   4 5 3 3 3 2 3 3 2 4 2 1 3 1 4   (11) passive     
#> ly integrated (12)   3 5 4 4 3 3 2 3 2 3 5 3 2 2 2   (12) excluded    
#>      isolated (13)   3 2 2 4 4 4 5 4 5 4 5 5 3 6 3   (13) sociable    
#>      balanced (14)   3 3 2 2 6 2 4 5 2 5 3 2 2 4 1   (14) get along wi
#> 
reorder(boeker, 12)
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                      elizabeth - 7 | 9 - george                       
#>                    therapist - 6 | | | 10 - karl                      
#>                      irene - 5 | | | | | 11 - kurt                    
#>           childhood self - 4 | | | | | | | 12 - father                
#>    self before illness - 3 | | | | | | | | | 13 - mother              
#>   self with delusion - 2 | | | | | | | | | | | 14 - ideal self        
#>    self as dreamer - 1 | | | | | | | | | | | | | 15 - self            
#>                      | | | | | | | | | | | | | | |                    
#>     wanderlust (1)   1 2 1 1 6 3 5 5 2 5 4 2 4 1 1   (1) home oriented
#>      introvert (2)   3 2 2 2 5 2 5 3 5 3 5 6 4 5 4   (2) extrovert    
#>     scientific (3)   3 3 2 4 6 1 4 3 5 4 4 3 5 1 2   (3) emotional    
#>       artistic (4)   4 4 4 4 3 3 4 2 3 5 3 4 2 1 5   (4) technical    
#>    quarrelsome (5)   5 5 5 6 1 4 4 2 5 2 5 5 5 5 5   (5) peaceful     
#> ocially minded (6)   3 3 3 3 6 2 3 1 4 5 4 2 3 1 2   (6) selfish      
#>        playful (7)   3 4 2 3 5 4 4 3 2 3 4 3 4 5 4   (7) serious      
#> cally oriented (8)   3 3 3 3 2 3 2 2 3 3 3 2 3 1 2   (8) depressed    
#>         dreamy (9)   2 3 3 3 6 4 4 5 4 5 4 5 3 5 4   (9) dispassionate
#>   open minded (10)   3 2 2 3 6 2 4 2 4 4 2 1 2 1 2   (10) indifferent 
#>    discursive (11)   4 5 3 3 3 2 3 3 2 4 2 1 3 1 4   (11) passive     
#> ly integrated (12)   3 5 4 4 3 3 2 3 2 3 5 3 2 2 2   (12) excluded    
#>      isolated (13)   3 2 2 4 4 4 5 4 5 4 5 5 3 6 3   (13) sociable    
#>      balanced (14)   3 3 2 2 6 2 4 5 2 5 3 2 2 4 1   (14) get along wi
#> 

# not reordering
reorder(boeker, NA)
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#>       balanced (1)   1 4 2 2 3 5 2 5 4 2 6 2 2 3 3   (1) get along wit
#>       isolated (2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2) sociable     
#> ely integrated (3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3) excluded     
#>     discursive (4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4) passive      
#>    open minded (5)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (5) indifferent  
#>         dreamy (6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6) dispassionate
#> cally oriented (7)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (7) depressed    
#>        playful (8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8) serious      
#> ocially minded (9)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (9) selfish      
#>   quarrelsome (10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10) peaceful    
#>      artistic (11)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (11) technical   
#>    scientific (12)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (12) emotional   
#>     introvert (13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13) extrovert   
#>    wanderlust (14)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (14) home oriente
#>