Order grid by angles between construct and/or elements in 2D.

The approach is to reorder the grid matrix by their polar angles on the first two principal components from a data reduction technique (here the biplot, i.e. SVD). The function reorder2d reorders the grid according to the angles between the x-axis and the element (construct) vectors derived from a 2D biplot solution. This approach is apt to identify circumplex structures in data indicated by the diagonal stripe in the display (see examples).

Usage

reorder2d(
  x,
  dim = c(1, 2),
  center = 1,
  normalize = 0,
  g = 0,
  h = 1 - g,
  rc = TRUE,
  re = TRUE,
  ...
)

Arguments

x

repgrid object.

dim

Dimension of 2D solution used to calculate angles (default c(1,2)).

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). The default is 1 (row centering).

normalize

A numeric value indicating along what direction (rows, columns) to normalize by standard deviations. 0 = none, 1= rows, 2 = columns (default is 0).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

rc

Logical. Reorder constructs by similarity (default TRUE).

re

Logical. Reorder elements by similarity (default TRUE).

...

Not evaluated.

Value

Reordered repgrid object.

Examples

x <- feixas2004
reorder2d(x) # reorder grid by angles in first two dimensions
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Non-grata                    
#>                       Father - 6 | 8 - Brother                    
#>                   Friend 1 - 5 | | | 9 - Mother                   
#>                 Friend 2 - 4 | | | | | 10 - Friend 4              
#>                 Cousin - 3 | | | | | | | 11 - Self now            
#>           Ideal Self - 2 | | | | | | | | | 12 - Friend 3          
#>          Boyfriend - 1 | | | | | | | | | | | 13 - Godmother       
#>                      | | | | | | | | | | | | |                    
#>   Hypocritical (1)   6 7 6 5 5 4 1 2 4 4 5 6 5   (1) Sincere      
#>  Lives to work (2)   6 7 7 6 6 2 1 2 1 6 5 6 6   (2) Works to live
#>    Blackmailer (3)   5 7 6 6 6 2 1 1 2 3 3 6 6   (3) Non blackmail
#>    Pessimistic (4)   7 7 4 6 3 5 2 2 1 2 1 6 3   (4) Optimistic   
#>    Introverted (5)   4 5 6 7 5 6 5 2 2 5 1 2 6   (5) Extroverted  
#>        Fearful (6)   4 5 3 6 5 6 5 2 2 5 2 2 4   (6) Enterprising 
#> pressed easily (7)   6 6 3 7 3 6 6 3 2 3 1 1 3   (7) Does not get 
#> Self-demanding (8)   2 4 6 6 5 6 3 2 6 5 1 5 4   (8) Takes it easy
#> d about others (9)   2 2 3 5 3 6 7 7 2 2 2 3 2   (9) Selfish      
#>     Sensitive (10)   1 1 4 5 4 7 7 7 5 4 1 1 3   (10) Materialisti
#> ciates others (11)   1 1 2 4 5 6 7 6 6 5 2 4 2   (11) Does not app
#>       Teasing (12)   4 3 3 4 3 1 6 6 7 6 2 3 5   (12) Touchy      
#>  good in thin (13)   1 1 3 2 5 4 7 6 6 5 6 6 3   (13) Sees only th
#> er the friend (14)   6 7 4 1 2 3 2 3 3 4 6 4 6   (14) Looks after 
#> onger than is (15)   5 6 6 4 2 1 2 2 3 5 6 7 6   (15) Natural     
#> is/her wishes (16)   4 7 3 6 3 1 1 1 2 2 6 7 4   (16) Tolerant wit
#> on Accessible (17)   4 7 3 4 2 2 1 1 2 2 5 6 5   (17) Accessible  
#>        Cheeky (18)   6 7 5 6 6 5 1 4 6 5 6 6 6   (18) Respectful  
#>    Avaricious (19)   7 7 3 5 5 1 1 1 1 6 6 6 3   (19) Generous    
#>    Aggressive (20)   7 7 6 6 4 2 2 2 4 3 6 6 6   (20) Calm        
#> 
reorder2d(x, rc = FALSE) # reorder elements only
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Non-grata                    
#>                       Father - 6 | 8 - Brother                    
#>                   Friend 1 - 5 | | | 9 - Mother                   
#>                 Friend 2 - 4 | | | | | 10 - Friend 4              
#>                 Cousin - 3 | | | | | | | 11 - Self now            
#>           Ideal Self - 2 | | | | | | | | | 12 - Friend 3          
#>          Boyfriend - 1 | | | | | | | | | | | 13 - Godmother       
#>                      | | | | | | | | | | | | |                    
#>    Pessimistic (1)   7 7 4 6 3 5 2 2 1 2 1 6 3   (1) Optimistic   
#> Self-demanding (2)   2 4 6 6 5 6 3 2 6 5 1 5 4   (2) Takes it easy
#>        Fearful (3)   4 5 3 6 5 6 5 2 2 5 2 2 4   (3) Enterprising 
#>  Lives to work (4)   6 7 7 6 6 2 1 2 1 6 5 6 6   (4) Works to live
#> his/her wishes (5)   4 7 3 6 3 1 1 1 2 2 6 7 4   (5) Tolerant with
#>        Teasing (6)   4 3 3 4 3 1 6 6 7 6 2 3 5   (6) Touchy       
#> eciates others (7)   1 1 2 4 5 6 7 6 6 5 2 4 2   (7) Does not appr
#>     Aggressive (8)   7 7 6 6 4 2 2 2 4 3 6 6 6   (8) Calm         
#> d about others (9)   2 2 3 5 3 6 7 7 2 2 2 3 2   (9) Selfish      
#>    Avaricious (10)   7 7 3 5 5 1 1 1 1 6 6 6 3   (10) Generous    
#>     Sensitive (11)   1 1 4 5 4 7 7 7 5 4 1 1 3   (11) Materialisti
#>        Cheeky (12)   6 7 5 6 6 5 1 4 6 5 6 6 6   (12) Respectful  
#>  Hypocritical (13)   6 7 6 5 5 4 1 2 4 4 5 6 5   (13) Sincere     
#>   Blackmailer (14)   5 7 6 6 6 2 1 1 2 3 3 6 6   (14) Non blackmai
#> onger than is (15)   5 6 6 4 2 1 2 2 3 5 6 7 6   (15) Natural     
#> er the friend (16)   6 7 4 1 2 3 2 3 3 4 6 4 6   (16) Looks after 
#> on Accessible (17)   4 7 3 4 2 2 1 1 2 2 5 6 5   (17) Accessible  
#>   Introverted (18)   4 5 6 7 5 6 5 2 2 5 1 2 6   (18) Extroverted 
#> ressed easily (19)   6 6 3 7 3 6 6 3 2 3 1 1 3   (19) Does not get
#>  good in thin (20)   1 1 3 2 5 4 7 6 6 5 6 6 3   (20) Sees only th
#> 
reorder2d(x, re = FALSE) # reorder constructs only
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>                                  7 - Friend 2                     
#>                     Friend 1 - 6 | 8 - Non-grata                  
#>                  Boyfriend - 5 | | | 9 - Friend 3                 
#>                  Brother - 4 | | | | | 10 - Cousin                
#>                 Father - 3 | | | | | | | 11 - Godmother           
#>               Mother - 2 | | | | | | | | | 12 - Friend 4          
#>           Self now - 1 | | | | | | | | | | | 13 - Ideal Self      
#>                      | | | | | | | | | | | | |                    
#>   Hypocritical (1)   5 4 4 2 6 5 5 1 6 6 5 4 7   (1) Sincere      
#>  Lives to work (2)   5 1 2 2 6 6 6 1 6 7 6 6 7   (2) Works to live
#>    Blackmailer (3)   3 2 2 1 5 6 6 1 6 6 6 3 7   (3) Non blackmail
#>    Pessimistic (4)   1 1 5 2 7 3 6 2 6 4 3 2 7   (4) Optimistic   
#>    Introverted (5)   1 2 6 2 4 5 7 5 2 6 6 5 5   (5) Extroverted  
#>        Fearful (6)   2 2 6 2 4 5 6 5 2 3 4 5 5   (6) Enterprising 
#> pressed easily (7)   1 2 6 3 6 3 7 6 1 3 3 3 6   (7) Does not get 
#> Self-demanding (8)   1 6 6 2 2 5 6 3 5 6 4 5 4   (8) Takes it easy
#> d about others (9)   2 2 6 7 2 3 5 7 3 3 2 2 2   (9) Selfish      
#>     Sensitive (10)   1 5 7 7 1 4 5 7 1 4 3 4 1   (10) Materialisti
#> ciates others (11)   2 6 6 6 1 5 4 7 4 2 2 5 1   (11) Does not app
#>       Teasing (12)   2 7 1 6 4 3 4 6 3 3 5 6 3   (12) Touchy      
#>  good in thin (13)   6 6 4 6 1 5 2 7 6 3 3 5 1   (13) Sees only th
#> er the friend (14)   6 3 3 3 6 2 1 2 4 4 6 4 7   (14) Looks after 
#> onger than is (15)   6 3 1 2 5 2 4 2 7 6 6 5 6   (15) Natural     
#> is/her wishes (16)   6 2 1 1 4 3 6 1 7 3 4 2 7   (16) Tolerant wit
#> on Accessible (17)   5 2 2 1 4 2 4 1 6 3 5 2 7   (17) Accessible  
#>        Cheeky (18)   6 6 5 4 6 6 6 1 6 5 6 5 7   (18) Respectful  
#>    Avaricious (19)   6 1 1 1 7 5 5 1 6 3 3 6 7   (19) Generous    
#>    Aggressive (20)   6 4 2 2 7 4 6 2 6 6 6 3 7   (20) Calm        
#>