Reverse constructs / swaps construct poles

Constructs are bipolar. They can be reversed without a change in meaning. The function swaps the constructs poles and re-adjusts ratings accordingly, i.e. it reversed selected constructs.

Usage

swapPoles(x, pos)

reverse(x, pos = 1L:nrow(x))

Arguments

x

A repgrid object.

pos

Row indexes of constructs to reverse..

Value

A repgrid object with reversed constructs.

Note

Please note that the scale of the rating grid has to be set in order to reverse constructs. If the scale is unknown no reversal occurs and an error is raised.

Examples

x <- boeker

reverse(x) # reverse all constructs
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#> with conflicts (1)   6 3 5 5 4 2 5 2 3 5 1 5 5 4 4   (1) balanced     
#>       sociable (2)   4 1 4 2 2 3 2 3 2 3 3 3 5 5 4   (2) isolated     
#>       excluded (3)   5 5 5 4 2 4 5 4 5 4 4 3 3 2 4   (3) closely integ
#>        passive (4)   3 6 4 6 5 3 5 4 4 5 4 4 4 2 3   (4) discursive   
#>    indifferent (5)   5 6 5 6 5 3 3 5 3 5 1 4 5 5 4   (5) open minded  
#>  dispassionate (6)   3 2 4 2 3 2 3 2 3 3 1 4 4 4 5   (6) dreamy       
#>      depressed (7)   5 6 4 5 4 4 4 5 5 4 5 4 4 4 4   (7) practically o
#>        serious (8)   3 2 3 4 3 4 5 4 3 3 2 4 5 3 4   (8) playful      
#>        selfish (9)   5 6 4 5 3 2 3 6 4 5 1 4 4 4 4   (9) socially mind
#>      peaceful (10)   2 2 2 2 2 5 2 5 3 3 6 1 2 2 2   (10) quarrelsome 
#>     technical (11)   2 6 5 3 4 2 4 5 3 4 4 3 3 3 3   (11) artistic    
#>     emotional (12)   5 6 2 4 3 3 2 4 3 6 1 3 5 4 4   (12) scientific  
#>     extrovert (13)   3 2 3 1 2 4 2 4 2 5 2 5 5 5 4   (13) introvert   
#> home oriented (14)   6 6 3 5 3 2 5 2 2 4 1 6 6 5 6   (14) wanderlust  
#> 
reverse(x, 1) # reverse construct 1
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#> with conflicts (1)   6 3 5 5 4 2 5 2 3 5 1 5 5 4 4   (1) balanced     
#>       isolated (2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2) sociable     
#> ely integrated (3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3) excluded     
#>     discursive (4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4) passive      
#>    open minded (5)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (5) indifferent  
#>         dreamy (6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6) dispassionate
#> cally oriented (7)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (7) depressed    
#>        playful (8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8) serious      
#> ocially minded (9)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (9) selfish      
#>   quarrelsome (10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10) peaceful    
#>      artistic (11)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (11) technical   
#>    scientific (12)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (12) emotional   
#>     introvert (13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13) extrovert   
#>    wanderlust (14)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (14) home oriente
#> 
reverse(x, 1:2) # reverse constructs 1 and 2
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#> with conflicts (1)   6 3 5 5 4 2 5 2 3 5 1 5 5 4 4   (1) balanced     
#>       sociable (2)   4 1 4 2 2 3 2 3 2 3 3 3 5 5 4   (2) isolated     
#> ely integrated (3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3) excluded     
#>     discursive (4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4) passive      
#>    open minded (5)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (5) indifferent  
#>         dreamy (6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6) dispassionate
#> cally oriented (7)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (7) depressed    
#>        playful (8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8) serious      
#> ocially minded (9)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (9) selfish      
#>   quarrelsome (10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10) peaceful    
#>      artistic (11)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (11) technical   
#>    scientific (12)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (12) emotional   
#>     introvert (13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13) extrovert   
#>    wanderlust (14)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (14) home oriente
#> 

# swapPoles will become deprecated, use reverse instead
swapPoles(x, 1) # swap construct poles of construct
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 6 (right pole)
#> 
#> RATINGS:
#>                                    8 - martin                         
#>                         george - 7 | 9 - elizabeth                    
#>                         karl - 6 | | | 10 - therapist                 
#>                       kurt - 5 | | | | | 11 - irene                   
#>                   father - 4 | | | | | | | 12 - childhood self        
#>                 mother - 3 | | | | | | | | | 13 - self before illness 
#>           ideal self - 2 | | | | | | | | | | | 14 - self with delusion
#>               self - 1 | | | | | | | | | | | | | 15 - self as dreamer 
#>                      | | | | | | | | | | | | | | |                    
#> with conflicts (1)   6 3 5 5 4 2 5 2 3 5 1 5 5 4 4   (1) balanced     
#>       isolated (2)   3 6 3 5 5 4 5 4 5 4 4 4 2 2 3   (2) sociable     
#> ely integrated (3)   2 2 2 3 5 3 2 3 2 3 3 4 4 5 3   (3) excluded     
#>     discursive (4)   4 1 3 1 2 4 2 3 3 2 3 3 3 5 4   (4) passive      
#>    open minded (5)   2 1 2 1 2 4 4 2 4 2 6 3 2 2 3   (5) indifferent  
#>         dreamy (6)   4 5 3 5 4 5 4 5 4 4 6 3 3 3 2   (6) dispassionate
#> cally oriented (7)   2 1 3 2 3 3 3 2 2 3 2 3 3 3 3   (7) depressed    
#>        playful (8)   4 5 4 3 4 3 2 3 4 4 5 3 2 4 3   (8) serious      
#> ocially minded (9)   2 1 3 2 4 5 4 1 3 2 6 3 3 3 3   (9) selfish      
#>   quarrelsome (10)   5 5 5 5 5 2 5 2 4 4 1 6 5 5 5   (10) peaceful    
#>      artistic (11)   5 1 2 4 3 5 3 2 4 3 3 4 4 4 4   (11) technical   
#>    scientific (12)   2 1 5 3 4 4 5 3 4 1 6 4 2 3 3   (12) emotional   
#>     introvert (13)   4 5 4 6 5 3 5 3 5 2 5 2 2 2 3   (13) extrovert   
#>    wanderlust (14)   1 1 4 2 4 5 2 5 5 3 6 1 1 2 1   (14) home oriente
#>