global settings for OpenRepGrid

global settings for OpenRepGrid

Usage

settings(...)

Arguments

...

Use parameter value pairs (par1=val1, par2=val2) to change a parameter. Use parameter names to request parameter's value ("par1", "par2").

Note

Currently the following parameters can be changed, ordered by topic. The default value is shown in the brackets at the end of a line.

show.scale: Show grid scale info? (TRUE)

show.meta: Show grid meta data? (TRUE)

show.trim: Number of chars to trim strings to (30)

show.cut: Maximum number of characters printed on the sides of a grid (20)

c.no: Print construct ID number? (TRUE)

e.no: Print element ID number? (TRUE)

Examples

if (FALSE) { # \dontrun{
# get current settings
settings()

# get some parameters
settings("show.scale", "show.meta")

# change parameters
bell2010

settings(show.meta = F)
bell2010

settings(show.scale = F, show.cut = 30)
bell2010
} # }