Calculate SSQ (accuracy) of biplot representation for elements and constructs.

Each construct and element are vectors in a multidimensional space. When reducing the representation to a lower dimensional space, a loss of information (sum-of-squares) will usually occur. The output of the function shows the proportion of sum-of-squares (SSQ) explained for the elements (constructs) and the amount explained by each principal component. This allows to assess which elements (construct) are represented how well in the current representation. Also it shows how much of the total variation is explained.

Usage

ssq(
  x,
  along = 2,
  center = 1,
  normalize = 0,
  g = 0,
  h = 1 - g,
  col.active = NA,
  col.passive = NA,
  ...
)

Arguments

x

repgrid object.

along

Numeric. Table of sum-of-squares (SSQ) for 1=constructs, 2=elements (default). Note that currently these calculations only make sense for biplot representations with g=1 and h=1 respectively.

center

Numeric. The type of centering to be performed. 0= no centering, 1= row mean centering (construct), 2= column mean centering (elements), 3= double-centering (construct and element means), 4= midpoint centering of rows (constructs). The default is 1 (row centering).

normalize

A numeric value indicating along what direction (rows, columns) to normalize by standard deviations. 0 = none, 1= rows, 2 = columns (default is 0).

g

Power of the singular value matrix assigned to the left singular vectors, i.e. the constructs.

h

Power of the singular value matrix assigned to the right singular vectors, i.e. the elements.

col.active

Columns (elements) that are no supplementary points, i.e. they are used in the SVD to find principal components. default is to use all elements.

col.passive

Columns (elements) that are supplementary points, i.e. they are NOT used in the SVD but projected into the component space afterwards. They do not determine the solution. Default is NA, i.e. no elements are set supplementary.

Value

A list containing three elements:

ssq.table: dataframe with sum-of-squares explained for element/construct by each dimension

ssq.table.cumsum: dataframe with cumulated sum-of-squares explained for element/construct number of dimensions

ssq.total: total sum-of-squares after pre-transforming grid matrix

Note

TODO: if g or h is not equal to 1 the SSQ does not measure accuracy of representation as currently the ssq of each point are set in contrast with the pre-transformed matrix.

Examples

# explained sum-of-squares for elements
ssq(bell2010)
#> 
#> Cumulated proportionof explained sum-of-squares for elements
#> 
#>                                                                     1D    2D
#> self                                                              0.08 59.22
#> closest friend of the same sex                                    0.38 35.04
#> the unhappiest person you know                                   38.26 42.08
#> A person of the opposite sex that you don't get along with        0.10 64.94
#> A teacher you respected                                          35.87 58.43
#> Mother (or the person who filled that kind of role)              85.10 85.34
#> A person of the opposite sex that you like more than you dislike  7.62 31.87
#> The most confident person you know                               86.90 88.75
#> A person you work well with (at job or on sports team etc)        2.76 64.62
#> A teacher you did not respect                                     1.18 59.89
#> TOTAL                                                            34.02 63.63
#>                                                                     3D    4D
#> self                                                             68.95 79.91
#> closest friend of the same sex                                   71.89 77.48
#> the unhappiest person you know                                   75.12 80.07
#> A person of the opposite sex that you don't get along with       76.38 96.73
#> A teacher you respected                                          58.50 97.25
#> Mother (or the person who filled that kind of role)              93.33 93.44
#> A person of the opposite sex that you like more than you dislike 54.54 93.82
#> The most confident person you know                               90.31 97.98
#> A person you work well with (at job or on sports team etc)       72.26 85.03
#> A teacher you did not respect                                    92.74 93.51
#> TOTAL                                                            79.58 91.19
#>                                                                     5D
#> self                                                             79.98
#> closest friend of the same sex                                   99.11
#> the unhappiest person you know                                   98.02
#> A person of the opposite sex that you don't get along with       97.60
#> A teacher you respected                                          97.53
#> Mother (or the person who filled that kind of role)              97.42
#> A person of the opposite sex that you like more than you dislike 93.83
#> The most confident person you know                               98.11
#> A person you work well with (at job or on sports team etc)       96.12
#> A teacher you did not respect                                    99.73
#> TOTAL                                                            97.19
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4

# explained sum-of-squares for constructs
ssq(bell2010, along = 1)
#> 
#> Cumulated proportionof explained sum-of-squares for constructs
#> 
#>                        1D   2D   3D   4D   5D
#> worried & tense      1.27 1.56 2.44 2.95 3.11
#> smart (academically) 0.74 1.73 1.93 1.93 4.09
#> loves sports         0.14 1.54 2.45 3.41 3.63
#> loves people         0.58 1.04 3.14 3.18 3.91
#> transparent          0.47 1.72 2.01 3.17 3.30
#> sensitive            0.89 0.90 2.57 2.68 5.02
#> fearless             1.32 1.53 1.81 2.71 2.78
#> gentle               1.32 1.98 2.12 2.14 2.18
#> loves to argue       0.26 0.63 0.84 3.87 5.76
#> TOTAL                0.73 1.47 2.20 2.93 3.67
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4

# save results
s <- ssq(bell2010)

# printing options
print(s)
#> 
#> Cumulated proportionof explained sum-of-squares for elements
#> 
#>                                                                     1D    2D
#> self                                                              0.08 59.22
#> closest friend of the same sex                                    0.38 35.04
#> the unhappiest person you know                                   38.26 42.08
#> A person of the opposite sex that you don't get along with        0.10 64.94
#> A teacher you respected                                          35.87 58.43
#> Mother (or the person who filled that kind of role)              85.10 85.34
#> A person of the opposite sex that you like more than you dislike  7.62 31.87
#> The most confident person you know                               86.90 88.75
#> A person you work well with (at job or on sports team etc)        2.76 64.62
#> A teacher you did not respect                                     1.18 59.89
#> TOTAL                                                            34.02 63.63
#>                                                                     3D    4D
#> self                                                             68.95 79.91
#> closest friend of the same sex                                   71.89 77.48
#> the unhappiest person you know                                   75.12 80.07
#> A person of the opposite sex that you don't get along with       76.38 96.73
#> A teacher you respected                                          58.50 97.25
#> Mother (or the person who filled that kind of role)              93.33 93.44
#> A person of the opposite sex that you like more than you dislike 54.54 93.82
#> The most confident person you know                               90.31 97.98
#> A person you work well with (at job or on sports team etc)       72.26 85.03
#> A teacher you did not respect                                    92.74 93.51
#> TOTAL                                                            79.58 91.19
#>                                                                     5D
#> self                                                             79.98
#> closest friend of the same sex                                   99.11
#> the unhappiest person you know                                   98.02
#> A person of the opposite sex that you don't get along with       97.60
#> A teacher you respected                                          97.53
#> Mother (or the person who filled that kind of role)              97.42
#> A person of the opposite sex that you like more than you dislike 93.83
#> The most confident person you know                               98.11
#> A person you work well with (at job or on sports team etc)       96.12
#> A teacher you did not respect                                    99.73
#> TOTAL                                                            97.19
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4
print(s, digits = 4)
#> 
#> Cumulated proportionof explained sum-of-squares for elements
#> 
#>                                                                       1D
#> self                                                              0.0836
#> closest friend of the same sex                                    0.3804
#> the unhappiest person you know                                   38.2615
#> A person of the opposite sex that you don't get along with        0.1026
#> A teacher you respected                                          35.8667
#> Mother (or the person who filled that kind of role)              85.0970
#> A person of the opposite sex that you like more than you dislike  7.6217
#> The most confident person you know                               86.9022
#> A person you work well with (at job or on sports team etc)        2.7614
#> A teacher you did not respect                                     1.1828
#> TOTAL                                                            34.0200
#>                                                                       2D
#> self                                                             59.2232
#> closest friend of the same sex                                   35.0418
#> the unhappiest person you know                                   42.0789
#> A person of the opposite sex that you don't get along with       64.9417
#> A teacher you respected                                          58.4313
#> Mother (or the person who filled that kind of role)              85.3414
#> A person of the opposite sex that you like more than you dislike 31.8658
#> The most confident person you know                               88.7542
#> A person you work well with (at job or on sports team etc)       64.6224
#> A teacher you did not respect                                    59.8890
#> TOTAL                                                            63.6341
#>                                                                       3D
#> self                                                             68.9476
#> closest friend of the same sex                                   71.8892
#> the unhappiest person you know                                   75.1178
#> A person of the opposite sex that you don't get along with       76.3800
#> A teacher you respected                                          58.5036
#> Mother (or the person who filled that kind of role)              93.3255
#> A person of the opposite sex that you like more than you dislike 54.5405
#> The most confident person you know                               90.3087
#> A person you work well with (at job or on sports team etc)       72.2573
#> A teacher you did not respect                                    92.7388
#> TOTAL                                                            79.5828
#>                                                                       4D
#> self                                                             79.9061
#> closest friend of the same sex                                   77.4804
#> the unhappiest person you know                                   80.0680
#> A person of the opposite sex that you don't get along with       96.7276
#> A teacher you respected                                          97.2529
#> Mother (or the person who filled that kind of role)              93.4384
#> A person of the opposite sex that you like more than you dislike 93.8156
#> The most confident person you know                               97.9777
#> A person you work well with (at job or on sports team etc)       85.0277
#> A teacher you did not respect                                    93.5145
#> TOTAL                                                            91.1857
#>                                                                       5D
#> self                                                             79.9832
#> closest friend of the same sex                                   99.1082
#> the unhappiest person you know                                   98.0241
#> A person of the opposite sex that you don't get along with       97.5974
#> A teacher you respected                                          97.5286
#> Mother (or the person who filled that kind of role)              97.4151
#> A person of the opposite sex that you like more than you dislike 93.8295
#> The most confident person you know                               98.1139
#> A person you work well with (at job or on sports team etc)       96.1209
#> A teacher you did not respect                                    99.7333
#> TOTAL                                                            97.1910
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4
print(s, dim = 3)
#> 
#> Cumulated proportionof explained sum-of-squares for elements
#> 
#>                                                                     1D    2D
#> self                                                              0.08 59.22
#> closest friend of the same sex                                    0.38 35.04
#> the unhappiest person you know                                   38.26 42.08
#> A person of the opposite sex that you don't get along with        0.10 64.94
#> A teacher you respected                                          35.87 58.43
#> Mother (or the person who filled that kind of role)              85.10 85.34
#> A person of the opposite sex that you like more than you dislike  7.62 31.87
#> The most confident person you know                               86.90 88.75
#> A person you work well with (at job or on sports team etc)        2.76 64.62
#> A teacher you did not respect                                     1.18 59.89
#> TOTAL                                                            34.02 63.63
#>                                                                     3D
#> self                                                             68.95
#> closest friend of the same sex                                   71.89
#> the unhappiest person you know                                   75.12
#> A person of the opposite sex that you don't get along with       76.38
#> A teacher you respected                                          58.50
#> Mother (or the person who filled that kind of role)              93.33
#> A person of the opposite sex that you like more than you dislike 54.54
#> The most confident person you know                               90.31
#> A person you work well with (at job or on sports team etc)       72.26
#> A teacher you did not respect                                    92.74
#> TOTAL                                                            79.58
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4
print(s, cumulated = FALSE)
#> 
#> Proportionof explained sum-of-squares for elements
#> 
#>                                                                     1D    2D
#> self                                                              0.08 59.14
#> closest friend of the same sex                                    0.38 34.66
#> the unhappiest person you know                                   38.26  3.82
#> A person of the opposite sex that you don't get along with        0.10 64.84
#> A teacher you respected                                          35.87 22.56
#> Mother (or the person who filled that kind of role)              85.10  0.24
#> A person of the opposite sex that you like more than you dislike  7.62 24.24
#> The most confident person you know                               86.90  1.85
#> A person you work well with (at job or on sports team etc)        2.76 61.86
#> A teacher you did not respect                                     1.18 58.71
#> TOTAL                                                            34.02 29.61
#>                                                                     3D    4D
#> self                                                              9.72 10.96
#> closest friend of the same sex                                   36.85  5.59
#> the unhappiest person you know                                   33.04  4.95
#> A person of the opposite sex that you don't get along with       11.44 20.35
#> A teacher you respected                                           0.07 38.75
#> Mother (or the person who filled that kind of role)               7.98  0.11
#> A person of the opposite sex that you like more than you dislike 22.67 39.28
#> The most confident person you know                                1.55  7.67
#> A person you work well with (at job or on sports team etc)        7.63 12.77
#> A teacher you did not respect                                    32.85  0.78
#> TOTAL                                                            15.95 11.60
#>                                                                     5D
#> self                                                              0.08
#> closest friend of the same sex                                   21.63
#> the unhappiest person you know                                   17.96
#> A person of the opposite sex that you don't get along with        0.87
#> A teacher you respected                                           0.28
#> Mother (or the person who filled that kind of role)               3.98
#> A person of the opposite sex that you like more than you dislike  0.01
#> The most confident person you know                                0.14
#> A person you work well with (at job or on sports team etc)       11.09
#> A teacher you did not respect                                     6.22
#> TOTAL                                                             6.01
#> 
#> Total sum-of-squares of pre-transformed  (i.e. centered and scaled) matrix: 136.4

# access results
names(s)
#> [1] "ssq.table"        "ssq.table.cumsum" "ssq.total"       
s$ssq.table
#>                                                                           1D
#> self                                                              0.08362054
#> closest friend of the same sex                                    0.38038232
#> the unhappiest person you know                                   38.26145953
#> A person of the opposite sex that you don't get along with        0.10255329
#> A teacher you respected                                          35.86673556
#> Mother (or the person who filled that kind of role)              85.09696929
#> A person of the opposite sex that you like more than you dislike  7.62166507
#> The most confident person you know                               86.90221665
#> A person you work well with (at job or on sports team etc)        2.76140523
#> A teacher you did not respect                                     1.18276971
#> TOTAL                                                            34.02002230
#>                                                                          2D
#> self                                                             59.1396100
#> closest friend of the same sex                                   34.6614168
#> the unhappiest person you know                                    3.8174694
#> A person of the opposite sex that you don't get along with       64.8391505
#> A teacher you respected                                          22.5645682
#> Mother (or the person who filled that kind of role)               0.2444566
#> A person of the opposite sex that you like more than you dislike 24.2441725
#> The most confident person you know                                1.8519711
#> A person you work well with (at job or on sports team etc)       61.8609668
#> A teacher you did not respect                                    58.7062201
#> TOTAL                                                            29.6140782
#>                                                                           3D
#> self                                                              9.72441608
#> closest friend of the same sex                                   36.84744260
#> the unhappiest person you know                                   33.03885773
#> A person of the opposite sex that you don't get along with       11.43834146
#> A teacher you respected                                           0.07233821
#> Mother (or the person who filled that kind of role)               7.98409667
#> A person of the opposite sex that you like more than you dislike 22.67464376
#> The most confident person you know                                1.55446887
#> A person you work well with (at job or on sports team etc)        7.63490233
#> A teacher you did not respect                                    32.84976087
#> TOTAL                                                            15.94873686
#>                                                                          4D
#> self                                                             10.9584461
#> closest friend of the same sex                                    5.5911733
#> the unhappiest person you know                                    4.9502001
#> A person of the opposite sex that you don't get along with       20.3475057
#> A teacher you respected                                          38.7493025
#> Mother (or the person who filled that kind of role)               0.1129045
#> A person of the opposite sex that you like more than you dislike 39.2750994
#> The most confident person you know                                7.6690124
#> A person you work well with (at job or on sports team etc)       12.7703941
#> A teacher you did not respect                                     0.7757358
#> TOTAL                                                            11.6028957
#>                                                                           5D
#> self                                                              0.07707218
#> closest friend of the same sex                                   21.62778661
#> the unhappiest person you know                                   17.95610428
#> A person of the opposite sex that you don't get along with        0.86983552
#> A teacher you respected                                           0.27562893
#> Mother (or the person who filled that kind of role)               3.97662592
#> A person of the opposite sex that you like more than you dislike  0.01392438
#> The most confident person you know                                0.13619358
#> A person you work well with (at job or on sports team etc)       11.09325675
#> A teacher you did not respect                                     6.21877220
#> TOTAL                                                             6.00521758
#>                                                                          6D
#> self                                                             1.05186111
#> closest friend of the same sex                                   0.01723136
#> the unhappiest person you know                                   0.68141656
#> A person of the opposite sex that you don't get along with       0.31766009
#> A teacher you respected                                          0.47378066
#> Mother (or the person who filled that kind of role)              2.35229778
#> A person of the opposite sex that you like more than you dislike 5.03938802
#> The most confident person you know                               0.85746621
#> A person you work well with (at job or on sports team etc)       3.23496091
#> A teacher you did not respect                                    0.06956441
#> TOTAL                                                            1.37880842
#>                                                                           7D
#> self                                                             15.29581505
#> closest friend of the same sex                                    0.11699863
#> the unhappiest person you know                                    0.78362788
#> A person of the opposite sex that you don't get along with        0.05059405
#> A teacher you respected                                           0.57869543
#> Mother (or the person who filled that kind of role)               0.20342498
#> A person of the opposite sex that you like more than you dislike  0.97139727
#> The most confident person you know                                0.51873308
#> A person you work well with (at job or on sports team etc)        0.61215668
#> A teacher you did not respect                                     0.11128456
#> TOTAL                                                             0.81365711
#>                                                                            8D
#> self                                                             3.3191535603
#> closest friend of the same sex                                   0.2244595854
#> the unhappiest person you know                                   0.3907095870
#> A person of the opposite sex that you don't get along with       2.0246491219
#> A teacher you respected                                          0.9083756481
#> Mother (or the person who filled that kind of role)              0.0206249222
#> A person of the opposite sex that you like more than you dislike 0.0156567400
#> The most confident person you know                               0.4927160709
#> A person you work well with (at job or on sports team etc)       0.0037691318
#> A teacher you did not respect                                    0.0005205033
#> TOTAL                                                            0.4876053468
#>                                                                           9D
#> self                                                             0.350005389
#> closest friend of the same sex                                   0.533108768
#> the unhappiest person you know                                   0.120154945
#> A person of the opposite sex that you don't get along with       0.009710340
#> A teacher you respected                                          0.510574854
#> Mother (or the person who filled that kind of role)              0.008599402
#> A person of the opposite sex that you like more than you dislike 0.144052867
#> The most confident person you know                               0.017222006
#> A person you work well with (at job or on sports team etc)       0.028188096
#> A teacher you did not respect                                    0.085371923
#> TOTAL                                                            0.128978420
s$ssq.table.cumsum
#>                                                                           1D
#> self                                                              0.08362054
#> closest friend of the same sex                                    0.38038232
#> the unhappiest person you know                                   38.26145953
#> A person of the opposite sex that you don't get along with        0.10255329
#> A teacher you respected                                          35.86673556
#> Mother (or the person who filled that kind of role)              85.09696929
#> A person of the opposite sex that you like more than you dislike  7.62166507
#> The most confident person you know                               86.90221665
#> A person you work well with (at job or on sports team etc)        2.76140523
#> A teacher you did not respect                                     1.18276971
#> TOTAL                                                            34.02002230
#>                                                                        2D
#> self                                                             59.22323
#> closest friend of the same sex                                   35.04180
#> the unhappiest person you know                                   42.07893
#> A person of the opposite sex that you don't get along with       64.94170
#> A teacher you respected                                          58.43130
#> Mother (or the person who filled that kind of role)              85.34143
#> A person of the opposite sex that you like more than you dislike 31.86584
#> The most confident person you know                               88.75419
#> A person you work well with (at job or on sports team etc)       64.62237
#> A teacher you did not respect                                    59.88899
#> TOTAL                                                            63.63410
#>                                                                        3D
#> self                                                             68.94765
#> closest friend of the same sex                                   71.88924
#> the unhappiest person you know                                   75.11779
#> A person of the opposite sex that you don't get along with       76.38005
#> A teacher you respected                                          58.50364
#> Mother (or the person who filled that kind of role)              93.32552
#> A person of the opposite sex that you like more than you dislike 54.54048
#> The most confident person you know                               90.30866
#> A person you work well with (at job or on sports team etc)       72.25727
#> A teacher you did not respect                                    92.73875
#> TOTAL                                                            79.58284
#>                                                                        4D
#> self                                                             79.90609
#> closest friend of the same sex                                   77.48042
#> the unhappiest person you know                                   80.06799
#> A person of the opposite sex that you don't get along with       96.72755
#> A teacher you respected                                          97.25294
#> Mother (or the person who filled that kind of role)              93.43843
#> A person of the opposite sex that you like more than you dislike 93.81558
#> The most confident person you know                               97.97767
#> A person you work well with (at job or on sports team etc)       85.02767
#> A teacher you did not respect                                    93.51449
#> TOTAL                                                            91.18573
#>                                                                        5D
#> self                                                             79.98316
#> closest friend of the same sex                                   99.10820
#> the unhappiest person you know                                   98.02409
#> A person of the opposite sex that you don't get along with       97.59739
#> A teacher you respected                                          97.52857
#> Mother (or the person who filled that kind of role)              97.41505
#> A person of the opposite sex that you like more than you dislike 93.82951
#> The most confident person you know                               98.11386
#> A person you work well with (at job or on sports team etc)       96.12093
#> A teacher you did not respect                                    99.73326
#> TOTAL                                                            97.19095
#>                                                                        6D
#> self                                                             81.03503
#> closest friend of the same sex                                   99.12543
#> the unhappiest person you know                                   98.70551
#> A person of the opposite sex that you don't get along with       97.91505
#> A teacher you respected                                          98.00235
#> Mother (or the person who filled that kind of role)              99.76735
#> A person of the opposite sex that you like more than you dislike 98.86889
#> The most confident person you know                               98.97133
#> A person you work well with (at job or on sports team etc)       99.35589
#> A teacher you did not respect                                    99.80282
#> TOTAL                                                            98.56976
#>                                                                        7D
#> self                                                             96.33084
#> closest friend of the same sex                                   99.24243
#> the unhappiest person you know                                   99.48914
#> A person of the opposite sex that you don't get along with       97.96564
#> A teacher you respected                                          98.58105
#> Mother (or the person who filled that kind of role)              99.97078
#> A person of the opposite sex that you like more than you dislike 99.84029
#> The most confident person you know                               99.49006
#> A person you work well with (at job or on sports team etc)       99.96804
#> A teacher you did not respect                                    99.91411
#> TOTAL                                                            99.38342
#>                                                                        8D  9D
#> self                                                             99.64999 100
#> closest friend of the same sex                                   99.46689 100
#> the unhappiest person you know                                   99.87985 100
#> A person of the opposite sex that you don't get along with       99.99029 100
#> A teacher you respected                                          99.48943 100
#> Mother (or the person who filled that kind of role)              99.99140 100
#> A person of the opposite sex that you like more than you dislike 99.85595 100
#> The most confident person you know                               99.98278 100
#> A person you work well with (at job or on sports team etc)       99.97181 100
#> A teacher you did not respect                                    99.91463 100
#> TOTAL                                                            99.87102 100
s$ssq.total
#> [1] 136.4