Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - charter then                   
#>                    essentially - 7 | 9 - by good man                  
#>                        flees - 6 | | | 10 - flees the flees           
#>      slipping owes depicts - 5 | | | | | 11 - activity realizes by I  
#>            before to him - 4 | | | | | | | 12 - activity money then aw
#>   the then she charter - 3 | | | | | | | | | 13 - who Mark without mai
#>           forces the - 2 | | | | | | | | | | | 14 - then realizes     
#>                 an - 1 | | | | | | | | | | | | | 15 - then by he reali
#>                      | | | | | | | | | | | | | | |                    
#> is airplane ch (1)   5 1 1 4 4 5 1 1 1 5 5 2 4 1 4   (1) of by owes Jo
#> alizes money I (2)   1 1 2 4 3 1 3 3 4 3 2 1 4 4 5   (2) to customer b
#> any flees away (3)   2 4 3 3 3 3 2 1 2 4 5 3 2 3 3   (3) he Harry to f
#> any Brownstone (4)   2 5 3 5 5 5 5 5 2 2 1 5 4 4 1   (4) Harry of he p
#>       before I (5)   5 3 2 3 5 2 1 1 1 4 3 5 3 3 4   (5) tricks I mone
#> tricks owes Jo (6)   2 3 3 3 2 2 3 3 4 2 3 4 5 5 5   (6) by away Harry
#>  initially his (7)   1 2 5 1 4 4 4 3 2 4 5 1 5 5 2   (7) activity real
#>            she (8)   1 2 2 5 1 1 5 3 4 3 5 3 2 3 4   (8) control by wh
#> ore owes slipp (9)   5 4 1 5 3 4 1 4 1 1 4 2 4 5 3   (9) has Johnson B
#> rces realizes (10)   4 2 3 5 2 3 1 5 4 3 3 3 4 3 2   (10) I airplane M
#> ovel Mr. good (11)   4 5 1 4 3 4 5 4 5 1 4 5 3 2 5   (11) into Johnson
#> Harry control (12)   1 3 3 1 5 3 3 3 1 4 5 2 4 1 3   (12) has novel ec
#>  realizes man (13)   2 2 3 4 3 5 3 3 1 4 2 5 5 2 2   (13) he owes depi
#> ipping by has (14)   5 4 5 1 1 5 1 2 4 5 3 1 1 2 3   (14) is money int
#> initially the (15)   2 3 1 2 5 4 3 2 5 5 1 4 1 2 5   (15) the by an fi
#>  he what Mark (16)   3 5 3 1 4 4 2 1 3 5 3 4 3 3 2   (16) is          
#> nto happened  (17)   4 2 4 2 1 2 3 5 1 3 4 2 3 1 3   (17) by Mr. has p
#> s mainland by (18)   5 3 3 5 1 3 2 1 1 1 3 2 1 5 5   (18) money slippi
#>      activity (19)   4 2 1 2 4 1 5 1 1 2 3 4 4 3 2   (19) economic    
#>     initially (20)   4 1 4 1 5 5 2 1 4 1 1 3 5 4 3   (20) forced econo
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - charter then                   
#>                    essentially - 7 | 9 - by good man                  
#>                        flees - 6 | | | 10 - flees the flees           
#>      slipping owes depicts - 5 | | | | | 11 - activity realizes by I  
#>            before to him - 4 | | | | | | | 12 - activity money then aw
#>   the then she charter - 3 | | | | | | | | | 13 - who Mark without mai
#>           forces the - 2 | | | | | | | | | | | 14 - then realizes     
#>                 an - 1 | | | | | | | | | | | | | 15 - then by he reali
#>                      | | | | | | | | | | | | | | |                    
#> is airplane ch (1)   5 1 1 4 4 5 1 1 1 5 5 2 4 1 4   (1) of by owes Jo
#> alizes money I (2)   1 1 2 4 3 1 3 3 4 3 2 1 4 4 5   (2) to customer b
#> any flees away (3)   2 4 3 3 3 3 2 1 2 4 5 3 2 3 3   (3) he Harry to f
#> any Brownstone (4)   2 5 3 5 5 5 5 5 2 2 1 5 4 4 1   (4) Harry of he p
#>       before I (5)   5 3 2 3 5 2 1 1 1 4 3 5 3 3 4   (5) tricks I mone
#> tricks owes Jo (6)   2 3 3 3 2 2 3 3 4 2 3 4 5 5 5   (6) by away Harry
#>  initially his (7)   1 2 5 1 4 4 4 3 2 4 5 1 5 5 2   (7) activity real
#>            she (8)   1 2 2 5 1 1 5 3 4 3 5 3 2 3 4   (8) control by wh
#> ore owes slipp (9)   5 4 1 5 3 4 1 4 1 1 4 2 4 5 3   (9) has Johnson B
#> rces realizes (10)   4 2 3 5 2 3 1 5 4 3 3 3 4 3 2   (10) I airplane M
#> ovel Mr. good (11)   4 5 1 4 3 4 5 4 5 1 4 5 3 2 5   (11) into Johnson
#> Harry control (12)   1 3 3 1 5 3 3 3 1 4 5 2 4 1 3   (12) has novel ec
#>  realizes man (13)   2 2 3 4 3 5 3 3 1 4 2 5 5 2 2   (13) he owes depi
#> ipping by has (14)   5 4 5 1 1 5 1 2 4 5 3 1 1 2 3   (14) is money int
#> initially the (15)   2 3 1 2 5 4 3 2 5 5 1 4 1 2 5   (15) the by an fi
#>  he what Mark (16)   3 5 3 1 4 4 2 1 3 5 3 4 3 3 2   (16) is          
#> nto happened  (17)   4 2 4 2 1 2 3 5 1 3 4 2 3 1 3   (17) by Mr. has p
#> s mainland by (18)   5 3 3 5 1 3 2 1 1 1 3 2 1 5 5   (18) money slippi
#>      activity (19)   4 2 1 2 4 1 5 1 1 2 3 4 4 3 2   (19) economic    
#>     initially (20)   4 1 4 1 5 5 2 1 4 1 1 3 5 4 3   (20) forced econo
#> s airplane ch (21)   5 1 1 4 4 5 1 1 1 5 5 2 4 1 4   (21) of by owes J
#> lizes money I (22)   1 1 2 4 3 1 3 3 4 3 2 1 4 4 5   (22) to customer 
#> ny flees away (23)   2 4 3 3 3 3 2 1 2 4 5 3 2 3 3   (23) he Harry to 
#> ny Brownstone (24)   2 5 3 5 5 5 5 5 2 2 1 5 4 4 1   (24) Harry of he 
#>      before I (25)   5 3 2 3 5 2 1 1 1 4 3 5 3 3 4   (25) tricks I mon
#> ricks owes Jo (26)   2 3 3 3 2 2 3 3 4 2 3 4 5 5 5   (26) by away Harr
#> initially his (27)   1 2 5 1 4 4 4 3 2 4 5 1 5 5 2   (27) activity rea
#>           she (28)   1 2 2 5 1 1 5 3 4 3 5 3 2 3 4   (28) control by w
#> re owes slipp (29)   5 4 1 5 3 4 1 4 1 1 4 2 4 5 3   (29) has Johnson 
#> rces realizes (30)   4 2 3 5 2 3 1 5 4 3 3 3 4 3 2   (30) I airplane M
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - charter then                   
#>                    essentially - 7 | 9 - by good man                  
#>                        flees - 6 | | | 10 - flees the flees           
#>      slipping owes depicts - 5 | | | | | 11 - activity realizes by I  
#>            before to him - 4 | | | | | | | 12 - activity money then aw
#>   the then she charter - 3 | | | | | | | | | 13 - who Mark without mai
#>           forces the - 2 | | | | | | | | | | | 14 - then realizes     
#>                 an - 1 | | | | | | | | | | | | | 15 - then by he reali
#>                      | | | | | | | | | | | | | | |                    
#> is airplane ch (1)   5 1 1 4 4 5 1 1 1 5 5 2 4 1 4   (1) of by owes Jo
#> alizes money I (2)   1 1 2 4 3 1 3 3 4 3 2 1 4 4 5   (2) to customer b
#> any flees away (3)   2 4 3 3 3 3 2 1 2 4 5 3 2 3 3   (3) he Harry to f
#> any Brownstone (4)   2 5 3 5 5 5 5 5 2 2 1 5 4 4 1   (4) Harry of he p
#>       before I (5)   5 3 2 3 5 2 1 1 1 4 3 5 3 3 4   (5) tricks I mone
#> tricks owes Jo (6)   2 3 3 3 2 2 3 3 4 2 3 4 5 5 5   (6) by away Harry
#>  initially his (7)   1 2 5 1 4 4 4 3 2 4 5 1 5 5 2   (7) activity real
#>            she (8)   1 2 2 5 1 1 5 3 4 3 5 3 2 3 4   (8) control by wh
#> ore owes slipp (9)   5 4 1 5 3 4 1 4 1 1 4 2 4 5 3   (9) has Johnson B
#> rces realizes (10)   4 2 3 5 2 3 1 5 4 3 3 3 4 3 2   (10) I airplane M
#> s airplane ch (11)   5 1 1 4 4 5 1 1 1 5 5 2 4 1 4   (11) of by owes J
#> lizes money I (12)   1 1 2 4 3 1 3 3 4 3 2 1 4 4 5   (12) to customer 
#> ny flees away (13)   2 4 3 3 3 3 2 1 2 4 5 3 2 3 3   (13) he Harry to 
#> ny Brownstone (14)   2 5 3 5 5 5 5 5 2 2 1 5 4 4 1   (14) Harry of he 
#>      before I (15)   5 3 2 3 5 2 1 1 1 4 3 5 3 3 4   (15) tricks I mon
#> ricks owes Jo (16)   2 3 3 3 2 2 3 3 4 2 3 4 5 5 5   (16) by away Harr
#> initially his (17)   1 2 5 1 4 4 4 3 2 4 5 1 5 5 2   (17) activity rea
#>           she (18)   1 2 2 5 1 1 5 3 4 3 5 3 2 3 4   (18) control by w
#> re owes slipp (19)   5 4 1 5 3 4 1 4 1 1 4 2 4 5 3   (19) has Johnson 
#> rces realizes (20)   4 2 3 5 2 3 1 5 4 3 3 3 4 3 2   (20) I airplane M
#> ovel Mr. good (21)   4 5 1 4 3 4 5 4 5 1 4 5 3 2 5   (21) into Johnson
#> Harry control (22)   1 3 3 1 5 3 3 3 1 4 5 2 4 1 3   (22) has novel ec
#>  realizes man (23)   2 2 3 4 3 5 3 3 1 4 2 5 5 2 2   (23) he owes depi
#> ipping by has (24)   5 4 5 1 1 5 1 2 4 5 3 1 1 2 3   (24) is money int
#> initially the (25)   2 3 1 2 5 4 3 2 5 5 1 4 1 2 5   (25) the by an fi
#>  he what Mark (26)   3 5 3 1 4 4 2 1 3 5 3 4 3 3 2   (26) is          
#> nto happened  (27)   4 2 4 2 1 2 3 5 1 3 4 2 3 1 3   (27) by Mr. has p
#> s mainland by (28)   5 3 3 5 1 3 2 1 1 1 3 2 1 5 5   (28) money slippi
#>      activity (29)   4 2 1 2 4 1 5 1 1 2 3 4 4 3 2   (29) economic    
#>     initially (30)   4 1 4 1 5 5 2 1 4 1 1 3 5 4 3   (30) forced econo
#>