Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he mainland                    
#>            mainland is depicts - 7 | 9 - airplane forced              
#>           forced blackmarket - 6 | | | 10 - essentially the before hap
#>                        has - 5 | | | | | 11 - paying novel            
#>     Mark depicts charter - 4 | | | | | | | 12 - his of realizes novel 
#>              is by his - 3 | | | | | | | | | 13 - to money            
#> is mainland realizes - 2 | | | | | | | | | | | 14 - Harry flees into f
#>           airplane - 1 | | | | | | | | | | | | | 15 - without charter 
#>                      | | | | | | | | | | | | | | |                    
#> he Mr. airplan (1)   1 4 2 5 5 4 4 3 5 3 4 5 2 5 4   (1) owes tricks f
#>       mainland (2)   4 4 2 2 3 2 1 1 5 3 3 3 3 2 3   (2) his owes by a
#> izes as before (3)   1 4 4 2 3 1 5 2 3 3 2 1 3 1 1   (3) then         
#> on Mr. him him (4)   1 3 3 4 4 1 3 4 4 2 4 2 4 2 2   (4) back fishing 
#> happened what  (5)   3 1 3 5 1 5 3 3 4 1 4 5 5 5 2   (5) forced back c
#>           good (6)   3 1 2 1 4 3 5 1 3 4 3 4 5 4 1   (6) Mr. owes Mark
#> ng airplane by (7)   3 4 5 5 1 2 4 1 1 1 2 5 5 1 1   (7) of away contr
#> od the his Mr. (8)   4 3 4 1 2 3 2 4 1 4 4 4 3 4 5   (8) initially the
#>            his (9)   1 4 5 1 5 5 4 4 1 5 2 5 1 2 4   (9) money of him 
#> back owes by  (10)   4 1 2 3 4 4 5 1 1 1 1 2 1 1 5   (10) him by she  
#>      by flees (11)   5 4 2 2 1 5 4 4 4 5 3 2 3 1 4   (11) activity sli
#>      by to by (12)   2 1 3 4 2 5 2 4 4 1 3 5 1 1 3   (12) what his an 
#> hout into she (13)   5 5 5 4 5 4 1 2 2 5 3 3 3 4 4   (13) into economi
#>  he any novel (14)   4 2 1 5 3 1 1 4 1 5 1 3 4 5 4   (14) forces owes 
#> what money an (15)   3 3 1 5 2 1 3 4 4 4 5 4 5 2 2   (15) of Mark     
#>       control (16)   1 1 5 4 3 1 2 3 1 4 3 2 2 4 3   (16) his customer
#> e essentially (17)   5 2 5 5 3 3 1 2 1 3 1 4 4 2 3   (17) slipping rea
#> on man mainla (18)   5 4 5 5 1 3 5 1 1 1 3 5 3 4 2   (18) she customer
#> stone fishing (19)   3 1 5 1 4 1 4 5 5 2 4 5 3 2 1   (19) Mark into be
#>  what control (20)   1 5 5 4 1 1 4 1 2 2 2 2 4 3 4   (20) by Johnson t
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he mainland                    
#>            mainland is depicts - 7 | 9 - airplane forced              
#>           forced blackmarket - 6 | | | 10 - essentially the before hap
#>                        has - 5 | | | | | 11 - paying novel            
#>     Mark depicts charter - 4 | | | | | | | 12 - his of realizes novel 
#>              is by his - 3 | | | | | | | | | 13 - to money            
#> is mainland realizes - 2 | | | | | | | | | | | 14 - Harry flees into f
#>           airplane - 1 | | | | | | | | | | | | | 15 - without charter 
#>                      | | | | | | | | | | | | | | |                    
#> he Mr. airplan (1)   1 4 2 5 5 4 4 3 5 3 4 5 2 5 4   (1) owes tricks f
#>       mainland (2)   4 4 2 2 3 2 1 1 5 3 3 3 3 2 3   (2) his owes by a
#> izes as before (3)   1 4 4 2 3 1 5 2 3 3 2 1 3 1 1   (3) then         
#> on Mr. him him (4)   1 3 3 4 4 1 3 4 4 2 4 2 4 2 2   (4) back fishing 
#> happened what  (5)   3 1 3 5 1 5 3 3 4 1 4 5 5 5 2   (5) forced back c
#>           good (6)   3 1 2 1 4 3 5 1 3 4 3 4 5 4 1   (6) Mr. owes Mark
#> ng airplane by (7)   3 4 5 5 1 2 4 1 1 1 2 5 5 1 1   (7) of away contr
#> od the his Mr. (8)   4 3 4 1 2 3 2 4 1 4 4 4 3 4 5   (8) initially the
#>            his (9)   1 4 5 1 5 5 4 4 1 5 2 5 1 2 4   (9) money of him 
#> back owes by  (10)   4 1 2 3 4 4 5 1 1 1 1 2 1 1 5   (10) him by she  
#>      by flees (11)   5 4 2 2 1 5 4 4 4 5 3 2 3 1 4   (11) activity sli
#>      by to by (12)   2 1 3 4 2 5 2 4 4 1 3 5 1 1 3   (12) what his an 
#> hout into she (13)   5 5 5 4 5 4 1 2 2 5 3 3 3 4 4   (13) into economi
#>  he any novel (14)   4 2 1 5 3 1 1 4 1 5 1 3 4 5 4   (14) forces owes 
#> what money an (15)   3 3 1 5 2 1 3 4 4 4 5 4 5 2 2   (15) of Mark     
#>       control (16)   1 1 5 4 3 1 2 3 1 4 3 2 2 4 3   (16) his customer
#> e essentially (17)   5 2 5 5 3 3 1 2 1 3 1 4 4 2 3   (17) slipping rea
#> on man mainla (18)   5 4 5 5 1 3 5 1 1 1 3 5 3 4 2   (18) she customer
#> stone fishing (19)   3 1 5 1 4 1 4 5 5 2 4 5 3 2 1   (19) Mark into be
#>  what control (20)   1 5 5 4 1 1 4 1 2 2 2 2 4 3 4   (20) by Johnson t
#> e Mr. airplan (21)   1 4 2 5 5 4 4 3 5 3 4 5 2 5 4   (21) owes tricks 
#>      mainland (22)   4 4 2 2 3 2 1 1 5 3 3 3 3 2 3   (22) his owes by 
#> zes as before (23)   1 4 4 2 3 1 5 2 3 3 2 1 3 1 1   (23) then        
#> n Mr. him him (24)   1 3 3 4 4 1 3 4 4 2 4 2 4 2 2   (24) back fishing
#> appened what  (25)   3 1 3 5 1 5 3 3 4 1 4 5 5 5 2   (25) forced back 
#>          good (26)   3 1 2 1 4 3 5 1 3 4 3 4 5 4 1   (26) Mr. owes Mar
#> g airplane by (27)   3 4 5 5 1 2 4 1 1 1 2 5 5 1 1   (27) of away cont
#> d the his Mr. (28)   4 3 4 1 2 3 2 4 1 4 4 4 3 4 5   (28) initially th
#>           his (29)   1 4 5 1 5 5 4 4 1 5 2 5 1 2 4   (29) money of him
#> back owes by  (30)   4 1 2 3 4 4 5 1 1 1 1 2 1 1 5   (30) him by she  
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he mainland                    
#>            mainland is depicts - 7 | 9 - airplane forced              
#>           forced blackmarket - 6 | | | 10 - essentially the before hap
#>                        has - 5 | | | | | 11 - paying novel            
#>     Mark depicts charter - 4 | | | | | | | 12 - his of realizes novel 
#>              is by his - 3 | | | | | | | | | 13 - to money            
#> is mainland realizes - 2 | | | | | | | | | | | 14 - Harry flees into f
#>           airplane - 1 | | | | | | | | | | | | | 15 - without charter 
#>                      | | | | | | | | | | | | | | |                    
#> he Mr. airplan (1)   1 4 2 5 5 4 4 3 5 3 4 5 2 5 4   (1) owes tricks f
#>       mainland (2)   4 4 2 2 3 2 1 1 5 3 3 3 3 2 3   (2) his owes by a
#> izes as before (3)   1 4 4 2 3 1 5 2 3 3 2 1 3 1 1   (3) then         
#> on Mr. him him (4)   1 3 3 4 4 1 3 4 4 2 4 2 4 2 2   (4) back fishing 
#> happened what  (5)   3 1 3 5 1 5 3 3 4 1 4 5 5 5 2   (5) forced back c
#>           good (6)   3 1 2 1 4 3 5 1 3 4 3 4 5 4 1   (6) Mr. owes Mark
#> ng airplane by (7)   3 4 5 5 1 2 4 1 1 1 2 5 5 1 1   (7) of away contr
#> od the his Mr. (8)   4 3 4 1 2 3 2 4 1 4 4 4 3 4 5   (8) initially the
#>            his (9)   1 4 5 1 5 5 4 4 1 5 2 5 1 2 4   (9) money of him 
#> back owes by  (10)   4 1 2 3 4 4 5 1 1 1 1 2 1 1 5   (10) him by she  
#> e Mr. airplan (11)   1 4 2 5 5 4 4 3 5 3 4 5 2 5 4   (11) owes tricks 
#>      mainland (12)   4 4 2 2 3 2 1 1 5 3 3 3 3 2 3   (12) his owes by 
#> zes as before (13)   1 4 4 2 3 1 5 2 3 3 2 1 3 1 1   (13) then        
#> n Mr. him him (14)   1 3 3 4 4 1 3 4 4 2 4 2 4 2 2   (14) back fishing
#> appened what  (15)   3 1 3 5 1 5 3 3 4 1 4 5 5 5 2   (15) forced back 
#>          good (16)   3 1 2 1 4 3 5 1 3 4 3 4 5 4 1   (16) Mr. owes Mar
#> g airplane by (17)   3 4 5 5 1 2 4 1 1 1 2 5 5 1 1   (17) of away cont
#> d the his Mr. (18)   4 3 4 1 2 3 2 4 1 4 4 4 3 4 5   (18) initially th
#>           his (19)   1 4 5 1 5 5 4 4 1 5 2 5 1 2 4   (19) money of him
#> back owes by  (20)   4 1 2 3 4 4 5 1 1 1 1 2 1 1 5   (20) him by she  
#>      by flees (21)   5 4 2 2 1 5 4 4 4 5 3 2 3 1 4   (21) activity sli
#>      by to by (22)   2 1 3 4 2 5 2 4 4 1 3 5 1 1 3   (22) what his an 
#> hout into she (23)   5 5 5 4 5 4 1 2 2 5 3 3 3 4 4   (23) into economi
#>  he any novel (24)   4 2 1 5 3 1 1 4 1 5 1 3 4 5 4   (24) forces owes 
#> what money an (25)   3 3 1 5 2 1 3 4 4 4 5 4 5 2 2   (25) of Mark     
#>       control (26)   1 1 5 4 3 1 2 3 1 4 3 2 2 4 3   (26) his customer
#> e essentially (27)   5 2 5 5 3 3 1 2 1 3 1 4 4 2 3   (27) slipping rea
#> on man mainla (28)   5 4 5 5 1 3 5 1 1 1 3 5 3 4 2   (28) she customer
#> stone fishing (29)   3 1 5 1 4 1 4 5 5 2 4 5 3 2 1   (29) Mark into be
#>  what control (30)   1 5 5 4 1 1 4 1 2 2 2 2 4 3 4   (30) by Johnson t
#>