Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Brownstone novel forced he     
#>                  by who an the - 7 | 9 - charter flees back the       
#>                      back by - 6 | | | 10 - he mainland essentially ma
#>      Johnson his away Mark - 5 | | | | | 11 - depicts                 
#>                     what - 4 | | | | | | | 12 - of beyond to without e
#>                   Mark - 3 | | | | | | | | | 13 - happened forces airp
#>    Harry I into good - 2 | | | | | | | | | | | 14 - flees without has 
#> cts by essentially - 1 | | | | | | | | | | | | | 15 - who             
#>                      | | | | | | | | | | | | | | |                    
#> picts realizes (1)   5 5 2 2 1 2 3 3 3 5 1 2 4 4 1   (1) into as mainl
#>  beyond tricks (2)   4 1 3 4 3 2 3 1 2 4 3 2 4 2 4   (2) activity main
#> omer by before (3)   4 1 4 3 3 1 1 5 1 3 4 2 3 4 3   (3) money money g
#> od has control (4)   2 4 3 3 4 5 3 5 3 2 3 5 2 2 4   (4) realizes slip
#> to I forced an (5)   4 1 4 4 3 4 2 2 2 5 3 3 4 5 2   (5) an Brownstone
#> realizes money (6)   4 2 2 1 3 4 1 2 3 3 3 5 4 2 5   (6) slipping airp
#> hnson airplane (7)   4 4 1 2 3 4 3 3 3 2 3 1 2 1 1   (7) into by his t
#>          Harry (8)   3 4 2 1 2 4 1 4 2 3 5 1 3 2 3   (8) realizes away
#>  airplane the  (9)   1 5 3 5 2 4 2 1 4 4 5 5 4 2 3   (9) his tricks   
#> e activity is (10)   3 2 2 4 2 4 2 3 4 2 4 2 4 2 3   (10) his mainland
#> hnson slippin (11)   5 5 4 5 4 5 1 3 1 5 2 3 1 2 4   (11) what depicts
#>  his slipping (12)   1 3 4 2 5 5 4 3 1 2 1 1 3 5 1   (12) Mr. activity
#> rownstone is  (13)   2 1 2 1 4 2 1 2 1 5 5 5 2 4 3   (13) I Harry     
#> nson mainland (14)   1 2 1 5 1 2 4 4 2 4 2 3 2 3 2   (14) is blackmark
#> ipping Johnso (15)   3 5 1 3 1 3 3 1 3 3 4 5 3 4 3   (15) fishing slip
#> s good his by (16)   4 3 4 2 1 4 3 5 2 2 3 1 4 4 5   (16) novel good i
#>  the any owes (17)   2 5 4 5 3 1 1 1 1 3 1 2 2 4 5   (17) any his the 
#>  airplane the (18)   2 5 2 5 1 4 1 4 1 2 4 4 5 2 5   (18) his the beyo
#> y initially b (19)   5 1 1 4 4 2 1 4 3 1 4 3 4 5 5   (19) back        
#>       Johnson (20)   2 2 1 1 4 4 1 4 3 1 1 1 2 5 4   (20) then without
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Brownstone novel forced he     
#>                  by who an the - 7 | 9 - charter flees back the       
#>                      back by - 6 | | | 10 - he mainland essentially ma
#>      Johnson his away Mark - 5 | | | | | 11 - depicts                 
#>                     what - 4 | | | | | | | 12 - of beyond to without e
#>                   Mark - 3 | | | | | | | | | 13 - happened forces airp
#>    Harry I into good - 2 | | | | | | | | | | | 14 - flees without has 
#> cts by essentially - 1 | | | | | | | | | | | | | 15 - who             
#>                      | | | | | | | | | | | | | | |                    
#> picts realizes (1)   5 5 2 2 1 2 3 3 3 5 1 2 4 4 1   (1) into as mainl
#>  beyond tricks (2)   4 1 3 4 3 2 3 1 2 4 3 2 4 2 4   (2) activity main
#> omer by before (3)   4 1 4 3 3 1 1 5 1 3 4 2 3 4 3   (3) money money g
#> od has control (4)   2 4 3 3 4 5 3 5 3 2 3 5 2 2 4   (4) realizes slip
#> to I forced an (5)   4 1 4 4 3 4 2 2 2 5 3 3 4 5 2   (5) an Brownstone
#> realizes money (6)   4 2 2 1 3 4 1 2 3 3 3 5 4 2 5   (6) slipping airp
#> hnson airplane (7)   4 4 1 2 3 4 3 3 3 2 3 1 2 1 1   (7) into by his t
#>          Harry (8)   3 4 2 1 2 4 1 4 2 3 5 1 3 2 3   (8) realizes away
#>  airplane the  (9)   1 5 3 5 2 4 2 1 4 4 5 5 4 2 3   (9) his tricks   
#> e activity is (10)   3 2 2 4 2 4 2 3 4 2 4 2 4 2 3   (10) his mainland
#> hnson slippin (11)   5 5 4 5 4 5 1 3 1 5 2 3 1 2 4   (11) what depicts
#>  his slipping (12)   1 3 4 2 5 5 4 3 1 2 1 1 3 5 1   (12) Mr. activity
#> rownstone is  (13)   2 1 2 1 4 2 1 2 1 5 5 5 2 4 3   (13) I Harry     
#> nson mainland (14)   1 2 1 5 1 2 4 4 2 4 2 3 2 3 2   (14) is blackmark
#> ipping Johnso (15)   3 5 1 3 1 3 3 1 3 3 4 5 3 4 3   (15) fishing slip
#> s good his by (16)   4 3 4 2 1 4 3 5 2 2 3 1 4 4 5   (16) novel good i
#>  the any owes (17)   2 5 4 5 3 1 1 1 1 3 1 2 2 4 5   (17) any his the 
#>  airplane the (18)   2 5 2 5 1 4 1 4 1 2 4 4 5 2 5   (18) his the beyo
#> y initially b (19)   5 1 1 4 4 2 1 4 3 1 4 3 4 5 5   (19) back        
#>       Johnson (20)   2 2 1 1 4 4 1 4 3 1 1 1 2 5 4   (20) then without
#> icts realizes (21)   5 5 2 2 1 2 3 3 3 5 1 2 4 4 1   (21) into as main
#> beyond tricks (22)   4 1 3 4 3 2 3 1 2 4 3 2 4 2 4   (22) activity mai
#> mer by before (23)   4 1 4 3 3 1 1 5 1 3 4 2 3 4 3   (23) money money 
#> d has control (24)   2 4 3 3 4 5 3 5 3 2 3 5 2 2 4   (24) realizes sli
#> o I forced an (25)   4 1 4 4 3 4 2 2 2 5 3 3 4 5 2   (25) an Brownston
#> ealizes money (26)   4 2 2 1 3 4 1 2 3 3 3 5 4 2 5   (26) slipping air
#> nson airplane (27)   4 4 1 2 3 4 3 3 3 2 3 1 2 1 1   (27) into by his 
#>         Harry (28)   3 4 2 1 2 4 1 4 2 3 5 1 3 2 3   (28) realizes awa
#> airplane the  (29)   1 5 3 5 2 4 2 1 4 4 5 5 4 2 3   (29) his tricks  
#> e activity is (30)   3 2 2 4 2 4 2 3 4 2 4 2 4 2 3   (30) his mainland
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Brownstone novel forced he     
#>                  by who an the - 7 | 9 - charter flees back the       
#>                      back by - 6 | | | 10 - he mainland essentially ma
#>      Johnson his away Mark - 5 | | | | | 11 - depicts                 
#>                     what - 4 | | | | | | | 12 - of beyond to without e
#>                   Mark - 3 | | | | | | | | | 13 - happened forces airp
#>    Harry I into good - 2 | | | | | | | | | | | 14 - flees without has 
#> cts by essentially - 1 | | | | | | | | | | | | | 15 - who             
#>                      | | | | | | | | | | | | | | |                    
#> picts realizes (1)   5 5 2 2 1 2 3 3 3 5 1 2 4 4 1   (1) into as mainl
#>  beyond tricks (2)   4 1 3 4 3 2 3 1 2 4 3 2 4 2 4   (2) activity main
#> omer by before (3)   4 1 4 3 3 1 1 5 1 3 4 2 3 4 3   (3) money money g
#> od has control (4)   2 4 3 3 4 5 3 5 3 2 3 5 2 2 4   (4) realizes slip
#> to I forced an (5)   4 1 4 4 3 4 2 2 2 5 3 3 4 5 2   (5) an Brownstone
#> realizes money (6)   4 2 2 1 3 4 1 2 3 3 3 5 4 2 5   (6) slipping airp
#> hnson airplane (7)   4 4 1 2 3 4 3 3 3 2 3 1 2 1 1   (7) into by his t
#>          Harry (8)   3 4 2 1 2 4 1 4 2 3 5 1 3 2 3   (8) realizes away
#>  airplane the  (9)   1 5 3 5 2 4 2 1 4 4 5 5 4 2 3   (9) his tricks   
#> e activity is (10)   3 2 2 4 2 4 2 3 4 2 4 2 4 2 3   (10) his mainland
#> icts realizes (11)   5 5 2 2 1 2 3 3 3 5 1 2 4 4 1   (11) into as main
#> beyond tricks (12)   4 1 3 4 3 2 3 1 2 4 3 2 4 2 4   (12) activity mai
#> mer by before (13)   4 1 4 3 3 1 1 5 1 3 4 2 3 4 3   (13) money money 
#> d has control (14)   2 4 3 3 4 5 3 5 3 2 3 5 2 2 4   (14) realizes sli
#> o I forced an (15)   4 1 4 4 3 4 2 2 2 5 3 3 4 5 2   (15) an Brownston
#> ealizes money (16)   4 2 2 1 3 4 1 2 3 3 3 5 4 2 5   (16) slipping air
#> nson airplane (17)   4 4 1 2 3 4 3 3 3 2 3 1 2 1 1   (17) into by his 
#>         Harry (18)   3 4 2 1 2 4 1 4 2 3 5 1 3 2 3   (18) realizes awa
#> airplane the  (19)   1 5 3 5 2 4 2 1 4 4 5 5 4 2 3   (19) his tricks  
#> e activity is (20)   3 2 2 4 2 4 2 3 4 2 4 2 4 2 3   (20) his mainland
#> hnson slippin (21)   5 5 4 5 4 5 1 3 1 5 2 3 1 2 4   (21) what depicts
#>  his slipping (22)   1 3 4 2 5 5 4 3 1 2 1 1 3 5 1   (22) Mr. activity
#> rownstone is  (23)   2 1 2 1 4 2 1 2 1 5 5 5 2 4 3   (23) I Harry     
#> nson mainland (24)   1 2 1 5 1 2 4 4 2 4 2 3 2 3 2   (24) is blackmark
#> ipping Johnso (25)   3 5 1 3 1 3 3 1 3 3 4 5 3 4 3   (25) fishing slip
#> s good his by (26)   4 3 4 2 1 4 3 5 2 2 3 1 4 4 5   (26) novel good i
#>  the any owes (27)   2 5 4 5 3 1 1 1 1 3 1 2 2 4 5   (27) any his the 
#>  airplane the (28)   2 5 2 5 1 4 1 4 1 2 4 4 5 2 5   (28) his the beyo
#> y initially b (29)   5 1 1 4 4 2 1 4 3 1 4 3 4 5 5   (29) back        
#>       Johnson (30)   2 2 1 1 4 4 1 4 3 1 1 1 2 5 4   (30) then without
#>