Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he then                        
#>                 he blackmarket - 7 | 9 - mainland economic back       
#> picts him I mainland realize - 6 | | | 10 - airplane Johnson he realiz
#>  essentially economic away - 5 | | | | | 11 - control blackmarket     
#>              is airplane - 4 | | | | | | | 12 - what depicts initially
#>                    his - 3 | | | | | | | | | 13 - beyond              
#> ks customer customer - 2 | | | | | | | | | | | 14 - airplane slipping 
#>    Brownstone what - 1 | | | | | | | | | | | | | 15 - without he by ma
#>                      | | | | | | | | | | | | | | |                    
#> y man charter  (1)   5 1 2 3 1 4 5 3 3 1 2 5 5 3 5   (1) realizes happ
#>           good (2)   2 2 5 4 3 5 3 1 4 5 2 5 4 4 5   (2) mainland     
#>             to (3)   2 4 1 5 2 1 3 1 1 5 1 1 4 2 1   (3) man is       
#> ishing blackma (4)   4 5 3 3 3 3 4 3 3 4 4 5 3 5 5   (4) I initially b
#> an by the away (5)   3 2 5 4 3 1 5 5 3 1 1 4 3 2 3   (5) novel forces 
#> by is then his (6)   5 4 3 1 4 2 4 1 1 4 5 4 5 1 4   (6) he has is    
#>  I he economic (7)   5 3 3 2 4 1 1 1 3 5 3 5 5 5 1   (7) who beyond th
#>  realizes him  (8)   2 3 4 1 2 4 4 3 5 4 5 4 1 5 4   (8) the as slippi
#> rket tricks th (9)   3 5 2 4 2 2 1 4 3 4 2 2 5 5 4   (9) mainland what
#> d Brownstone  (10)   2 3 1 3 5 1 2 5 4 1 5 3 4 2 4   (10) Johnson     
#> Mark airplane (11)   5 5 1 4 4 3 4 5 3 3 2 2 1 2 5   (11) she happened
#> nland by Brow (12)   3 5 1 1 3 2 1 1 3 4 1 2 1 2 4   (12) an          
#> ket has money (13)   5 2 1 5 5 3 2 3 5 4 2 1 1 5 1   (13) him Brownsto
#> before the of (14)   5 5 3 1 2 4 5 4 2 2 1 3 3 2 1   (14) man into hap
#> rownstone who (15)   3 1 1 1 4 2 1 1 1 1 3 1 1 2 4   (15) activity    
#> k forces what (16)   4 5 2 4 1 3 4 1 5 5 5 5 1 1 1   (16) money withou
#>        forced (17)   4 5 2 5 1 1 3 5 3 4 5 1 3 5 5   (17) his he Johns
#> ackmarket tri (18)   3 1 1 4 3 5 4 1 4 4 5 4 5 4 4   (18) Brownstone c
#>         Harry (19)   4 4 2 5 3 5 1 4 2 4 4 2 3 4 3   (19) Johnson back
#> his the the d (20)   1 5 5 2 3 3 3 2 2 4 5 3 4 2 4   (20) forced by th
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he then                        
#>                 he blackmarket - 7 | 9 - mainland economic back       
#> picts him I mainland realize - 6 | | | 10 - airplane Johnson he realiz
#>  essentially economic away - 5 | | | | | 11 - control blackmarket     
#>              is airplane - 4 | | | | | | | 12 - what depicts initially
#>                    his - 3 | | | | | | | | | 13 - beyond              
#> ks customer customer - 2 | | | | | | | | | | | 14 - airplane slipping 
#>    Brownstone what - 1 | | | | | | | | | | | | | 15 - without he by ma
#>                      | | | | | | | | | | | | | | |                    
#> y man charter  (1)   5 1 2 3 1 4 5 3 3 1 2 5 5 3 5   (1) realizes happ
#>           good (2)   2 2 5 4 3 5 3 1 4 5 2 5 4 4 5   (2) mainland     
#>             to (3)   2 4 1 5 2 1 3 1 1 5 1 1 4 2 1   (3) man is       
#> ishing blackma (4)   4 5 3 3 3 3 4 3 3 4 4 5 3 5 5   (4) I initially b
#> an by the away (5)   3 2 5 4 3 1 5 5 3 1 1 4 3 2 3   (5) novel forces 
#> by is then his (6)   5 4 3 1 4 2 4 1 1 4 5 4 5 1 4   (6) he has is    
#>  I he economic (7)   5 3 3 2 4 1 1 1 3 5 3 5 5 5 1   (7) who beyond th
#>  realizes him  (8)   2 3 4 1 2 4 4 3 5 4 5 4 1 5 4   (8) the as slippi
#> rket tricks th (9)   3 5 2 4 2 2 1 4 3 4 2 2 5 5 4   (9) mainland what
#> d Brownstone  (10)   2 3 1 3 5 1 2 5 4 1 5 3 4 2 4   (10) Johnson     
#> Mark airplane (11)   5 5 1 4 4 3 4 5 3 3 2 2 1 2 5   (11) she happened
#> nland by Brow (12)   3 5 1 1 3 2 1 1 3 4 1 2 1 2 4   (12) an          
#> ket has money (13)   5 2 1 5 5 3 2 3 5 4 2 1 1 5 1   (13) him Brownsto
#> before the of (14)   5 5 3 1 2 4 5 4 2 2 1 3 3 2 1   (14) man into hap
#> rownstone who (15)   3 1 1 1 4 2 1 1 1 1 3 1 1 2 4   (15) activity    
#> k forces what (16)   4 5 2 4 1 3 4 1 5 5 5 5 1 1 1   (16) money withou
#>        forced (17)   4 5 2 5 1 1 3 5 3 4 5 1 3 5 5   (17) his he Johns
#> ackmarket tri (18)   3 1 1 4 3 5 4 1 4 4 5 4 5 4 4   (18) Brownstone c
#>         Harry (19)   4 4 2 5 3 5 1 4 2 4 4 2 3 4 3   (19) Johnson back
#> his the the d (20)   1 5 5 2 3 3 3 2 2 4 5 3 4 2 4   (20) forced by th
#>  man charter  (21)   5 1 2 3 1 4 5 3 3 1 2 5 5 3 5   (21) realizes hap
#>          good (22)   2 2 5 4 3 5 3 1 4 5 2 5 4 4 5   (22) mainland    
#>            to (23)   2 4 1 5 2 1 3 1 1 5 1 1 4 2 1   (23) man is      
#> shing blackma (24)   4 5 3 3 3 3 4 3 3 4 4 5 3 5 5   (24) I initially 
#> n by the away (25)   3 2 5 4 3 1 5 5 3 1 1 4 3 2 3   (25) novel forces
#> y is then his (26)   5 4 3 1 4 2 4 1 1 4 5 4 5 1 4   (26) he has is   
#> I he economic (27)   5 3 3 2 4 1 1 1 3 5 3 5 5 5 1   (27) who beyond t
#> realizes him  (28)   2 3 4 1 2 4 4 3 5 4 5 4 1 5 4   (28) the as slipp
#> ket tricks th (29)   3 5 2 4 2 2 1 4 3 4 2 2 5 5 4   (29) mainland wha
#> d Brownstone  (30)   2 3 1 3 5 1 2 5 4 1 5 3 4 2 4   (30) Johnson     
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - he then                        
#>                 he blackmarket - 7 | 9 - mainland economic back       
#> picts him I mainland realize - 6 | | | 10 - airplane Johnson he realiz
#>  essentially economic away - 5 | | | | | 11 - control blackmarket     
#>              is airplane - 4 | | | | | | | 12 - what depicts initially
#>                    his - 3 | | | | | | | | | 13 - beyond              
#> ks customer customer - 2 | | | | | | | | | | | 14 - airplane slipping 
#>    Brownstone what - 1 | | | | | | | | | | | | | 15 - without he by ma
#>                      | | | | | | | | | | | | | | |                    
#> y man charter  (1)   5 1 2 3 1 4 5 3 3 1 2 5 5 3 5   (1) realizes happ
#>           good (2)   2 2 5 4 3 5 3 1 4 5 2 5 4 4 5   (2) mainland     
#>             to (3)   2 4 1 5 2 1 3 1 1 5 1 1 4 2 1   (3) man is       
#> ishing blackma (4)   4 5 3 3 3 3 4 3 3 4 4 5 3 5 5   (4) I initially b
#> an by the away (5)   3 2 5 4 3 1 5 5 3 1 1 4 3 2 3   (5) novel forces 
#> by is then his (6)   5 4 3 1 4 2 4 1 1 4 5 4 5 1 4   (6) he has is    
#>  I he economic (7)   5 3 3 2 4 1 1 1 3 5 3 5 5 5 1   (7) who beyond th
#>  realizes him  (8)   2 3 4 1 2 4 4 3 5 4 5 4 1 5 4   (8) the as slippi
#> rket tricks th (9)   3 5 2 4 2 2 1 4 3 4 2 2 5 5 4   (9) mainland what
#> d Brownstone  (10)   2 3 1 3 5 1 2 5 4 1 5 3 4 2 4   (10) Johnson     
#>  man charter  (11)   5 1 2 3 1 4 5 3 3 1 2 5 5 3 5   (11) realizes hap
#>          good (12)   2 2 5 4 3 5 3 1 4 5 2 5 4 4 5   (12) mainland    
#>            to (13)   2 4 1 5 2 1 3 1 1 5 1 1 4 2 1   (13) man is      
#> shing blackma (14)   4 5 3 3 3 3 4 3 3 4 4 5 3 5 5   (14) I initially 
#> n by the away (15)   3 2 5 4 3 1 5 5 3 1 1 4 3 2 3   (15) novel forces
#> y is then his (16)   5 4 3 1 4 2 4 1 1 4 5 4 5 1 4   (16) he has is   
#> I he economic (17)   5 3 3 2 4 1 1 1 3 5 3 5 5 5 1   (17) who beyond t
#> realizes him  (18)   2 3 4 1 2 4 4 3 5 4 5 4 1 5 4   (18) the as slipp
#> ket tricks th (19)   3 5 2 4 2 2 1 4 3 4 2 2 5 5 4   (19) mainland wha
#> d Brownstone  (20)   2 3 1 3 5 1 2 5 4 1 5 3 4 2 4   (20) Johnson     
#> Mark airplane (21)   5 5 1 4 4 3 4 5 3 3 2 2 1 2 5   (21) she happened
#> nland by Brow (22)   3 5 1 1 3 2 1 1 3 4 1 2 1 2 4   (22) an          
#> ket has money (23)   5 2 1 5 5 3 2 3 5 4 2 1 1 5 1   (23) him Brownsto
#> before the of (24)   5 5 3 1 2 4 5 4 2 2 1 3 3 2 1   (24) man into hap
#> rownstone who (25)   3 1 1 1 4 2 1 1 1 1 3 1 1 2 4   (25) activity    
#> k forces what (26)   4 5 2 4 1 3 4 1 5 5 5 5 1 1 1   (26) money withou
#>        forced (27)   4 5 2 5 1 1 3 5 3 4 5 1 3 5 5   (27) his he Johns
#> ackmarket tri (28)   3 1 1 4 3 5 4 1 4 4 5 4 5 4 4   (28) Brownstone c
#>         Harry (29)   4 4 2 5 3 5 1 4 2 4 4 2 3 4 3   (29) Johnson back
#> his the the d (30)   1 5 5 2 3 3 3 2 2 4 5 3 4 2 4   (30) forced by th
#>