Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an as by                       
#>               who without away - 7 | 9 - fishing                      
#> happened tricks him he novel - 6 | | | 10 - to charter the the        
#> shing into economic beyond - 5 | | | | | 11 - money                   
#>          by happened his - 4 | | | | | | | 12 - airplane man charter t
#>  the realizes airplane - 3 | | | | | | | | | 13 - she an has he       
#> inland good Mr. depi - 2 | | | | | | | | | | | 14 - any forced by     
#>        his depicts - 1 | | | | | | | | | | | | | 15 - forces          
#>                      | | | | | | | | | | | | | | |                    
#> thout the into (1)   4 2 1 1 1 5 3 4 2 4 1 2 4 2 3   (1) Harry into   
#> ts is by beyon (2)   3 4 4 4 4 4 5 5 4 2 2 1 5 1 5   (2) to man the   
#>  man initially (3)   2 1 5 3 5 5 3 2 5 3 3 4 4 1 5   (3) initially for
#> e essentially  (4)   5 3 4 4 3 4 3 5 3 3 1 1 2 5 3   (4) of he I she a
#> es then any of (5)   4 4 2 1 2 5 3 4 4 2 3 3 1 4 4   (5) of paying    
#> the I novel by (6)   1 5 4 4 5 4 1 5 3 4 4 3 2 1 3   (6) owes forces f
#>   him happened (7)   4 4 1 1 3 4 5 4 1 3 4 1 5 4 3   (7) forced       
#> ipping what sh (8)   3 1 1 4 4 1 2 3 5 4 2 3 3 5 3   (8) essentially i
#> s airplane Bro (9)   4 1 1 1 5 1 3 3 3 3 3 2 1 3 3   (9) activity man 
#>   blackmarket (10)   4 3 4 4 1 3 5 3 3 5 3 2 2 3 4   (10) flees as I J
#>  customer owe (11)   1 5 5 4 4 5 5 1 1 3 2 1 5 1 1   (11) man tricks f
#> away beyond f (12)   1 5 4 5 2 3 2 2 2 2 4 1 5 4 2   (12) blackmarket 
#> rk charter he (13)   4 2 2 1 2 1 2 3 2 4 5 4 4 5 3   (13) has activity
#> oney has as B (14)   3 4 3 3 4 4 1 1 3 4 2 5 1 2 4   (14) customer cha
#> economic what (15)   3 3 3 4 3 1 2 3 3 1 3 5 2 2 5   (15) by          
#>   he back the (16)   1 1 1 4 1 4 4 5 5 4 1 2 2 1 3   (16) what slippin
#> n Mr. without (17)   4 5 5 3 5 2 5 2 2 3 2 2 3 2 5   (17) who away Mr.
#> ping by mainl (18)   4 1 1 3 2 4 4 3 2 2 5 5 3 5 4   (18) Johnson he m
#> rter happened (19)   4 4 3 5 4 4 3 4 5 5 2 1 1 5 3   (19) has depicts 
#>          Mark (20)   5 5 1 3 5 5 5 1 3 1 5 1 1 5 3   (20) slipping an 
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an as by                       
#>               who without away - 7 | 9 - fishing                      
#> happened tricks him he novel - 6 | | | 10 - to charter the the        
#> shing into economic beyond - 5 | | | | | 11 - money                   
#>          by happened his - 4 | | | | | | | 12 - airplane man charter t
#>  the realizes airplane - 3 | | | | | | | | | 13 - she an has he       
#> inland good Mr. depi - 2 | | | | | | | | | | | 14 - any forced by     
#>        his depicts - 1 | | | | | | | | | | | | | 15 - forces          
#>                      | | | | | | | | | | | | | | |                    
#> thout the into (1)   4 2 1 1 1 5 3 4 2 4 1 2 4 2 3   (1) Harry into   
#> ts is by beyon (2)   3 4 4 4 4 4 5 5 4 2 2 1 5 1 5   (2) to man the   
#>  man initially (3)   2 1 5 3 5 5 3 2 5 3 3 4 4 1 5   (3) initially for
#> e essentially  (4)   5 3 4 4 3 4 3 5 3 3 1 1 2 5 3   (4) of he I she a
#> es then any of (5)   4 4 2 1 2 5 3 4 4 2 3 3 1 4 4   (5) of paying    
#> the I novel by (6)   1 5 4 4 5 4 1 5 3 4 4 3 2 1 3   (6) owes forces f
#>   him happened (7)   4 4 1 1 3 4 5 4 1 3 4 1 5 4 3   (7) forced       
#> ipping what sh (8)   3 1 1 4 4 1 2 3 5 4 2 3 3 5 3   (8) essentially i
#> s airplane Bro (9)   4 1 1 1 5 1 3 3 3 3 3 2 1 3 3   (9) activity man 
#>   blackmarket (10)   4 3 4 4 1 3 5 3 3 5 3 2 2 3 4   (10) flees as I J
#>  customer owe (11)   1 5 5 4 4 5 5 1 1 3 2 1 5 1 1   (11) man tricks f
#> away beyond f (12)   1 5 4 5 2 3 2 2 2 2 4 1 5 4 2   (12) blackmarket 
#> rk charter he (13)   4 2 2 1 2 1 2 3 2 4 5 4 4 5 3   (13) has activity
#> oney has as B (14)   3 4 3 3 4 4 1 1 3 4 2 5 1 2 4   (14) customer cha
#> economic what (15)   3 3 3 4 3 1 2 3 3 1 3 5 2 2 5   (15) by          
#>   he back the (16)   1 1 1 4 1 4 4 5 5 4 1 2 2 1 3   (16) what slippin
#> n Mr. without (17)   4 5 5 3 5 2 5 2 2 3 2 2 3 2 5   (17) who away Mr.
#> ping by mainl (18)   4 1 1 3 2 4 4 3 2 2 5 5 3 5 4   (18) Johnson he m
#> rter happened (19)   4 4 3 5 4 4 3 4 5 5 2 1 1 5 3   (19) has depicts 
#>          Mark (20)   5 5 1 3 5 5 5 1 3 1 5 1 1 5 3   (20) slipping an 
#> hout the into (21)   4 2 1 1 1 5 3 4 2 4 1 2 4 2 3   (21) Harry into  
#> s is by beyon (22)   3 4 4 4 4 4 5 5 4 2 2 1 5 1 5   (22) to man the  
#> man initially (23)   2 1 5 3 5 5 3 2 5 3 3 4 4 1 5   (23) initially fo
#>  essentially  (24)   5 3 4 4 3 4 3 5 3 3 1 1 2 5 3   (24) of he I she 
#> s then any of (25)   4 4 2 1 2 5 3 4 4 2 3 3 1 4 4   (25) of paying   
#> he I novel by (26)   1 5 4 4 5 4 1 5 3 4 4 3 2 1 3   (26) owes forces 
#>  him happened (27)   4 4 1 1 3 4 5 4 1 3 4 1 5 4 3   (27) forced      
#> pping what sh (28)   3 1 1 4 4 1 2 3 5 4 2 3 3 5 3   (28) essentially 
#>  airplane Bro (29)   4 1 1 1 5 1 3 3 3 3 3 2 1 3 3   (29) activity man
#>   blackmarket (30)   4 3 4 4 1 3 5 3 3 5 3 2 2 3 4   (30) flees as I J
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an as by                       
#>               who without away - 7 | 9 - fishing                      
#> happened tricks him he novel - 6 | | | 10 - to charter the the        
#> shing into economic beyond - 5 | | | | | 11 - money                   
#>          by happened his - 4 | | | | | | | 12 - airplane man charter t
#>  the realizes airplane - 3 | | | | | | | | | 13 - she an has he       
#> inland good Mr. depi - 2 | | | | | | | | | | | 14 - any forced by     
#>        his depicts - 1 | | | | | | | | | | | | | 15 - forces          
#>                      | | | | | | | | | | | | | | |                    
#> thout the into (1)   4 2 1 1 1 5 3 4 2 4 1 2 4 2 3   (1) Harry into   
#> ts is by beyon (2)   3 4 4 4 4 4 5 5 4 2 2 1 5 1 5   (2) to man the   
#>  man initially (3)   2 1 5 3 5 5 3 2 5 3 3 4 4 1 5   (3) initially for
#> e essentially  (4)   5 3 4 4 3 4 3 5 3 3 1 1 2 5 3   (4) of he I she a
#> es then any of (5)   4 4 2 1 2 5 3 4 4 2 3 3 1 4 4   (5) of paying    
#> the I novel by (6)   1 5 4 4 5 4 1 5 3 4 4 3 2 1 3   (6) owes forces f
#>   him happened (7)   4 4 1 1 3 4 5 4 1 3 4 1 5 4 3   (7) forced       
#> ipping what sh (8)   3 1 1 4 4 1 2 3 5 4 2 3 3 5 3   (8) essentially i
#> s airplane Bro (9)   4 1 1 1 5 1 3 3 3 3 3 2 1 3 3   (9) activity man 
#>   blackmarket (10)   4 3 4 4 1 3 5 3 3 5 3 2 2 3 4   (10) flees as I J
#> hout the into (11)   4 2 1 1 1 5 3 4 2 4 1 2 4 2 3   (11) Harry into  
#> s is by beyon (12)   3 4 4 4 4 4 5 5 4 2 2 1 5 1 5   (12) to man the  
#> man initially (13)   2 1 5 3 5 5 3 2 5 3 3 4 4 1 5   (13) initially fo
#>  essentially  (14)   5 3 4 4 3 4 3 5 3 3 1 1 2 5 3   (14) of he I she 
#> s then any of (15)   4 4 2 1 2 5 3 4 4 2 3 3 1 4 4   (15) of paying   
#> he I novel by (16)   1 5 4 4 5 4 1 5 3 4 4 3 2 1 3   (16) owes forces 
#>  him happened (17)   4 4 1 1 3 4 5 4 1 3 4 1 5 4 3   (17) forced      
#> pping what sh (18)   3 1 1 4 4 1 2 3 5 4 2 3 3 5 3   (18) essentially 
#>  airplane Bro (19)   4 1 1 1 5 1 3 3 3 3 3 2 1 3 3   (19) activity man
#>   blackmarket (20)   4 3 4 4 1 3 5 3 3 5 3 2 2 3 4   (20) flees as I J
#>  customer owe (21)   1 5 5 4 4 5 5 1 1 3 2 1 5 1 1   (21) man tricks f
#> away beyond f (22)   1 5 4 5 2 3 2 2 2 2 4 1 5 4 2   (22) blackmarket 
#> rk charter he (23)   4 2 2 1 2 1 2 3 2 4 5 4 4 5 3   (23) has activity
#> oney has as B (24)   3 4 3 3 4 4 1 1 3 4 2 5 1 2 4   (24) customer cha
#> economic what (25)   3 3 3 4 3 1 2 3 3 1 3 5 2 2 5   (25) by          
#>   he back the (26)   1 1 1 4 1 4 4 5 5 4 1 2 2 1 3   (26) what slippin
#> n Mr. without (27)   4 5 5 3 5 2 5 2 2 3 2 2 3 2 5   (27) who away Mr.
#> ping by mainl (28)   4 1 1 3 2 4 4 3 2 2 5 5 3 5 4   (28) Johnson he m
#> rter happened (29)   4 4 3 5 4 4 3 4 5 5 2 1 1 5 3   (29) has depicts 
#>          Mark (30)   5 5 1 3 5 5 5 1 3 1 5 1 1 5 3   (30) slipping an 
#>