Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - forces tricks                  
#>                           Mark - 7 | 9 - money forces                 
#>                 by initially - 6 | | | 10 - back fishing              
#>  Mark forces economic then - 5 | | | | | 11 - control he              
#>       has Harry to Harry - 4 | | | | | | | 12 - beyond charter paying 
#> y the tricks what owes - 3 | | | | | | | | | 13 - Mr. novel Johnson Mr
#>      Harry the is to - 2 | | | | | | | | | | | 14 - before he an      
#>    depicts I money - 1 | | | | | | | | | | | | | 15 - before his his t
#>                      | | | | | | | | | | | | | | |                    
#> alizes Johnson (1)   5 2 4 3 3 1 5 1 2 2 3 1 3 5 3   (1) Brownstone   
#>         before (2)   3 3 1 3 1 2 1 2 4 2 5 3 4 4 1   (2) she away befo
#> s his without  (3)   5 4 2 4 4 4 4 4 5 2 3 1 5 5 2   (3) control flees
#> lackmarket who (4)   5 5 5 5 5 3 2 5 4 1 5 4 3 5 3   (4) paying I his 
#>         forced (5)   2 2 2 2 3 1 5 1 1 5 1 4 5 2 3   (5) economic the 
#> blackmarket hi (6)   1 5 2 1 1 5 1 1 1 2 3 2 1 1 5   (6) depicts by Br
#> m forces the a (7)   3 2 2 4 2 4 5 4 4 5 4 1 3 3 5   (7) he tricks he 
#>  blackmarket b (8)   4 3 1 4 4 2 4 3 4 2 4 1 5 3 3   (8) I Harry him  
#> e then fishing (9)   1 1 4 5 5 1 2 1 3 2 1 1 1 3 3   (9) beyond Harry 
#> then airplane (10)   3 1 4 5 3 4 2 1 2 1 4 4 4 3 3   (10) airplane hap
#>       fishing (11)   4 5 2 4 2 1 4 1 4 5 1 1 4 1 1   (11) what        
#> en airplane w (12)   1 1 2 4 4 1 3 3 2 4 1 4 2 4 3   (12) Johnson     
#>            is (13)   2 3 1 3 1 5 1 4 5 4 3 5 2 4 1   (13) who customer
#> he as control (14)   3 4 2 2 2 3 5 4 5 2 3 2 4 1 2   (14) economic he 
#>        she by (15)   5 3 3 1 4 1 3 3 4 3 4 3 2 3 3   (15) blackmarket 
#> picts she the (16)   1 4 2 4 5 1 3 1 1 3 1 3 3 3 1   (16) paying fishi
#>  realizes cha (17)   3 2 2 2 3 2 1 3 2 4 3 3 1 3 5   (17) what charter
#> ing by away b (18)   5 1 2 1 2 5 4 5 3 3 3 1 4 1 1   (18) an the by is
#> cts novel she (19)   2 2 4 1 4 1 4 1 4 4 5 2 3 5 4   (19) Harry initia
#> c novel she I (20)   2 4 1 1 2 3 5 1 1 5 1 2 2 4 3   (20) without has 
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - forces tricks                  
#>                           Mark - 7 | 9 - money forces                 
#>                 by initially - 6 | | | 10 - back fishing              
#>  Mark forces economic then - 5 | | | | | 11 - control he              
#>       has Harry to Harry - 4 | | | | | | | 12 - beyond charter paying 
#> y the tricks what owes - 3 | | | | | | | | | 13 - Mr. novel Johnson Mr
#>      Harry the is to - 2 | | | | | | | | | | | 14 - before he an      
#>    depicts I money - 1 | | | | | | | | | | | | | 15 - before his his t
#>                      | | | | | | | | | | | | | | |                    
#> alizes Johnson (1)   5 2 4 3 3 1 5 1 2 2 3 1 3 5 3   (1) Brownstone   
#>         before (2)   3 3 1 3 1 2 1 2 4 2 5 3 4 4 1   (2) she away befo
#> s his without  (3)   5 4 2 4 4 4 4 4 5 2 3 1 5 5 2   (3) control flees
#> lackmarket who (4)   5 5 5 5 5 3 2 5 4 1 5 4 3 5 3   (4) paying I his 
#>         forced (5)   2 2 2 2 3 1 5 1 1 5 1 4 5 2 3   (5) economic the 
#> blackmarket hi (6)   1 5 2 1 1 5 1 1 1 2 3 2 1 1 5   (6) depicts by Br
#> m forces the a (7)   3 2 2 4 2 4 5 4 4 5 4 1 3 3 5   (7) he tricks he 
#>  blackmarket b (8)   4 3 1 4 4 2 4 3 4 2 4 1 5 3 3   (8) I Harry him  
#> e then fishing (9)   1 1 4 5 5 1 2 1 3 2 1 1 1 3 3   (9) beyond Harry 
#> then airplane (10)   3 1 4 5 3 4 2 1 2 1 4 4 4 3 3   (10) airplane hap
#>       fishing (11)   4 5 2 4 2 1 4 1 4 5 1 1 4 1 1   (11) what        
#> en airplane w (12)   1 1 2 4 4 1 3 3 2 4 1 4 2 4 3   (12) Johnson     
#>            is (13)   2 3 1 3 1 5 1 4 5 4 3 5 2 4 1   (13) who customer
#> he as control (14)   3 4 2 2 2 3 5 4 5 2 3 2 4 1 2   (14) economic he 
#>        she by (15)   5 3 3 1 4 1 3 3 4 3 4 3 2 3 3   (15) blackmarket 
#> picts she the (16)   1 4 2 4 5 1 3 1 1 3 1 3 3 3 1   (16) paying fishi
#>  realizes cha (17)   3 2 2 2 3 2 1 3 2 4 3 3 1 3 5   (17) what charter
#> ing by away b (18)   5 1 2 1 2 5 4 5 3 3 3 1 4 1 1   (18) an the by is
#> cts novel she (19)   2 2 4 1 4 1 4 1 4 4 5 2 3 5 4   (19) Harry initia
#> c novel she I (20)   2 4 1 1 2 3 5 1 1 5 1 2 2 4 3   (20) without has 
#> lizes Johnson (21)   5 2 4 3 3 1 5 1 2 2 3 1 3 5 3   (21) Brownstone  
#>        before (22)   3 3 1 3 1 2 1 2 4 2 5 3 4 4 1   (22) she away bef
#>  his without  (23)   5 4 2 4 4 4 4 4 5 2 3 1 5 5 2   (23) control flee
#> ackmarket who (24)   5 5 5 5 5 3 2 5 4 1 5 4 3 5 3   (24) paying I his
#>        forced (25)   2 2 2 2 3 1 5 1 1 5 1 4 5 2 3   (25) economic the
#> lackmarket hi (26)   1 5 2 1 1 5 1 1 1 2 3 2 1 1 5   (26) depicts by B
#>  forces the a (27)   3 2 2 4 2 4 5 4 4 5 4 1 3 3 5   (27) he tricks he
#> blackmarket b (28)   4 3 1 4 4 2 4 3 4 2 4 1 5 3 3   (28) I Harry him 
#>  then fishing (29)   1 1 4 5 5 1 2 1 3 2 1 1 1 3 3   (29) beyond Harry
#> then airplane (30)   3 1 4 5 3 4 2 1 2 1 4 4 4 3 3   (30) airplane hap
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - forces tricks                  
#>                           Mark - 7 | 9 - money forces                 
#>                 by initially - 6 | | | 10 - back fishing              
#>  Mark forces economic then - 5 | | | | | 11 - control he              
#>       has Harry to Harry - 4 | | | | | | | 12 - beyond charter paying 
#> y the tricks what owes - 3 | | | | | | | | | 13 - Mr. novel Johnson Mr
#>      Harry the is to - 2 | | | | | | | | | | | 14 - before he an      
#>    depicts I money - 1 | | | | | | | | | | | | | 15 - before his his t
#>                      | | | | | | | | | | | | | | |                    
#> alizes Johnson (1)   5 2 4 3 3 1 5 1 2 2 3 1 3 5 3   (1) Brownstone   
#>         before (2)   3 3 1 3 1 2 1 2 4 2 5 3 4 4 1   (2) she away befo
#> s his without  (3)   5 4 2 4 4 4 4 4 5 2 3 1 5 5 2   (3) control flees
#> lackmarket who (4)   5 5 5 5 5 3 2 5 4 1 5 4 3 5 3   (4) paying I his 
#>         forced (5)   2 2 2 2 3 1 5 1 1 5 1 4 5 2 3   (5) economic the 
#> blackmarket hi (6)   1 5 2 1 1 5 1 1 1 2 3 2 1 1 5   (6) depicts by Br
#> m forces the a (7)   3 2 2 4 2 4 5 4 4 5 4 1 3 3 5   (7) he tricks he 
#>  blackmarket b (8)   4 3 1 4 4 2 4 3 4 2 4 1 5 3 3   (8) I Harry him  
#> e then fishing (9)   1 1 4 5 5 1 2 1 3 2 1 1 1 3 3   (9) beyond Harry 
#> then airplane (10)   3 1 4 5 3 4 2 1 2 1 4 4 4 3 3   (10) airplane hap
#> lizes Johnson (11)   5 2 4 3 3 1 5 1 2 2 3 1 3 5 3   (11) Brownstone  
#>        before (12)   3 3 1 3 1 2 1 2 4 2 5 3 4 4 1   (12) she away bef
#>  his without  (13)   5 4 2 4 4 4 4 4 5 2 3 1 5 5 2   (13) control flee
#> ackmarket who (14)   5 5 5 5 5 3 2 5 4 1 5 4 3 5 3   (14) paying I his
#>        forced (15)   2 2 2 2 3 1 5 1 1 5 1 4 5 2 3   (15) economic the
#> lackmarket hi (16)   1 5 2 1 1 5 1 1 1 2 3 2 1 1 5   (16) depicts by B
#>  forces the a (17)   3 2 2 4 2 4 5 4 4 5 4 1 3 3 5   (17) he tricks he
#> blackmarket b (18)   4 3 1 4 4 2 4 3 4 2 4 1 5 3 3   (18) I Harry him 
#>  then fishing (19)   1 1 4 5 5 1 2 1 3 2 1 1 1 3 3   (19) beyond Harry
#> then airplane (20)   3 1 4 5 3 4 2 1 2 1 4 4 4 3 3   (20) airplane hap
#>       fishing (21)   4 5 2 4 2 1 4 1 4 5 1 1 4 1 1   (21) what        
#> en airplane w (22)   1 1 2 4 4 1 3 3 2 4 1 4 2 4 3   (22) Johnson     
#>            is (23)   2 3 1 3 1 5 1 4 5 4 3 5 2 4 1   (23) who customer
#> he as control (24)   3 4 2 2 2 3 5 4 5 2 3 2 4 1 2   (24) economic he 
#>        she by (25)   5 3 3 1 4 1 3 3 4 3 4 3 2 3 3   (25) blackmarket 
#> picts she the (26)   1 4 2 4 5 1 3 1 1 3 1 3 3 3 1   (26) paying fishi
#>  realizes cha (27)   3 2 2 2 3 2 1 3 2 4 3 3 1 3 5   (27) what charter
#> ing by away b (28)   5 1 2 1 2 5 4 5 3 3 3 1 4 1 1   (28) an the by is
#> cts novel she (29)   2 2 4 1 4 1 4 1 4 4 5 2 3 5 4   (29) Harry initia
#> c novel she I (30)   2 4 1 1 2 3 5 1 1 5 1 2 2 4 3   (30) without has 
#>