Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of elements:  15 
#> Number of constructs:  20 
#> Preferred poles defined:  18/20 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                         8 - is customer paying                 
#>                                 the - 7 | 9 - novel his him depicts the        
#>            beyond Johnson Johnson - 6 | | | 10 - Harry                         
#>                   slipping then - 5 | | | | | 11 - him the                     
#>               by of good away - 4 | | | | | | | 12 - money an initially beyond 
#>                  charter to - 3 | | | | | | | | | 13 - novel he                
#> without airplane back the - 2 | | | | | | | | | | | 14 - the                   
#>    has initially by the - 1 | | | | | | | | | | | | | 15 - Johnson away essenti
#>                           | | | | | | | | | | | | | | |                        
#> as money airplane  (-1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1+) who then into th
#>                 of (.2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2.) money control   
#>      economic into (+3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3-) any             
#>           the away (-4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4+) the depicts     
#> rces slipping as w (+5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5-) essentially main
#> essentially his fi (-6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6+) is him he       
#>  man airplane owes (-7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7+) realizes mainlan
#>            she the (-8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8+) activity forces 
#> s flees charter he (+9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9-) realizes control
#> forces his before (+10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10-) the            
#> irplane who flees (+11)   3 2 2 2 4 5 2 4 2 5 5 1 1 2 3   (11-) of charter his 
#> kmarket into Brow (+12)   4 5 5 5 5 5 2 2 1 4 4 3 2 2 4   (12-) happened the ha
#> entially what has (+13)   2 5 4 1 4 4 4 4 1 3 3 4 5 4 4   (13-) an good        
#>               the (+14)   3 5 2 5 1 1 3 3 4 3 5 2 5 3 2   (14-) depicts as the 
#>                as (-15)   5 5 5 4 2 3 2 5 3 2 1 1 1 2 5   (15+) charter initial
#>               she (.16)   1 3 1 1 3 5 3 3 2 3 1 2 1 3 5   (16.) before initiall
#>        an by then (+17)   3 1 3 4 4 4 3 2 2 5 4 1 1 3 4   (17-) man she mainlan
#> ity Johnson as as (+18)   1 3 4 2 3 3 3 3 4 4 4 1 2 3 4   (18-) Mark           
#> rol man customer  (+19)   1 4 5 3 5 1 1 3 3 1 3 1 1 1 2   (19-) customer withou
#>       is slipping (+20)   2 3 3 2 4 3 2 4 2 4 4 2 1 2 3   (20-) mainland any ow
#> 
#>  
#>  
#>  
#>  
#> h
#>  
#>  
#> a
#>  
#> e
#>  
#>  
#>  
#> l
#>  
#> d
#> t
#>  
#>  
#> b
#> s
#>  
#>  
#> l
#> y
#> d
#>  
#> t
#> e
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of elements:  15 
#> Number of constructs:  30 
#> Preferred poles defined:  27/30 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                         8 - is customer paying                 
#>                                 the - 7 | 9 - novel his him depicts the        
#>            beyond Johnson Johnson - 6 | | | 10 - Harry                         
#>                   slipping then - 5 | | | | | 11 - him the                     
#>               by of good away - 4 | | | | | | | 12 - money an initially beyond 
#>                  charter to - 3 | | | | | | | | | 13 - novel he                
#> without airplane back the - 2 | | | | | | | | | | | 14 - the                   
#>    has initially by the - 1 | | | | | | | | | | | | | 15 - Johnson away essenti
#>                           | | | | | | | | | | | | | | |                        
#> as money airplane  (-1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1+) who then into th
#>                 of (.2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2.) money control   
#>      economic into (+3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3-) any             
#>           the away (-4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4+) the depicts     
#> rces slipping as w (+5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5-) essentially main
#> essentially his fi (-6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6+) is him he       
#>  man airplane owes (-7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7+) realizes mainlan
#>            she the (-8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8+) activity forces 
#> s flees charter he (+9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9-) realizes control
#> forces his before (+10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10-) the            
#> irplane who flees (+11)   3 2 2 2 4 5 2 4 2 5 5 1 1 2 3   (11-) of charter his 
#> kmarket into Brow (+12)   4 5 5 5 5 5 2 2 1 4 4 3 2 2 4   (12-) happened the ha
#> entially what has (+13)   2 5 4 1 4 4 4 4 1 3 3 4 5 4 4   (13-) an good        
#>               the (+14)   3 5 2 5 1 1 3 3 4 3 5 2 5 3 2   (14-) depicts as the 
#>                as (-15)   5 5 5 4 2 3 2 5 3 2 1 1 1 2 5   (15+) charter initial
#>               she (.16)   1 3 1 1 3 5 3 3 2 3 1 2 1 3 5   (16.) before initiall
#>        an by then (+17)   3 1 3 4 4 4 3 2 2 5 4 1 1 3 4   (17-) man she mainlan
#> ity Johnson as as (+18)   1 3 4 2 3 3 3 3 4 4 4 1 2 3 4   (18-) Mark           
#> rol man customer  (+19)   1 4 5 3 5 1 1 3 3 1 3 1 1 1 2   (19-) customer withou
#>       is slipping (+20)   2 3 3 2 4 3 2 4 2 4 4 2 1 2 3   (20-) mainland any ow
#> s money airplane  (-21)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (21+) who then into t
#>                of (.22)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (22.) money control  
#>     economic into (+23)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (23-) any            
#>          the away (-24)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (24+) the depicts    
#> ces slipping as w (+25)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (25-) essentially mai
#> ssentially his fi (-26)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (26+) is him he      
#> man airplane owes (-27)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (27+) realizes mainla
#>           she the (-28)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (28+) activity forces
#>  flees charter he (+29)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (29-) realizes contro
#> forces his before (+30)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (30-) the            
#> 
#>  
#>  
#>  
#>  
#> h
#>  
#>  
#> a
#>  
#> e
#>  
#>  
#>  
#> l
#>  
#> d
#> t
#>  
#>  
#> b
#> s
#>  
#>  
#> l
#> y
#> d
#>  
#> t
#> e
#> h
#>  
#>  
#>  
#> n
#>  
#> n
#>  
#> l
#>  
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of elements:  15 
#> Number of constructs:  30 
#> Preferred poles defined:  27/30 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                         8 - is customer paying                 
#>                                 the - 7 | 9 - novel his him depicts the        
#>            beyond Johnson Johnson - 6 | | | 10 - Harry                         
#>                   slipping then - 5 | | | | | 11 - him the                     
#>               by of good away - 4 | | | | | | | 12 - money an initially beyond 
#>                  charter to - 3 | | | | | | | | | 13 - novel he                
#> without airplane back the - 2 | | | | | | | | | | | 14 - the                   
#>    has initially by the - 1 | | | | | | | | | | | | | 15 - Johnson away essenti
#>                           | | | | | | | | | | | | | | |                        
#> as money airplane  (-1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1+) who then into th
#>                 of (.2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2.) money control   
#>      economic into (+3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3-) any             
#>           the away (-4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4+) the depicts     
#> rces slipping as w (+5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5-) essentially main
#> essentially his fi (-6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6+) is him he       
#>  man airplane owes (-7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7+) realizes mainlan
#>            she the (-8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8+) activity forces 
#> s flees charter he (+9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9-) realizes control
#> forces his before (+10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10-) the            
#> s money airplane  (-11)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (11+) who then into t
#>                of (.12)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (12.) money control  
#>     economic into (+13)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (13-) any            
#>          the away (-14)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (14+) the depicts    
#> ces slipping as w (+15)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (15-) essentially mai
#> ssentially his fi (-16)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (16+) is him he      
#> man airplane owes (-17)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (17+) realizes mainla
#>           she the (-18)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (18+) activity forces
#>  flees charter he (+19)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (19-) realizes contro
#> forces his before (+20)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (20-) the            
#> irplane who flees (+21)   3 2 2 2 4 5 2 4 2 5 5 1 1 2 3   (21-) of charter his 
#> kmarket into Brow (+22)   4 5 5 5 5 5 2 2 1 4 4 3 2 2 4   (22-) happened the ha
#> entially what has (+23)   2 5 4 1 4 4 4 4 1 3 3 4 5 4 4   (23-) an good        
#>               the (+24)   3 5 2 5 1 1 3 3 4 3 5 2 5 3 2   (24-) depicts as the 
#>                as (-25)   5 5 5 4 2 3 2 5 3 2 1 1 1 2 5   (25+) charter initial
#>               she (.26)   1 3 1 1 3 5 3 3 2 3 1 2 1 3 5   (26.) before initiall
#>        an by then (+27)   3 1 3 4 4 4 3 2 2 5 4 1 1 3 4   (27-) man she mainlan
#> ity Johnson as as (+28)   1 3 4 2 3 3 3 3 4 4 4 1 2 3 4   (28-) Mark           
#> rol man customer  (+29)   1 4 5 3 5 1 1 3 3 1 3 1 1 1 2   (29-) customer withou
#>       is slipping (+30)   2 3 3 2 4 3 2 4 2 4 4 2 1 2 3   (30-) mainland any ow
#> 
#>  
#>  
#>  
#>  
#> h
#>  
#>  
#> a
#>  
#> e
#>  
#>  
#>  
#> l
#>  
#> d
#> t
#>  
#>  
#> h
#>  
#>  
#>  
#> n
#>  
#> n
#>  
#> l
#>  
#> b
#> s
#>  
#>  
#> l
#> y
#> d
#>  
#> t
#> e
#> 
#> Poles: (+) = preferred , (-) = non-preferred , (/) = none , (.) = not defined