Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - control forces tricks into     
#>                            the - 7 | 9 - forced has depicts fishing co
#>  good by is mainland fishing - 6 | | | 10 - economic paying           
#>                   by Mr. I - 5 | | | | | 11 - happened slipping mainla
#> Johnson the charter then - 4 | | | | | | | 12 - initially essentially 
#>             Brownstone - 3 | | | | | | | | | 13 - into money          
#>  charter tricks back - 2 | | | | | | | | | | | 14 - the by without    
#>           airplane - 1 | | | | | | | | | | | | | 15 - then            
#>                      | | | | | | | | | | | | | | |                    
#> s what the man (1)   2 2 4 4 4 5 5 3 3 4 4 1 1 4 1   (1) has Mark     
#> ckmarket him M (2)   3 1 1 2 1 1 2 1 2 5 1 1 4 4 5   (2) depicts novel
#>  is he into co (3)   3 5 3 2 2 3 4 1 4 4 4 5 1 5 4   (3) activity nove
#> efore control  (4)   2 1 2 2 2 3 3 3 5 2 5 3 3 5 3   (4) is has an ini
#> arter who forc (5)   4 5 2 3 4 4 3 5 1 1 3 5 2 3 1   (5) Harry paying 
#>  an Mr. beyond (6)   3 3 1 1 2 5 2 2 4 5 2 5 2 4 2   (6) man by he tri
#>         by any (7)   3 3 1 4 3 5 1 2 1 3 5 2 5 2 5   (7) as paying    
#>        charter (8)   2 2 5 5 4 1 5 5 2 1 1 2 2 5 3   (8) I of has owes
#>  Brownstone wh (9)   3 5 4 1 1 3 1 5 5 3 5 2 1 3 4   (9) him essential
#> ackmarket the (10)   5 1 4 5 1 4 4 2 4 3 5 4 3 3 4   (10) into his the
#> before to man (11)   4 1 3 5 5 5 3 4 2 5 1 5 2 1 1   (11) Mr. happened
#> owes before h (12)   3 4 2 1 3 4 5 2 1 5 2 2 3 1 4   (12) depicts happ
#>  his realizes (13)   5 1 4 2 3 5 2 2 2 2 1 4 3 4 1   (13) blackmarket 
#>  his owes him (14)   1 4 1 3 2 1 4 5 5 2 5 3 4 3 1   (14) forces black
#>  is tricks ai (15)   2 2 2 1 3 5 4 1 5 4 1 5 4 4 4   (15) before      
#>  I economic c (16)   5 5 5 4 1 5 1 5 5 5 3 3 4 1 1   (16) realizes is 
#> vel who airpl (17)   4 5 2 2 4 2 2 1 3 3 3 5 2 5 5   (17) what man hap
#> ownstone he h (18)   5 1 3 1 3 5 2 2 5 3 1 5 3 2 1   (18) forced any a
#>  as blackmark (19)   5 4 1 4 2 1 3 4 1 1 3 2 3 1 2   (19) an owes     
#> er he Mr. any (20)   4 3 3 1 5 2 3 1 2 4 3 4 4 3 4   (20) by activity 
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - control forces tricks into     
#>                            the - 7 | 9 - forced has depicts fishing co
#>  good by is mainland fishing - 6 | | | 10 - economic paying           
#>                   by Mr. I - 5 | | | | | 11 - happened slipping mainla
#> Johnson the charter then - 4 | | | | | | | 12 - initially essentially 
#>             Brownstone - 3 | | | | | | | | | 13 - into money          
#>  charter tricks back - 2 | | | | | | | | | | | 14 - the by without    
#>           airplane - 1 | | | | | | | | | | | | | 15 - then            
#>                      | | | | | | | | | | | | | | |                    
#> s what the man (1)   2 2 4 4 4 5 5 3 3 4 4 1 1 4 1   (1) has Mark     
#> ckmarket him M (2)   3 1 1 2 1 1 2 1 2 5 1 1 4 4 5   (2) depicts novel
#>  is he into co (3)   3 5 3 2 2 3 4 1 4 4 4 5 1 5 4   (3) activity nove
#> efore control  (4)   2 1 2 2 2 3 3 3 5 2 5 3 3 5 3   (4) is has an ini
#> arter who forc (5)   4 5 2 3 4 4 3 5 1 1 3 5 2 3 1   (5) Harry paying 
#>  an Mr. beyond (6)   3 3 1 1 2 5 2 2 4 5 2 5 2 4 2   (6) man by he tri
#>         by any (7)   3 3 1 4 3 5 1 2 1 3 5 2 5 2 5   (7) as paying    
#>        charter (8)   2 2 5 5 4 1 5 5 2 1 1 2 2 5 3   (8) I of has owes
#>  Brownstone wh (9)   3 5 4 1 1 3 1 5 5 3 5 2 1 3 4   (9) him essential
#> ackmarket the (10)   5 1 4 5 1 4 4 2 4 3 5 4 3 3 4   (10) into his the
#> before to man (11)   4 1 3 5 5 5 3 4 2 5 1 5 2 1 1   (11) Mr. happened
#> owes before h (12)   3 4 2 1 3 4 5 2 1 5 2 2 3 1 4   (12) depicts happ
#>  his realizes (13)   5 1 4 2 3 5 2 2 2 2 1 4 3 4 1   (13) blackmarket 
#>  his owes him (14)   1 4 1 3 2 1 4 5 5 2 5 3 4 3 1   (14) forces black
#>  is tricks ai (15)   2 2 2 1 3 5 4 1 5 4 1 5 4 4 4   (15) before      
#>  I economic c (16)   5 5 5 4 1 5 1 5 5 5 3 3 4 1 1   (16) realizes is 
#> vel who airpl (17)   4 5 2 2 4 2 2 1 3 3 3 5 2 5 5   (17) what man hap
#> ownstone he h (18)   5 1 3 1 3 5 2 2 5 3 1 5 3 2 1   (18) forced any a
#>  as blackmark (19)   5 4 1 4 2 1 3 4 1 1 3 2 3 1 2   (19) an owes     
#> er he Mr. any (20)   4 3 3 1 5 2 3 1 2 4 3 4 4 3 4   (20) by activity 
#>  what the man (21)   2 2 4 4 4 5 5 3 3 4 4 1 1 4 1   (21) has Mark    
#> kmarket him M (22)   3 1 1 2 1 1 2 1 2 5 1 1 4 4 5   (22) depicts nove
#> is he into co (23)   3 5 3 2 2 3 4 1 4 4 4 5 1 5 4   (23) activity nov
#> fore control  (24)   2 1 2 2 2 3 3 3 5 2 5 3 3 5 3   (24) is has an in
#> rter who forc (25)   4 5 2 3 4 4 3 5 1 1 3 5 2 3 1   (25) Harry paying
#> an Mr. beyond (26)   3 3 1 1 2 5 2 2 4 5 2 5 2 4 2   (26) man by he tr
#>        by any (27)   3 3 1 4 3 5 1 2 1 3 5 2 5 2 5   (27) as paying   
#>       charter (28)   2 2 5 5 4 1 5 5 2 1 1 2 2 5 3   (28) I of has owe
#> Brownstone wh (29)   3 5 4 1 1 3 1 5 5 3 5 2 1 3 4   (29) him essentia
#> ackmarket the (30)   5 1 4 5 1 4 4 2 4 3 5 4 3 3 4   (30) into his the
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - control forces tricks into     
#>                            the - 7 | 9 - forced has depicts fishing co
#>  good by is mainland fishing - 6 | | | 10 - economic paying           
#>                   by Mr. I - 5 | | | | | 11 - happened slipping mainla
#> Johnson the charter then - 4 | | | | | | | 12 - initially essentially 
#>             Brownstone - 3 | | | | | | | | | 13 - into money          
#>  charter tricks back - 2 | | | | | | | | | | | 14 - the by without    
#>           airplane - 1 | | | | | | | | | | | | | 15 - then            
#>                      | | | | | | | | | | | | | | |                    
#> s what the man (1)   2 2 4 4 4 5 5 3 3 4 4 1 1 4 1   (1) has Mark     
#> ckmarket him M (2)   3 1 1 2 1 1 2 1 2 5 1 1 4 4 5   (2) depicts novel
#>  is he into co (3)   3 5 3 2 2 3 4 1 4 4 4 5 1 5 4   (3) activity nove
#> efore control  (4)   2 1 2 2 2 3 3 3 5 2 5 3 3 5 3   (4) is has an ini
#> arter who forc (5)   4 5 2 3 4 4 3 5 1 1 3 5 2 3 1   (5) Harry paying 
#>  an Mr. beyond (6)   3 3 1 1 2 5 2 2 4 5 2 5 2 4 2   (6) man by he tri
#>         by any (7)   3 3 1 4 3 5 1 2 1 3 5 2 5 2 5   (7) as paying    
#>        charter (8)   2 2 5 5 4 1 5 5 2 1 1 2 2 5 3   (8) I of has owes
#>  Brownstone wh (9)   3 5 4 1 1 3 1 5 5 3 5 2 1 3 4   (9) him essential
#> ackmarket the (10)   5 1 4 5 1 4 4 2 4 3 5 4 3 3 4   (10) into his the
#>  what the man (11)   2 2 4 4 4 5 5 3 3 4 4 1 1 4 1   (11) has Mark    
#> kmarket him M (12)   3 1 1 2 1 1 2 1 2 5 1 1 4 4 5   (12) depicts nove
#> is he into co (13)   3 5 3 2 2 3 4 1 4 4 4 5 1 5 4   (13) activity nov
#> fore control  (14)   2 1 2 2 2 3 3 3 5 2 5 3 3 5 3   (14) is has an in
#> rter who forc (15)   4 5 2 3 4 4 3 5 1 1 3 5 2 3 1   (15) Harry paying
#> an Mr. beyond (16)   3 3 1 1 2 5 2 2 4 5 2 5 2 4 2   (16) man by he tr
#>        by any (17)   3 3 1 4 3 5 1 2 1 3 5 2 5 2 5   (17) as paying   
#>       charter (18)   2 2 5 5 4 1 5 5 2 1 1 2 2 5 3   (18) I of has owe
#> Brownstone wh (19)   3 5 4 1 1 3 1 5 5 3 5 2 1 3 4   (19) him essentia
#> ackmarket the (20)   5 1 4 5 1 4 4 2 4 3 5 4 3 3 4   (20) into his the
#> before to man (21)   4 1 3 5 5 5 3 4 2 5 1 5 2 1 1   (21) Mr. happened
#> owes before h (22)   3 4 2 1 3 4 5 2 1 5 2 2 3 1 4   (22) depicts happ
#>  his realizes (23)   5 1 4 2 3 5 2 2 2 2 1 4 3 4 1   (23) blackmarket 
#>  his owes him (24)   1 4 1 3 2 1 4 5 5 2 5 3 4 3 1   (24) forces black
#>  is tricks ai (25)   2 2 2 1 3 5 4 1 5 4 1 5 4 4 4   (25) before      
#>  I economic c (26)   5 5 5 4 1 5 1 5 5 5 3 3 4 1 1   (26) realizes is 
#> vel who airpl (27)   4 5 2 2 4 2 2 1 3 3 3 5 2 5 5   (27) what man hap
#> ownstone he h (28)   5 1 3 1 3 5 2 2 5 3 1 5 3 2 1   (28) forced any a
#>  as blackmark (29)   5 4 1 4 2 1 3 4 1 1 3 2 3 1 2   (29) an owes     
#> er he Mr. any (30)   4 3 3 1 5 2 3 1 2 4 3 4 4 3 4   (30) by activity 
#>