Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - his realizes                   
#>            of charter economic - 7 | 9 - blackmarket his activity     
#>                           as - 6 | | | 10 - realizes blackmarket airpl
#> money has owes charter she - 5 | | | | | 11 - who fishing before man  
#> ed charter tricks Harry  - 4 | | | | | | | 12 - mainland Mark customer
#>                     of - 3 | | | | | | | | | 13 - money man money     
#>            paying is - 2 | | | | | | | | | | | 14 - Harry who the blac
#> icts money mainlan - 1 | | | | | | | | | | | | | 15 - Mr. novel mainla
#>                      | | | | | | | | | | | | | | |                    
#> and his tricks (1)   3 3 2 2 1 4 5 3 3 3 4 5 2 5 5   (1) what forced o
#> ainland of man (2)   3 4 5 4 2 1 3 5 3 3 4 2 3 5 5   (2) blackmarket f
#>       the owes (3)   5 3 5 2 5 4 3 4 1 3 5 3 4 1 2   (3) as without bl
#>  man owes into (4)   5 4 4 1 4 1 5 2 3 4 4 3 2 2 3   (4) the I Brownst
#>        charter (5)   4 1 2 5 4 4 1 5 3 3 3 5 5 2 1   (5) to           
#> er blackmarket (6)   1 3 3 4 4 4 3 4 5 2 4 5 4 3 1   (6) what         
#>      by before (7)   2 1 1 1 2 1 3 3 1 2 1 5 4 3 3   (7) he then reali
#>  Mr. paying he (8)   4 1 2 3 4 1 5 3 4 5 4 3 4 3 4   (8) good by      
#> e as activity  (9)   4 2 3 3 3 4 2 4 1 3 3 2 4 4 2   (9) into back the
#>  Mr. the forc (10)   3 5 3 1 5 3 2 2 1 2 2 4 4 4 2   (10) beyond      
#>  essentially  (11)   1 3 3 1 2 5 4 5 2 1 3 3 5 5 1   (11) Johnson     
#>           who (12)   2 1 1 5 5 3 1 2 1 1 3 3 1 4 4   (12) by of as mai
#> rces any back (13)   5 2 1 5 4 5 1 1 2 1 4 3 2 3 2   (13) what without
#> ssentially he (14)   5 5 2 4 5 3 3 1 5 5 2 5 3 5 5   (14) essentially 
#> plane depicts (15)   1 3 4 4 2 4 1 4 4 2 2 2 5 3 4   (15) man owes he 
#>      happened (16)   2 2 2 4 1 3 3 5 3 2 2 1 3 1 5   (16) novel blackm
#>  mainland him (17)   4 2 2 3 1 5 4 4 1 3 1 5 5 4 2   (17) blackmarket 
#>       control (18)   4 2 4 2 3 4 1 5 4 4 2 4 2 5 3   (18) back money o
#>            is (19)   1 2 2 1 2 2 5 2 1 3 2 5 1 3 4   (19) realizes cus
#> ng initially  (20)   3 1 5 2 1 2 1 1 5 4 3 4 1 3 4   (20) by forces re
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - his realizes                   
#>            of charter economic - 7 | 9 - blackmarket his activity     
#>                           as - 6 | | | 10 - realizes blackmarket airpl
#> money has owes charter she - 5 | | | | | 11 - who fishing before man  
#> ed charter tricks Harry  - 4 | | | | | | | 12 - mainland Mark customer
#>                     of - 3 | | | | | | | | | 13 - money man money     
#>            paying is - 2 | | | | | | | | | | | 14 - Harry who the blac
#> icts money mainlan - 1 | | | | | | | | | | | | | 15 - Mr. novel mainla
#>                      | | | | | | | | | | | | | | |                    
#> and his tricks (1)   3 3 2 2 1 4 5 3 3 3 4 5 2 5 5   (1) what forced o
#> ainland of man (2)   3 4 5 4 2 1 3 5 3 3 4 2 3 5 5   (2) blackmarket f
#>       the owes (3)   5 3 5 2 5 4 3 4 1 3 5 3 4 1 2   (3) as without bl
#>  man owes into (4)   5 4 4 1 4 1 5 2 3 4 4 3 2 2 3   (4) the I Brownst
#>        charter (5)   4 1 2 5 4 4 1 5 3 3 3 5 5 2 1   (5) to           
#> er blackmarket (6)   1 3 3 4 4 4 3 4 5 2 4 5 4 3 1   (6) what         
#>      by before (7)   2 1 1 1 2 1 3 3 1 2 1 5 4 3 3   (7) he then reali
#>  Mr. paying he (8)   4 1 2 3 4 1 5 3 4 5 4 3 4 3 4   (8) good by      
#> e as activity  (9)   4 2 3 3 3 4 2 4 1 3 3 2 4 4 2   (9) into back the
#>  Mr. the forc (10)   3 5 3 1 5 3 2 2 1 2 2 4 4 4 2   (10) beyond      
#>  essentially  (11)   1 3 3 1 2 5 4 5 2 1 3 3 5 5 1   (11) Johnson     
#>           who (12)   2 1 1 5 5 3 1 2 1 1 3 3 1 4 4   (12) by of as mai
#> rces any back (13)   5 2 1 5 4 5 1 1 2 1 4 3 2 3 2   (13) what without
#> ssentially he (14)   5 5 2 4 5 3 3 1 5 5 2 5 3 5 5   (14) essentially 
#> plane depicts (15)   1 3 4 4 2 4 1 4 4 2 2 2 5 3 4   (15) man owes he 
#>      happened (16)   2 2 2 4 1 3 3 5 3 2 2 1 3 1 5   (16) novel blackm
#>  mainland him (17)   4 2 2 3 1 5 4 4 1 3 1 5 5 4 2   (17) blackmarket 
#>       control (18)   4 2 4 2 3 4 1 5 4 4 2 4 2 5 3   (18) back money o
#>            is (19)   1 2 2 1 2 2 5 2 1 3 2 5 1 3 4   (19) realizes cus
#> ng initially  (20)   3 1 5 2 1 2 1 1 5 4 3 4 1 3 4   (20) by forces re
#> nd his tricks (21)   3 3 2 2 1 4 5 3 3 3 4 5 2 5 5   (21) what forced 
#> inland of man (22)   3 4 5 4 2 1 3 5 3 3 4 2 3 5 5   (22) blackmarket 
#>      the owes (23)   5 3 5 2 5 4 3 4 1 3 5 3 4 1 2   (23) as without b
#> man owes into (24)   5 4 4 1 4 1 5 2 3 4 4 3 2 2 3   (24) the I Browns
#>       charter (25)   4 1 2 5 4 4 1 5 3 3 3 5 5 2 1   (25) to          
#> r blackmarket (26)   1 3 3 4 4 4 3 4 5 2 4 5 4 3 1   (26) what        
#>     by before (27)   2 1 1 1 2 1 3 3 1 2 1 5 4 3 3   (27) he then real
#> Mr. paying he (28)   4 1 2 3 4 1 5 3 4 5 4 3 4 3 4   (28) good by     
#>  as activity  (29)   4 2 3 3 3 4 2 4 1 3 3 2 4 4 2   (29) into back th
#>  Mr. the forc (30)   3 5 3 1 5 3 2 2 1 2 2 4 4 4 2   (30) beyond      
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - his realizes                   
#>            of charter economic - 7 | 9 - blackmarket his activity     
#>                           as - 6 | | | 10 - realizes blackmarket airpl
#> money has owes charter she - 5 | | | | | 11 - who fishing before man  
#> ed charter tricks Harry  - 4 | | | | | | | 12 - mainland Mark customer
#>                     of - 3 | | | | | | | | | 13 - money man money     
#>            paying is - 2 | | | | | | | | | | | 14 - Harry who the blac
#> icts money mainlan - 1 | | | | | | | | | | | | | 15 - Mr. novel mainla
#>                      | | | | | | | | | | | | | | |                    
#> and his tricks (1)   3 3 2 2 1 4 5 3 3 3 4 5 2 5 5   (1) what forced o
#> ainland of man (2)   3 4 5 4 2 1 3 5 3 3 4 2 3 5 5   (2) blackmarket f
#>       the owes (3)   5 3 5 2 5 4 3 4 1 3 5 3 4 1 2   (3) as without bl
#>  man owes into (4)   5 4 4 1 4 1 5 2 3 4 4 3 2 2 3   (4) the I Brownst
#>        charter (5)   4 1 2 5 4 4 1 5 3 3 3 5 5 2 1   (5) to           
#> er blackmarket (6)   1 3 3 4 4 4 3 4 5 2 4 5 4 3 1   (6) what         
#>      by before (7)   2 1 1 1 2 1 3 3 1 2 1 5 4 3 3   (7) he then reali
#>  Mr. paying he (8)   4 1 2 3 4 1 5 3 4 5 4 3 4 3 4   (8) good by      
#> e as activity  (9)   4 2 3 3 3 4 2 4 1 3 3 2 4 4 2   (9) into back the
#>  Mr. the forc (10)   3 5 3 1 5 3 2 2 1 2 2 4 4 4 2   (10) beyond      
#> nd his tricks (11)   3 3 2 2 1 4 5 3 3 3 4 5 2 5 5   (11) what forced 
#> inland of man (12)   3 4 5 4 2 1 3 5 3 3 4 2 3 5 5   (12) blackmarket 
#>      the owes (13)   5 3 5 2 5 4 3 4 1 3 5 3 4 1 2   (13) as without b
#> man owes into (14)   5 4 4 1 4 1 5 2 3 4 4 3 2 2 3   (14) the I Browns
#>       charter (15)   4 1 2 5 4 4 1 5 3 3 3 5 5 2 1   (15) to          
#> r blackmarket (16)   1 3 3 4 4 4 3 4 5 2 4 5 4 3 1   (16) what        
#>     by before (17)   2 1 1 1 2 1 3 3 1 2 1 5 4 3 3   (17) he then real
#> Mr. paying he (18)   4 1 2 3 4 1 5 3 4 5 4 3 4 3 4   (18) good by     
#>  as activity  (19)   4 2 3 3 3 4 2 4 1 3 3 2 4 4 2   (19) into back th
#>  Mr. the forc (20)   3 5 3 1 5 3 2 2 1 2 2 4 4 4 2   (20) beyond      
#>  essentially  (21)   1 3 3 1 2 5 4 5 2 1 3 3 5 5 1   (21) Johnson     
#>           who (22)   2 1 1 5 5 3 1 2 1 1 3 3 1 4 4   (22) by of as mai
#> rces any back (23)   5 2 1 5 4 5 1 1 2 1 4 3 2 3 2   (23) what without
#> ssentially he (24)   5 5 2 4 5 3 3 1 5 5 2 5 3 5 5   (24) essentially 
#> plane depicts (25)   1 3 4 4 2 4 1 4 4 2 2 2 5 3 4   (25) man owes he 
#>      happened (26)   2 2 2 4 1 3 3 5 3 2 2 1 3 1 5   (26) novel blackm
#>  mainland him (27)   4 2 2 3 1 5 4 4 1 3 1 5 5 4 2   (27) blackmarket 
#>       control (28)   4 2 4 2 3 4 1 5 4 4 2 4 2 5 3   (28) back money o
#>            is (29)   1 2 2 1 2 2 5 2 1 3 2 5 1 3 4   (29) realizes cus
#> ng initially  (30)   3 1 5 2 1 2 1 1 5 4 3 4 1 3 4   (30) by forces re
#>