Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an                             
#>               an airplane I is - 7 | 9 - away                         
#>                fishing I any - 6 | | | 10 - depicts the his control   
#>           has good Johnson - 5 | | | | | 11 - I                       
#>      she realizes beyond - 4 | | | | | | | 12 - blackmarket the he bey
#>   before activity what - 3 | | | | | | | | | 13 - beyond Mr. by has mo
#>        Mr. by tricks - 2 | | | | | | | | | | | 14 - the by of Brownsto
#>     as his without - 1 | | | | | | | | | | | | | 15 - man control forc
#>                      | | | | | | | | | | | | | | |                    
#>  control money (1)   2 4 2 4 4 3 4 3 4 1 3 1 3 2 1   (1) Harry of     
#> ntially econom (2)   4 3 5 3 5 2 5 3 5 5 4 1 4 4 4   (2) the essential
#> s away away re (3)   1 1 4 5 2 3 1 1 4 4 1 2 4 1 2   (3) by then black
#> d of Mr. novel (4)   2 4 5 5 4 5 3 4 4 4 3 4 1 2 3   (4) flees activit
#>  the Harry the (5)   5 2 2 3 4 5 2 5 3 4 5 3 2 1 1   (5) then he befor
#> pened of away  (6)   4 1 4 3 1 1 4 3 2 2 3 5 1 5 3   (6) flees before 
#> ho depicts she (7)   5 5 1 5 3 3 4 5 1 5 3 5 2 4 5   (7) airplane cont
#> ay what what a (8)   3 3 1 2 3 1 2 3 1 4 1 2 3 3 5   (8) forces by by 
#> ket initially  (9)   1 1 2 5 2 5 5 3 4 3 1 5 1 2 3   (9) I what he goo
#> ckmarket Mark (10)   4 4 2 5 2 1 2 3 5 1 4 4 5 5 2   (10) beyond as re
#> f slipping as (11)   3 4 5 3 5 4 2 4 2 2 1 4 3 4 2   (11) flees an int
#>        before (12)   1 2 4 2 2 5 2 5 4 2 1 4 2 4 5   (12) has         
#> into economic (13)   3 2 3 2 2 3 4 4 4 4 1 5 2 2 3   (13) flees       
#>         flees (14)   2 2 1 3 3 4 4 4 3 5 2 4 2 2 2   (14) forces initi
#> she initially (15)   1 5 2 1 4 5 4 4 1 4 2 1 5 1 1   (15) him essentia
#> mic blackmark (16)   2 2 5 2 3 3 5 2 3 3 3 2 2 5 3   (16) any essentia
#> as slipping p (17)   5 5 1 5 1 2 5 4 1 4 3 4 3 5 5   (17) control tric
#>         money (18)   3 5 3 4 2 5 5 5 4 3 5 2 2 3 2   (18) novel       
#> into economic (19)   5 5 2 4 2 2 3 5 5 1 5 5 2 4 2   (19) depicts cont
#> land happened (20)   3 3 3 3 4 1 5 1 4 2 1 1 2 4 5   (20) tricks slipp
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an                             
#>               an airplane I is - 7 | 9 - away                         
#>                fishing I any - 6 | | | 10 - depicts the his control   
#>           has good Johnson - 5 | | | | | 11 - I                       
#>      she realizes beyond - 4 | | | | | | | 12 - blackmarket the he bey
#>   before activity what - 3 | | | | | | | | | 13 - beyond Mr. by has mo
#>        Mr. by tricks - 2 | | | | | | | | | | | 14 - the by of Brownsto
#>     as his without - 1 | | | | | | | | | | | | | 15 - man control forc
#>                      | | | | | | | | | | | | | | |                    
#>  control money (1)   2 4 2 4 4 3 4 3 4 1 3 1 3 2 1   (1) Harry of     
#> ntially econom (2)   4 3 5 3 5 2 5 3 5 5 4 1 4 4 4   (2) the essential
#> s away away re (3)   1 1 4 5 2 3 1 1 4 4 1 2 4 1 2   (3) by then black
#> d of Mr. novel (4)   2 4 5 5 4 5 3 4 4 4 3 4 1 2 3   (4) flees activit
#>  the Harry the (5)   5 2 2 3 4 5 2 5 3 4 5 3 2 1 1   (5) then he befor
#> pened of away  (6)   4 1 4 3 1 1 4 3 2 2 3 5 1 5 3   (6) flees before 
#> ho depicts she (7)   5 5 1 5 3 3 4 5 1 5 3 5 2 4 5   (7) airplane cont
#> ay what what a (8)   3 3 1 2 3 1 2 3 1 4 1 2 3 3 5   (8) forces by by 
#> ket initially  (9)   1 1 2 5 2 5 5 3 4 3 1 5 1 2 3   (9) I what he goo
#> ckmarket Mark (10)   4 4 2 5 2 1 2 3 5 1 4 4 5 5 2   (10) beyond as re
#> f slipping as (11)   3 4 5 3 5 4 2 4 2 2 1 4 3 4 2   (11) flees an int
#>        before (12)   1 2 4 2 2 5 2 5 4 2 1 4 2 4 5   (12) has         
#> into economic (13)   3 2 3 2 2 3 4 4 4 4 1 5 2 2 3   (13) flees       
#>         flees (14)   2 2 1 3 3 4 4 4 3 5 2 4 2 2 2   (14) forces initi
#> she initially (15)   1 5 2 1 4 5 4 4 1 4 2 1 5 1 1   (15) him essentia
#> mic blackmark (16)   2 2 5 2 3 3 5 2 3 3 3 2 2 5 3   (16) any essentia
#> as slipping p (17)   5 5 1 5 1 2 5 4 1 4 3 4 3 5 5   (17) control tric
#>         money (18)   3 5 3 4 2 5 5 5 4 3 5 2 2 3 2   (18) novel       
#> into economic (19)   5 5 2 4 2 2 3 5 5 1 5 5 2 4 2   (19) depicts cont
#> land happened (20)   3 3 3 3 4 1 5 1 4 2 1 1 2 4 5   (20) tricks slipp
#> control money (21)   2 4 2 4 4 3 4 3 4 1 3 1 3 2 1   (21) Harry of    
#> tially econom (22)   4 3 5 3 5 2 5 3 5 5 4 1 4 4 4   (22) the essentia
#>  away away re (23)   1 1 4 5 2 3 1 1 4 4 1 2 4 1 2   (23) by then blac
#>  of Mr. novel (24)   2 4 5 5 4 5 3 4 4 4 3 4 1 2 3   (24) flees activi
#> the Harry the (25)   5 2 2 3 4 5 2 5 3 4 5 3 2 1 1   (25) then he befo
#> ened of away  (26)   4 1 4 3 1 1 4 3 2 2 3 5 1 5 3   (26) flees before
#> o depicts she (27)   5 5 1 5 3 3 4 5 1 5 3 5 2 4 5   (27) airplane con
#> y what what a (28)   3 3 1 2 3 1 2 3 1 4 1 2 3 3 5   (28) forces by by
#> et initially  (29)   1 1 2 5 2 5 5 3 4 3 1 5 1 2 3   (29) I what he go
#> ckmarket Mark (30)   4 4 2 5 2 1 2 3 5 1 4 4 5 5 2   (30) beyond as re
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - an                             
#>               an airplane I is - 7 | 9 - away                         
#>                fishing I any - 6 | | | 10 - depicts the his control   
#>           has good Johnson - 5 | | | | | 11 - I                       
#>      she realizes beyond - 4 | | | | | | | 12 - blackmarket the he bey
#>   before activity what - 3 | | | | | | | | | 13 - beyond Mr. by has mo
#>        Mr. by tricks - 2 | | | | | | | | | | | 14 - the by of Brownsto
#>     as his without - 1 | | | | | | | | | | | | | 15 - man control forc
#>                      | | | | | | | | | | | | | | |                    
#>  control money (1)   2 4 2 4 4 3 4 3 4 1 3 1 3 2 1   (1) Harry of     
#> ntially econom (2)   4 3 5 3 5 2 5 3 5 5 4 1 4 4 4   (2) the essential
#> s away away re (3)   1 1 4 5 2 3 1 1 4 4 1 2 4 1 2   (3) by then black
#> d of Mr. novel (4)   2 4 5 5 4 5 3 4 4 4 3 4 1 2 3   (4) flees activit
#>  the Harry the (5)   5 2 2 3 4 5 2 5 3 4 5 3 2 1 1   (5) then he befor
#> pened of away  (6)   4 1 4 3 1 1 4 3 2 2 3 5 1 5 3   (6) flees before 
#> ho depicts she (7)   5 5 1 5 3 3 4 5 1 5 3 5 2 4 5   (7) airplane cont
#> ay what what a (8)   3 3 1 2 3 1 2 3 1 4 1 2 3 3 5   (8) forces by by 
#> ket initially  (9)   1 1 2 5 2 5 5 3 4 3 1 5 1 2 3   (9) I what he goo
#> ckmarket Mark (10)   4 4 2 5 2 1 2 3 5 1 4 4 5 5 2   (10) beyond as re
#> control money (11)   2 4 2 4 4 3 4 3 4 1 3 1 3 2 1   (11) Harry of    
#> tially econom (12)   4 3 5 3 5 2 5 3 5 5 4 1 4 4 4   (12) the essentia
#>  away away re (13)   1 1 4 5 2 3 1 1 4 4 1 2 4 1 2   (13) by then blac
#>  of Mr. novel (14)   2 4 5 5 4 5 3 4 4 4 3 4 1 2 3   (14) flees activi
#> the Harry the (15)   5 2 2 3 4 5 2 5 3 4 5 3 2 1 1   (15) then he befo
#> ened of away  (16)   4 1 4 3 1 1 4 3 2 2 3 5 1 5 3   (16) flees before
#> o depicts she (17)   5 5 1 5 3 3 4 5 1 5 3 5 2 4 5   (17) airplane con
#> y what what a (18)   3 3 1 2 3 1 2 3 1 4 1 2 3 3 5   (18) forces by by
#> et initially  (19)   1 1 2 5 2 5 5 3 4 3 1 5 1 2 3   (19) I what he go
#> ckmarket Mark (20)   4 4 2 5 2 1 2 3 5 1 4 4 5 5 2   (20) beyond as re
#> f slipping as (21)   3 4 5 3 5 4 2 4 2 2 1 4 3 4 2   (21) flees an int
#>        before (22)   1 2 4 2 2 5 2 5 4 2 1 4 2 4 5   (22) has         
#> into economic (23)   3 2 3 2 2 3 4 4 4 4 1 5 2 2 3   (23) flees       
#>         flees (24)   2 2 1 3 3 4 4 4 3 5 2 4 2 2 2   (24) forces initi
#> she initially (25)   1 5 2 1 4 5 4 4 1 4 2 1 5 1 1   (25) him essentia
#> mic blackmark (26)   2 2 5 2 3 3 5 2 3 3 3 2 2 5 3   (26) any essentia
#> as slipping p (27)   5 5 1 5 1 2 5 4 1 4 3 4 3 5 5   (27) control tric
#>         money (28)   3 5 3 4 2 5 5 5 4 3 5 2 2 3 2   (28) novel       
#> into economic (29)   5 5 2 4 2 2 3 5 5 1 5 5 2 4 2   (29) depicts cont
#> land happened (30)   3 3 3 3 4 1 5 1 4 2 1 1 2 4 5   (30) tricks slipp
#>