Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the without                    
#>        as beyond novel has any - 7 | 9 - airplane                     
#>      beyond good customer of - 6 | | | 10 - as mainland               
#>                     paying - 5 | | | | | 11 - activity                
#> money before before Mark - 4 | | | | | | | 12 - of tricks Brownstone o
#> ng mainland without he - 3 | | | | | | | | | 13 - Brownstone          
#>    his forced beyond - 2 | | | | | | | | | | | 14 - the essentially fo
#> ity activity witho - 1 | | | | | | | | | | | | | 15 - he flees blackma
#>                      | | | | | | | | | | | | | | |                    
#> re essentially (1)   1 4 2 5 3 4 2 4 3 2 1 4 3 5 2   (1) happened forc
#> tivity as back (2)   4 5 4 3 5 3 2 1 1 3 2 5 2 4 2   (2) fishing his b
#>  Johnson Harry (3)   3 4 3 2 5 3 5 4 5 2 2 2 4 5 3   (3) has mainland 
#>  essentially o (4)   4 1 3 5 4 5 5 1 5 5 2 2 4 4 1   (4) who slipping 
#> appened fishin (5)   5 2 5 3 4 3 5 3 1 1 1 4 2 3 4   (5) by she novel 
#> n forced beyon (6)   1 2 5 5 2 4 1 3 5 1 3 2 3 2 2   (6) money mainlan
#> m forced his m (7)   1 5 4 1 4 4 5 5 4 4 3 4 2 2 1   (7) she money    
#>   Mark she she (8)   4 4 5 1 5 2 3 2 4 4 3 4 4 1 2   (8) is Mark happe
#>       back the (9)   2 3 1 4 1 1 3 3 5 4 5 2 3 4 2   (9) forced       
#>   fishing his (10)   3 5 4 5 1 5 1 5 2 4 2 4 3 4 5   (10) his realizes
#> owes happened (11)   1 2 4 3 2 2 3 4 2 2 2 3 2 2 3   (11) without payi
#> rced realizes (12)   3 2 3 2 4 4 4 4 1 5 1 1 3 3 4   (12) owes then hi
#> plane an Mark (13)   4 2 4 5 2 2 3 2 3 1 3 1 5 2 2   (13) he I forces 
#>      Mark who (14)   3 4 3 2 2 2 5 3 2 4 4 1 2 1 4   (14) without has 
#> mer beyond an (15)   1 3 1 2 2 2 2 1 5 2 2 5 2 3 5   (15) to the reali
#> the as him to (16)   1 4 3 5 1 2 5 2 2 2 1 2 1 4 2   (16) is owes Brow
#>           the (17)   5 2 5 5 3 3 5 3 4 1 2 1 2 2 4   (17) away economi
#>    Brownstone (18)   4 3 5 1 3 2 5 5 1 3 1 5 5 1 2   (18) forces his a
#> l to happened (19)   1 2 4 1 5 2 3 4 4 4 3 2 5 5 3   (19) I good as Mr
#> Brownstone in (20)   3 2 5 5 2 1 1 4 1 3 1 3 2 3 4   (20) before I Mr.
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the without                    
#>        as beyond novel has any - 7 | 9 - airplane                     
#>      beyond good customer of - 6 | | | 10 - as mainland               
#>                     paying - 5 | | | | | 11 - activity                
#> money before before Mark - 4 | | | | | | | 12 - of tricks Brownstone o
#> ng mainland without he - 3 | | | | | | | | | 13 - Brownstone          
#>    his forced beyond - 2 | | | | | | | | | | | 14 - the essentially fo
#> ity activity witho - 1 | | | | | | | | | | | | | 15 - he flees blackma
#>                      | | | | | | | | | | | | | | |                    
#> re essentially (1)   1 4 2 5 3 4 2 4 3 2 1 4 3 5 2   (1) happened forc
#> tivity as back (2)   4 5 4 3 5 3 2 1 1 3 2 5 2 4 2   (2) fishing his b
#>  Johnson Harry (3)   3 4 3 2 5 3 5 4 5 2 2 2 4 5 3   (3) has mainland 
#>  essentially o (4)   4 1 3 5 4 5 5 1 5 5 2 2 4 4 1   (4) who slipping 
#> appened fishin (5)   5 2 5 3 4 3 5 3 1 1 1 4 2 3 4   (5) by she novel 
#> n forced beyon (6)   1 2 5 5 2 4 1 3 5 1 3 2 3 2 2   (6) money mainlan
#> m forced his m (7)   1 5 4 1 4 4 5 5 4 4 3 4 2 2 1   (7) she money    
#>   Mark she she (8)   4 4 5 1 5 2 3 2 4 4 3 4 4 1 2   (8) is Mark happe
#>       back the (9)   2 3 1 4 1 1 3 3 5 4 5 2 3 4 2   (9) forced       
#>   fishing his (10)   3 5 4 5 1 5 1 5 2 4 2 4 3 4 5   (10) his realizes
#> owes happened (11)   1 2 4 3 2 2 3 4 2 2 2 3 2 2 3   (11) without payi
#> rced realizes (12)   3 2 3 2 4 4 4 4 1 5 1 1 3 3 4   (12) owes then hi
#> plane an Mark (13)   4 2 4 5 2 2 3 2 3 1 3 1 5 2 2   (13) he I forces 
#>      Mark who (14)   3 4 3 2 2 2 5 3 2 4 4 1 2 1 4   (14) without has 
#> mer beyond an (15)   1 3 1 2 2 2 2 1 5 2 2 5 2 3 5   (15) to the reali
#> the as him to (16)   1 4 3 5 1 2 5 2 2 2 1 2 1 4 2   (16) is owes Brow
#>           the (17)   5 2 5 5 3 3 5 3 4 1 2 1 2 2 4   (17) away economi
#>    Brownstone (18)   4 3 5 1 3 2 5 5 1 3 1 5 5 1 2   (18) forces his a
#> l to happened (19)   1 2 4 1 5 2 3 4 4 4 3 2 5 5 3   (19) I good as Mr
#> Brownstone in (20)   3 2 5 5 2 1 1 4 1 3 1 3 2 3 4   (20) before I Mr.
#> e essentially (21)   1 4 2 5 3 4 2 4 3 2 1 4 3 5 2   (21) happened for
#> ivity as back (22)   4 5 4 3 5 3 2 1 1 3 2 5 2 4 2   (22) fishing his 
#> Johnson Harry (23)   3 4 3 2 5 3 5 4 5 2 2 2 4 5 3   (23) has mainland
#> essentially o (24)   4 1 3 5 4 5 5 1 5 5 2 2 4 4 1   (24) who slipping
#> ppened fishin (25)   5 2 5 3 4 3 5 3 1 1 1 4 2 3 4   (25) by she novel
#>  forced beyon (26)   1 2 5 5 2 4 1 3 5 1 3 2 3 2 2   (26) money mainla
#>  forced his m (27)   1 5 4 1 4 4 5 5 4 4 3 4 2 2 1   (27) she money   
#>  Mark she she (28)   4 4 5 1 5 2 3 2 4 4 3 4 4 1 2   (28) is Mark happ
#>      back the (29)   2 3 1 4 1 1 3 3 5 4 5 2 3 4 2   (29) forced      
#>   fishing his (30)   3 5 4 5 1 5 1 5 2 4 2 4 3 4 5   (30) his realizes
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the without                    
#>        as beyond novel has any - 7 | 9 - airplane                     
#>      beyond good customer of - 6 | | | 10 - as mainland               
#>                     paying - 5 | | | | | 11 - activity                
#> money before before Mark - 4 | | | | | | | 12 - of tricks Brownstone o
#> ng mainland without he - 3 | | | | | | | | | 13 - Brownstone          
#>    his forced beyond - 2 | | | | | | | | | | | 14 - the essentially fo
#> ity activity witho - 1 | | | | | | | | | | | | | 15 - he flees blackma
#>                      | | | | | | | | | | | | | | |                    
#> re essentially (1)   1 4 2 5 3 4 2 4 3 2 1 4 3 5 2   (1) happened forc
#> tivity as back (2)   4 5 4 3 5 3 2 1 1 3 2 5 2 4 2   (2) fishing his b
#>  Johnson Harry (3)   3 4 3 2 5 3 5 4 5 2 2 2 4 5 3   (3) has mainland 
#>  essentially o (4)   4 1 3 5 4 5 5 1 5 5 2 2 4 4 1   (4) who slipping 
#> appened fishin (5)   5 2 5 3 4 3 5 3 1 1 1 4 2 3 4   (5) by she novel 
#> n forced beyon (6)   1 2 5 5 2 4 1 3 5 1 3 2 3 2 2   (6) money mainlan
#> m forced his m (7)   1 5 4 1 4 4 5 5 4 4 3 4 2 2 1   (7) she money    
#>   Mark she she (8)   4 4 5 1 5 2 3 2 4 4 3 4 4 1 2   (8) is Mark happe
#>       back the (9)   2 3 1 4 1 1 3 3 5 4 5 2 3 4 2   (9) forced       
#>   fishing his (10)   3 5 4 5 1 5 1 5 2 4 2 4 3 4 5   (10) his realizes
#> e essentially (11)   1 4 2 5 3 4 2 4 3 2 1 4 3 5 2   (11) happened for
#> ivity as back (12)   4 5 4 3 5 3 2 1 1 3 2 5 2 4 2   (12) fishing his 
#> Johnson Harry (13)   3 4 3 2 5 3 5 4 5 2 2 2 4 5 3   (13) has mainland
#> essentially o (14)   4 1 3 5 4 5 5 1 5 5 2 2 4 4 1   (14) who slipping
#> ppened fishin (15)   5 2 5 3 4 3 5 3 1 1 1 4 2 3 4   (15) by she novel
#>  forced beyon (16)   1 2 5 5 2 4 1 3 5 1 3 2 3 2 2   (16) money mainla
#>  forced his m (17)   1 5 4 1 4 4 5 5 4 4 3 4 2 2 1   (17) she money   
#>  Mark she she (18)   4 4 5 1 5 2 3 2 4 4 3 4 4 1 2   (18) is Mark happ
#>      back the (19)   2 3 1 4 1 1 3 3 5 4 5 2 3 4 2   (19) forced      
#>   fishing his (20)   3 5 4 5 1 5 1 5 2 4 2 4 3 4 5   (20) his realizes
#> owes happened (21)   1 2 4 3 2 2 3 4 2 2 2 3 2 2 3   (21) without payi
#> rced realizes (22)   3 2 3 2 4 4 4 4 1 5 1 1 3 3 4   (22) owes then hi
#> plane an Mark (23)   4 2 4 5 2 2 3 2 3 1 3 1 5 2 2   (23) he I forces 
#>      Mark who (24)   3 4 3 2 2 2 5 3 2 4 4 1 2 1 4   (24) without has 
#> mer beyond an (25)   1 3 1 2 2 2 2 1 5 2 2 5 2 3 5   (25) to the reali
#> the as him to (26)   1 4 3 5 1 2 5 2 2 2 1 2 1 4 2   (26) is owes Brow
#>           the (27)   5 2 5 5 3 3 5 3 4 1 2 1 2 2 4   (27) away economi
#>    Brownstone (28)   4 3 5 1 3 2 5 5 1 3 1 5 5 1 2   (28) forces his a
#> l to happened (29)   1 2 4 1 5 2 3 4 4 4 3 2 5 5 3   (29) I good as Mr
#> Brownstone in (30)   3 2 5 5 2 1 1 4 1 3 1 3 2 3 4   (30) before I Mr.
#>