Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - fishing by                     
#>    novel to owes man initially - 7 | 9 - mainland essentially has who 
#>             he his as the an - 6 | | | 10 - mainland                  
#>                     forces - 5 | | | | | 11 - he good economic he the 
#>                 the then - 4 | | | | | | | 12 - into                  
#>  slipping without John - 3 | | | | | | | | | 13 - money novel without 
#> cks Brownstone has h - 2 | | | | | | | | | | | 14 - good Mark         
#> e good blackmarket - 1 | | | | | | | | | | | | | 15 - the             
#>                      | | | | | | | | | | | | | | |                    
#> way paying she (1)   3 2 4 1 1 1 2 4 2 5 5 3 1 2 4   (1) fishing man a
#>         before (2)   2 5 1 3 2 5 2 4 3 4 5 2 1 2 2   (2) he man econom
#> ny the he back (3)   4 3 4 1 3 5 3 3 5 2 1 5 3 1 4   (3) owes realizes
#>             by (4)   5 2 5 2 2 3 5 2 2 3 1 2 1 5 5   (4) any happened 
#> rry tricks nov (5)   3 3 4 3 3 5 5 1 3 1 1 4 2 5 2   (5) paying Browns
#> es paying into (6)   4 2 4 4 2 3 3 3 4 1 5 2 1 4 1   (6) I by flees be
#> paying Johnson (7)   3 2 1 5 5 5 3 4 1 1 5 5 3 1 2   (7) money by beyo
#>  is back Mr. s (8)   5 5 5 2 3 4 1 4 5 4 2 1 2 4 1   (8) he an mainlan
#> irplane Johnso (9)   1 4 5 1 4 4 4 5 4 1 4 2 4 2 4   (9) has tricks   
#> ry then by ai (10)   5 3 3 3 2 2 1 4 4 4 5 1 3 5 3   (10) fishing     
#> d initially I (11)   2 3 2 3 5 3 4 5 4 4 2 4 3 1 5   (11) before      
#> lackmarket sh (12)   2 1 4 5 3 1 3 4 5 4 2 5 2 4 2   (12) he mainland 
#> control befor (13)   2 1 4 3 4 5 3 5 1 1 5 5 4 5 5   (13) man away rea
#>     initially (14)   3 4 4 3 4 4 1 5 2 3 5 3 1 1 4   (14) depicts the 
#> s away tricks (15)   3 1 2 5 4 3 4 2 3 2 1 3 2 3 3   (15) paying of aw
#>         by is (16)   2 3 4 4 5 3 4 3 4 5 1 2 3 1 1   (16) his blackmar
#> y as slipping (17)   3 2 2 3 5 2 2 5 1 4 5 5 1 3 2   (17) by his the b
#> itially initi (18)   2 3 5 2 1 4 1 1 5 3 3 1 5 4 5   (18) charter then
#>            an (19)   3 2 3 1 5 5 3 1 1 3 5 5 3 2 2   (19) the an what 
#> tone the good (20)   2 2 5 5 3 4 4 1 2 3 2 3 2 1 1   (20) tricks flees
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - fishing by                     
#>    novel to owes man initially - 7 | 9 - mainland essentially has who 
#>             he his as the an - 6 | | | 10 - mainland                  
#>                     forces - 5 | | | | | 11 - he good economic he the 
#>                 the then - 4 | | | | | | | 12 - into                  
#>  slipping without John - 3 | | | | | | | | | 13 - money novel without 
#> cks Brownstone has h - 2 | | | | | | | | | | | 14 - good Mark         
#> e good blackmarket - 1 | | | | | | | | | | | | | 15 - the             
#>                      | | | | | | | | | | | | | | |                    
#> way paying she (1)   3 2 4 1 1 1 2 4 2 5 5 3 1 2 4   (1) fishing man a
#>         before (2)   2 5 1 3 2 5 2 4 3 4 5 2 1 2 2   (2) he man econom
#> ny the he back (3)   4 3 4 1 3 5 3 3 5 2 1 5 3 1 4   (3) owes realizes
#>             by (4)   5 2 5 2 2 3 5 2 2 3 1 2 1 5 5   (4) any happened 
#> rry tricks nov (5)   3 3 4 3 3 5 5 1 3 1 1 4 2 5 2   (5) paying Browns
#> es paying into (6)   4 2 4 4 2 3 3 3 4 1 5 2 1 4 1   (6) I by flees be
#> paying Johnson (7)   3 2 1 5 5 5 3 4 1 1 5 5 3 1 2   (7) money by beyo
#>  is back Mr. s (8)   5 5 5 2 3 4 1 4 5 4 2 1 2 4 1   (8) he an mainlan
#> irplane Johnso (9)   1 4 5 1 4 4 4 5 4 1 4 2 4 2 4   (9) has tricks   
#> ry then by ai (10)   5 3 3 3 2 2 1 4 4 4 5 1 3 5 3   (10) fishing     
#> d initially I (11)   2 3 2 3 5 3 4 5 4 4 2 4 3 1 5   (11) before      
#> lackmarket sh (12)   2 1 4 5 3 1 3 4 5 4 2 5 2 4 2   (12) he mainland 
#> control befor (13)   2 1 4 3 4 5 3 5 1 1 5 5 4 5 5   (13) man away rea
#>     initially (14)   3 4 4 3 4 4 1 5 2 3 5 3 1 1 4   (14) depicts the 
#> s away tricks (15)   3 1 2 5 4 3 4 2 3 2 1 3 2 3 3   (15) paying of aw
#>         by is (16)   2 3 4 4 5 3 4 3 4 5 1 2 3 1 1   (16) his blackmar
#> y as slipping (17)   3 2 2 3 5 2 2 5 1 4 5 5 1 3 2   (17) by his the b
#> itially initi (18)   2 3 5 2 1 4 1 1 5 3 3 1 5 4 5   (18) charter then
#>            an (19)   3 2 3 1 5 5 3 1 1 3 5 5 3 2 2   (19) the an what 
#> tone the good (20)   2 2 5 5 3 4 4 1 2 3 2 3 2 1 1   (20) tricks flees
#> ay paying she (21)   3 2 4 1 1 1 2 4 2 5 5 3 1 2 4   (21) fishing man 
#>        before (22)   2 5 1 3 2 5 2 4 3 4 5 2 1 2 2   (22) he man econo
#> y the he back (23)   4 3 4 1 3 5 3 3 5 2 1 5 3 1 4   (23) owes realize
#>            by (24)   5 2 5 2 2 3 5 2 2 3 1 2 1 5 5   (24) any happened
#> ry tricks nov (25)   3 3 4 3 3 5 5 1 3 1 1 4 2 5 2   (25) paying Brown
#> s paying into (26)   4 2 4 4 2 3 3 3 4 1 5 2 1 4 1   (26) I by flees b
#> aying Johnson (27)   3 2 1 5 5 5 3 4 1 1 5 5 3 1 2   (27) money by bey
#> is back Mr. s (28)   5 5 5 2 3 4 1 4 5 4 2 1 2 4 1   (28) he an mainla
#> rplane Johnso (29)   1 4 5 1 4 4 4 5 4 1 4 2 4 2 4   (29) has tricks  
#> ry then by ai (30)   5 3 3 3 2 2 1 4 4 4 5 1 3 5 3   (30) fishing     
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - fishing by                     
#>    novel to owes man initially - 7 | 9 - mainland essentially has who 
#>             he his as the an - 6 | | | 10 - mainland                  
#>                     forces - 5 | | | | | 11 - he good economic he the 
#>                 the then - 4 | | | | | | | 12 - into                  
#>  slipping without John - 3 | | | | | | | | | 13 - money novel without 
#> cks Brownstone has h - 2 | | | | | | | | | | | 14 - good Mark         
#> e good blackmarket - 1 | | | | | | | | | | | | | 15 - the             
#>                      | | | | | | | | | | | | | | |                    
#> way paying she (1)   3 2 4 1 1 1 2 4 2 5 5 3 1 2 4   (1) fishing man a
#>         before (2)   2 5 1 3 2 5 2 4 3 4 5 2 1 2 2   (2) he man econom
#> ny the he back (3)   4 3 4 1 3 5 3 3 5 2 1 5 3 1 4   (3) owes realizes
#>             by (4)   5 2 5 2 2 3 5 2 2 3 1 2 1 5 5   (4) any happened 
#> rry tricks nov (5)   3 3 4 3 3 5 5 1 3 1 1 4 2 5 2   (5) paying Browns
#> es paying into (6)   4 2 4 4 2 3 3 3 4 1 5 2 1 4 1   (6) I by flees be
#> paying Johnson (7)   3 2 1 5 5 5 3 4 1 1 5 5 3 1 2   (7) money by beyo
#>  is back Mr. s (8)   5 5 5 2 3 4 1 4 5 4 2 1 2 4 1   (8) he an mainlan
#> irplane Johnso (9)   1 4 5 1 4 4 4 5 4 1 4 2 4 2 4   (9) has tricks   
#> ry then by ai (10)   5 3 3 3 2 2 1 4 4 4 5 1 3 5 3   (10) fishing     
#> ay paying she (11)   3 2 4 1 1 1 2 4 2 5 5 3 1 2 4   (11) fishing man 
#>        before (12)   2 5 1 3 2 5 2 4 3 4 5 2 1 2 2   (12) he man econo
#> y the he back (13)   4 3 4 1 3 5 3 3 5 2 1 5 3 1 4   (13) owes realize
#>            by (14)   5 2 5 2 2 3 5 2 2 3 1 2 1 5 5   (14) any happened
#> ry tricks nov (15)   3 3 4 3 3 5 5 1 3 1 1 4 2 5 2   (15) paying Brown
#> s paying into (16)   4 2 4 4 2 3 3 3 4 1 5 2 1 4 1   (16) I by flees b
#> aying Johnson (17)   3 2 1 5 5 5 3 4 1 1 5 5 3 1 2   (17) money by bey
#> is back Mr. s (18)   5 5 5 2 3 4 1 4 5 4 2 1 2 4 1   (18) he an mainla
#> rplane Johnso (19)   1 4 5 1 4 4 4 5 4 1 4 2 4 2 4   (19) has tricks  
#> ry then by ai (20)   5 3 3 3 2 2 1 4 4 4 5 1 3 5 3   (20) fishing     
#> d initially I (21)   2 3 2 3 5 3 4 5 4 4 2 4 3 1 5   (21) before      
#> lackmarket sh (22)   2 1 4 5 3 1 3 4 5 4 2 5 2 4 2   (22) he mainland 
#> control befor (23)   2 1 4 3 4 5 3 5 1 1 5 5 4 5 5   (23) man away rea
#>     initially (24)   3 4 4 3 4 4 1 5 2 3 5 3 1 1 4   (24) depicts the 
#> s away tricks (25)   3 1 2 5 4 3 4 2 3 2 1 3 2 3 3   (25) paying of aw
#>         by is (26)   2 3 4 4 5 3 4 3 4 5 1 2 3 1 1   (26) his blackmar
#> y as slipping (27)   3 2 2 3 5 2 2 5 1 4 5 5 1 3 2   (27) by his the b
#> itially initi (28)   2 3 5 2 1 4 1 1 5 3 3 1 5 4 5   (28) charter then
#>            an (29)   3 2 3 1 5 5 3 1 1 3 5 5 3 2 2   (29) the an what 
#> tone the good (30)   2 2 5 5 3 4 4 1 2 3 2 3 2 1 1   (30) tricks flees
#>