Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - back who good is any           
#>  realizes essentially slipping - 7 | 9 - control blackmarket his the  
#>                      away is - 6 | | | 10 - control                   
#>       charter what fishing - 5 | | | | | 11 - by he                   
#>    paying as essentially - 4 | | | | | | | 12 - essentially           
#> land Harry blackmarket - 3 | | | | | | | | | 13 - of customer         
#> kmarket what fishing - 2 | | | | | | | | | | | 14 - economic control b
#>      novel good as - 1 | | | | | | | | | | | | | 15 - airplane activit
#>                      | | | | | | | | | | | | | | |                    
#> omic his owes  (1)   2 5 3 5 1 5 2 2 5 5 1 3 1 1 2   (1) I            
#> im activity an (2)   3 2 4 5 5 2 5 5 3 1 2 4 2 1 3   (2) any as       
#>   the back the (3)   1 5 4 1 4 3 4 4 5 2 1 2 1 4 1   (3) man what into
#>   Mark Mark by (4)   4 4 5 4 2 4 2 1 3 2 2 3 2 1 5   (4) Brownstone   
#>       economic (5)   4 4 3 1 5 5 4 4 4 4 4 5 2 4 5   (5) to money into
#> omic Mark esse (6)   1 1 5 4 4 2 3 4 3 1 4 5 1 3 2   (6) an his econom
#> ne essentially (7)   3 3 4 1 4 4 1 3 2 2 3 3 3 1 4   (7) she          
#>          money (8)   5 5 2 4 3 5 3 5 1 2 3 3 4 5 5   (8) man fishing a
#> Johnson by bef (9)   4 4 3 4 4 1 5 1 2 2 4 5 5 5 5   (9) what she what
#> rry back the  (10)   1 5 4 3 1 5 2 3 2 5 1 3 3 4 3   (10) before      
#>  control beyo (11)   1 3 1 2 1 1 5 1 5 2 5 3 1 1 1   (11) Brownstone f
#>           him (12)   2 1 5 4 1 4 2 4 4 4 4 5 1 5 5   (12) owes        
#> fishing Johns (13)   4 1 2 5 5 1 2 2 2 4 1 2 1 2 4   (13) control Mark
#> ity any he th (14)   1 4 1 4 1 2 4 4 3 1 2 5 2 2 4   (14) his forced h
#>      money he (15)   2 1 3 1 2 2 1 4 3 4 1 3 4 4 1   (15) without with
#>          good (16)   5 2 3 5 2 3 2 4 2 2 2 2 1 2 1   (16) Johnson blac
#> ing initially (17)   1 3 4 1 4 3 5 4 1 2 4 3 5 5 1   (17) Johnson slip
#> d who realize (18)   3 3 1 1 4 4 3 1 1 2 2 4 2 3 1   (18) Mr. has Mr. 
#>  he initially (19)   2 3 1 4 3 2 1 3 4 2 5 3 4 4 4   (19) the by slipp
#> who novel Mr. (20)   3 5 5 1 2 1 1 4 5 3 3 2 5 1 2   (20) is his trick
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - back who good is any           
#>  realizes essentially slipping - 7 | 9 - control blackmarket his the  
#>                      away is - 6 | | | 10 - control                   
#>       charter what fishing - 5 | | | | | 11 - by he                   
#>    paying as essentially - 4 | | | | | | | 12 - essentially           
#> land Harry blackmarket - 3 | | | | | | | | | 13 - of customer         
#> kmarket what fishing - 2 | | | | | | | | | | | 14 - economic control b
#>      novel good as - 1 | | | | | | | | | | | | | 15 - airplane activit
#>                      | | | | | | | | | | | | | | |                    
#> omic his owes  (1)   2 5 3 5 1 5 2 2 5 5 1 3 1 1 2   (1) I            
#> im activity an (2)   3 2 4 5 5 2 5 5 3 1 2 4 2 1 3   (2) any as       
#>   the back the (3)   1 5 4 1 4 3 4 4 5 2 1 2 1 4 1   (3) man what into
#>   Mark Mark by (4)   4 4 5 4 2 4 2 1 3 2 2 3 2 1 5   (4) Brownstone   
#>       economic (5)   4 4 3 1 5 5 4 4 4 4 4 5 2 4 5   (5) to money into
#> omic Mark esse (6)   1 1 5 4 4 2 3 4 3 1 4 5 1 3 2   (6) an his econom
#> ne essentially (7)   3 3 4 1 4 4 1 3 2 2 3 3 3 1 4   (7) she          
#>          money (8)   5 5 2 4 3 5 3 5 1 2 3 3 4 5 5   (8) man fishing a
#> Johnson by bef (9)   4 4 3 4 4 1 5 1 2 2 4 5 5 5 5   (9) what she what
#> rry back the  (10)   1 5 4 3 1 5 2 3 2 5 1 3 3 4 3   (10) before      
#>  control beyo (11)   1 3 1 2 1 1 5 1 5 2 5 3 1 1 1   (11) Brownstone f
#>           him (12)   2 1 5 4 1 4 2 4 4 4 4 5 1 5 5   (12) owes        
#> fishing Johns (13)   4 1 2 5 5 1 2 2 2 4 1 2 1 2 4   (13) control Mark
#> ity any he th (14)   1 4 1 4 1 2 4 4 3 1 2 5 2 2 4   (14) his forced h
#>      money he (15)   2 1 3 1 2 2 1 4 3 4 1 3 4 4 1   (15) without with
#>          good (16)   5 2 3 5 2 3 2 4 2 2 2 2 1 2 1   (16) Johnson blac
#> ing initially (17)   1 3 4 1 4 3 5 4 1 2 4 3 5 5 1   (17) Johnson slip
#> d who realize (18)   3 3 1 1 4 4 3 1 1 2 2 4 2 3 1   (18) Mr. has Mr. 
#>  he initially (19)   2 3 1 4 3 2 1 3 4 2 5 3 4 4 4   (19) the by slipp
#> who novel Mr. (20)   3 5 5 1 2 1 1 4 5 3 3 2 5 1 2   (20) is his trick
#> mic his owes  (21)   2 5 3 5 1 5 2 2 5 5 1 3 1 1 2   (21) I           
#> m activity an (22)   3 2 4 5 5 2 5 5 3 1 2 4 2 1 3   (22) any as      
#>  the back the (23)   1 5 4 1 4 3 4 4 5 2 1 2 1 4 1   (23) man what int
#>  Mark Mark by (24)   4 4 5 4 2 4 2 1 3 2 2 3 2 1 5   (24) Brownstone  
#>      economic (25)   4 4 3 1 5 5 4 4 4 4 4 5 2 4 5   (25) to money int
#> mic Mark esse (26)   1 1 5 4 4 2 3 4 3 1 4 5 1 3 2   (26) an his econo
#> e essentially (27)   3 3 4 1 4 4 1 3 2 2 3 3 3 1 4   (27) she         
#>         money (28)   5 5 2 4 3 5 3 5 1 2 3 3 4 5 5   (28) man fishing 
#> ohnson by bef (29)   4 4 3 4 4 1 5 1 2 2 4 5 5 5 5   (29) what she wha
#> rry back the  (30)   1 5 4 3 1 5 2 3 2 5 1 3 3 4 3   (30) before      
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - back who good is any           
#>  realizes essentially slipping - 7 | 9 - control blackmarket his the  
#>                      away is - 6 | | | 10 - control                   
#>       charter what fishing - 5 | | | | | 11 - by he                   
#>    paying as essentially - 4 | | | | | | | 12 - essentially           
#> land Harry blackmarket - 3 | | | | | | | | | 13 - of customer         
#> kmarket what fishing - 2 | | | | | | | | | | | 14 - economic control b
#>      novel good as - 1 | | | | | | | | | | | | | 15 - airplane activit
#>                      | | | | | | | | | | | | | | |                    
#> omic his owes  (1)   2 5 3 5 1 5 2 2 5 5 1 3 1 1 2   (1) I            
#> im activity an (2)   3 2 4 5 5 2 5 5 3 1 2 4 2 1 3   (2) any as       
#>   the back the (3)   1 5 4 1 4 3 4 4 5 2 1 2 1 4 1   (3) man what into
#>   Mark Mark by (4)   4 4 5 4 2 4 2 1 3 2 2 3 2 1 5   (4) Brownstone   
#>       economic (5)   4 4 3 1 5 5 4 4 4 4 4 5 2 4 5   (5) to money into
#> omic Mark esse (6)   1 1 5 4 4 2 3 4 3 1 4 5 1 3 2   (6) an his econom
#> ne essentially (7)   3 3 4 1 4 4 1 3 2 2 3 3 3 1 4   (7) she          
#>          money (8)   5 5 2 4 3 5 3 5 1 2 3 3 4 5 5   (8) man fishing a
#> Johnson by bef (9)   4 4 3 4 4 1 5 1 2 2 4 5 5 5 5   (9) what she what
#> rry back the  (10)   1 5 4 3 1 5 2 3 2 5 1 3 3 4 3   (10) before      
#> mic his owes  (11)   2 5 3 5 1 5 2 2 5 5 1 3 1 1 2   (11) I           
#> m activity an (12)   3 2 4 5 5 2 5 5 3 1 2 4 2 1 3   (12) any as      
#>  the back the (13)   1 5 4 1 4 3 4 4 5 2 1 2 1 4 1   (13) man what int
#>  Mark Mark by (14)   4 4 5 4 2 4 2 1 3 2 2 3 2 1 5   (14) Brownstone  
#>      economic (15)   4 4 3 1 5 5 4 4 4 4 4 5 2 4 5   (15) to money int
#> mic Mark esse (16)   1 1 5 4 4 2 3 4 3 1 4 5 1 3 2   (16) an his econo
#> e essentially (17)   3 3 4 1 4 4 1 3 2 2 3 3 3 1 4   (17) she         
#>         money (18)   5 5 2 4 3 5 3 5 1 2 3 3 4 5 5   (18) man fishing 
#> ohnson by bef (19)   4 4 3 4 4 1 5 1 2 2 4 5 5 5 5   (19) what she wha
#> rry back the  (20)   1 5 4 3 1 5 2 3 2 5 1 3 3 4 3   (20) before      
#>  control beyo (21)   1 3 1 2 1 1 5 1 5 2 5 3 1 1 1   (21) Brownstone f
#>           him (22)   2 1 5 4 1 4 2 4 4 4 4 5 1 5 5   (22) owes        
#> fishing Johns (23)   4 1 2 5 5 1 2 2 2 4 1 2 1 2 4   (23) control Mark
#> ity any he th (24)   1 4 1 4 1 2 4 4 3 1 2 5 2 2 4   (24) his forced h
#>      money he (25)   2 1 3 1 2 2 1 4 3 4 1 3 4 4 1   (25) without with
#>          good (26)   5 2 3 5 2 3 2 4 2 2 2 2 1 2 1   (26) Johnson blac
#> ing initially (27)   1 3 4 1 4 3 5 4 1 2 4 3 5 5 1   (27) Johnson slip
#> d who realize (28)   3 3 1 1 4 4 3 1 1 2 2 4 2 3 1   (28) Mr. has Mr. 
#>  he initially (29)   2 3 1 4 3 2 1 3 4 2 5 3 4 4 4   (29) the by slipp
#> who novel Mr. (30)   3 5 5 1 2 1 1 4 5 3 3 2 5 1 2   (30) is his trick
#>