Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - as                             
#>               fishing good she - 7 | 9 - then                         
#> ssentially activity the by I - 6 | | | 10 - who                       
#>                   has away - 5 | | | | | 11 - forces                  
#>                novel Mr. - 4 | | | | | | | 12 - by tricks             
#>                  flees - 3 | | | | | | | | | 13 - owes                
#>                 then - 2 | | | | | | | | | | | 14 - him depicts       
#> hat blackmarket to - 1 | | | | | | | | | | | | | 15 - activity by who 
#>                      | | | | | | | | | | | | | | |                    
#> man owes I him (1)   1 1 5 5 3 4 4 1 2 2 3 2 3 1 2   (1) she          
#>            Mr. (2)   4 1 3 2 5 5 1 2 4 5 1 5 1 2 3   (2) him control p
#> t he back befo (3)   2 1 2 5 3 2 3 2 3 1 4 3 5 5 3   (3) by his the co
#> ces I slipping (4)   5 3 2 5 1 2 2 1 1 1 4 1 4 2 1   (4) Brownstone wh
#> Johnson paying (5)   1 3 2 1 5 1 2 1 5 3 1 2 2 5 1   (5) Harry good   
#> owes Harry wha (6)   5 3 5 1 5 2 5 3 3 5 2 5 4 1 3   (6) Johnson      
#> ontrol to back (7)   1 3 5 1 5 5 4 4 2 5 5 4 4 4 3   (7) before depict
#>          novel (8)   2 4 1 2 2 1 4 3 4 3 5 5 5 5 1   (8) mainland then
#> ially happened (9)   4 5 4 4 1 2 1 5 3 4 3 2 1 3 5   (9) novel initial
#> nd blackmarke (10)   3 2 1 4 5 3 4 2 3 5 3 1 4 2 1   (10) Mark blackma
#>          what (11)   1 5 5 5 5 3 4 3 5 3 2 1 4 2 3   (11) Brownstone a
#> the paying he (12)   3 3 5 3 3 3 2 1 3 4 3 3 5 4 2   (12) essentially 
#>  without Mark (13)   4 1 3 5 3 3 2 4 1 2 1 3 5 3 4   (13) forces then 
#> ey to control (14)   4 5 3 5 1 5 5 4 5 4 5 2 5 5 4   (14) Mark by back
#> ack by into t (15)   1 2 2 4 1 5 4 3 2 5 5 4 4 2 5   (15) an beyond I 
#>   activity by (16)   4 3 5 4 1 4 5 3 3 4 1 1 2 2 2   (16) back his hap
#> plane essenti (17)   4 5 2 1 3 1 1 5 5 1 2 2 5 3 4   (17) blackmarket 
#> ying Johnson  (18)   1 4 4 5 5 2 2 1 4 1 5 3 5 5 5   (18) she         
#> at economic t (19)   2 2 1 2 5 5 4 3 2 5 3 3 5 1 3   (19) as airplane 
#> rced economic (20)   2 4 5 1 1 1 2 1 4 3 3 3 2 2 3   (20) mainland rea
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - as                             
#>               fishing good she - 7 | 9 - then                         
#> ssentially activity the by I - 6 | | | 10 - who                       
#>                   has away - 5 | | | | | 11 - forces                  
#>                novel Mr. - 4 | | | | | | | 12 - by tricks             
#>                  flees - 3 | | | | | | | | | 13 - owes                
#>                 then - 2 | | | | | | | | | | | 14 - him depicts       
#> hat blackmarket to - 1 | | | | | | | | | | | | | 15 - activity by who 
#>                      | | | | | | | | | | | | | | |                    
#> man owes I him (1)   1 1 5 5 3 4 4 1 2 2 3 2 3 1 2   (1) she          
#>            Mr. (2)   4 1 3 2 5 5 1 2 4 5 1 5 1 2 3   (2) him control p
#> t he back befo (3)   2 1 2 5 3 2 3 2 3 1 4 3 5 5 3   (3) by his the co
#> ces I slipping (4)   5 3 2 5 1 2 2 1 1 1 4 1 4 2 1   (4) Brownstone wh
#> Johnson paying (5)   1 3 2 1 5 1 2 1 5 3 1 2 2 5 1   (5) Harry good   
#> owes Harry wha (6)   5 3 5 1 5 2 5 3 3 5 2 5 4 1 3   (6) Johnson      
#> ontrol to back (7)   1 3 5 1 5 5 4 4 2 5 5 4 4 4 3   (7) before depict
#>          novel (8)   2 4 1 2 2 1 4 3 4 3 5 5 5 5 1   (8) mainland then
#> ially happened (9)   4 5 4 4 1 2 1 5 3 4 3 2 1 3 5   (9) novel initial
#> nd blackmarke (10)   3 2 1 4 5 3 4 2 3 5 3 1 4 2 1   (10) Mark blackma
#>          what (11)   1 5 5 5 5 3 4 3 5 3 2 1 4 2 3   (11) Brownstone a
#> the paying he (12)   3 3 5 3 3 3 2 1 3 4 3 3 5 4 2   (12) essentially 
#>  without Mark (13)   4 1 3 5 3 3 2 4 1 2 1 3 5 3 4   (13) forces then 
#> ey to control (14)   4 5 3 5 1 5 5 4 5 4 5 2 5 5 4   (14) Mark by back
#> ack by into t (15)   1 2 2 4 1 5 4 3 2 5 5 4 4 2 5   (15) an beyond I 
#>   activity by (16)   4 3 5 4 1 4 5 3 3 4 1 1 2 2 2   (16) back his hap
#> plane essenti (17)   4 5 2 1 3 1 1 5 5 1 2 2 5 3 4   (17) blackmarket 
#> ying Johnson  (18)   1 4 4 5 5 2 2 1 4 1 5 3 5 5 5   (18) she         
#> at economic t (19)   2 2 1 2 5 5 4 3 2 5 3 3 5 1 3   (19) as airplane 
#> rced economic (20)   2 4 5 1 1 1 2 1 4 3 3 3 2 2 3   (20) mainland rea
#> an owes I him (21)   1 1 5 5 3 4 4 1 2 2 3 2 3 1 2   (21) she         
#>           Mr. (22)   4 1 3 2 5 5 1 2 4 5 1 5 1 2 3   (22) him control 
#>  he back befo (23)   2 1 2 5 3 2 3 2 3 1 4 3 5 5 3   (23) by his the c
#> es I slipping (24)   5 3 2 5 1 2 2 1 1 1 4 1 4 2 1   (24) Brownstone w
#> ohnson paying (25)   1 3 2 1 5 1 2 1 5 3 1 2 2 5 1   (25) Harry good  
#> wes Harry wha (26)   5 3 5 1 5 2 5 3 3 5 2 5 4 1 3   (26) Johnson     
#> ntrol to back (27)   1 3 5 1 5 5 4 4 2 5 5 4 4 4 3   (27) before depic
#>         novel (28)   2 4 1 2 2 1 4 3 4 3 5 5 5 5 1   (28) mainland the
#> ally happened (29)   4 5 4 4 1 2 1 5 3 4 3 2 1 3 5   (29) novel initia
#> nd blackmarke (30)   3 2 1 4 5 3 4 2 3 5 3 1 4 2 1   (30) Mark blackma
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - as                             
#>               fishing good she - 7 | 9 - then                         
#> ssentially activity the by I - 6 | | | 10 - who                       
#>                   has away - 5 | | | | | 11 - forces                  
#>                novel Mr. - 4 | | | | | | | 12 - by tricks             
#>                  flees - 3 | | | | | | | | | 13 - owes                
#>                 then - 2 | | | | | | | | | | | 14 - him depicts       
#> hat blackmarket to - 1 | | | | | | | | | | | | | 15 - activity by who 
#>                      | | | | | | | | | | | | | | |                    
#> man owes I him (1)   1 1 5 5 3 4 4 1 2 2 3 2 3 1 2   (1) she          
#>            Mr. (2)   4 1 3 2 5 5 1 2 4 5 1 5 1 2 3   (2) him control p
#> t he back befo (3)   2 1 2 5 3 2 3 2 3 1 4 3 5 5 3   (3) by his the co
#> ces I slipping (4)   5 3 2 5 1 2 2 1 1 1 4 1 4 2 1   (4) Brownstone wh
#> Johnson paying (5)   1 3 2 1 5 1 2 1 5 3 1 2 2 5 1   (5) Harry good   
#> owes Harry wha (6)   5 3 5 1 5 2 5 3 3 5 2 5 4 1 3   (6) Johnson      
#> ontrol to back (7)   1 3 5 1 5 5 4 4 2 5 5 4 4 4 3   (7) before depict
#>          novel (8)   2 4 1 2 2 1 4 3 4 3 5 5 5 5 1   (8) mainland then
#> ially happened (9)   4 5 4 4 1 2 1 5 3 4 3 2 1 3 5   (9) novel initial
#> nd blackmarke (10)   3 2 1 4 5 3 4 2 3 5 3 1 4 2 1   (10) Mark blackma
#> an owes I him (11)   1 1 5 5 3 4 4 1 2 2 3 2 3 1 2   (11) she         
#>           Mr. (12)   4 1 3 2 5 5 1 2 4 5 1 5 1 2 3   (12) him control 
#>  he back befo (13)   2 1 2 5 3 2 3 2 3 1 4 3 5 5 3   (13) by his the c
#> es I slipping (14)   5 3 2 5 1 2 2 1 1 1 4 1 4 2 1   (14) Brownstone w
#> ohnson paying (15)   1 3 2 1 5 1 2 1 5 3 1 2 2 5 1   (15) Harry good  
#> wes Harry wha (16)   5 3 5 1 5 2 5 3 3 5 2 5 4 1 3   (16) Johnson     
#> ntrol to back (17)   1 3 5 1 5 5 4 4 2 5 5 4 4 4 3   (17) before depic
#>         novel (18)   2 4 1 2 2 1 4 3 4 3 5 5 5 5 1   (18) mainland the
#> ally happened (19)   4 5 4 4 1 2 1 5 3 4 3 2 1 3 5   (19) novel initia
#> nd blackmarke (20)   3 2 1 4 5 3 4 2 3 5 3 1 4 2 1   (20) Mark blackma
#>          what (21)   1 5 5 5 5 3 4 3 5 3 2 1 4 2 3   (21) Brownstone a
#> the paying he (22)   3 3 5 3 3 3 2 1 3 4 3 3 5 4 2   (22) essentially 
#>  without Mark (23)   4 1 3 5 3 3 2 4 1 2 1 3 5 3 4   (23) forces then 
#> ey to control (24)   4 5 3 5 1 5 5 4 5 4 5 2 5 5 4   (24) Mark by back
#> ack by into t (25)   1 2 2 4 1 5 4 3 2 5 5 4 4 2 5   (25) an beyond I 
#>   activity by (26)   4 3 5 4 1 4 5 3 3 4 1 1 2 2 2   (26) back his hap
#> plane essenti (27)   4 5 2 1 3 1 1 5 5 1 2 2 5 3 4   (27) blackmarket 
#> ying Johnson  (28)   1 4 4 5 5 2 2 1 4 1 5 3 5 5 5   (28) she         
#> at economic t (29)   2 2 1 2 5 5 4 3 2 5 3 3 5 1 3   (29) as airplane 
#> rced economic (30)   2 4 5 1 1 1 2 1 4 3 3 3 2 2 3   (30) mainland rea
#>