Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Johnson                        
#>                             is - 7 | 9 - by blackmarket blackmarket th
#>                the initially - 6 | | | 10 - activity control          
#>               what by what - 5 | | | | | 11 - man forces              
#>         mainland by good - 4 | | | | | | | 12 - airplane he           
#>                  Harry - 3 | | | | | | | | | 13 - any                 
#>      customer beyond - 2 | | | | | | | | | | | 14 - Harry             
#> ry an airplane Mr. - 1 | | | | | | | | | | | | | 15 - I owes          
#>                      | | | | | | | | | | | | | | |                    
#>  forces good b (1)   3 3 4 2 3 3 4 2 4 5 4 1 3 2 4   (1) who airplane 
#>  flees charter (2)   3 4 3 3 2 2 2 2 1 1 2 2 1 4 5   (2) fishing initi
#> I man has away (3)   5 5 1 5 2 2 5 2 2 5 1 3 3 4 4   (3) Johnson initi
#>        without (4)   4 4 5 2 4 4 2 3 1 3 5 4 5 1 1   (4) back an I owe
#> nstone has his (5)   5 3 4 2 4 5 5 5 3 5 4 5 4 1 2   (5) novel        
#> zes airplane w (6)   4 3 3 1 5 5 1 2 1 1 4 2 2 5 2   (6) flees then by
#> ntially forces (7)   4 3 3 2 1 3 2 3 3 1 2 1 5 5 5   (7) any Mark is t
#>         forced (8)   5 1 4 1 3 5 5 5 3 3 1 1 4 1 4   (8) she good cont
#>         has he (9)   1 2 5 2 3 4 2 4 2 5 3 2 4 1 3   (9) away         
#> away by money (10)   4 3 2 1 4 3 1 2 1 4 3 4 1 5 3   (10) fishing depi
#>        beyond (11)   2 3 3 5 5 4 5 2 1 4 4 2 5 3 3   (11) what without
#> harter realiz (12)   1 2 2 1 5 2 5 3 3 2 1 3 3 2 1   (12) the flees th
#> ohnson fishin (13)   5 5 5 2 4 4 3 2 1 3 4 4 5 5 5   (13) any essentia
#> Mark him befo (14)   2 2 1 3 4 1 2 3 4 3 2 4 2 1 2   (14) Mr.         
#>      mainland (15)   2 1 2 4 3 1 3 4 2 4 1 5 3 2 4   (15) the         
#> k blackmarket (16)   3 2 1 5 5 5 4 5 3 5 3 1 1 1 5   (16) man by who i
#> harter before (17)   1 4 5 3 5 1 1 4 3 3 5 1 4 5 2   (17) the as airpl
#> y Mark away w (18)   1 5 3 5 4 1 3 5 3 1 2 4 3 3 1   (18) he Mr. of   
#>          good (19)   1 1 4 1 4 1 4 4 1 1 4 2 4 4 3   (19) the by witho
#> ntrol fishing (20)   2 4 5 1 2 3 2 1 5 4 4 4 5 5 3   (20) then who cus
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Johnson                        
#>                             is - 7 | 9 - by blackmarket blackmarket th
#>                the initially - 6 | | | 10 - activity control          
#>               what by what - 5 | | | | | 11 - man forces              
#>         mainland by good - 4 | | | | | | | 12 - airplane he           
#>                  Harry - 3 | | | | | | | | | 13 - any                 
#>      customer beyond - 2 | | | | | | | | | | | 14 - Harry             
#> ry an airplane Mr. - 1 | | | | | | | | | | | | | 15 - I owes          
#>                      | | | | | | | | | | | | | | |                    
#>  forces good b (1)   3 3 4 2 3 3 4 2 4 5 4 1 3 2 4   (1) who airplane 
#>  flees charter (2)   3 4 3 3 2 2 2 2 1 1 2 2 1 4 5   (2) fishing initi
#> I man has away (3)   5 5 1 5 2 2 5 2 2 5 1 3 3 4 4   (3) Johnson initi
#>        without (4)   4 4 5 2 4 4 2 3 1 3 5 4 5 1 1   (4) back an I owe
#> nstone has his (5)   5 3 4 2 4 5 5 5 3 5 4 5 4 1 2   (5) novel        
#> zes airplane w (6)   4 3 3 1 5 5 1 2 1 1 4 2 2 5 2   (6) flees then by
#> ntially forces (7)   4 3 3 2 1 3 2 3 3 1 2 1 5 5 5   (7) any Mark is t
#>         forced (8)   5 1 4 1 3 5 5 5 3 3 1 1 4 1 4   (8) she good cont
#>         has he (9)   1 2 5 2 3 4 2 4 2 5 3 2 4 1 3   (9) away         
#> away by money (10)   4 3 2 1 4 3 1 2 1 4 3 4 1 5 3   (10) fishing depi
#>        beyond (11)   2 3 3 5 5 4 5 2 1 4 4 2 5 3 3   (11) what without
#> harter realiz (12)   1 2 2 1 5 2 5 3 3 2 1 3 3 2 1   (12) the flees th
#> ohnson fishin (13)   5 5 5 2 4 4 3 2 1 3 4 4 5 5 5   (13) any essentia
#> Mark him befo (14)   2 2 1 3 4 1 2 3 4 3 2 4 2 1 2   (14) Mr.         
#>      mainland (15)   2 1 2 4 3 1 3 4 2 4 1 5 3 2 4   (15) the         
#> k blackmarket (16)   3 2 1 5 5 5 4 5 3 5 3 1 1 1 5   (16) man by who i
#> harter before (17)   1 4 5 3 5 1 1 4 3 3 5 1 4 5 2   (17) the as airpl
#> y Mark away w (18)   1 5 3 5 4 1 3 5 3 1 2 4 3 3 1   (18) he Mr. of   
#>          good (19)   1 1 4 1 4 1 4 4 1 1 4 2 4 4 3   (19) the by witho
#> ntrol fishing (20)   2 4 5 1 2 3 2 1 5 4 4 4 5 5 3   (20) then who cus
#> forces good b (21)   3 3 4 2 3 3 4 2 4 5 4 1 3 2 4   (21) who airplane
#> flees charter (22)   3 4 3 3 2 2 2 2 1 1 2 2 1 4 5   (22) fishing init
#>  man has away (23)   5 5 1 5 2 2 5 2 2 5 1 3 3 4 4   (23) Johnson init
#>       without (24)   4 4 5 2 4 4 2 3 1 3 5 4 5 1 1   (24) back an I ow
#> stone has his (25)   5 3 4 2 4 5 5 5 3 5 4 5 4 1 2   (25) novel       
#> es airplane w (26)   4 3 3 1 5 5 1 2 1 1 4 2 2 5 2   (26) flees then b
#> tially forces (27)   4 3 3 2 1 3 2 3 3 1 2 1 5 5 5   (27) any Mark is 
#>        forced (28)   5 1 4 1 3 5 5 5 3 3 1 1 4 1 4   (28) she good con
#>        has he (29)   1 2 5 2 3 4 2 4 2 5 3 2 4 1 3   (29) away        
#> away by money (30)   4 3 2 1 4 3 1 2 1 4 3 4 1 5 3   (30) fishing depi
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - Johnson                        
#>                             is - 7 | 9 - by blackmarket blackmarket th
#>                the initially - 6 | | | 10 - activity control          
#>               what by what - 5 | | | | | 11 - man forces              
#>         mainland by good - 4 | | | | | | | 12 - airplane he           
#>                  Harry - 3 | | | | | | | | | 13 - any                 
#>      customer beyond - 2 | | | | | | | | | | | 14 - Harry             
#> ry an airplane Mr. - 1 | | | | | | | | | | | | | 15 - I owes          
#>                      | | | | | | | | | | | | | | |                    
#>  forces good b (1)   3 3 4 2 3 3 4 2 4 5 4 1 3 2 4   (1) who airplane 
#>  flees charter (2)   3 4 3 3 2 2 2 2 1 1 2 2 1 4 5   (2) fishing initi
#> I man has away (3)   5 5 1 5 2 2 5 2 2 5 1 3 3 4 4   (3) Johnson initi
#>        without (4)   4 4 5 2 4 4 2 3 1 3 5 4 5 1 1   (4) back an I owe
#> nstone has his (5)   5 3 4 2 4 5 5 5 3 5 4 5 4 1 2   (5) novel        
#> zes airplane w (6)   4 3 3 1 5 5 1 2 1 1 4 2 2 5 2   (6) flees then by
#> ntially forces (7)   4 3 3 2 1 3 2 3 3 1 2 1 5 5 5   (7) any Mark is t
#>         forced (8)   5 1 4 1 3 5 5 5 3 3 1 1 4 1 4   (8) she good cont
#>         has he (9)   1 2 5 2 3 4 2 4 2 5 3 2 4 1 3   (9) away         
#> away by money (10)   4 3 2 1 4 3 1 2 1 4 3 4 1 5 3   (10) fishing depi
#> forces good b (11)   3 3 4 2 3 3 4 2 4 5 4 1 3 2 4   (11) who airplane
#> flees charter (12)   3 4 3 3 2 2 2 2 1 1 2 2 1 4 5   (12) fishing init
#>  man has away (13)   5 5 1 5 2 2 5 2 2 5 1 3 3 4 4   (13) Johnson init
#>       without (14)   4 4 5 2 4 4 2 3 1 3 5 4 5 1 1   (14) back an I ow
#> stone has his (15)   5 3 4 2 4 5 5 5 3 5 4 5 4 1 2   (15) novel       
#> es airplane w (16)   4 3 3 1 5 5 1 2 1 1 4 2 2 5 2   (16) flees then b
#> tially forces (17)   4 3 3 2 1 3 2 3 3 1 2 1 5 5 5   (17) any Mark is 
#>        forced (18)   5 1 4 1 3 5 5 5 3 3 1 1 4 1 4   (18) she good con
#>        has he (19)   1 2 5 2 3 4 2 4 2 5 3 2 4 1 3   (19) away        
#> away by money (20)   4 3 2 1 4 3 1 2 1 4 3 4 1 5 3   (20) fishing depi
#>        beyond (21)   2 3 3 5 5 4 5 2 1 4 4 2 5 3 3   (21) what without
#> harter realiz (22)   1 2 2 1 5 2 5 3 3 2 1 3 3 2 1   (22) the flees th
#> ohnson fishin (23)   5 5 5 2 4 4 3 2 1 3 4 4 5 5 5   (23) any essentia
#> Mark him befo (24)   2 2 1 3 4 1 2 3 4 3 2 4 2 1 2   (24) Mr.         
#>      mainland (25)   2 1 2 4 3 1 3 4 2 4 1 5 3 2 4   (25) the         
#> k blackmarket (26)   3 2 1 5 5 5 4 5 3 5 3 1 1 1 5   (26) man by who i
#> harter before (27)   1 4 5 3 5 1 1 4 3 3 5 1 4 5 2   (27) the as airpl
#> y Mark away w (28)   1 5 3 5 4 1 3 5 3 1 2 4 3 3 1   (28) he Mr. of   
#>          good (29)   1 1 4 1 4 1 4 4 1 1 4 2 4 4 3   (29) the by witho
#> ntrol fishing (30)   2 4 5 1 2 3 2 1 5 4 4 4 5 5 3   (30) then who cus
#>