Concatenate the constructs of two or more grids.

I.e. the constructs are combined to form one long grid. The girds must have the same set of elements and an identical scale range. The order of the elements may differ.

Usage

bindConstructs(..., index = FALSE)

Arguments

...

One or more repgrid objects or a list containing repgrid object.

index

TODO. Logical (default TRUE). Whether to add an index at the end of each construct name so it remains clear from which grid each construct came.

Value

repgrid object with concatenated constructs.

Details

This function can be used in order to analyze multiple grids as one 'big grid' (eg. Slater, 1977, chap. 11).

References

Slater, P. (1977). The measurement of intrapersonal space by grid technique. London: Wiley.

Examples

a <- randomGrid()
b <- randomGrid()
elements(b) <- rev(elements(a)) # reverse elements
bindConstructs(a, b)
#> 
#> META DATA:
#> Number of constructs:  20 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the charter to by of           
#>   the an without airplane back - 7 | 9 - good away                    
#> onomic before has initially  - 6 | | | 10 - slipping then beyond      
#>                    without - 5 | | | | | 11 - Johnson Johnson the     
#> nitially who slipping he - 4 | | | | | | | 12 - is customer paying nov
#> lly slipping he an man - 3 | | | | | | | | | 13 - him depicts the Harr
#> rter initially Brown - 2 | | | | | | | | | | | 14 - the money an initi
#>  Harry the without - 1 | | | | | | | | | | | | | 15 - beyond happened 
#>                      | | | | | | | | | | | | | | |                    
#> alizes blackma (1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1) airplane econ
#>             as (2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2) control any t
#> orced mainland (3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3) essentially m
#> away away into (4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4) Mr. forced mo
#> as what is awa (5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5) him he       
#>            any (6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6) realizes     
#> ssentially his (7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7) mainland then
#> ishing without (8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8) before activi
#> e owes she the (9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9) I tricks who 
#> er he paying  (10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10) depicts any 
#>     then back (11)   4 5 2 4 2 5 5 1 1 2 3 2 1 5 5   (11) his beyond  
#> lees mainland (12)   5 5 4 2 1 4 4 3 2 2 4 3 2 2 2   (12) man him happ
#> f is the esse (13)   4 4 2 4 1 3 3 4 5 4 4 4 5 5 5   (13) the has by i
#>  what has the (14)   1 1 4 3 4 3 5 2 5 3 2 2 5 4 1   (14) depicts as t
#>  activity Joh (15)   2 3 3 5 3 2 1 1 1 2 5 3 5 2 5   (15) back man bef
#> e control man (16)   3 5 4 3 2 3 1 2 1 3 5 5 5 5 4   (16) initially   
#>      customer (17)   4 4 1 2 2 5 4 1 1 3 4 1 3 1 1   (17) an economic 
#>         Harry (18)   3 3 1 3 4 4 4 1 2 3 4 3 1 3 4   (18) mainland Mar
#>            is (19)   5 1 2 3 3 1 3 1 1 1 2 1 3 4 2   (19) I by mainlan
#> ack initially (20)   4 3 1 4 2 4 4 2 1 2 3 1 4 5 3   (20) Mr.         
#> 
bindConstructs(a, b, a)
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the charter to by of           
#>   the an without airplane back - 7 | 9 - good away                    
#> onomic before has initially  - 6 | | | 10 - slipping then beyond      
#>                    without - 5 | | | | | 11 - Johnson Johnson the     
#> nitially who slipping he - 4 | | | | | | | 12 - is customer paying nov
#> lly slipping he an man - 3 | | | | | | | | | 13 - him depicts the Harr
#> rter initially Brown - 2 | | | | | | | | | | | 14 - the money an initi
#>  Harry the without - 1 | | | | | | | | | | | | | 15 - beyond happened 
#>                      | | | | | | | | | | | | | | |                    
#> alizes blackma (1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1) airplane econ
#>             as (2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2) control any t
#> orced mainland (3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3) essentially m
#> away away into (4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4) Mr. forced mo
#> as what is awa (5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5) him he       
#>            any (6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6) realizes     
#> ssentially his (7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7) mainland then
#> ishing without (8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8) before activi
#> e owes she the (9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9) I tricks who 
#> er he paying  (10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10) depicts any 
#>     then back (11)   4 5 2 4 2 5 5 1 1 2 3 2 1 5 5   (11) his beyond  
#> lees mainland (12)   5 5 4 2 1 4 4 3 2 2 4 3 2 2 2   (12) man him happ
#> f is the esse (13)   4 4 2 4 1 3 3 4 5 4 4 4 5 5 5   (13) the has by i
#>  what has the (14)   1 1 4 3 4 3 5 2 5 3 2 2 5 4 1   (14) depicts as t
#>  activity Joh (15)   2 3 3 5 3 2 1 1 1 2 5 3 5 2 5   (15) back man bef
#> e control man (16)   3 5 4 3 2 3 1 2 1 3 5 5 5 5 4   (16) initially   
#>      customer (17)   4 4 1 2 2 5 4 1 1 3 4 1 3 1 1   (17) an economic 
#>         Harry (18)   3 3 1 3 4 4 4 1 2 3 4 3 1 3 4   (18) mainland Mar
#>            is (19)   5 1 2 3 3 1 3 1 1 1 2 1 3 4 2   (19) I by mainlan
#> ack initially (20)   4 3 1 4 2 4 4 2 1 2 3 1 4 5 3   (20) Mr.         
#> lizes blackma (21)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (21) airplane eco
#>            as (22)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (22) control any 
#> rced mainland (23)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (23) essentially 
#> way away into (24)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (24) Mr. forced m
#> s what is awa (25)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (25) him he      
#>           any (26)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (26) realizes    
#> sentially his (27)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (27) mainland the
#> shing without (28)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (28) before activ
#>  owes she the (29)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (29) I tricks who
#> er he paying  (30)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (30) depicts any 
#> 

# using lists of repgrid objects
bindConstructs(a, list(a, b))
#> 
#> META DATA:
#> Number of constructs:  30 
#> Number of elements:  15 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 5 (right pole)
#> 
#> RATINGS:
#>                                    8 - the charter to by of           
#>   the an without airplane back - 7 | 9 - good away                    
#> onomic before has initially  - 6 | | | 10 - slipping then beyond      
#>                    without - 5 | | | | | 11 - Johnson Johnson the     
#> nitially who slipping he - 4 | | | | | | | 12 - is customer paying nov
#> lly slipping he an man - 3 | | | | | | | | | 13 - him depicts the Harr
#> rter initially Brown - 2 | | | | | | | | | | | 14 - the money an initi
#>  Harry the without - 1 | | | | | | | | | | | | | 15 - beyond happened 
#>                      | | | | | | | | | | | | | | |                    
#> alizes blackma (1)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (1) airplane econ
#>             as (2)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (2) control any t
#> orced mainland (3)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (3) essentially m
#> away away into (4)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (4) Mr. forced mo
#> as what is awa (5)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (5) him he       
#>            any (6)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (6) realizes     
#> ssentially his (7)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (7) mainland then
#> ishing without (8)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (8) before activi
#> e owes she the (9)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (9) I tricks who 
#> er he paying  (10)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (10) depicts any 
#> lizes blackma (11)   2 1 5 3 2 5 2 4 4 3 4 1 1 5 4   (11) airplane eco
#>            as (12)   2 1 1 2 1 3 1 3 5 1 5 1 4 3 3   (12) control any 
#> rced mainland (13)   4 3 4 4 3 4 2 2 2 4 1 5 2 2 1   (13) essentially 
#> way away into (14)   1 3 1 4 3 4 3 1 4 1 1 5 2 3 4   (14) Mr. forced m
#> s what is awa (15)   4 2 2 2 2 2 2 5 3 4 2 5 5 3 1   (15) him he      
#>           any (16)   3 2 4 4 3 2 5 2 5 3 4 1 5 1 5   (16) realizes    
#> sentially his (17)   2 5 3 2 3 1 1 1 2 2 3 5 2 2 3   (17) mainland the
#> shing without (18)   2 4 5 4 4 5 4 4 4 3 5 5 3 4 4   (18) before activ
#>  owes she the (19)   4 1 5 2 5 4 3 3 2 2 3 2 1 5 5   (19) I tricks who
#> er he paying  (20)   4 4 1 5 4 1 1 1 4 2 3 2 5 2 1   (20) depicts any 
#>     then back (21)   4 5 2 4 2 5 5 1 1 2 3 2 1 5 5   (21) his beyond  
#> lees mainland (22)   5 5 4 2 1 4 4 3 2 2 4 3 2 2 2   (22) man him happ
#> f is the esse (23)   4 4 2 4 1 3 3 4 5 4 4 4 5 5 5   (23) the has by i
#>  what has the (24)   1 1 4 3 4 3 5 2 5 3 2 2 5 4 1   (24) depicts as t
#>  activity Joh (25)   2 3 3 5 3 2 1 1 1 2 5 3 5 2 5   (25) back man bef
#> e control man (26)   3 5 4 3 2 3 1 2 1 3 5 5 5 5 4   (26) initially   
#>      customer (27)   4 4 1 2 2 5 4 1 1 3 4 1 3 1 1   (27) an economic 
#>         Harry (28)   3 3 1 3 4 4 4 1 2 3 4 3 1 3 4   (28) mainland Mar
#>            is (29)   5 1 2 3 3 1 3 1 1 1 2 1 3 4 2   (29) I by mainlan
#> ack initially (30)   4 3 1 4 2 4 4 2 1 2 3 1 4 5 3   (30) Mr.         
#>