biplotDraw is the workhorse doing the drawing of a 2D biplot.

When the number of elements and constructs differs to a large extent, the absolute values of the coordinates for either constructs or elements will be much smaller or greater. This is an inherent property of the biplot. In the case it is not necessary to be able to read off the original data entries from the plot, the axes for elements and constructs can be scaled separately. The proportional projection values will stay unaffected. the absolute will change though. For grid interpretation the absolute values are usually oh no importance. Thus, there is an optional argument normalize which is FALSE as a default which rescales the axes so the longest construct and element vector will be set to the length of 1.

Usage

biplotDraw(
  x,
  inner.positioning = TRUE,
  outer.positioning = TRUE,
  c.labels.inside = F,
  flipaxes = c(F, F),
  strokes.x = 0.1,
  strokes.y = 0.1,
  offsetting = TRUE,
  offset.labels = 0,
  offset.e = 1,
  axis.ext = 0.1,
  mai = c(0.2, 1.5, 0.2, 1.5),
  rect.margins = c(0.01, 0.01),
  srt = 45,
  cex.pos = 0.7,
  xpd = TRUE,
  c.lines = TRUE,
  col.c.lines = grey(0.9),
  zoom = 1,
  ...
)

Arguments

x

repgrid object.

inner.positioning

Logical. Whether to calculate positions to minimize overplotting of elements and construct labels (default isTRUE). Note that the positioning may slow down the plotting.

outer.positioning

Logical. Whether to calculate positions to minimize overplotting of of construct labels on the outer borders (default isTRUE). Note that the positioning may slow down the plotting.

c.labels.inside

Logical. Whether to print construct labels next to the points. Can be useful during inspection of the plot (default FALSE).

flipaxes

Logical vector of length two. Whether x and y axes are reversed (default is c(F,F)).

strokes.x

Length of outer strokes in x direction in NDC.

strokes.y

Length of outer strokes in y direction in NDC.

offsetting

Do offsetting? (TODO)

offset.labels

Offsetting parameter for labels (TODO).

offset.e

offsetting parameter for elements (TODO).

axis.ext

Axis extension factor (default is .1). A bigger value will zoom out the plot.

mai

Margins available for plotting the labels in inch (default is c(.2, 1.5, .2, 1.5)).

rect.margins

Vector of length two (default is c(.07, .07)). Two values specifying the additional horizontal and vertical margin around each label.

srt

Angle to rotate construct label text. Only used in case offsetting=FALSE.

cex.pos

Cex parameter used during positioning of labels if prompted. Does usually not have to be changed by user.

xpd

Logical (default is TRUE). Whether to extend text labels over figure region. Usually not needed by the user.

c.lines

Logical. Whether construct lines from the center of the biplot to the surrounding box are drawn (default is FALSE).

col.c.lines

The color of the construct lines from the center to the borders of the plot (default is gray(.9)).

zoom

Scaling factor for all vectors. Can be used to zoom the plot in and out (default 1).

...

Not evaluated.

Value

Invisible return of dataframe used during construction of plot (useful for developers).