Grid data from Bell and McGorry (1992).

The grid data set is used in Bell's technical report "Using SPSS to Analyse Repertory Grid Data" (1997, p. 6). Originally, the data comes from a study by Bell and McGorry (1992).

References

Bell, R. C. (1977). Using SPSS to Analyse Repertory Grid Data. Technical Report, University of Melbourne.

Bell, R. C., & McGorry, P. (1992). The analysis of repertory grids used to monitor the perceptions of recovering psychotic patients. In A. Thomson & P. Cummins (Eds.), European Perspectives in Personal Construct Psychology (p. 137-150). Lincoln, UK: European Personal Construct Association.

Examples

bellmcgorry1992
#> 
#> META DATA:
#> Number of constructs:  14 
#> Number of elements:  14 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from 1 (left pole) to 7 (right pole)
#> 
#> RATINGS:
#>           person with diabetes - 7 8 - person with cancer           
#>                 AIDS patient - 6 | | 9 - person under stress        
#>             average person - 5 | | | | 10 - myself as I usually am  
#>       convicted criminal - 4 | | | | | | 11 - myself as I am now    
#>    psychiatric patient - 3 | | | | | | | | 12 - myself as I will be 
#> n with schizophrenia - 2 | | | | | | | | | | 13 - myself as the staf
#> manic depressive i - 1 | | | | | | | | | | | | 14 - my ideal self   
#>                      | | | | | | | | | | | | | |                    
#>           good (1)   3 5 3 7 4 2 1 1 7 1 1 1 2 1   (1)              
#>     dependable (2)   4 5 4 7 3 3 1 2 7 2 1 1 3 1   (2)              
#>           safe (3)   6 7 4 7 3 7 1 2 6 1 1 1 2 1   (3)              
#>    clearheaded (4)   7 7 4 7 2 7 2 4 7 1 1 1 2 1   (4)              
#>         stable (5)   7 7 5 7 2 7 1 4 7 2 2 1 3 1   (5)              
#>    predictable (6)   7 7 5 7 6 5 4 4 7 7 7 3 7 7   (6)              
#>    intelligent (7)   4 5 4 7 2 1 2 2 7 1 1 1 2 1   (7)              
#>           free (8)   7 7 5 7 1 7 1 1 1 1 7 1 7 1   (8)              
#>        healthy (9)   6 7 4 7 1 7 7 7 7 1 4 1 4 1   (9)              
#>        honest (10)   5 5 4 7 1 1 3 4 4 1 1 1 3 1   (10)             
#>      rational (11)   5 7 5 6 1 2 2 4 5 1 1 1 3 1   (11)             
#>   independent (12)   5 6 5 7 1 5 3 4 5 1 1 1 2 1   (12)             
#>          calm (13)   6 5 5 7 2 7 3 4 7 1 1 1 1 1   (13)             
#>    understood (14)   6 7 6 7 1 7 3 4 7 3 7 1 7 1   (14)             
#>