Grid data from Raeithel (1998).

Grid data to demonstrate the use of Bertin diagrams (Raeithel, 1998, p. 223). The context of its administration is unknown.

References

Raeithel, A. (1998). Kooperative Modellproduktion von Professionellen und Klienten. Erlaeutert am Beispiel des Repertory Grid. In A. Raeithel (1998). Selbstorganisation, Kooperation, Zeichenprozess. Arbeiten zu einer kulturwissenschaftlichen, anwendungsbezogenen Psychologie (p. 209-254). Opladen: Westdeutscher Verlag.

Examples

raeithel
#> 
#> META DATA:
#> Number of constructs:  13 
#> Number of elements:  13 
#> 
#> SCALE INFO:
#> The grid is rated on a scale from -1 (left pole) to 1 (right pole)
#> 
#> RATINGS:
#>                                        7 - Mutter                              
#>                             Virus - 6  |  8 - Schwester                        
#>                     Ich vorher - 5  |  |  |  9 - Partner                       
#>                        Arzt - 4  |  |  |  |  |  10 - Vater                     
#>                    Ideal - 3  |  |  |  |  |  |  |  11 - Neg. Person            
#>             Therapeut - 2  |  |  |  |  |  |  |  |  |  12 - Ich                 
#>             Freund - 1  |  |  |  |  |  |  |  |  |  |  |  13 - Freundin         
#>                      |  |  |  |  |  |  |  |  |  |  |  |  |                     
#>   charakterlos (1)    1  1  1  1  1 -1 -1 -1  1 -1 -1  1  1   (1) Charakter hab
#> uninteressiert (2)    1  1  1  1  1  1 -1 -1  1  1 -1  1  1   (2) interessiert 
#> h gehen lassen (3)    1  1  1  1  1  1 -1 -1  1 -1 -1  1  1   (3) zielstrebig  
#>     unerfahren (4)    1  1  1  1  1  1 -1 -1 -1  0 -1  1  1   (4) erfahren     
#>   abschaetzbar (5)   -1  1  1  1  1  1  1 -1 -1 -1 -1  1  0   (5) unberechenbar
#>          mutig (6)   -1 -1 -1 -1  1  1  1  1 -1  1  1  1 -1   (6) feige        
#>  leichtfuessig (7)   -1 -1 -1 -1 -1 -1  1  1 -1  1  0  1 -1   (7) schwerfaellig
#>          kuehl (8)    1  1  1  0 -1 -1 -1  1  1  1  1  0 -1   (8) warmherzig   
#> h eigennuetzig (9)    1  1  1  1 -1 -1 -1  0  1  1 -1 -1  0   (9) lebensbejahen
#>     angepasst (10)    1 -1  1  0  1  1 -1 -1 -1 -1 -1 -1 -1   (10) unangepasst 
#>         offen (11)   -1 -1 -1 -1  1  1  1 -1 -1 -1  1 -1 -1   (11) hinterhaelti
#>  unvorsichtig (12)   -1 -1  1  1 -1  1  0  1 -1  1 -1  1 -1   (12) disziplinier
#> undschaftlich (13)    0 -1 -1  1 -1  0  0 -1  0  1  0  0 -1   (13) vaeterlich a
#>