Calculate the correlations between elements.

Note that simple element correlations as a measure of similarity are flawed as they are not invariant to construct reflection (Mackay, 1992; Bell, 2010). A correlation index invariant to construct reflection is Cohen's rc measure (1969), which can be calculated using the argument rc=TRUE which is the default option.

Usage

elementCor(x, rc = TRUE, method = "pearson", trim = 20, index = TRUE)

Arguments

x

repgrid object.

rc

Use Cohen's rc which is invariant to construct reflection (see description above). It is used as the default.

method

A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall" or "spearman", can be abbreviated. The default is "pearson".

trim

The number of characters a construct is trimmed to (default is 20). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs with long names.

index

Whether to print the number of the construct.

Value

matrix of element correlations

References

Bell, R. C. (2010). A note on aligning constructs. Personal Construct Theory & Practice, (7), 42-48.

Cohen, J. (1969). rc: A profile similarity coefficient invariant over variable reflection. Psychological Bulletin, 71(4), 281-284.

Mackay, N. (1992). Identification, Reflection, and Correlation: Problems In The Bases Of Repertory Grid Measures. International Journal of Personal Construct Psychology, 5(1), 57-75.

See also

Examples

elementCor(mackay1992) # Cohen's rc
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                         1     2     3     4     5     6
#> (1) Self            1      0.59  0.28  0.29 -0.61 -0.62
#> (2) Ideal self      2           -0.04 -0.20 -0.38 -0.81
#> (3) Mother          3                  0.63  0.37 -0.43
#> (4) Father          4                        0.00  0.00
#> (5) Spouse          5                              0.00
#> (6) Disliked person 6                                  
elementCor(mackay1992, rc = FALSE) # PM correlation
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  pearson 
#> Note: Standard correlations are not invariant to scale reflection.
#> 
#>                         1     2     3     4     5     6
#> (1) Self            1     -0.37 -0.19  0.32 -0.75  0.96
#> (2) Ideal self      2           -0.71 -0.56 -0.26 -0.30
#> (3) Mother          3                  0.63  0.57 -0.12
#> (4) Father          4                        0.05  0.38
#> (5) Spouse          5                             -0.74
#> (6) Disliked person 6                                  
elementCor(mackay1992, rc = FALSE, method = "spearman") # Spearman correlation
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  spearman 
#> Note: Standard correlations are not invariant to scale reflection.
#> 
#>                         1     2     3     4     5     6
#> (1) Self            1     -0.25  0.01  0.64 -0.77  0.98
#> (2) Ideal self      2           -0.70 -0.57 -0.29 -0.19
#> (3) Mother          3                  0.55  0.45  0.06
#> (4) Father          4                       -0.14  0.65
#> (5) Spouse          5                             -0.74
#> (6) Disliked person 6                                  

# format output
elementCor(mackay1992, trim = 6)
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                1     2     3     4     5     6
#> (1) Self   1      0.59  0.28  0.29 -0.61 -0.62
#> (2) Ideal  2           -0.04 -0.20 -0.38 -0.81
#> (3) Mother 3                  0.63  0.37 -0.43
#> (4) Father 4                        0.00  0.00
#> (5) Spouse 5                              0.00
#> (6) Dislik 6                                  
elementCor(mackay1992, index = FALSE, trim = 6)
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>            1     2     3     4     5     6
#> Self   1      0.59  0.28  0.29 -0.61 -0.62
#> Ideal  2           -0.04 -0.20 -0.38 -0.81
#> Mother 3                  0.63  0.37 -0.43
#> Father 4                        0.00  0.00
#> Spouse 5                              0.00
#> Dislik 6                                  

# save as object for further processing
r <- elementCor(mackay1992)
r
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                         1     2     3     4     5     6
#> (1) Self            1      0.59  0.28  0.29 -0.61 -0.62
#> (2) Ideal self      2           -0.04 -0.20 -0.38 -0.81
#> (3) Mother          3                  0.63  0.37 -0.43
#> (4) Father          4                        0.00  0.00
#> (5) Spouse          5                              0.00
#> (6) Disliked person 6                                  

# change output of object
print(r, digits = 5)
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                            1        2        3        4        5        6
#> (1) Self            1         0.58761  0.28090  0.29096 -0.61237 -0.61828
#> (2) Ideal self      2                 -0.03745 -0.20365 -0.38100 -0.81461
#> (3) Mother          3                           0.63049  0.36860 -0.43346
#> (4) Father          4                                    0.00000  0.00000
#> (5) Spouse          5                                             0.00000
#> (6) Disliked person 6                                                    
print(r, col.index = FALSE)
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                     (1) Self (2) Ideal self (3) Mother (4) Father (5) Spouse
#> (1) Self                               0.59       0.28       0.29      -0.61
#> (2) Ideal self                                   -0.04      -0.20      -0.38
#> (3) Mother                                                   0.63       0.37
#> (4) Father                                                              0.00
#> (5) Spouse                                                                  
#> (6) Disliked person                                                         
#>                     (6) Disliked person
#> (1) Self                          -0.62
#> (2) Ideal self                    -0.81
#> (3) Mother                        -0.43
#> (4) Father                         0.00
#> (5) Spouse                         0.00
#> (6) Disliked person                    
print(r, upper = FALSE)
#> 
#> ############################
#> Correlation between elements
#> ############################
#> 
#> Type of correlation:  Cohens's rc (invariant to scale reflection) 
#> 
#>                           1     2     3     4     5     6
#> (1) Self            1  1.00  0.59  0.28  0.29 -0.61 -0.62
#> (2) Ideal self      2  0.59  1.00 -0.04 -0.20 -0.38 -0.81
#> (3) Mother          3  0.28 -0.04  1.00  0.63  0.37 -0.43
#> (4) Father          4  0.29 -0.20  0.63  1.00  0.00  0.00
#> (5) Spouse          5 -0.61 -0.38  0.37  0.00  1.00  0.00
#> (6) Disliked person 6 -0.62 -0.81 -0.43  0.00  0.00  1.00

# accessing elements of the correlation matrix
r[1, 3]
#> [1] 0.2809003