Intensity index

Calculate intensity index.

Usage

indexIntensity(x, rc = FALSE, trim = 30)

Arguments

x

A repgrid object.

rc

Whether to use Cohen's rc for the calculation of inter-element correlations. See elementCor() for further explanations of this measure.

trim

The number of characters a construct is trimmed to (default is 30). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs or elements with long names.

Value

An object of class indexIntensity containing a list with the following elements:

c.int: Intensity scores by construct. e.int: Intensity scores by element. c.int.mean: Average intensity score for constructs. e.int.mean: Average intensity score for elements. total.int: Total intensity score.

Details

The Intensity index has been suggested by Bannister (1960) as a measure of the amount of construct linkage. Bannister suggested that the score reflects the degree of organization of the construct system under investigation (Bannister & Mair, 1968). The index resulted from his and his colleagues work on construction systems of patient suffering schizophrenic thought disorder. The concept of intensity has a theoretical connection to the notion of "tight" and "loose" construing as proposed by Kelly (1991). While tight constructs lead to unvarying prediction, loose constructs allow for varying predictions. Bannister hypothesized that schizophrenic thought disorder is liked to a process of extremely loose construing leading to a loss of predictive power of the subject's construct system. The Intensity score as a structural measure is thought to reflect this type of system disintegration (Bannister, 1960).

Implementation as in the Gridcor program and explained on the correspoding help pages: "... the sum of the squared values of the correlations of each construct with the rest of the constructs, averaged by the total number of constructs minus one. This process is repeated with each element, and the overall Intensity is calculated by averaging the intensity scores of constructs and elements." (Gridcor manual). Currently the total is calculated as the unweighted average of all single scores (for elements and construct).

Development

TODO: Results have not been tested against other programs' results.

References

Bannister, D. (1960). Conceptual structure in thought-disordered schizophrenics. The Journal of mental science, 106, 1230-49.

Examples

indexIntensity(bell2010)
#> 
#> ################
#> Intensity index
#> ################
#> 
#> Total intensity: 0.19 
#> 
#> 
#> Average intensity of constructs: 0.16 
#> 
#> Itensity by construct:
#>                                     intensity
#> 1 relaxed - worried & tense              0.18
#> 2 not so smart (a - smart (academic      0.14
#> 3 dislikes sports - loves sports         0.05
#> 4 not interactive - loves people         0.15
#> 5 not transparent - transparent          0.13
#> 6 insensitive - sensitive                0.16
#> 7 fearful&timid - fearless               0.25
#> 8 rough - gentle                         0.25
#> 9 accept as it is - loves to argue       0.12
#> 
#> 
#> Average intensity of elements: 0.14 
#> 
#> Itensity by element:
#>                                       intensity
#> 1 (1) self                                 0.16
#> 2 (2) closest friend of the same sex       0.16
#> 3 (3) the unhappiest person you know       0.04
#> 4 (4) A person of the opposite sex t       0.14
#> 5 (5) A teacher you respected              0.19
#> 6 (6) Mother (or the person who fill       0.10
#> 7 (7) A person of the opposite sex t       0.13
#> 8 (8) The most confident person you        0.22
#> 9 (9) A person you work well with (a       0.14
#> 10 (10) A teacher you did not respect      0.09
indexIntensity(bell2010, trim = NA)
#> 
#> ################
#> Intensity index
#> ################
#> 
#> Total intensity: 0.19 
#> 
#> 
#> Average intensity of constructs: 0.16 
#> 
#> Itensity by construct:
#>                                                      intensity
#> 1 relaxed - worried & tense                               0.18
#> 2 not so smart (academically) - smart (academically)      0.14
#> 3 dislikes sports - loves sports                          0.05
#> 4 not interactive - loves people                          0.15
#> 5 not transparent - transparent                           0.13
#> 6 insensitive - sensitive                                 0.16
#> 7 fearful&timid - fearless                                0.25
#> 8 rough - gentle                                          0.25
#> 9 accept as it is - loves to argue                        0.12
#> 
#> 
#> Average intensity of elements: 0.14 
#> 
#> Itensity by element:
#>                                                                        intensity
#> 1 (1) self                                                                  0.16
#> 2 (2) closest friend of the same sex                                        0.16
#> 3 (3) the unhappiest person you know                                        0.04
#> 4 (4) A person of the opposite sex that you don't get along with            0.14
#> 5 (5) A teacher you respected                                               0.19
#> 6 (6) Mother (or the person who filled that kind of role)                   0.10
#> 7 (7) A person of the opposite sex that you like more than you dislike      0.13
#> 8 (8) The most confident person you know                                    0.22
#> 9 (9) A person you work well with (at job or on sports team etc)            0.14
#> 10 (10) A teacher you did not respect                                       0.09

# using Cohen's rc for element correlations
indexIntensity(bell2010, rc = TRUE)
#> 
#> ################
#> Intensity index
#> ################
#> 
#> Total intensity: 0.25 
#> 
#> 
#> Average intensity of constructs: 0.16 
#> 
#> Itensity by construct:
#>                                     intensity
#> 1 relaxed - worried & tense              0.18
#> 2 not so smart (a - smart (academic      0.14
#> 3 dislikes sports - loves sports         0.05
#> 4 not interactive - loves people         0.15
#> 5 not transparent - transparent          0.13
#> 6 insensitive - sensitive                0.16
#> 7 fearful&timid - fearless               0.25
#> 8 rough - gentle                         0.25
#> 9 accept as it is - loves to argue       0.12
#> 
#> 
#> Average intensity of elements: 0.26 
#> 
#> Itensity by element:
#>                                       intensity
#> 1 (1) self                                 0.39
#> 2 (2) closest friend of the same sex       0.31
#> 3 (3) the unhappiest person you know       0.27
#> 4 (4) A person of the opposite sex t       0.29
#> 5 (5) A teacher you respected              0.24
#> 6 (6) Mother (or the person who fill       0.16
#> 7 (7) A person of the opposite sex t       0.32
#> 8 (8) The most confident person you        0.24
#> 9 (9) A person you work well with (a       0.24
#> 10 (10) A teacher you did not respect      0.12

# save output
x <- indexIntensity(bell2010)
x
#> 
#> ################
#> Intensity index
#> ################
#> 
#> Total intensity: 0.19 
#> 
#> 
#> Average intensity of constructs: 0.16 
#> 
#> Itensity by construct:
#>                                     intensity
#> 1 relaxed - worried & tense              0.18
#> 2 not so smart (a - smart (academic      0.14
#> 3 dislikes sports - loves sports         0.05
#> 4 not interactive - loves people         0.15
#> 5 not transparent - transparent          0.13
#> 6 insensitive - sensitive                0.16
#> 7 fearful&timid - fearless               0.25
#> 8 rough - gentle                         0.25
#> 9 accept as it is - loves to argue       0.12
#> 
#> 
#> Average intensity of elements: 0.14 
#> 
#> Itensity by element:
#>                                       intensity
#> 1 (1) self                                 0.16
#> 2 (2) closest friend of the same sex       0.16
#> 3 (3) the unhappiest person you know       0.04
#> 4 (4) A person of the opposite sex t       0.14
#> 5 (5) A teacher you respected              0.19
#> 6 (6) Mother (or the person who fill       0.10
#> 7 (7) A person of the opposite sex t       0.13
#> 8 (8) The most confident person you        0.22
#> 9 (9) A person you work well with (a       0.14
#> 10 (10) A teacher you did not respect      0.09

# printing options
print(x, digits = 4)
#> 
#> ################
#> Intensity index
#> ################
#> 
#> Total intensity: 0.1894 
#> 
#> 
#> Average intensity of constructs: 0.1588 
#> 
#> Itensity by construct:
#>                                     intensity
#> 1 relaxed - worried & tense            0.1828
#> 2 not so smart (a - smart (academic    0.1410
#> 3 dislikes sports - loves sports       0.0518
#> 4 not interactive - loves people       0.1488
#> 5 not transparent - transparent        0.1320
#> 6 insensitive - sensitive              0.1560
#> 7 fearful&timid - fearless             0.2492
#> 8 rough - gentle                       0.2474
#> 9 accept as it is - loves to argue     0.1204
#> 
#> 
#> Average intensity of elements: 0.1358 
#> 
#> Itensity by element:
#>                                       intensity
#> 1 (1) self                               0.1553
#> 2 (2) closest friend of the same sex     0.1616
#> 3 (3) the unhappiest person you know     0.0410
#> 4 (4) A person of the opposite sex t     0.1362
#> 5 (5) A teacher you respected            0.1872
#> 6 (6) Mother (or the person who fill     0.0961
#> 7 (7) A person of the opposite sex t     0.1291
#> 8 (8) The most confident person you      0.2153
#> 9 (9) A person you work well with (a     0.1434
#> 10 (10) A teacher you did not respect    0.0924

# accessing the objects' content
x$c.int
#>         relaxed - worried & tense not so smart (a - smart (academic 
#>                         0.1827637                         0.1409762 
#>    dislikes sports - loves sports    not interactive - loves people 
#>                         0.0518083                         0.1487900 
#>     not transparent - transparent           insensitive - sensitive 
#>                         0.1320374                         0.1559959 
#>          fearful&timid - fearless                    rough - gentle 
#>                         0.2491723                         0.2474163 
#>  accept as it is - loves to argue 
#>                         0.1203857 
x$e.int
#>                           (1) self (2) closest friend of the same sex 
#>                         0.15532806                         0.16164286 
#> (3) the unhappiest person you know (4) A person of the opposite sex t 
#>                         0.04102202                         0.13620017 
#>        (5) A teacher you respected (6) Mother (or the person who fill 
#>                         0.18719718                         0.09610145 
#> (7) A person of the opposite sex t (8) The most confident person you  
#>                         0.12910356                         0.21527250 
#> (9) A person you work well with (a (10) A teacher you did not respect 
#>                         0.14342080                         0.09241641 
x$c.int.mean
#> [1] 0.1588162
x$e.int.mean
#> [1] 0.1357705
x$total.int
#> [1] 0.1893525