Descriptive statistics for constructs and elements

Several descriptive measures for constructs and elements.

Usage

statsElements(x, index = TRUE, trim = 20)

statsConstructs(x, index = T, trim = 20)

Arguments

x

repgrid object.

index

Whether to print the number of the element.

trim

The number of characters an element or a construct is trimmed to (default is 20). If NA no trimming occurs. Trimming simply saves space when displaying correlation of constructs or elements with long names.

Value

A dataframe containing the following measures is returned invisibly (see psych::describe()):

item name

item number

number of valid cases

mean standard deviation

trimmed mean (default .1)

median (standard or interpolated)

mad: median absolute deviation (from the median)

minimum

maximum

skew

kurtosis

standard error

Note

Note that standard deviation and variance are estimations, i.e. including Bessel's correction. For more info type ?describe.

Examples

statsConstructs(fbb2003)
#> 
#> ####################################
#> Desriptive statistics for constructs
#> ####################################
#> 
#>                             vars n mean   sd median trimmed  mad min max range
#> (1) clever - not bright        1 8 3.75 2.31    4.0    3.75 2.97   1   7     6
#> (2) disorganiz - organized     2 8 4.00 1.77    4.5    4.00 2.22   2   6     4
#> (3) listens - doesn't he       3 8 3.50 2.14    3.0    3.50 2.22   1   7     6
#> (4) no clear v - clear view    4 8 4.38 1.60    4.0    4.38 1.48   3   7     4
#> (5) understand - no underst    5 8 3.50 1.85    2.5    3.50 0.74   2   6     4
#> (6) ambitious - no ambitio     6 8 4.50 1.51    4.5    4.50 2.22   3   7     4
#> (7) respected - not respec     7 8 3.25 1.75    3.0    3.25 1.48   1   6     5
#> (8) distant - warm             8 8 4.12 1.96    4.0    4.12 1.48   1   7     6
#> (9) rather agg - not aggres    9 8 3.62 1.92    3.0    3.62 2.22   1   7     6
#>                              skew kurtosis   se
#> (1) clever - not bright      0.02    -1.84 0.82
#> (2) disorganiz - organized  -0.13    -1.96 0.63
#> (3) listens - doesn't he     0.35    -1.40 0.76
#> (4) no clear v - clear view  0.38    -1.68 0.56
#> (5) understand - no underst  0.41    -1.90 0.65
#> (6) ambitious - no ambitio   0.33    -1.58 0.53
#> (7) respected - not respec   0.23    -1.67 0.62
#> (8) distant - warm          -0.05    -1.46 0.69
#> (9) rather agg - not aggres  0.36    -1.25 0.68
statsConstructs(fbb2003, trim = 10)
#> 
#> ####################################
#> Desriptive statistics for constructs
#> ####################################
#> 
#>                   vars n mean   sd median trimmed  mad min max range  skew
#> (1) cleve - not b    1 8 3.75 2.31    4.0    3.75 2.97   1   7     6  0.02
#> (2) disor - organ    2 8 4.00 1.77    4.5    4.00 2.22   2   6     4 -0.13
#> (3) liste - doesn    3 8 3.50 2.14    3.0    3.50 2.22   1   7     6  0.35
#> (4) no cl - clear    4 8 4.38 1.60    4.0    4.38 1.48   3   7     4  0.38
#> (5) under - no un    5 8 3.50 1.85    2.5    3.50 0.74   2   6     4  0.41
#> (6) ambit - no am    6 8 4.50 1.51    4.5    4.50 2.22   3   7     4  0.33
#> (7) respe - not r    7 8 3.25 1.75    3.0    3.25 1.48   1   6     5  0.23
#> (8) dista - warm     8 8 4.12 1.96    4.0    4.12 1.48   1   7     6 -0.05
#> (9) rathe - not a    9 8 3.62 1.92    3.0    3.62 2.22   1   7     6  0.36
#>                   kurtosis   se
#> (1) cleve - not b    -1.84 0.82
#> (2) disor - organ    -1.96 0.63
#> (3) liste - doesn    -1.40 0.76
#> (4) no cl - clear    -1.68 0.56
#> (5) under - no un    -1.90 0.65
#> (6) ambit - no am    -1.58 0.53
#> (7) respe - not r    -1.67 0.62
#> (8) dista - warm     -1.46 0.69
#> (9) rathe - not a    -1.25 0.68
statsConstructs(fbb2003, trim = 10, index = FALSE)
#> 
#> ####################################
#> Desriptive statistics for constructs
#> ####################################
#> 
#>               vars n mean   sd median trimmed  mad min max range  skew kurtosis
#> cleve - not b    1 8 3.75 2.31    4.0    3.75 2.97   1   7     6  0.02    -1.84
#> disor - organ    2 8 4.00 1.77    4.5    4.00 2.22   2   6     4 -0.13    -1.96
#> liste - doesn    3 8 3.50 2.14    3.0    3.50 2.22   1   7     6  0.35    -1.40
#> no cl - clear    4 8 4.38 1.60    4.0    4.38 1.48   3   7     4  0.38    -1.68
#> under - no un    5 8 3.50 1.85    2.5    3.50 0.74   2   6     4  0.41    -1.90
#> ambit - no am    6 8 4.50 1.51    4.5    4.50 2.22   3   7     4  0.33    -1.58
#> respe - not r    7 8 3.25 1.75    3.0    3.25 1.48   1   6     5  0.23    -1.67
#> dista - warm     8 8 4.12 1.96    4.0    4.12 1.48   1   7     6 -0.05    -1.46
#> rathe - not a    9 8 3.62 1.92    3.0    3.62 2.22   1   7     6  0.36    -1.25
#>                 se
#> cleve - not b 0.82
#> disor - organ 0.63
#> liste - doesn 0.76
#> no cl - clear 0.56
#> under - no un 0.65
#> ambit - no am 0.53
#> respe - not r 0.62
#> dista - warm  0.69
#> rathe - not a 0.68

statsElements(fbb2003)
#> 
#> ##################################
#> Desriptive statistics for elements
#> ##################################
#> 
#>                          vars n mean   sd median trimmed  mad min max range
#> (1) self                    1 9 3.44 1.81      3    3.44 1.48   1   6     5
#> (2) my father               2 9 3.00 1.87      3    3.00 1.48   1   6     5
#> (3) an old flame            3 9 4.89 1.45      5    4.89 1.48   3   7     4
#> (4) an ethical person       4 9 3.11 0.93      3    3.11 0.00   2   5     3
#> (5) my mother               5 9 4.11 1.69      5    4.11 2.97   2   7     5
#> (6) a rejected teacher      6 9 4.33 2.35      5    4.33 2.97   1   7     6
#> (7) as I would love to b    7 9 3.44 2.35      3    3.44 2.97   1   7     6
#> (8) a pitied person         8 9 4.44 1.42      5    4.44 1.48   2   7     5
#>                           skew kurtosis   se
#> (1) self                  0.30    -1.60 0.60
#> (2) my father             0.61    -1.22 0.62
#> (3) an old flame         -0.05    -1.72 0.48
#> (4) an ethical person     0.65    -0.54 0.31
#> (5) my mother             0.12    -1.44 0.56
#> (6) a rejected teacher   -0.15    -1.89 0.78
#> (7) as I would love to b  0.19    -1.82 0.78
#> (8) a pitied person      -0.02    -0.75 0.47
statsElements(fbb2003, trim = 10)
#> 
#> ##################################
#> Desriptive statistics for elements
#> ##################################
#> 
#>                vars n mean   sd median trimmed  mad min max range  skew
#> (1) self          1 9 3.44 1.81      3    3.44 1.48   1   6     5  0.30
#> (2) my father     2 9 3.00 1.87      3    3.00 1.48   1   6     5  0.61
#> (3) an old fla    3 9 4.89 1.45      5    4.89 1.48   3   7     4 -0.05
#> (4) an ethical    4 9 3.11 0.93      3    3.11 0.00   2   5     3  0.65
#> (5) my mother     5 9 4.11 1.69      5    4.11 2.97   2   7     5  0.12
#> (6) a rejected    6 9 4.33 2.35      5    4.33 2.97   1   7     6 -0.15
#> (7) as I would    7 9 3.44 2.35      3    3.44 2.97   1   7     6  0.19
#> (8) a pitied p    8 9 4.44 1.42      5    4.44 1.48   2   7     5 -0.02
#>                kurtosis   se
#> (1) self          -1.60 0.60
#> (2) my father     -1.22 0.62
#> (3) an old fla    -1.72 0.48
#> (4) an ethical    -0.54 0.31
#> (5) my mother     -1.44 0.56
#> (6) a rejected    -1.89 0.78
#> (7) as I would    -1.82 0.78
#> (8) a pitied p    -0.75 0.47
statsElements(fbb2003, trim = 10, index = FALSE)
#> 
#> ##################################
#> Desriptive statistics for elements
#> ##################################
#> 
#>            vars n mean   sd median trimmed  mad min max range  skew kurtosis
#> self          1 9 3.44 1.81      3    3.44 1.48   1   6     5  0.30    -1.60
#> my father     2 9 3.00 1.87      3    3.00 1.48   1   6     5  0.61    -1.22
#> an old fla    3 9 4.89 1.45      5    4.89 1.48   3   7     4 -0.05    -1.72
#> an ethical    4 9 3.11 0.93      3    3.11 0.00   2   5     3  0.65    -0.54
#> my mother     5 9 4.11 1.69      5    4.11 2.97   2   7     5  0.12    -1.44
#> a rejected    6 9 4.33 2.35      5    4.33 2.97   1   7     6 -0.15    -1.89
#> as I would    7 9 3.44 2.35      3    3.44 2.97   1   7     6  0.19    -1.82
#> a pitied p    8 9 4.44 1.42      5    4.44 1.48   2   7     5 -0.02    -0.75
#>              se
#> self       0.60
#> my father  0.62
#> an old fla 0.48
#> an ethical 0.31
#> my mother  0.56
#> a rejected 0.78
#> as I would 0.78
#> a pitied p 0.47

# save the access the results
d <- statsElements(fbb2003)
d
#> 
#> ##################################
#> Desriptive statistics for elements
#> ##################################
#> 
#>                          vars n mean   sd median trimmed  mad min max range
#> (1) self                    1 9 3.44 1.81      3    3.44 1.48   1   6     5
#> (2) my father               2 9 3.00 1.87      3    3.00 1.48   1   6     5
#> (3) an old flame            3 9 4.89 1.45      5    4.89 1.48   3   7     4
#> (4) an ethical person       4 9 3.11 0.93      3    3.11 0.00   2   5     3
#> (5) my mother               5 9 4.11 1.69      5    4.11 2.97   2   7     5
#> (6) a rejected teacher      6 9 4.33 2.35      5    4.33 2.97   1   7     6
#> (7) as I would love to b    7 9 3.44 2.35      3    3.44 2.97   1   7     6
#> (8) a pitied person         8 9 4.44 1.42      5    4.44 1.48   2   7     5
#>                           skew kurtosis   se
#> (1) self                  0.30    -1.60 0.60
#> (2) my father             0.61    -1.22 0.62
#> (3) an old flame         -0.05    -1.72 0.48
#> (4) an ethical person     0.65    -0.54 0.31
#> (5) my mother             0.12    -1.44 0.56
#> (6) a rejected teacher   -0.15    -1.89 0.78
#> (7) as I would love to b  0.19    -1.82 0.78
#> (8) a pitied person      -0.02    -0.75 0.47
d["mean"]
#> 
#> ##################################
#> Desriptive statistics for elements
#> ##################################
#> 
#>                          mean
#> (1) self                 3.44
#> (2) my father            3.00
#> (3) an old flame         4.89
#> (4) an ethical person    3.11
#> (5) my mother            4.11
#> (6) a rejected teacher   4.33
#> (7) as I would love to b 3.44
#> (8) a pitied person      4.44
d[2, "mean"] # mean rating of 2nd element
#> [1] 3

d <- statsConstructs(fbb2003)
d
#> 
#> ####################################
#> Desriptive statistics for constructs
#> ####################################
#> 
#>                             vars n mean   sd median trimmed  mad min max range
#> (1) clever - not bright        1 8 3.75 2.31    4.0    3.75 2.97   1   7     6
#> (2) disorganiz - organized     2 8 4.00 1.77    4.5    4.00 2.22   2   6     4
#> (3) listens - doesn't he       3 8 3.50 2.14    3.0    3.50 2.22   1   7     6
#> (4) no clear v - clear view    4 8 4.38 1.60    4.0    4.38 1.48   3   7     4
#> (5) understand - no underst    5 8 3.50 1.85    2.5    3.50 0.74   2   6     4
#> (6) ambitious - no ambitio     6 8 4.50 1.51    4.5    4.50 2.22   3   7     4
#> (7) respected - not respec     7 8 3.25 1.75    3.0    3.25 1.48   1   6     5
#> (8) distant - warm             8 8 4.12 1.96    4.0    4.12 1.48   1   7     6
#> (9) rather agg - not aggres    9 8 3.62 1.92    3.0    3.62 2.22   1   7     6
#>                              skew kurtosis   se
#> (1) clever - not bright      0.02    -1.84 0.82
#> (2) disorganiz - organized  -0.13    -1.96 0.63
#> (3) listens - doesn't he     0.35    -1.40 0.76
#> (4) no clear v - clear view  0.38    -1.68 0.56
#> (5) understand - no underst  0.41    -1.90 0.65
#> (6) ambitious - no ambitio   0.33    -1.58 0.53
#> (7) respected - not respec   0.23    -1.67 0.62
#> (8) distant - warm          -0.05    -1.46 0.69
#> (9) rather agg - not aggres  0.36    -1.25 0.68
d["sd"]
#> 
#> ####################################
#> Desriptive statistics for constructs
#> ####################################
#> 
#>                               sd
#> (1) clever - not bright     2.31
#> (2) disorganiz - organized  1.77
#> (3) listens - doesn't he    2.14
#> (4) no clear v - clear view 1.60
#> (5) understand - no underst 1.85
#> (6) ambitious - no ambitio  1.51
#> (7) respected - not respec  1.75
#> (8) distant - warm          1.96
#> (9) rather agg - not aggres 1.92
d[1, "sd"] # sd of ratings on first construct
#> [1] 2.31455